Codec.Extras.SerialiseViaFlat

Constructors

SerialiseViaFlat

Fields

unSerialiseViaFlat :: a

Constructors

DeserialiseFailureInfo

Fields

dfOffset :: ByteOffset

dfReason :: DeserialiseFailureReason

data DeserialiseFailureReason

The reason of the cbor failure as a datatype, not as a plain string.

Constructors

EndOfInput
Not enough input provided

ExpectedBytes
The bytes inside the input are malformed.

OtherReason String
This is either a cbor failure that we (plutus) are not aware of, or an underlying flat failure. We use whatever message cborg or flat returns.