Data.RandomAccessList.SkewBinary

Documentation

data RAList a

A strict list of complete binary trees accompanied by their size. The trees appear in >=-size order. Note: this list is strict in its spine, unlike the Prelude list

Constructors

Nil

Bundled Patterns

pattern Cons :: a -> RAList a -> RAList a

type Item (RAList a)

Instance details

Defined in Data.RandomAccessList.SkewBinary

type Item (RAList a) = Item (AsRAL (RAList a))

type Element (RAList a)

Instance details

Defined in Data.RandomAccessList.SkewBinary

type Element (RAList a) = a

contIndexZero :: forall a b. b -> (a -> b) -> RAList a -> Word64 -> b

contIndexOne :: forall a b. b -> (a -> b) -> RAList a -> Word64 -> b

safeIndexZero :: RAList a -> Word64 -> Maybe a

unsafeIndexZero :: RAList a -> Word64 -> a

safeIndexOne :: RAList a -> Word64 -> Maybe a

unsafeIndexOne :: RAList a -> Word64 -> a

null :: RAList a -> Bool

uncons :: RAList a -> Maybe (a, RAList a)