Data.RandomAccessList.SkewBinarySlab

Documentation

data RAList a

A strict list of complete binary trees accompanied by their node size. The trees appear in >=-node size order. Note: this list is strict in its spine, unlike the Prelude list

Constructors

Nil

Bundled Patterns

pattern Cons :: a -> RAList a -> RAList a