ErrorCode

Documentation

class HasErrorCode a where

Assigns an error-code to data-constructors (values) of error types. Note, when implementing this method you're only supposed to look at the outermost constructor (whnf) of the a value to decide for the error-code.

Methods

errorCode :: a -> ErrorCode

newtype ErrorCode

A wrapper to Natural so as to override the pretty instance of Natural with zero padding

Constructors

ErrorCode Natural