PlutusCore.Annotation

Documentation

data Ann

An annotation type used during the compilation.

Constructors

Ann

Fields

annInline :: Inline

annSrcSpans :: SrcSpans

data SrcSpan

The span between two source locations.

This corresponds roughly to the SrcSpan used by GHC, but we define our own version so we don't have to depend on ghc to use it.

The line and column numbers are 1-based, and the unit is Unicode code point (or Char).

Constructors

SrcSpan

Fields

srcSpanFile :: FilePath

srcSpanSLine :: Int

srcSpanSCol :: Int

srcSpanELine :: Int

srcSpanECol :: Int

Same as GHC's SrcSpan, srcSpanECol is usually one more than the column of the last character of the thing this SrcSpan is for (unless the last character is the line break).

newtype SrcSpans

Constructors

SrcSpans

Fields

unSrcSpans :: Set SrcSpan

newtype InlineHints name a

Constructors

InlineHints

Fields

shouldInline :: a -> name -> Bool

data Inline

Constructors

AlwaysInline
When calling PlutusIR.Compiler.Definitions.defineTerm to add a new term definition, if we annotation the var on the LHS of the definition with AlwaysInline, the inliner will always inline that var.

This is currently used to ensure builtin functions such as trace (when the remove-trace flag is on and trace is rewritten to const) are inlined, because the inliner would otherwise not inline them. To achieve that, we annotate the definition with AlwaysInline when defining trace, i.e., trace AlwaysInline = _ a -> a.

MayInline

annAlwaysInline :: Ann

Create an Ann with AlwaysInline.

annMayInline :: Ann

Create an Ann with MayInline.

data SourcePos

Constructors

SourcePos

Fields

sourceName :: FilePath

sourceLine :: !Pos

sourceColumn :: !Pos

data Pos

addSrcSpan :: SrcSpan -> Ann -> Ann

Add an extra SrcSpan to existing SrcSpans of Ann

lineInSrcSpan :: Pos -> SrcSpan -> Bool

Tells if a line (positive integer) falls inside a SrcSpan.