PlutusCore.Bitwise.Convert

Description

Implementations for conversion primops from Integer to ByteString and back again.

Documentation

integerToByteStringWrapper :: Bool -> Integer -> Integer -> BuiltinResult ByteString

Wrapper for integerToByteString to make it more convenient to define as a builtin.

byteStringToIntegerWrapper :: Bool -> ByteString -> Integer

Wrapper for byteStringToInteger to make it more convenient to define as a builtin.

data IntegerToByteStringError

Structured type to help indicate conversion errors.

Constructors

NegativeInput

NotEnoughDigits

integerToByteStringMaximumOutputLength :: Integer

integerToByteString :: ByteOrder -> Int -> Integer -> Either IntegerToByteStringError ByteString

Conversion from Integer to ByteString, as per CIP-0087.

For performance and clarity, the endianness argument uses ByteOrder, and the length argument is an Int.

byteStringToInteger :: ByteOrder -> ByteString -> Integer

Conversion from ByteString to Integer, as per CIP-0087.

For clarity, the stated endianness argument uses ByteOrder.