PlutusCore.Builtin

Description

Reexports from modules from the Builtin folder.

Documentation

class Typeable (a :: k)

The class Typeable allows a concrete representation of a type to be calculated.

Minimal complete definition

typeRep#

data TypeScheme val (args :: [Type]) res where

The type of type schemes of built-in functions. args is a list of types of arguments, res is the resulting type. E.g. Text -> Bool -> Integer is encoded as TypeScheme val [Text, Bool] Integer.

Constructors

TypeSchemeResult :: (Typeable res, KnownTypeAst TyName (UniOf val) res, MakeKnown val res) => TypeScheme val '[] res

TypeSchemeArrow :: (Typeable arg, KnownTypeAst TyName (UniOf val) arg, MakeKnown val arg, ReadKnown val arg) => TypeScheme val args res -> TypeScheme val (arg ': args) res infixr 9

TypeSchemeAll :: (KnownSymbol text, KnownNat uniq, KnownKind kind) => Proxy '(text, uniq, kind) -> TypeScheme val args res -> TypeScheme val args res

argProxy :: TypeScheme val (arg ': args) res -> Proxy arg

typeSchemeToType :: TypeScheme val args res -> Type TyName (UniOf val) ()

Convert a TypeScheme to the corresponding Kind.

class (uni `Everywhere` ImplementedKnownTypeAst uni, uni `Everywhere` ImplementedReadKnownIn uni, uni `Everywhere` ImplementedMakeKnownIn uni) => TestTypesFromTheUniverseAreAllKnown uni

An instance of this class not having any constraints ensures that every type (according to Everywhere) from the universe has 'KnownTypeAst, ReadKnownIn and MakeKnownIn instances.

data BuiltinRuntime val

A BuiltinRuntime represents a possibly partial builtin application, including an empty builtin application (i.e. just the builtin with no arguments).

Applying or type-instantiating a builtin peels off the corresponding constructor from its BuiltinRuntime.

BuiltinCostedResult contains the cost (an ExBudget) and the result (a BuiltinResult val) of the builtin application. The cost is stored strictly, since the evaluator is going to look at it and the result is stored lazily, since it's not supposed to be forced before accounting for the cost of the application. If the cost exceeds the available budget, the evaluator discards the result of the builtin application without ever forcing it and terminates with evaluation failure. Allowing the user to compute something that they don't have the budget for would be a major bug.

Evaluators that ignore the entire concept of costing (e.g. the CK machine) may of course force the result of the builtin application unconditionally.

Constructors

BuiltinCostedResult ExBudgetStream ~(BuiltinResult val)

BuiltinExpectArgument (val -> BuiltinRuntime val)

BuiltinExpectForce (BuiltinRuntime val)

data BuiltinsRuntime fun val

A data wrapper around a function returning the BuiltinRuntime of a built-in function. We use data rather than newtype, because GHC is able to see through newtypes and may break carefully set up optimizations, see https://github.com/IntersectMBO/plutus/pull/4914#issuecomment-1396306606

Using data may make things more expensive, however it was verified at the time of writing that the wrapper is removed before the CEK machine starts, leaving the stored function to be used directly.

In order for lookups to be efficient the BuiltinRuntimes need to be cached, i.e. pulled out of the function statically. See makeBuiltinMeaning for how we achieve that.

Constructors

BuiltinsRuntime

Fields

unBuiltinsRuntime :: fun -> BuiltinRuntime val

builtinRuntimeFailure :: BuiltinError -> BuiltinRuntime val

lookupBuiltin :: fun -> BuiltinsRuntime fun val -> BuiltinRuntime val

Look up the runtime info of a built-in function during evaluation.

data EvaluationError structural operational

The type of errors that can occur during evaluation. There are two kinds of errors:

Structural ones -- these are errors that are indicative of the _structure_ of the program being wrong. For example, a free variable was encountered during evaluation, a non-function was applied to an argument or tailList was applied to a non-list.

Operational ones -- these are errors that are indicative of the _logic_ of the program being wrong. For example, error was executed, tailList was applied to an empty list or evaluation ran out of gas.

On the chain both of these are just regular failures and we don't distinguish between them there: if a script fails, it fails, it doesn't matter what the reason was. However in the tests it does matter why the failure occurred: a structural error may indicate that the test was written incorrectly while an operational error may be entirely expected.

In other words, structural errors are "runtime type errors" and operational errors are regular runtime errors. Which means that evaluating an (erased) well-typed program should never produce a structural error, only an operational one. This creates a sort of "runtime type system" for UPLC and it would be great to stick to it and enforce in tests etc, but we currently don't.

Constructors

StructuralEvaluationError !structural

OperationalEvaluationError !operational

class AsEvaluationError r structural operational | r -> structural operational where

Minimal complete definition

_EvaluationError

Methods

_EvaluationError :: Prism' r (EvaluationError structural operational)

_StructuralEvaluationError :: Prism' r structural

_OperationalEvaluationError :: Prism' r operational

newtype UnliftingError

The error message part of an UnliftingEvaluationError.

Constructors

MkUnliftingError

Fields

unUnliftingError :: Text

newtype UnliftingEvaluationError

When unlifting of a PLC term into a Haskell value fails, this error is thrown.

Constructors

MkUnliftingEvaluationError

Fields

unUnliftingEvaluationError :: EvaluationError UnliftingError UnliftingError

data BuiltinError

The type of errors that readKnown and makeKnown can return.

Constructors

BuiltinUnliftingEvaluationError UnliftingEvaluationError

BuiltinEvaluationFailure

data BuiltinResult a

The monad that makeKnown runs in. Equivalent to ExceptT BuiltinError (Writer (DList Text)), except optimized in two ways:

everything is strict

has the BuiltinSuccess constructor that is used for returning a value with no logs attached, which is the most common case for us, so it helps a lot not to construct and deconstruct a redundant tuple

Moving from ExceptT BuiltinError (Writer (DList Text)) to this data type gave us a speedup of 8% of total evaluation time.

Logs are represented as a DList, because we don't particularly care about the efficiency of logging, since there's no logging on the chain and builtins don't emit much anyway. Otherwise we'd have to use text-builder or text-builder-linear or something of this sort.

Constructors

BuiltinSuccess a

BuiltinSuccessWithLogs (DList Text) a

BuiltinFailure (DList Text) BuiltinError

class AsUnliftingEvaluationError r where

Minimal complete definition

_UnliftingEvaluationError

Methods

_UnliftingEvaluationError :: Prism' r UnliftingEvaluationError

_MkUnliftingEvaluationError :: Prism' r (EvaluationError UnliftingError UnliftingError)

class AsUnliftingError r where

Minimal complete definition

_UnliftingError

Methods

_UnliftingError :: Prism' r UnliftingError

_MkUnliftingError :: Prism' r Text

class AsBuiltinError r where

Minimal complete definition

_BuiltinError

Methods

_BuiltinError :: Prism' r BuiltinError

_BuiltinUnliftingEvaluationError :: Prism' r UnliftingEvaluationError

_BuiltinEvaluationFailure :: Prism' r ()

class AsBuiltinResult r a | r -> a where

Minimal complete definition

_BuiltinResult

Methods

_BuiltinResult :: Prism' r (BuiltinResult a)

_BuiltinSuccess :: Prism' r a

_BuiltinSuccessWithLogs :: Prism' r (DList Text, a)

_BuiltinFailure :: Prism' r (DList Text, BuiltinError)

_UnliftingErrorVia :: Pretty err => err -> Prism' err UnliftingError

Construct a prism focusing on the *EvaluationFailure part of err by taking that *EvaluationFailure and

pretty-printing and embedding it into an UnliftingError for the setter part of the prism

returning it directly for the opposite direction (there's no other way to convert an UnliftingError to an evaluation failure, since the latter doesn't carry any content)

This is useful for providing AsUnliftingError instances for types such as CkUserError and CekUserError.

_StructuralUnliftingError :: AsBuiltinError err => Prism' err UnliftingError

_OperationalUnliftingError :: AsBuiltinError err => Prism' err UnliftingError

throwNotAConstant :: MonadError BuiltinError m => m void

throwUnderTypeError :: MonadError BuiltinError m => m void

emit :: Text -> BuiltinResult ()

Add a log line to the logs.

withLogs :: DList Text -> BuiltinResult a -> BuiltinResult a

Prepend logs to a BuiltinResult computation.

throwing :: MonadError e m => AReview e t -> t -> m x

throwing_ :: MonadError e m => AReview e () -> m x

newtype Opaque val (rep :: Type)

The AST of a value with a Plutus type attached to it. The type is for the Plutus type checker to look at. Opaque can appear in the type of the denotation of a builtin.

Constructors

Opaque

Fields

unOpaque :: val

newtype SomeConstant uni (rep :: Type)

For unlifting from the Constant constructor when the stored value is of a monomorphic built-in type

The rep parameter specifies how the type looks on the PLC side (i.e. just like with Opaque val rep).

Constructors

SomeConstant

Fields

unSomeConstant :: Some (ValueOf uni)

data TyNameRep (kind :: Type)

Representation of a type variable: its name and unique and an implicit kind.

Constructors

TyNameRep Symbol Nat

data family TyVarRep (name :: TyNameRep kind) :: kind

Representation of an intrinsically-kinded type variable: a name.

data family TyAppRep (fun :: dom -> cod) (arg :: dom) :: cod

Representation of an intrinsically-kinded type application: a function and an argument.

data family TyForallRep (name :: TyNameRep kind) (a :: Type) :: Type

Representation of of an intrinsically-kinded universal quantifier: a bound name and a body.

data family BuiltinHead x

For annotating an uninstantiated built-in type, so that it gets handled by the right instance or type family.

type family ElaborateBuiltin uni x

Take an iterated application of a built-in type and elaborate every function application inside of it to TyAppRep and annotate the head with BuiltinHead.

The idea is that we don't need to process built-in types manually if we simply add some annotations for instance resolution to look for. Think what we'd have to do manually for, say, ToHoles: traverse the spine of the application and collect all the holes into a list, which is troubling, because type applications are left-nested and lists are right-nested, so we'd have to use accumulators or an explicit Reverse type family. And then we also have KnownTypeAst and ToBinds, so handling built-in types in a special way for each of those would be a hassle, especially given the fact that type-level Haskell is not exactly good at computing things. With the ElaborateBuiltin approach we get KnownTypeAst, ToHoles and ToBinds for free.

We make this an open type family, so that elaboration is customizable for each universe.

type family AllElaboratedArgs constr x where ...

Take a constraint and use it to constrain every argument of a possibly 0-ary elaborated application of a built-in type.

Equations

AllElaboratedArgs constr (f `TyAppRep` x) = (constr x, AllElaboratedArgs constr f)

AllElaboratedArgs _ (BuiltinHead _) = ()

class AllElaboratedArgs constr (ElaborateBuiltin uni x) => AllBuiltinArgs uni constr x

Take a constraint and use it to constrain every argument of a possibly 0-ary application of a built-in type.

type family FoldArgs args res where ...

Turn a list of Haskell types args into a functional type ending in res.

>>> :set -XDataKinds
>>> :kind! FoldArgs [(), Bool] Integer
FoldArgs [(), Bool] Integer :: *
= () -> Bool -> Integer

Equations

FoldArgs '[] res = res

FoldArgs (arg ': args) res = arg -> FoldArgs args res

data BuiltinMeaning val cost

The meaning of a built-in function consists of its type represented as a TypeScheme, its Haskell denotation and its uninstantiated runtime denotation.

The TypeScheme of a built-in function is used for example for

computing the PLC type of the function to be used during type checking

getting arity information

generating arbitrary values to apply the function to in tests

The denotation is lazy, so that we don't need to worry about a builtin being bottom (happens in tests). The production path is not affected by that, since only runtime denotations are used for evaluation.

Constructors

forall args res. BuiltinMeaning (TypeScheme val args res) ~(FoldArgs args res) (cost -> BuiltinRuntime val)

type HasMeaningIn uni val = (Typeable val, ExMemoryUsage val, HasConstantIn uni val)

Constraints available when defining a built-in function.

class (Typeable uni, Typeable fun, Bounded fun, Enum fun, Ix fun, Default (BuiltinSemanticsVariant fun)) => ToBuiltinMeaning uni fun where

A type class for "each function from a set of built-in functions has a BuiltinMeaning".

Associated Types

type CostingPart uni fun

The cost part of BuiltinMeaning.

data BuiltinSemanticsVariant fun

See Note [Builtin semantics variants]

Methods

toBuiltinMeaning :: HasMeaningIn uni val => BuiltinSemanticsVariant fun -> fun -> BuiltinMeaning val (CostingPart uni fun)

Get the BuiltinMeaning of a built-in function.

withTypeSchemeOfBuiltinFunction :: forall val fun r. (ToBuiltinMeaning (UniOf val) fun, ExMemoryUsage val, Typeable val, HasConstant val) => BuiltinSemanticsVariant fun -> fun -> (forall args res. TypeScheme val args res -> r) -> r

Feed the TypeScheme of the given built-in function to the continuation.

typeOfBuiltinFunction :: forall uni fun. ToBuiltinMeaning uni fun => BuiltinSemanticsVariant fun -> fun -> Type TyName uni ()

Get the type of a built-in function.

type family GetArgs a where ...

Chop a function type to get a list of its argument types.

Equations

GetArgs (a -> b) = a ': GetArgs b

GetArgs _ = '[]

class KnownMonotype val args res where

A class that allows us to derive a monotype for a builtin. We could've computed the runtime denotation from the TypeScheme and the denotation of the builtin, but not statically (due to unfolding not working for recursive functions and TypeScheme being recursive, i.e. requiring the conversion function to be recursive), and so it would cause us to retain a lot of evaluation-irrelevant stuff in the constructors of BuiltinRuntime, which has to make evaluation slower (we didn't check) and certainly makes the generated Core much harder to read. Technically speaking, we could get a RuntimeScheme from the TypeScheme and the denotation statically if we changed the definition of TypeScheme and made it a singleton, but then the conversion function would have to become a class anyway and we'd just replicate what we have here, except in a much more complicated way.

Methods

knownMonotype :: TypeScheme val args res

toMonoF :: ReadKnownM (FoldArgs args res, FoldArgs args ExBudgetStream) -> BuiltinRuntime val

Convert the denotation of a builtin to its runtime counterpart . The argument is in ReadKnownM, because that's what deferred unlifting amounts to: passing the action returning the builtin application around until full saturation, which is when the action actually gets run.

class KnownMonotype val args res => KnownPolytype (binds :: [Some TyNameRep]) val args res where

A class that allows us to derive a polytype for a builtin.

Methods

knownPolytype :: TypeScheme val args res

toPolyF :: ReadKnownM (FoldArgs args res, FoldArgs args ExBudgetStream) -> BuiltinRuntime val

Convert the denotation of a builtin to its runtime counterpart. The argument is in ReadKnownM, because that's what we need to do: passing the action returning the builtin application around until full saturation, which is when the action actually gets run.

type family ThrowOnBothEmpty binds args isBuiltin a where ...

Ensure a built-in function is not nullary and throw a nice error otherwise.

Equations

ThrowOnBothEmpty '[] '[] 'True a = TypeError (('Text "A built-in function must take at least one type or term argument" ':$$: ((('Text "\8216" ':<>: 'ShowType a) ':<>: 'Text "\8217 is a built-in type") ':<>: 'Text " so you can embed any of its values as a constant")) ':$$: 'Text "If you still want a built-in function, add a dummy \8216()\8217 argument")

ThrowOnBothEmpty '[] '[] 'False a = TypeError ('Text "A built-in function must take at least one type or term argument" ':$$: 'Text "To fix this error add a dummy \8216()\8217 argument")

ThrowOnBothEmpty _ _ _ _ = ()

class MakeBuiltinMeaning a val where

A function turned into a type class with exactly one fully general instance. We can't package up the constraints of makeBuiltinMeaning (see the instance) into a type or class synonym, because they contain a bunch of variables defined by ~ or determined via functional dependencies and neither class nor type definitions can handle that (see https://gitlab.haskell.org/ghc/ghc/-/issues/7100). Inlining three lines of constraints whenever we need to call makeBuiltinMeaning over a non-concrete type is a bad option and this abstraction is free anyway, hence its existence.

The a type variable goes first, because makeBuiltinMeaning @A is a common pattern.

Methods

makeBuiltinMeaning :: a -> (cost -> FoldArgs (GetArgs a) ExBudgetStream) -> BuiltinMeaning val cost

Construct the meaning for a built-in function by automatically deriving its TypeScheme, given

the denotation of the builtin

an uninstantiated costing function

toBuiltinRuntime :: cost -> BuiltinMeaning val cost -> BuiltinRuntime val

Convert a BuiltinMeaning to a BuiltinRuntime given a cost model.

toBuiltinsRuntime :: (cost ~ CostingPart uni fun, ToBuiltinMeaning uni fun, HasMeaningIn uni val) => BuiltinSemanticsVariant fun -> cost -> BuiltinsRuntime fun val

Calculate runtime info for all built-in functions given meanings of builtins (as a constraint), the semantics variant of the set of builtins and a cost model.

data TyNameRep (kind :: Type)

Representation of a type variable: its name and unique and an implicit kind.

Constructors

TyNameRep Symbol Nat

data family TyVarRep (name :: TyNameRep kind) :: kind

Representation of an intrinsically-kinded type variable: a name.

data family TyAppRep (fun :: dom -> cod) (arg :: dom) :: cod

Representation of an intrinsically-kinded type application: a function and an argument.

data family TyForallRep (name :: TyNameRep kind) (a :: Type) :: Type

Representation of of an intrinsically-kinded universal quantifier: a bound name and a body.

data Hole

The kind of holes.

data family RepHole x

A hole in the Rep context.

type HasTermLevel uni = Includes uni

Specifies that the given type is a built-in one and its values can be embedded into a Term.

type HasTypeLevel uni x = KnownTypeAst Void uni (ElaborateBuiltin uni x)

Specifies that the given type is a built-in one and can be embedded into a Kind.

type HasTypeAndTermLevel uni x = (uni `HasTypeLevel` x, uni `HasTermLevel` x)

The product of HasTypeLevel and HasTermLevel.

mkTyBuiltin :: forall a (x :: a) uni ann tyname. uni `HasTypeLevel` x => ann -> Type tyname uni ann

Convert a Haskell representation of a possibly 0-ary application of a built-in type to arbitrary types implementing KnownTypeAst.

data family TypeHole a

A hole in the Type context.

type KnownBuiltinTypeAst tyname uni x = AllBuiltinArgs uni (KnownTypeAst tyname uni) x

A constraint for "a is a KnownTypeAst by means of being included in uni".

class KnownTypeAst tyname uni x where

This class allows one to convert the type-level Haskell representation of a Plutus type into the corresponding Plutus type. Associated type families are needed to help elaboration.

Depending on the universe converting a Haskell type to a Plutus team can give different results (e.g. Int can be a built-in type instead of being encoded via built-in Integer), hence this class takes a uni argument. Plus, elaboration is universe-specific too.

Minimal complete definition

Nothing

Associated Types

type IsBuiltin uni x :: Bool

Whether x is a built-in type.

type IsBuiltin uni x = IsBuiltin uni (ElaborateBuiltin uni x)

type ToHoles uni x :: [Hole]

Return every part of the type that can be a to-be-instantiated type variable. For example, in Integer there's no such types and in (a, b) it's the two arguments (a and b) and the same applies to a -> b (to mention a type that is not built-in).

type ToHoles uni x = ToHoles uni (ElaborateBuiltin uni x)

type ToBinds uni (acc :: [Some TyNameRep]) x :: [Some TyNameRep]

Collect all unique variables (a variable consists of a textual name, a unique and a kind) in an accumulator and return the accumulator once a leaf is reached.

type ToBinds uni acc x = ToBinds uni acc (ElaborateBuiltin uni x)

Methods

typeAst :: Type tyname uni ()

The Plutus counterpart of the x type.

default typeAst :: KnownTypeAst tyname uni (ElaborateBuiltin uni x) => Type tyname uni ()

toTypeAst :: forall a tyname uni (x :: a) proxy. KnownTypeAst tyname uni x => proxy x -> Type tyname uni ()

Return the Plutus counterpart of the x type.

type family Insert x xs where ...

Insert x into xs unless it's already there.

Equations

Insert x '[] = '[x]

Insert x (x : xs) = x ': xs

Insert x (y : xs) = y ': Insert x xs

type family Delete x xs where ...

Delete the first x from a list. Which is okay since we only ever put things in once.

Equations

Delete _ '[] = '[]

Delete x (x ': xs) = xs

Delete x (y ': xs) = y ': Delete x xs

data BuiltinError

The type of errors that readKnown and makeKnown can return.

throwBuiltinErrorWithCause :: (MonadError (ErrorWithCause err cause) m, AsUnliftingEvaluationError err, AsEvaluationFailure err) => cause -> BuiltinError -> m void

Attach a cause to a BuiltinError and throw that. Note that an evaluator might require the cause to be computed lazily for best performance on the happy path, hence this function must not force its first argument. TODO: wrap cause in Lazy once we have it.

type KnownBuiltinTypeIn uni val a = (HasConstantIn uni val, PrettyParens (SomeTypeIn uni), GEq uni, uni `HasTermLevel` a)

A constraint for "a is a ReadKnownIn and MakeKnownIn by means of being included in uni".

type KnownBuiltinType val a = KnownBuiltinTypeIn (UniOf val) val a

A constraint for "a is a ReadKnownIn and MakeKnownIn by means of being included in UniOf term".

data BuiltinResult a

The monad that makeKnown runs in. Equivalent to ExceptT BuiltinError (Writer (DList Text)), except optimized in two ways:

everything is strict

has the BuiltinSuccess constructor that is used for returning a value with no logs attached, which is the most common case for us, so it helps a lot not to construct and deconstruct a redundant tuple

Moving from ExceptT BuiltinError (Writer (DList Text)) to this data type gave us a speedup of 8% of total evaluation time.

Logs are represented as a DList, because we don't particularly care about the efficiency of logging, since there's no logging on the chain and builtins don't emit much anyway. Otherwise we'd have to use text-builder or text-builder-linear or something of this sort.

Constructors

BuiltinSuccess a

BuiltinSuccessWithLogs (DList Text) a

BuiltinFailure (DList Text) BuiltinError

type ReadKnownM = Either BuiltinError

The monad that readKnown runs in.

class uni ~ UniOf val => MakeKnownIn uni val a where

Minimal complete definition

Nothing

Methods

makeKnown :: a -> BuiltinResult val

Convert a Haskell value to the corresponding PLC value. The inverse of readKnown.

default makeKnown :: KnownBuiltinType val a => a -> BuiltinResult val

readKnownConstant :: forall val a. KnownBuiltinType val a => val -> ReadKnownM a

Convert a constant embedded into a PLC term to the corresponding Haskell value.

type MakeKnown val = MakeKnownIn (UniOf val) val

class uni ~ UniOf val => ReadKnownIn uni val a where

Minimal complete definition

Nothing

Methods

readKnown :: val -> ReadKnownM a

Convert a PLC value to the corresponding Haskell value. The inverse of makeKnown.

default readKnown :: KnownBuiltinType val a => val -> ReadKnownM a

type ReadKnown val = ReadKnownIn (UniOf val) val

makeKnownOrFail :: MakeKnownIn uni val a => a -> EvaluationResult val

Same as makeKnown, but allows for neither emitting nor storing the cause of a failure.

readKnownSelf :: (ReadKnown val a, AsUnliftingEvaluationError err, AsEvaluationFailure err) => val -> Either (ErrorWithCause err val) a

Same as readKnown, but the cause of a potential failure is the provided term itself.

data SingKind k where

The type of singletonized Haskell kinds representing Plutus kinds. Indexing by a Haskell kind allows us to avoid an error call in the ToKind instance of DefaultUni and not worry about proxies in the type of knownKind.

Constructors

SingType :: SingKind Type

SingKindArrow :: SingKind k -> SingKind l -> SingKind (k -> l) infixr 5

withSingKind :: Kind ann -> (forall k. SingKind k -> r) -> r

Feed the SingKind version of the given Kind to the given continuation.

class KnownKind k where

For reifying Haskell kinds representing Plutus kinds at the term level.

Methods

knownKind :: SingKind k

bringKnownKind :: SingKind k -> (KnownKind k => r) -> r

Satisfy the KnownKind constraint of a continuation using the given SingKind.

withKnownKind :: Kind ann -> (forall k. KnownKind k => Proxy k -> r) -> r

class ToKind (uni :: Type -> Type) where

For computing the Plutus kind of a built-in type. See kindOfBuiltinType.

Methods

toSingKind :: uni (Esc (a :: k)) -> SingKind k

Reify the kind of a type from the universe at the term level.

demoteKind :: SingKind k -> Kind ()

Convert a reified Haskell kind to a Plutus kind.

kindOfBuiltinType :: ToKind uni => uni (Esc a) -> Kind ()

Compute the kind of a type from a universe.

data BuiltinError

The type of errors that readKnown and makeKnown can return.

Constructors

BuiltinUnliftingEvaluationError UnliftingEvaluationError

BuiltinEvaluationFailure

throwNotAConstant :: MonadError BuiltinError m => m void

class HasConstant term where

Ensures that term has a Constant-like constructor to lift values to and unlift values from.

Methods

asConstant :: term -> Either BuiltinError (Some (ValueOf (UniOf term)))

Unwrap from a Constant-like constructor throwing an UnliftingError if the provided term is not a wrapped Haskell value.

fromConstant :: Some (ValueOf (UniOf term)) -> term

Wrap a Haskell value as a term.

type HasConstantIn uni term = (UniOf term ~ uni, HasConstant term)

Ensures that term has a Constant-like constructor to lift values to and unlift values from and connects term and its uni.

fromValueOf :: forall a term. HasConstant term => UniOf term (Esc a) -> a -> term

Wrap a Haskell value (given its explicit type tag) as a term.

fromValue :: forall a term. (HasConstant term, UniOf term `HasTermLevel` a) => a -> term

Wrap a Haskell value (provided its type is in the universe) as a term.