PlutusCore.Builtin.Debug

Description

This module helps to visualize and debug the TypeScheme inference machinery from the Meaning module.

Documentation

elaborateDebug :: forall a j. ElaborateFromTo 0 j (Term TyName Name DefaultUni DefaultFun ()) a => a -> a

Instantiate type variables in the type of a value using ElaborateFromTo. Example usages:

>>> :t elaborateDebug False
elaborateDebug False :: Bool
>>> :t elaborateDebug $ \_ -> ()
elaborateDebug $ \_ -> ()
  :: Opaque
       (Term TyName Name DefaultUni DefaultFun ())
       (TyVarRep ('TyNameRep "a" 0))
     -> ()
>>> :t elaborateDebug $ Just ()
<interactive>:1:1: error:
    • No instance for ‘KnownTypeAst DefaultUni (Maybe ())’
      Which means
        ‘Maybe ()’
      is neither a built-in type, nor one of the control types.
      If it can be represented in terms of one of the built-in types
        then go add the instance (you may need a ‘KnownTypeIn’ one too)
        alongside the instance for the built-in type.
      Otherwise you may need to add a new built-in type
        (provided you're doing something that can be supported in principle)
    • In the expression: elaborateDebug $ Just ()
>>> :t elaborateDebug $ 0 + 42
<interactive>:1:18: error:
    • Built-in functions are not allowed to have constraints
      To fix this error instantiate all constrained type variables
    • In the second argument of ‘($)’, namely ‘0 + 42’
      In the expression: elaborateDebug $ 0 + 42

module PlutusCore.Name

module PlutusCore.Default

module PlutusCore.Core

newtype Opaque val (rep :: Type)

The denotation of a term whose PLC type is encoded in rep (for example a type variable or an application of a type variable). I.e. the denotation of such a term is the term itself. This is because we have parametricity in Haskell, so we can't inspect a value whose type is, say, a bound variable, so we never need to convert such a term from Plutus Core to Haskell and back and instead can keep it intact.

Constructors

Opaque

Fields

unOpaque :: val

newtype SomeConstant uni (rep :: Type)

For unlifting from the Constant constructor when the stored value is of a monomorphic built-in type

The rep parameter specifies how the type looks on the PLC side (i.e. just like with Opaque val rep).

Constructors

SomeConstant

Fields

unSomeConstant :: Some (ValueOf uni)

data TyNameRep (kind :: Type)

Representation of a type variable: its name and unique and an implicit kind.

Constructors

TyNameRep Symbol Nat

data family TyVarRep (name :: TyNameRep kind) :: kind

Representation of an intrinsically-kinded type variable: a name.

data family TyAppRep (fun :: dom -> cod) (arg :: dom) :: cod

Representation of an intrinsically-kinded type application: a function and an argument.

data family TyForallRep (name :: TyNameRep kind) (a :: Type) :: Type

Representation of of an intrinsically-kinded universal quantifier: a bound name and a body.

module PlutusCore.Builtin.Elaborate