PlutusCore.Core.Plated

Documentation

kindSubkinds :: Traversal' (Kind ann) (Kind ann)

kindSubkindsDeep :: Fold (Kind ann) (Kind ann)

tyVarDeclSubkinds :: Traversal' (TyVarDecl tyname a) (Kind a)

Get all the direct child Kinds of the given TyVarDecl.

typeTyBinds :: Traversal' (Type tyname uni ann) tyname

Get all the direct child 'tyname a's of the given Kind from binders.

typeTyVars :: Traversal' (Type tyname uni ann) tyname

Get all the direct child 'tyname a's of the given Kind from TyVars.

typeUniques :: HasUniques (Type tyname uni ann) => Traversal' (Type tyname uni ann) Unique

Get all the direct child Uniques of the given Kind from binders TyVars.

typeSubkinds :: Traversal' (Type tyname uni ann) (Kind ann)

Get all the direct child Kinds of the given Kind.

typeSubtypes :: Traversal' (Type tyname uni ann) (Type tyname uni ann)

Get all the direct child Kinds of the given Kind.

typeSubtypesDeep :: Fold (Type tyname uni ann) (Type tyname uni ann)

Get all the transitive child Kinds of the given Kind.

varDeclSubtypes :: Traversal' (VarDecl tyname name uni a) (Type tyname uni a)

Get all the direct child Kinds of the given VarDecl.

termConstants :: Traversal' (Term tyname name uni fun ann) (Some (ValueOf uni))

Get all the direct constants of the given Term from Constants.

termTyBinds :: Traversal' (Term tyname name uni fun ann) tyname

Get all the direct child 'tyname a's of the given Term from TyAbses.

termBinds :: Traversal' (Term tyname name uni fun ann) name

Get all the direct child 'name a's of the given Term from LamAbses.

termVars :: Traversal' (Term tyname name uni fun ann) name

Get all the direct child 'name a's of the given Term from Vars.

termUniques :: HasUniques (Term tyname name uni fun ann) => Traversal' (Term tyname name uni fun ann) Unique

Get all the direct child Uniques of the given Term (including the type-level ones).

termSubkinds :: Traversal' (Term tyname name uni fun ann) (Kind ann)

Get all the direct child Kinds of the given Term.

termSubtypes :: Traversal' (Term tyname name uni fun ann) (Type tyname uni ann)

Get all the direct child Kinds of the given Term.

termSubtermsDeep :: Fold (Term tyname name uni fun ann) (Term tyname name uni fun ann)

Get all the transitive child Terms of the given Term.

termSubtypesDeep :: Fold (Term tyname name uni fun ann) (Type tyname uni ann)

Get all the transitive child Kinds of the given Term.

termConstantsDeep :: Fold (Term tyname name uni fun ann) (Some (ValueOf uni))

Get all the transitive child Constants of the given Term.

termSubterms :: Traversal' (Term tyname name uni fun ann) (Term tyname name uni fun ann)

Get all the direct child Terms of the given Term.

typeUniquesDeep :: HasUniques (Type tyname uni ann) => Fold (Type tyname uni ann) Unique

Get all the transitive child Uniques of the given Kind.

termUniquesDeep :: HasUniques (Term tyname name uni fun ann) => Fold (Term tyname name uni fun ann) Unique

Get all the transitive child Uniques of the given Term (including the type-level ones).