PlutusCore.DataFilePaths

Description

Various file paths used in plutus-core, currently all to do with the cost model.

Documentation

costModelDataDir :: FilePath

builtinCostModelFile :: FilePath

cekMachineCostsFile :: FilePath

rModelFile :: FilePath

The file containing the R models: only needed for cost-model-test.

benchingResultsFile :: FilePath

The file containing the benchmark results for the built-in functions: only needed for cost-model-test.