PlutusCore.Error

Documentation

data ParseError

An error encountered during parsing.

Constructors

UnknownBuiltinType Text SourcePos

BuiltinTypeNotAStar Text SourcePos

UnknownBuiltinFunction Text SourcePos

InvalidBuiltinConstant Text Text SourcePos

class AsParseError r where

Minimal complete definition

_ParseError

Methods

_ParseError :: Prism' r ParseError

_UnknownBuiltinType :: Prism' r (Text, SourcePos)

_BuiltinTypeNotAStar :: Prism' r (Text, SourcePos)

_UnknownBuiltinFunction :: Prism' r (Text, SourcePos)

_InvalidBuiltinConstant :: Prism' r (Text, Text, SourcePos)

data NormCheckError tyname name uni fun ann

Constructors

BadType ann (Type tyname uni ann) Text

BadTerm ann (Term tyname name uni fun ann) Text

class AsNormCheckError r tyname name uni fun ann | r -> tyname name uni fun ann where

Minimal complete definition

_NormCheckError

Methods

_NormCheckError :: Prism' r (NormCheckError tyname name uni fun ann)

_BadType :: Prism' r (ann, Type tyname uni ann, Text)

_BadTerm :: Prism' r (ann, Term tyname name uni fun ann, Text)

data UniqueError ann

Constructors

MultiplyDefined Unique ann ann

IncoherentUsage Unique ann ann

FreeVariable Unique ann

class AsUniqueError r ann | r -> ann where

Minimal complete definition

_UniqueError

Methods

_UniqueError :: Prism' r (UniqueError ann)

_MultiplyDefined :: Prism' r (Unique, ann, ann)

_IncoherentUsage :: Prism' r (Unique, ann, ann)

_FreeVariable :: Prism' r (Unique, ann)

data TypeError term uni fun ann

Constructors

KindMismatch ann (Type TyName uni ()) (Kind ()) (Kind ())

TypeMismatch ann term (Type TyName uni ()) (Normalized (Type TyName uni ()))

FreeTypeVariableE ann TyName

FreeVariableE ann Name

UnknownBuiltinFunctionE ann fun

class AsTypeError r term uni fun ann | r -> term uni fun ann where

Minimal complete definition

_TypeError

Methods

_TypeError :: Prism' r (TypeError term uni fun ann)

_KindMismatch :: Prism' r (ann, Type TyName uni (), Kind (), Kind ())

_TypeMismatch :: Prism' r (ann, term, Type TyName uni (), Normalized (Type TyName uni ()))

_FreeTypeVariableE :: Prism' r (ann, TyName)

_FreeVariableE :: Prism' r (ann, Name)

_UnknownBuiltinFunctionE :: Prism' r (ann, fun)

data FreeVariableError

We cannot do a correct translation to or from de Bruijn indices if the program is not well-scoped. So we throw an error in such a case.

Constructors

FreeUnique Unique

FreeIndex Index

class AsFreeVariableError r where

Minimal complete definition

_FreeVariableError

Methods

_FreeVariableError :: Prism' r FreeVariableError

_FreeUnique :: Prism' r Unique

_FreeIndex :: Prism' r Index

data Error uni fun ann

Constructors

ParseErrorE ParseError

UniqueCoherencyErrorE (UniqueError ann)

TypeErrorE (TypeError (Term TyName Name uni fun ()) uni fun ann)

NormCheckErrorE (NormCheckError TyName Name uni fun ann)

FreeVariableErrorE FreeVariableError

class AsError r uni fun ann | r -> uni fun ann where

Minimal complete definition

_Error

Methods

_Error :: Prism' r (Error uni fun ann)

_ParseErrorE :: Prism' r ParseError

_UniqueCoherencyErrorE :: Prism' r (UniqueError ann)

_TypeErrorE :: Prism' r (TypeError (Term TyName Name uni fun ()) uni fun ann)

_NormCheckErrorE :: Prism' r (NormCheckError TyName Name uni fun ann)

_FreeVariableErrorE :: Prism' r FreeVariableError

throwingEither :: MonadError e m => AReview e t -> Either t a -> m a

Lifts an Either into an error context where we can embed the Left value into the error.

class Ord a => ShowErrorComponent a where

The type class defines how to print a custom component of ParseError.

Since: megaparsec-5.0.0

Methods

showErrorComponent :: a -> String

Pretty-print a component of ParseError.

Orphan instances

Pretty SourcePos

Instance details

Methods

pretty :: SourcePos -> Doc ann

prettyList :: [SourcePos] -> Doc ann