PlutusCore.Evaluation.Machine.Ck

Description

The CK machine.

Documentation

data EvaluationResult a

The parameterized type of results various evaluation engines return. On the PLC side this becomes (via makeKnown) either a call to Error or a value of the PLC counterpart of type a.

Constructors

EvaluationSuccess !a

EvaluationFailure

type CkEvaluationException uni fun = EvaluationException CkUserError (MachineError fun) (Term TyName Name uni fun ())

The CK machine-specific EvaluationException.

type CkM uni fun s = ReaderT (CkEnv uni fun s) (ExceptT (CkEvaluationException uni fun) (ST s))

data CkValue uni fun

extractEvaluationResult :: Either (EvaluationException user internal term) a -> Either (ErrorWithCause internal term) (EvaluationResult a)

Turn any UserEvaluationError into an EvaluationFailure.

runCk :: BuiltinsRuntime fun (CkValue uni fun) -> Bool -> Term TyName Name uni fun () -> (Either (CkEvaluationException uni fun) (Term TyName Name uni fun ()), [Text])

evaluateCk :: BuiltinsRuntime fun (CkValue uni fun) -> Term TyName Name uni fun () -> (Either (CkEvaluationException uni fun) (Term TyName Name uni fun ()), [Text])

Evaluate a term using the CK machine with logging enabled.

evaluateCkNoEmit :: BuiltinsRuntime fun (CkValue uni fun) -> Term TyName Name uni fun () -> Either (CkEvaluationException uni fun) (Term TyName Name uni fun ())

Evaluate a term using the CK machine with logging disabled.

unsafeEvaluateCk :: ThrowableBuiltins uni fun => BuiltinsRuntime fun (CkValue uni fun) -> Term TyName Name uni fun () -> (EvaluationResult (Term TyName Name uni fun ()), [Text])

Evaluate a term using the CK machine with logging enabled. May throw a CkEvaluationException.

unsafeEvaluateCkNoEmit :: ThrowableBuiltins uni fun => BuiltinsRuntime fun (CkValue uni fun) -> Term TyName Name uni fun () -> EvaluationResult (Term TyName Name uni fun ())

Evaluate a term using the CK machine with logging disabled. May throw a CkEvaluationException.

readKnownCk :: ReadKnown (Term TyName Name uni fun ()) a => BuiltinsRuntime fun (CkValue uni fun) -> Term TyName Name uni fun () -> Either (CkEvaluationException uni fun) a

Unlift a value using the CK machine.