PlutusCore.Evaluation.Machine.CostingFun.Core

Documentation

data CostingFun model

Constructors

CostingFun

Fields

costingFunCpu :: model

costingFunMemory :: model

data ModelAddedSizes

s * (x + y) + I

Constructors

ModelAddedSizes

Fields

modelAddedSizesIntercept :: CostingInteger

modelAddedSizesSlope :: CostingInteger

data ModelSubtractedSizes

s * (x - y) + I

Constructors

ModelSubtractedSizes

Fields

modelSubtractedSizesIntercept :: CostingInteger

modelSubtractedSizesSlope :: CostingInteger

modelSubtractedSizesMinimum :: CostingInteger

data ModelConstantOrLinear

if p then s*x else c; p depends on usage

Constructors

ModelConstantOrLinear

Fields

modelConstantOrLinearConstant :: CostingInteger

modelConstantOrLinearIntercept :: CostingInteger

modelConstantOrLinearSlope :: CostingInteger

data ModelConstantOrTwoArguments

if p then f(x,y) else c; p depends on usage

Constructors

ModelConstantOrTwoArguments

Fields

modelConstantOrTwoArgumentsConstant :: CostingInteger

modelConstantOrTwoArgumentsModel :: ModelTwoArguments

data ModelLinearSize

Constructors

ModelLinearSize

Fields

modelLinearSizeIntercept :: CostingInteger

modelLinearSizeSlope :: CostingInteger

data ModelMultipliedSizes

s * (x * y) + I

Constructors

ModelMultipliedSizes

Fields

modelMultipliedSizesIntercept :: CostingInteger

modelMultipliedSizesSlope :: CostingInteger

data ModelMinSize

s * min(x, y) + I

Constructors

ModelMinSize

Fields

modelMinSizeIntercept :: CostingInteger

modelMinSizeSlope :: CostingInteger

data ModelMaxSize

s * max(x, y) + I

Constructors

ModelMaxSize

Fields

modelMaxSizeIntercept :: CostingInteger

modelMaxSizeSlope :: CostingInteger

data ModelOneArgument

Constructors

ModelOneArgumentConstantCost CostingInteger

ModelOneArgumentLinearCost ModelLinearSize

data ModelTwoArguments

Constructors

ModelTwoArgumentsConstantCost CostingInteger

ModelTwoArgumentsLinearInX ModelLinearSize

ModelTwoArgumentsLinearInY ModelLinearSize

ModelTwoArgumentsAddedSizes ModelAddedSizes

ModelTwoArgumentsSubtractedSizes ModelSubtractedSizes

ModelTwoArgumentsMultipliedSizes ModelMultipliedSizes

ModelTwoArgumentsMinSize ModelMinSize

ModelTwoArgumentsMaxSize ModelMaxSize

ModelTwoArgumentsLinearOnDiagonal ModelConstantOrLinear

ModelTwoArgumentsConstAboveDiagonal ModelConstantOrTwoArguments

ModelTwoArgumentsConstBelowDiagonal ModelConstantOrTwoArguments

data ModelThreeArguments

Constructors

ModelThreeArgumentsConstantCost CostingInteger

ModelThreeArgumentsAddedSizes ModelAddedSizes

ModelThreeArgumentsLinearInX ModelLinearSize

ModelThreeArgumentsLinearInY ModelLinearSize

ModelThreeArgumentsLinearInZ ModelLinearSize

data ModelFourArguments

Constructors

ModelFourArgumentsConstantCost CostingInteger

data ModelFiveArguments

Constructors

ModelFiveArgumentsConstantCost CostingInteger

data ModelSixArguments

Constructors

ModelSixArgumentsConstantCost CostingInteger

runCostingFunOneArgument :: ExMemoryUsage a1 => CostingFun ModelOneArgument -> a1 -> ExBudget

runCostingFunTwoArguments :: (ExMemoryUsage a1, ExMemoryUsage a2) => CostingFun ModelTwoArguments -> a1 -> a2 -> ExBudget

runCostingFunThreeArguments :: (ExMemoryUsage a1, ExMemoryUsage a2, ExMemoryUsage a3) => CostingFun ModelThreeArguments -> a1 -> a2 -> a3 -> ExBudget

runCostingFunFourArguments :: (ExMemoryUsage a1, ExMemoryUsage a2, ExMemoryUsage a3, ExMemoryUsage a4) => CostingFun ModelFourArguments -> a1 -> a2 -> a3 -> a4 -> ExBudget

runCostingFunFiveArguments :: (ExMemoryUsage a1, ExMemoryUsage a2, ExMemoryUsage a3, ExMemoryUsage a4, ExMemoryUsage a5) => CostingFun ModelFiveArguments -> a1 -> a2 -> a3 -> a4 -> a5 -> ExBudget

runCostingFunSixArguments :: (ExMemoryUsage a1, ExMemoryUsage a2, ExMemoryUsage a3, ExMemoryUsage a4, ExMemoryUsage a5, ExMemoryUsage a6) => CostingFun ModelSixArguments -> a1 -> a2 -> a3 -> a4 -> a5 -> a6 -> ExBudget

class Eq a => Hashable a

The class of types that can be converted to a hash value.

Minimal implementation: hashWithSalt.

Note: the hash is not guaranteed to be stable across library versions, operating systems or architectures. For stable hashing use named hashes: SHA256, CRC32 etc.

If you are looking for Hashable instance in time package, check time-compat