PlutusCore.Evaluation.Machine.ExBudgetingDefaults

Documentation

defaultBuiltinsRuntimeForSemanticsVariant :: HasMeaningIn DefaultUni term => BuiltinSemanticsVariant DefaultFun -> BuiltinsRuntime DefaultFun term

defaultBuiltinsRuntime :: HasMeaningIn DefaultUni term => BuiltinsRuntime DefaultFun term

defaultCekCostModel :: CostModel CekMachineCosts BuiltinCostModel

The default cost model, including both builtin costs and machine step costs. Note that this is not necessarily the cost model in use on the chain at any given time. The definitive values used for calculating on-chain costs are protocol parameters which are part of the state of the chain; in practice these will usually have been obtained from the contents of the JSON files at some point in the past, but we do not guarantee this. During on-chain evaluation the ledger passes a cost model to the Plutus Core evaluator using the mkEvaluationContext functions in PlutusLedgerApi.

defaultCekMachineCosts :: CekMachineCosts

Default costs for CEK machine instructions.

defaultCekParameters :: Typeable ann => MachineParameters CekMachineCosts DefaultFun (CekValue DefaultUni DefaultFun ann)

defaultCostModelParams :: Maybe CostModelParams

The default cost model data. This is exposed to the ledger, so let's not confuse anybody by mentioning the CEK machine

defaultBuiltinCostModel :: BuiltinCostModel

The default cost model for built-in functions.

unitCekMachineCosts :: CekMachineCosts

unitCekParameters :: Typeable ann => MachineParameters CekMachineCosts DefaultFun (CekValue DefaultUni DefaultFun ann)