PlutusCore.Evaluation.Machine.ExMemoryUsage

Documentation

data CostRose

A lazy tree of costs. Convenient for calculating the costs of values of built-in types, because they may have arbitrary branching (in particular a Data object can contain a list of Data objects inside of it).

CostRose gets collapsed to a lazy linear structure down the pipeline, so that we can stream the costs to the outside where, say, the CEK machine picks them up one by one and handles somehow (in particular, subtracts from the remaining budget).

Constructors

CostRose !CostingInteger ![CostRose]

singletonRose :: CostingInteger -> CostRose

Create a CostRose containing a single cost.

class ExMemoryUsage a where

Methods

memoryUsage :: a -> CostRose

flattenCostRose :: CostRose -> CostStream

Collapse a CostRose to a lazy linear stream of costs. Retrieving the next element takes O(1) time in the worst case regardless of the recursion pattern of the given CostRose.