PlutusCore.Flat.Encoder.Prim

Description

Encoding Primitives

Documentation

eBits16F :: NumBits -> Word16 -> Prim

Encode up to 9 bits

eBitsF :: NumBits -> Word8 -> Prim

Encode up to 8 bits.

eFloatF :: Float -> Prim

eDoubleF :: Double -> Prim

eUTF16F :: Text -> Prim

Encode text as UTF16 and encode the result as an array of bytes

eUTF8F :: Text -> Prim

Encode text as UTF8 and encode the result as an array of bytes

eCharF :: Char -> Prim

eNaturalF :: Natural -> Prim

eIntegerF :: Integer -> Prim

eInt64F :: Int64 -> Prim

eInt32F :: Int32 -> Prim

eIntF :: Int -> Prim

eInt16F :: Int16 -> Prim

eInt8F :: Int8 -> Prim

eWordF :: Word -> Prim

eWord64F :: Word64 -> Prim

eWord32F :: Word32 -> Prim

eWord16F :: Word16 -> Prim

eBytesF :: ByteString -> Prim

eLazyBytesF :: ByteString -> Prim

Encode a Lazy ByteString

eShortBytesF :: ShortByteString -> Prim

eWord8F :: Word8 -> Prim

eFillerF :: Prim

>>> enc $ eTrueF >=> eFillerF
"10000001"

>>> enc eFillerF
"00000001"

eBoolF :: Bool -> Prim

eTrueF :: Prim

>>> enc eTrueF
"1"

eFalseF :: Prim

>>> enc eFalseF
"0"

varWordF :: (Bits t, Integral t) => t -> Prim

updateWord8 :: Word8 -> S -> Prim

>>> enc $ \s-> eWord8F 0 s >>= updateWord8 255 s
"11111111"

>>> enc $ \s0 -> eTrueF s0 >>= \s1 -> eWord8F 255 s1 >>= eWord8F 255 >>= updateWord8 0 s1
"10000000 01111111 1"

>>> enc $ \s0 -> eFalseF s0 >>= \s1 -> eWord8F 0 s1 >>= updateWord8 255 s1
"01111111 1"

>>> enc $ \s0 -> eFalseF s0 >>= \s1 -> eWord8F 0 s1 >>= updateWord8 255 s1 >>= eFalseF
"01111111 10"

>>> enc $ \s0 -> eTrueF s0 >>= \s1 -> eWord8F 255 s1 >>= eTrueF >>= updateWord8 0 s1 >>= eTrueF
"10000000 011"

Since: 0.5

w7l :: (Bits t, Integral t) => t -> [Word8]

Exported for testing only

eWord32BEF :: Word32 -> Prim

eWord64BEF :: Word64 -> Prim

eWord32E :: (Word32 -> Word32) -> Word32 -> Prim

eWord64E :: (Word64 -> Word64) -> Word64 -> Prim