PlutusCore.Flat.Filler

Description

Pre-value and post-value byte alignments

Documentation

data Filler

A meaningless sequence of 0 bits terminated with a 1 bit (easier to implement than the reverse)

Used to align encoded values at byte/word boundaries.

Constructors

FillerBit !Filler

FillerEnd

fillerLength :: Num a => Filler -> a

Length of a filler in bits

data PreAligned a

A Pre aligned value, a value preceded by a filler

Useful to prealign ByteArrays, Texts and any structure that can be encoded more efficiently when byte aligned.

Constructors

PreAligned

Fields

preFiller :: Filler

preValue :: a

preAligned :: a -> PreAligned a

Pre align a value

data PostAligned a

A Post aligned value, a value followed by a filler

Useful to complete the encoding of a top-level value

Constructors

PostAligned

Fields

postValue :: a

postFiller :: Filler

postAligned :: a -> PostAligned a

Post align a value

preAlignedDecoder :: Get b -> Get b

Decode a value assuming that is PreAligned

Since: 0.5

postAlignedDecoder :: Get b -> Get b

Decode a value assuming that is PostAligned