PlutusCore.Flat.Types

Description

Common Types

Documentation

type NumBits = Int

Number of bits

module Data.Word

module Data.Int

data Natural

Natural number

Invariant: numbers <= 0xffffffffffffffff use the NS constructor

KnownNat n => HasResolution (n :: Nat)
For example, Fixed 1000 will give you a Fixed with a resolution of 1000.

Instance details

Defined in Data.Fixed

Methods

resolution :: p n -> Integer

TestCoercion SNat
Since: base-4.18.0.0

Instance details

Defined in GHC.TypeNats

Methods

testCoercion :: forall (a :: k) (b :: k). SNat a -> SNat b -> Maybe (Coercion a b)

TestEquality SNat
Since: base-4.18.0.0

Instance details

Defined in GHC.TypeNats

Methods

testEquality :: forall (a :: k) (b :: k). SNat a -> SNat b -> Maybe (a :~: b)

ZigZag Integer Natural

Instance details

Defined in PlutusCore.Flat.Data.ZigZag

Methods

zigZag :: Integer -> Natural

zagZig :: Natural -> Integer

Lift Natural

Instance details

Defined in Language.Haskell.TH.Syntax

Methods

lift :: Quote m => Natural -> m Exp

liftTyped :: forall (m :: Type -> Type). Quote m => Natural -> Code m Natural

type Compare (a :: Natural) (b :: Natural)

Instance details

Defined in Data.Type.Ord

type Compare (a :: Natural) (b :: Natural) = CmpNat a b

data ShortByteString

A compact representation of a Word8 vector.

It has a lower memory overhead than a ByteString and does not contribute to heap fragmentation. It can be converted to or from a ByteString (at the cost of copying the string data). It supports very few other operations.

Flat Text
Text (and Data.Text.Lazy) is encoded as a byte aligned array of bytes corresponding to its UTF8 encoding.

>>> tt $ T.pack ""
(True,[1,0])

>>> tt $ T.pack "aaa"
(True,[1,3,97,97,97,0])

>>> tt $ T.pack "¢¢¢"
(True,[1,6,194,162,194,162,194,162,0])

>>> tt $ T.pack "日日日"
(True,[1,9,230,151,165,230,151,165,230,151,165,0])

>>> tt $ T.pack "𐍈𐍈𐍈"
(True,[1,12,240,144,141,136,240,144,141,136,240,144,141,136,0])

Strict and Lazy Text have the same encoding:

>>> tst (T.pack "abc") == tst (TL.pack "abc")
True

Instance details

Defined in PlutusCore.Flat.Instances.Text

Methods

encode :: Text -> Encoding

decode :: Get Text

size :: Text -> NumBits -> NumBits

Index Text = Int