PlutusCore.Generators.Hedgehog.Builtin

Documentation

class GenTypedTerm uni where

This class exists so we can provide an ad-hoc typed term generator for various universes. We usually rely-on a universe-specific generator for well-typed constants within that universe.

TODO: Move this to PlutusIR.Generators.AST, and merge genConstant with genConstant.

Methods

genTypedTerm :: forall (a :: DefaultUni fun ())

class GenArbitraryTerm uni where

This class exists so we can provide an ad-hoc arbitrary term generator for various universes.

Methods

genArbitraryTerm :: forall fun. (DefaultUni