PlutusCore.Generators.Hedgehog.Denotation

Description

This module defines tools for associating PLC terms with their corresponding Haskell values.

data DenotationContextMember term res

A member of a DenotationContext. object is existentially quantified, so the only thing that can be done with it, is turning it into a Term using _denotationToTerm.

Constructors

forall object. DenotationContextMember (Denotation term object res)

newtype DenotationContext term

A context of DenotationContextMembers. Each row is a mapping from a type to a list of things that can return that type. For example it can contain a mapping from integer to 1. a bound variable of type integer 2. a bound variable of functional type with the result being integer 3. the Builtin which returns an integer.

Constructors

DenotationContext

Fields

unDenotationContext :: DMap TypeRep (Compose List (DenotationContextMember term))

lookupInContext :: forall a term. TypeRep a -> DenotationContext term -> [DenotationContextMember term a]

Look up a list of Denotations from DenotationContext each of which has a type that ends in the same type as the one that the DefaultFuns.

DEPRECATED: No need to update for a new builtin. Outdated, since we moved to quickcheck generators.