PlutusCore.Generators.Hedgehog.Interesting

Description

Sample generators used for tests.

Foldable (TermOf term)

Instance details

Defined in PlutusCore.Generators.Hedgehog.TypedBuiltinGen

Methods

fold :: Monoid m => TermOf term m -> m

foldMap :: Monoid m => (a -> m) -> TermOf term a -> m

foldMap' :: Monoid m => (a -> m) -> TermOf term a -> m

foldr :: (a -> b -> b) -> b -> TermOf term a -> b

foldr' :: (a -> b -> b) -> b -> TermOf term a -> b

foldl :: (b -> a -> b) -> b -> TermOf term a -> b

foldl' :: (b -> a -> b) -> b -> TermOf term a -> b

foldr1 :: (a -> a -> a) -> TermOf term a -> a

foldl1 :: (a -> a -> a) -> TermOf term a -> a

toList :: TermOf term a -> [a]

null :: TermOf term a -> Bool

length :: TermOf term a -> Int

elem :: Eq a => a -> TermOf term a -> Bool

maximum :: Ord a => TermOf term a -> a

minimum :: Ord a => TermOf term a -> a

sum :: Num a => TermOf term a -> a

product :: Num a => TermOf term a -> a

Traversable (TermOf term)

Instance details

Defined in PlutusCore.Generators.Hedgehog.TypedBuiltinGen

Methods

traverse :: Applicative f => (a -> f b) -> TermOf term a -> f (TermOf term b)

sequenceA :: Applicative f => TermOf term (f a) -> f (TermOf term a)

mapM :: Monad m => (a -> m b) -> TermOf term a -> m (TermOf term b)

sequence :: Monad m => TermOf term (m a) -> m (TermOf term a)

Functor (TermOf term)

Instance details

Defined in PlutusCore.Generators.Hedgehog.TypedBuiltinGen

Methods

fmap :: (a -> b) -> TermOf term a -> TermOf term b

(<$) :: a -> TermOf term b -> TermOf term a

genOverapplication :: TermGen Integer

Generates application of a builtin that returns a function, immediately saturated afterwards.

ifThenElse {integer -> integer -> integer} (lessThanInteger i j) addInteger subtractInteger i j
    == if i < j then i + j else i - j 

factorial :: DefaultFun ()

i -> product [1 :: Integer .. i] as a PLC term.

\(i : integer) -> product (enumFromTo 1 i) 

genFactorial :: TermGen Integer

Generate a term that computes the factorial of an integer and return it along with the factorial of the corresponding DefaultFun ()

The naive exponential fibonacci function as a PLC term.

\(i0 : integer) ->
    fix {integer} {integer}
        (\(rec : integer -> integer) (i : integer) ->
                ifThenElse {integer}
                    (lessThanEqInteger i 1)
                    (\(u : unit) -> i)
                    (\(u : unit) -> addInteger
                        (rec (subtractInteger i 1))
                        (rec (subtractInteger i 2)))
        i0 

genNaiveFib :: TermGen Integer

Generate a term that computes the ith Fibonacci number and return it along with the corresponding Integer computed on the Haskell side.

genNatRoundtrip :: TermGen Integer

Generate an Integer, turn it into a Scott-encoded PLC Nat (see Nat), turn that Nat into the corresponding PLC integer using a fold (see FoldNat) defined in terms of generic fix (see Fix) and return the result along with the original DefaultFun ()

sumNat as a PLC term.

genScottListSum :: TermGen Integer

Generate a list of Integers, turn it into a Scott-encoded PLC List (see List), sum elements of the list (see Sum) and return it along with the sum of the original list.

genIfIntegers :: TermGen Integer

Generate a boolean and two integers and check whether if b then i1 else i2 means the same thing in Haskell and PLC. Terms are generated using genTermLoose.

fromInterestingTermGens :: (forall a. KnownType (DefaultFun ()) a => String -> TermGen a -> c) -> [c]

Apply a function to all interesting generators and collect the results.