PlutusCore.Generators.Hedgehog.Utils

Description

Utilities used in modules from the TestSupport folder.

forAllNoShow :: PropertyT m a

Generate a value, but do not show it in case an error occurs. A supplied generator has access to the Pretty class for getting its String representation. A supplied generator has access to the PrettyPlc constraint for getting its String representation. A supplied generator has access to the PrettyPlc err => err -> String

Pretty-print a PLC error.