PlutusCore.Generators.NEAT.Spec

Description

This file contains the tests and some associated machinery but not the generators.

data Options

Options for parallel conjunction strategies. Only matters for predicates using the Cool data type instead of Bool.

Constructors

D
Sequential

O
Optimal Short-circuiting

F
Parallel (fair)

OF
Optimal Short-circuiting and fairness

OS
Optimal Short-circuiting and choice-subset detection

OSF
Subset choice short-circuiting

packAssertion :: DefaultFun ())

newtype GenDepth

Search depth, measured in program size

Constructors

GenDepth

Fields

unGenDepth :: Parser GenMode