PlutusCore.Generators.QuickCheck.GenTm

Documentation

type GenTm = GenT (Reader GenEnv)

Term generators carry around a context to know e.g. what types and terms are in scope.

data GenEnv

Constructors

GenEnv

Fields

geAstSize :: DefaultUni ())

Datatype context. TODO: it's a little weird, 'cause it maps the datatype name to the datatype and the datatype introduces multiple names. It's probably fine to have something like that, though.

geTypes :: TypeCtx

Type context

geTerms :: DefaultFun ())

A custom user-controlled generator for terms - useful for situations where we want to e.g. generate custom strings for coverage or test some specific pattern that generates a special case for the compiler.

geCustomFreq :: Int

How often do we use the custom generator - values in the range of 10-30 are usually reasonable.

geDebug :: Bool

Are we currently running in debug-mode (to debug our generators)

suchThatMap :: Monad m => GenT m a -> (a -> Maybe b) -> GenT m b

Create a generator that runs the given generator and applies the given function to produced values until the result is a Just y, returning the y.

withDebug :: GenTm a -> GenTm a

Run a generator in debug-mode.

runGenTm :: GenTm a -> Gen a

Run a `GenTm generator in a top-level empty context where we are allowed to generate datatypes.

runGenTmCustom :: DefaultFun ())) -> GenTm a -> Gen a

Run a GenTm generator with a plug-in custom generator for terms that is included with the other generators.

deliver :: Monad m => GenT m (Maybe a) -> GenT m a

Create a generator that runs the given generator until the result is a Just x, returning the x.

withNoEscape :: GenTm a -> GenTm a

Don't allow types to escape from a generator.

onAstSize :: (Int -> Int) -> GenTm a -> GenTm a

Map a function over the generator size

ifAstSizeZero :: GenTm a -> GenTm a -> GenTm a

Default to the first generator if the size is zero (or negative), use the second generator otherwise.

withAstSize :: Int -> GenTm a -> GenTm a

Locally set the size in a generator

astSizeSplit_ :: Int -> Int -> GenTm a -> GenTm b -> GenTm (a, b)

Split the size between two generators in the ratio specified by the first two arguments.

astSizeSplit :: Int -> Int -> GenTm a -> (a -> GenTm b) -> GenTm (a, b)

Split the size between two generators in the ratio specified by the first two arguments and use the result of the first generator in the second.

getUniques :: GenTm (Name]

Generate one likely fresh name per string in the input list. Note that this may not give you a fresh name, if it happens to generate a name that was removed from the terms map in bindTyName (due to referencing a now-shadowed type variable) but is still in the scope.

genLikelyFreshTyName :: Name

Generate a name that likely overlaps with existing names on purpose. If there are no existing names, generate a fresh name. This function doesn't distinguish between the type- and term-level scopes, hence it may generate a TyName and not clashing with any previously generated Name. Which is a good thing, because we want to check that Plutus is able to handle such spurious name clashes. Generating weird stuff is useful for a testing machinery! We don't need any "definitely non-fresh name" anyway, since we get enough non-fresh names already.

genMaybeFreshName :: Name

Generate a fresh name most (roughly 75%) of the time and otherwise generate an already bound name. When there are no bound names generate a fresh name.

genMaybeFreshTyName :: TyName -> GenTm a -> GenTm a

Remember that we have generated a type name locally but don't bind it. Useful for non-recursive definitions where we want to control name overlap.

bindTmName :: DefaultFun ()) -> GenTm a -> GenTm a

Bind multiple bindings

class (Random a => (a, a) -> GenT m a

runGenT :: GenT m a -> Gen (m a)

frequency :: MonadGen m => [(Int, m a)] -> m a

Chooses one of the given generators, with a weighted random distribution. The input list must be non-empty.

suchThat :: MonadGen m => m a -> (a -> Bool) -> m a

Generates a value that satisfies a predicate.

suchThatMaybe :: MonadGen m => m a -> (a -> Bool) -> m (Maybe a)

Tries to generate a value that satisfies a predicate.

listOf :: MonadGen m => m a -> m [a]

Generates a list of random length. The maximum length depends on the size parameter.

listOf1 :: MonadGen m => m a -> m [a]

Generates a non-empty list of random length. The maximum length depends on the size parameter.

vectorOf :: MonadGen m => Int -> m a -> m [a]

Generates a list of the given length.

oneof :: MonadGen m => [m a] -> m a

Randomly uses one of the given generators. The input list must be non-empty.

elements :: MonadGen m => [a] -> m a

Generates one of the given values. The input list must be non-empty.

growingElements :: MonadGen m => [a] -> m a

Takes a list of elements of increasing size, and chooses among an initial segment of the list. The size of this initial segment increases with the size parameter. The input list must be non-empty.

oneofMay :: MonadGen m => [m a] -> m (Maybe a)

Randomly uses one of the given generators.

elementsMay :: MonadGen m => [a] -> m (Maybe a)

Generates one of the given values.

growingElementsMay :: MonadGen m => [a] -> m (Maybe a)

Takes a list of elements of increasing size, and chooses among an initial segment of the list. The size of this initial segment increases with the size parameter.

class Arbitrary a where

Minimal complete definition

arbitrary

Methods

arbitrary :: Gen a

shrink :: a -> [a]

(Arbitrary a, Arbitrary b) => Arbitrary (a, b)

Instance details

Defined in Test.QuickCheck.Arbitrary

Methods

arbitrary :: Gen (a, b)

shrink :: (a, b) -> [(a, b)]

(CoArbitrary a, Arbitrary b) => Arbitrary (a -> b)

Instance details

Defined in Test.QuickCheck.Arbitrary

Methods

arbitrary :: Gen (a -> b)

shrink :: (a -> b) -> [a -> b]

Arbitrary (a b c) => Arbitrary (Random a => (a, a) -> Gen a

Testable prop => Testable (Gen prop)

Instance details

Defined in Test.QuickCheck.Property

Methods

property :: Gen prop -> Property

propertyForAllShrinkShow :: Gen a -> (a -> [a]) -> (a -> [String]) -> (a -> Gen prop) -> Property

Orphan instances

MonadReader r m => MonadReader r (GenT m)

Instance details

Methods

ask :: GenT m r

local :: (r -> r) -> GenT m a -> GenT m a

reader :: (r -> a) -> GenT m a