PlutusCore.Generators.QuickCheck.Substitutions

Generators for substitutions

fvTypeBag :: TyName Int

Get the free type variables in a type along with how many times they occur. The elements of the map are guaranteed to be non-zero.

genSubst :: TypeCtx -> Gen TypeSub

Generate a type substitution mapping some of the variables in the given context to some arbitrary types (valid in the context).

shrinkSubst :: TypeCtx -> TypeSub -> [TypeSub]