PlutusCore.MkPlc

Documentation

class TermLike term tyname name uni fun | term -> tyname name uni fun where

A final encoding for Term, to allow PLC terms to be used transparently as PIR terms.

Minimal complete definition

var, tyAbs, lamAbs, apply, constant, builtin, tyInst, unwrap, iWrap, error

Methods

var :: ann -> name -> term ann

tyAbs :: ann -> tyname -> Kind ann -> term ann -> term ann

lamAbs :: ann -> name -> Type tyname uni ann -> term ann -> term ann

apply :: ann -> term ann -> term ann -> term ann

constant :: ann -> Some (ValueOf uni) -> term ann

builtin :: ann -> fun -> term ann

tyInst :: ann -> term ann -> Type tyname uni ann -> term ann

unwrap :: ann -> term ann -> term ann

iWrap :: ann -> Type tyname uni ann -> Type tyname uni ann -> term ann -> term ann

error :: ann -> Type tyname uni ann -> term ann

termLet :: ann -> TermDef term tyname name uni ann -> term ann -> term ann

typeLet :: ann -> TypeDef tyname uni ann -> term ann -> term ann

type family UniOf a :: Type -> Type

Extract the universe from a type.

mkTyBuiltinOf :: forall k (a :: k) uni tyname ann. ann -> uni (Esc a) -> Type tyname uni ann

Embed a type (given its explicit type tag) into a PLC type.

mkTyBuiltin :: forall k (a :: k) uni tyname ann. uni `Contains` a => ann -> Type tyname uni ann

Embed a type (provided it's in the universe) into a PLC type.

mkConstantOf :: forall a uni fun term tyname name ann. TermLike term tyname name uni fun => ann -> uni (Esc a) -> a -> term ann

Embed a Haskell value (given its explicit type tag) into a PLC term.

mkConstant :: forall a uni fun term tyname name ann. (TermLike term tyname name uni fun, uni `Includes` a) => ann -> a -> term ann

Embed a Haskell value (provided its type is in the universe) into a PLC term.

data VarDecl tyname name uni ann

A "variable declaration", i.e. a name and a type for a variable.

Constructors

VarDecl

Fields

_varDeclAnn :: ann

_varDeclName :: name

_varDeclType :: Type tyname uni ann

data TyVarDecl tyname ann

A "type variable declaration", i.e. a name and a kind for a type variable.

Constructors

TyVarDecl

Fields

_tyVarDeclAnn :: ann

_tyVarDeclName :: tyname

_tyVarDeclKind :: Kind ann

data TyDecl tyname uni ann

A "type declaration", i.e. a kind for a type.

Constructors

TyDecl

Fields

_tyDeclAnn :: ann

_tyDeclType :: Type tyname uni ann

_tyDeclKind :: Kind ann

mkVar :: TermLike term tyname name uni fun => ann -> VarDecl tyname name uni ann -> term ann

Make a Var referencing the given VarDecl.

mkTyVar :: ann -> TyVarDecl tyname ann -> Type tyname uni ann

Make a TyVar referencing the given TyVarDecl.

tyDeclVar :: TyVarDecl tyname ann -> TyDecl tyname uni ann

data Def var val

A definition. Pretty much just a pair with more descriptive names.

Constructors

Def

Fields

defVar :: var

defVal :: val

embed :: TermLike term tyname name uni fun => Term tyname name uni fun ann -> term ann

type TermDef term tyname name uni ann = Def (VarDecl tyname name uni ann) (term ann)

A term definition as a variable.

type TypeDef tyname uni ann = Def (TyVarDecl tyname ann) (Type tyname uni ann)

A type definition as a type variable.

data FunctionType tyname uni ann

The type of a PLC function.

Constructors

FunctionType

Fields

_functionTypeAnn :: ann

An annotation.

_functionTypeDom :: Type tyname uni ann

The domain of a function.

_functionTypeCod :: Type tyname uni ann

The codomain of the function.

data FunctionDef term tyname name uni fun ann

A PLC function.

Constructors

FunctionDef

Fields

_functionDefAnn :: ann

An annotation.

_functionDefName :: name

The name of a function.

_functionDefType :: FunctionType tyname uni ann

The type of the function.

_functionDefTerm :: term ann

The definition of the function.

functionTypeToType :: FunctionType tyname uni ann -> Type tyname uni ann

Convert a FunctionType to the corresponding Kind.

functionDefToType :: FunctionDef term tyname name uni fun ann -> Type tyname uni ann

Get the type of a FunctionDef.

functionDefVarDecl :: FunctionDef term tyname name uni fun ann -> VarDecl tyname name uni ann

Convert a FunctionDef to a VarDecl. I.e. ignore the actual term.

mkFunctionDef :: ann -> name -> Type tyname uni ann -> term ann -> Maybe (FunctionDef term tyname name uni fun ann)

Make a FunctionDef. Return Nothing if the provided type is not functional.

mkImmediateLamAbs

Arguments

:: TermLike term tyname name uni fun

=> ann

-> TermDef term tyname name uni ann

-> term ann
The body of the let, possibly referencing the name.

-> term ann

Make a "let-binding" for a term as an immediately applied lambda abstraction.

mkImmediateTyAbs

Arguments

:: TermLike term tyname name uni fun

=> ann

-> TypeDef tyname uni ann

-> term ann
The body of the let, possibly referencing the name.

-> term ann

Make a "let-binding" for a type as an immediately instantiated type abstraction. Note: the body must be a value.

mkIterTyForall :: [TyVarDecl tyname ann] -> Type tyname uni ann -> Type tyname uni ann

Universally quantify a list of names.

mkIterTyLam :: [TyVarDecl tyname ann] -> Type tyname uni ann -> Type tyname uni ann

Lambda abstract a list of names.

mkIterApp

Arguments

:: TermLike term tyname name uni fun

=> ann

-> term ann
f

-> [term ann]
[ x0 ... xn ]

-> term ann
[f x0 ... xn ]

Make an iterated application.

mkIterTyFun :: ann -> [Type tyname uni ann] -> Type tyname uni ann -> Type tyname uni ann

Make an iterated function type.

mkIterLamAbs :: TermLike term tyname name uni fun => [VarDecl tyname name uni ann] -> term ann -> term ann

Lambda abstract a list of names.

mkIterInst

Arguments

:: TermLike term tyname name uni fun

=> ann

-> term ann
a

-> [Type tyname uni ann]
 [ x0 ... xn ]

-> term ann
{ a x0 ... xn }

Make an iterated instantiation.

mkIterTyAbs :: TermLike term tyname name uni fun => [TyVarDecl tyname ann] -> term ann -> term ann

Type abstract a list of names.

mkIterTyApp

Arguments

:: ann

-> Type tyname uni ann
f

-> [Type tyname uni ann]
[ x0 ... xn ]

-> Type tyname uni ann
[ f x0 ... xn ]

Make an iterated type application.

mkIterKindArrow :: ann -> [Kind ann] -> Kind ann -> Kind ann

Make an iterated function kind.