PlutusCore.Name

Types

data Name

A Name represents variables/names in Plutus Core.

Constructors

Name

Fields

_nameText :: Text

The identifier name, for use in error messages.

_nameUnique :: Unique

A Unique assigned to the name, allowing for cheap comparisons in the compiler.

newtype TyName

We use a newtype to enforce separation between names used for types and those used for terms.

Constructors

TyName

Fields

unTyName :: Name

data Named a

Constructors

Named

Fields

_namedString :: Text

_namedValue :: a

newtype Unique

A unique identifier

Constructors

Unique

Fields

unUnique :: Int

newtype TypeUnique

The unique of a type-level name.

Constructors

TypeUnique

Fields

unTypeUnique :: Unique

newtype TermUnique

The unique of a term-level name.

Constructors

TermUnique

Fields

unTermUnique :: Unique

class HasText a where

Types which have a textual name attached to them.

Methods

theText :: Lens' a Text

class Coercible unique Unique => HasUnique a unique | a -> unique where

Types which have a Unique attached to them, mostly names.

Minimal complete definition

Nothing

Methods

unique :: Lens' a unique

The default implementation of HasUnique for newtypes.

default unique :: (Wrapped a, HasUnique (Unwrapped a) unique', Coercible unique' unique) => Lens' a unique

theUnique :: HasUnique name unique => Lens' name Unique

A lens focused on the Unique of a name.

newtype UniqueMap unique a

A mapping from uniques to values of type a.

Constructors

UniqueMap

Fields

unUniqueMap :: IntMap a

Functions

insertByUnique :: Coercible unique Unique => unique -> a -> UniqueMap unique a -> UniqueMap unique a

Insert a value by a unique.

insertByName :: HasUnique name unique => name -> a -> UniqueMap unique a -> UniqueMap unique a

Insert a value by the unique of a name.

insertByNameIndex :: (HasUnique name unique1, Coercible unique2 Unique) => name -> a -> UniqueMap unique2 a -> UniqueMap unique2 a

Insert a value by the index of the unique of a name. Unlike insertByUnique and insertByName, this function does not provide any static guarantees, so you can for example insert by a type-level name in a map from term-level uniques.

insertNamed :: (HasText name, HasUnique name unique) => name -> a -> UniqueMap unique (Named a) -> UniqueMap unique (Named a)

Insert a named value by the index of the unique of the name.

fromFoldable :: Foldable f => (i -> a -> UniqueMap unique a -> UniqueMap unique a) -> f (i, a) -> UniqueMap unique a

Convert a Foldable into a UniqueMap using the given insertion function.

fromUniques :: Foldable f => Coercible Unique unique => f (unique, a) -> UniqueMap unique a

Convert a Foldable with uniques into a UniqueMap.

fromNames :: Foldable f => HasUnique name unique => f (name, a) -> UniqueMap unique a

Convert a Foldable with names into a UniqueMap.

lookupUnique :: Coercible unique Unique => unique -> UniqueMap unique a -> Maybe a

Look up a value by a unique.

lookupName :: HasUnique name unique => name -> UniqueMap unique a -> Maybe a

Look up a value by the unique of a name.

lookupNameIndex :: (HasUnique name unique1, Coercible unique2 Unique) => name -> UniqueMap unique2 a -> Maybe a

Look up a value by the index of the unique of a name. Unlike lookupUnique and lookupName, this function does not provide any static guarantees, so you can for example look up a type-level name in a map from term-level uniques.

mapNameString :: (Text -> Text) -> Name -> Name

Apply a function to the string representation of a Name.

mapTyNameString :: (Text -> Text) -> TyName -> TyName

Apply a function to the string representation of a TyName.

isEmpty :: UniqueMap unique a -> Bool