PlutusCore.Parser

Description

Parsers for PLC terms in DefaultUni.

Documentation

parseProgram :: ByteString -> Either (ParseErrorBundle Text ParseError) (Program TyName Name DefaultUni DefaultFun SourcePos)

Parse a PLC program. The resulting program will have fresh names. The underlying monad must be capable of handling any parse errors.

parseTerm :: ByteString -> Either (ParseErrorBundle Text ParseError) (Term TyName Name DefaultUni DefaultFun SourcePos)

Parse a PLC term. The resulting program will have fresh names. The underlying monad must be capable of handling any parse errors.

parseType :: ByteString -> Either (ParseErrorBundle Text ParseError) (Type TyName DefaultUni SourcePos)

Parse a PLC type. The resulting program will have fresh names. The underlying monad must be capable of handling any parse errors.

data ParseError

An error encountered during parsing.

Constructors

UnknownBuiltinType Text SourcePos

BuiltinTypeNotAStar Text SourcePos

UnknownBuiltinFunction Text SourcePos

InvalidBuiltinConstant Text Text SourcePos