PlutusCore.Pretty

Basic types and functions

data Doc ann

The abstract data type Doc ann represents pretty documents that have been annotated with data of type ann.

More specifically, a value of type Doc represents a non-empty set of possible layouts of a document. The layout functions select one of these possibilities, taking into account things like the width of the output document.

The annotation is an arbitrary piece of data associated with (part of) a document. Annotations may be used by the rendering backends in order to display output differently, such as

color information (e.g. when rendering to the terminal)

mouseover text (e.g. when rendering to rich HTML)

whether to show something or not (to allow simple or detailed versions)

The simplest way to display a Doc is via the Show class.

>>> putStrLn (show (vsep ["hello", "world"]))
hello
world

class Pretty a where

Overloaded conversion to Doc.

Laws:

output should be pretty. :-)

Minimal complete definition

pretty

Methods

pretty :: a -> Doc ann

>>> pretty 1 <+> pretty "hello" <+> pretty 1.234
1 hello 1.234

prettyList :: [a] -> Doc ann

prettyList is only used to define the instance Pretty a => Pretty [a]. In normal circumstances only the pretty function is used.

>>> prettyList [1, 23, 456]
[1, 23, 456]

class PrettyBy config a where

A class for pretty-printing values in a configurable manner.

A basic example:

>>> data Case = UpperCase | LowerCase
>>> data D = D
>>> instance PrettyBy Case D where prettyBy UpperCase D = "D"; prettyBy LowerCase D = "d"
>>> prettyBy UpperCase D
D
>>> prettyBy LowerCase D
d

The library provides instances for common types like Integer or Bool, so you can't define your own PrettyBy SomeConfig Integer instance. And for the same reason you should not define instances like PrettyBy SomeAnotherConfig a for universally quantified a, because such an instance would overlap with the existing ones. Take for example

>>> data ViaShow = ViaShow
>>> instance Show a => PrettyBy ViaShow a where prettyBy ViaShow = pretty . show

with such an instance prettyBy ViaShow (1 :: Int) throws an error about overlapping instances:

• Overlapping instances for PrettyBy ViaShow Int
    arising from a use of ‘prettyBy’
  Matching instances:
    instance PrettyDefaultBy config Int => PrettyBy config Int
    instance [safe] Show a => PrettyBy ViaShow a

There's a newtype provided specifically for the purpose of defining a PrettyBy instance for any a: PrettyAny. Read its docs for details on when you might want to use it.

The PrettyBy instance for common types is defined in a way that allows to override default pretty-printing behaviour, read the docs of HasPrettyDefaults for details.

Minimal complete definition

Nothing

Methods

prettyBy :: config -> a -> Doc ann

Pretty-print a value of type a the way a config specifies it. The default implementation of prettyBy is in terms of pretty, defaultPrettyFunctorBy or defaultPrettyBifunctorBy depending on the kind of the data type that you're providing an instance for. For example, the default implementation of prettyBy for a monomorphic type is going to be "ignore the config and call pretty over the value":

>>> newtype N = N Int deriving newtype (Pretty)
>>> instance PrettyBy () N
>>> prettyBy () (N 42)
42

The default implementation of prettyBy for a Functor is going to be in terms of defaultPrettyFunctorBy:

>>> newtype N a = N a deriving stock (Functor) deriving newtype (Pretty)
>>> instance PrettyBy () a => PrettyBy () (N a)
>>> prettyBy () (N (42 :: Int))
42

It's fine for the data type to have a phantom parameter as long as the data type is still a Functor (i.e. the parameter has to be of kind Type). Then defaultPrettyFunctorBy is used again:

>>> newtype N a = N Int deriving stock (Functor) deriving newtype (Pretty)
>>> instance PrettyBy () (N b)
>>> prettyBy () (N 42)
42

If the data type has a single parameter of any other kind, then it's not a functor and so like in the monomorphic case pretty is used:

>>> newtype N (b :: Bool) = N Int deriving newtype (Pretty)
>>> instance PrettyBy () (N b)
>>> prettyBy () (N 42)
42

Same applies to a data type with two parameters: if both the parameters are of kind Type, then the data type is assumed to be a Bifunctor and hence defaultPrettyBifunctorBy is used. If the right parameter is of kind Type and the left parameter is of any other kind, then we fallback to assuming the data type is a Functor and defining prettyBy as defaultPrettyFunctorBy. If both the parameters are not of kind Type, we fallback to implementing prettyBy in terms of pretty like in the monomorphic case.

Note that in all those cases a Pretty instance for the data type has to already exist, so that we can derive a PrettyBy one in terms of it. If it doesn't exist or if your data type is not supported (for example, if it has three or more parameters of kind Type), then you'll need to provide the implementation manually.

prettyListBy :: config -> [a] -> Doc ann

prettyListBy is used to define the default PrettyBy instance for [a] and NonEmpty a. In normal circumstances only the prettyBy function is used. The default implementation of prettyListBy is in terms of defaultPrettyFunctorBy.

newtype IgnorePrettyConfig a

A newtype wrapper around a whose point is to provide a PrettyBy config instance for anything that has a Pretty instance.

Constructors

IgnorePrettyConfig

Fields

unIgnorePrettyConfig :: a

data AttachPrettyConfig config a

A config together with some value. The point is to provide a Pretty instance for anything that has a PrettyBy config instance.

Constructors

AttachPrettyConfig !config !a

class Render str where

A class for rendering Docs as string types.

Methods

render :: Doc ann -> str

Render a Doc as a string type.

display :: forall str a. (Pretty a, Render str) => a -> str

Pretty-print and render a value as a string type.

displayBy :: forall str a config. (PrettyBy config a, Render str) => config -> a -> str

Pretty-print and render a value as a string type in a configurable way.

Defaults

prettyPlcDef :: PrettyPlc a => a -> Doc ann

Pretty-print a value in the default mode using the classic view.

displayPlcDef :: (PrettyPlc a, Render str) => a -> str

Render a value to String in the default mode using the classic view.

prettyPlcDebug :: PrettyPlc a => a -> Doc ann

Pretty-print a value in the debug mode using the classic view.

displayPlcDebug :: (PrettyPlc a, Render str) => a -> str

Render a value to String in the debug mode using the classic view.

Global configuration

data CondensedErrors

Whether to pretty-print PLC errors in full or with some information omitted.

Constructors

CondensedErrorsYes

CondensedErrorsNo

newtype PrettyConfigPlcOptions

Options for pretty-printing PLC entities.

Constructors

PrettyConfigPlcOptions

Fields

_pcpoCondensedErrors :: CondensedErrors

data PrettyConfigPlcStrategy

Strategy for pretty-printing PLC entities.

Constructors

PrettyConfigPlcClassic (PrettyConfigClassic PrettyConfigName)

PrettyConfigPlcReadable (PrettyConfigReadable PrettyConfigName)

data PrettyConfigPlc

Global configuration used for pretty-printing PLC entities.

Constructors

PrettyConfigPlc

Fields

_pcpOptions :: PrettyConfigPlcOptions

_pcpStrategy :: PrettyConfigPlcStrategy

type PrettyPlc = PrettyBy PrettyConfigPlc

The "pretty-printable PLC entity" constraint.

defPrettyConfigPlcOptions :: PrettyConfigPlcOptions

The PrettyConfigPlcOptions used by default: print errors in full.

defPrettyConfigPlcClassic :: PrettyConfigPlcOptions -> PrettyConfigPlc

The PrettyConfigPlc used by default: use the classic view and print neither Uniques, nor name attachments.

debugPrettyConfigPlcClassic :: PrettyConfigPlcOptions -> PrettyConfigPlc

The PrettyConfigPlc used for debugging: use the classic view and print Uniques, but not name attachments.

defPrettyConfigPlcReadable :: PrettyConfigPlcOptions -> PrettyConfigPlc

The PrettyConfigPlc used by default and for readability: use the refined view and print neither Uniques, nor name attachments.

debugPrettyConfigPlcReadable :: PrettyConfigPlcOptions -> PrettyConfigPlc

The PrettyConfigPlc used for debugging and readability: use the refined view and print Uniques, but not name attachments.

Custom functions for PLC types.

prettyPlcClassicDef :: PrettyPlc a => a -> Doc ann

Pretty-print a PLC value in the default mode using the classic view.

prettyPlcClassicDebug :: PrettyPlc a => a -> Doc ann

Pretty-print a PLC value in the debug mode using the classic view.

prettyPlcReadableDef :: PrettyPlc a => a -> Doc ann

Pretty-print a PLC value in the default mode using the readable view.

prettyPlcReadableDebug :: PrettyPlc a => a -> Doc ann

Pretty-print a PLC value in the debug mode using the readable view.

prettyPlcCondensedErrorBy :: PrettyPlc a => (PrettyConfigPlcOptions -> PrettyConfigPlc) -> a -> Doc ann

Pretty-print a PLC value using the condensed way (see CondensedErrors) of pretty-printing PLC errors (in case there are any).

displayPlcCondensedErrorClassic :: (PrettyPlc a, Render str) => a -> str

Render an error to String in the condensed manner using the classic view.

Names

newtype PrettyConfigName

A config that determines how to pretty-print a PLC name.

Constructors

PrettyConfigName

Fields

_pcnShowsUnique :: Bool

Whether to show the Unique of a name or not.

class HasPrettyConfigName config where

A class of configs from which a PrettyConfigName can be extracted.

Methods

toPrettyConfigName :: config -> PrettyConfigName

defPrettyConfigName :: PrettyConfigName

The PrettyConfigName used by default: print neither Uniques, nor name attachments.

debugPrettyConfigName :: PrettyConfigName

The PrettyConfigName used for debugging: print Uniques, but not name attachments.

Classic view

data PrettyConfigClassic configName

Configuration for the classic pretty-printing.

Constructors

PrettyConfigClassic

Fields

_pccConfigName :: configName

How to pretty-print names.

_pccDisplayAnn :: Bool

Whether to display annotations.

type PrettyClassicBy configName = PrettyBy (PrettyConfigClassic configName)

The "classically pretty-printable" constraint.

type PrettyClassic = PrettyClassicBy PrettyConfigName

consAnnIf :: Pretty ann => PrettyConfigClassic configName -> ann -> [Doc dann] -> [Doc dann]

Add a pretty-printed annotation to a list of Docs if the given config enables pretty-printing of annotations.

prettyClassicDef :: PrettyClassic a => a -> Doc ann

Pretty-print a value in the default mode using the classic view.

prettyClassicDebug :: PrettyClassic a => a -> Doc ann

Pretty-print a value in the debug mode using the classic view.

Readable view

data ShowKinds

Constructors

ShowKindsYes

ShowKindsNo

data PrettyConfigReadable configName

Configuration for the readable pretty-printing.

Constructors

PrettyConfigReadable

Fields

_pcrConfigName :: configName

_pcrRenderContext :: RenderContext

_pcrShowKinds :: ShowKinds

type PrettyReadableBy configName = PrettyBy (PrettyConfigReadable configName)

The "readably pretty-printable" constraint.

type PrettyReadable = PrettyReadableBy PrettyConfigName

topPrettyConfigReadable :: configName -> ShowKinds -> PrettyConfigReadable configName

A PrettyConfigReadable with the fixity specified to topFixity.

botPrettyConfigReadable :: configName -> ShowKinds -> PrettyConfigReadable configName

A PrettyConfigReadable with the fixity specified to botFixity.

Utils

prettyBytes :: ByteString -> Doc ann

data ConstConfig

Constructors

ConstConfig

type PrettyConst = PrettyBy ConstConfig

prettyConst :: PrettyConst a => a -> Doc ann