PlutusCore.Rename

Description

The user-facing API of the renamer.

Documentation

class Rename a where

The class of things that can be renamed. I.e. things that are capable of satisfying the global uniqueness condition.

Methods

rename :: MonadQuote m => a -> m a

Rename Uniques so that they're globally unique. In case there are any free variables, they must be left untouched and bound variables must not get renamed to free ones. Must always assign new names to bound variables, so that rename can be used for alpha-renaming as well.

data Dupable a

Dupable a is isomorphic to a, but the only way to extract the a is via liftDupable which renames the stored value along the way. This type is used whenever

preserving global uniqueness is required

some value may be used multiple times

so we annotate such a value with Dupable and call liftDupable at each usage, which ensures global conditions is preserved.

liftDupable :: (MonadQuote m, Rename a) => Dupable a -> m a

Extract the value stored in a Dupable a and rename it.