PlutusCore.Subst

Documentation

substTyVarA :: Applicative f => (tyname -> f (Maybe (Type tyname uni ann))) -> Type tyname uni ann -> f (Type tyname uni ann)

Applicatively replace a type variable using the given function.

substVarA :: Applicative f => (name -> f (Maybe (Term tyname name uni fun ann))) -> Term tyname name uni fun ann -> f (Term tyname name uni fun ann)

Applicatively replace a variable using the given function.

substTyVar :: (tyname -> Maybe (Type tyname uni ann)) -> Type tyname uni ann -> Type tyname uni ann

Replace a type variable using the given function.

substVar :: (name -> Maybe (Term tyname name uni fun ann)) -> Term tyname name uni fun ann -> Term tyname name uni fun ann

Replace a variable using the given function.

termSubstNamesM :: Monad m => (name -> m (Maybe (Term tyname name uni fun ann))) -> Term tyname name uni fun ann -> m (Term tyname name uni fun ann)

Naively monadically substitute names using the given function (i.e. do not substitute binders).

termSubstTyNamesM :: Monad m => (tyname -> m (Maybe (Type tyname uni ann))) -> Term tyname name uni fun ann -> m (Term tyname name uni fun ann)

Naively monadically substitute type names using the given function (i.e. do not substitute binders).

typeSubstTyNamesM :: Monad m => (tyname -> m (Maybe (Type tyname uni ann))) -> Type tyname uni ann -> m (Type tyname uni ann)

Naively monadically substitute type names (i.e. do not substitute binders). INLINE is important here because the function is too polymorphic (determined from profiling)

termSubstNames :: (name -> Maybe (Term tyname name uni fun ann)) -> Term tyname name uni fun ann -> Term tyname name uni fun ann

Naively substitute names using the given function (i.e. do not substitute binders).

termSubstTyNames :: (tyname -> Maybe (Type tyname uni ann)) -> Term tyname name uni fun ann -> Term tyname name uni fun ann

Naively substitute type names using the given function (i.e. do not substitute binders).

typeSubstTyNames :: (tyname -> Maybe (Type tyname uni ann)) -> Type tyname uni ann -> Type tyname uni ann

Naively substitute type names (i.e. do not substitute binders).

typeSubstClosedType :: Eq tyname => tyname -> Type tyname uni () -> Type tyname uni () -> Type tyname uni ()

Substitute the given closed Kind for the given type variable in the given Kind. Does not descend under binders that bind the same variable as the one we're substituting for (since from there that variable is no longer free). The resulting Term may and likely will not satisfy global uniqueness.

termSubstClosedType :: Eq tyname => tyname -> Type tyname uni () -> Term tyname name uni fun () -> Term tyname name uni fun ()

Substitute the given closed Kind for the given type variable in the given Term. Does not descend under binders that bind the same variable as the one we're substituting for (since from there that variable is no longer free). The resulting Term may and likely will not satisfy global uniqueness.

termSubstClosedTerm :: Eq name => name -> Term tyname name uni fun () -> Term tyname name uni fun () -> Term tyname name uni fun ()

Substitute the given closed Term for the given term variable in the given Term. Does not descend under binders that bind the same variable as the one we're substituting for (since from there that variable is no longer free). The resulting Term may and likely will not satisfy global uniqueness.

fvTerm :: Ord name => Traversal' (Term tyname name uni fun ann) name

Get all the free term variables in a term.

ftvTerm :: Ord tyname => Traversal' (Term tyname name uni fun ann) tyname

Get all the free type variables in a term.

ftvTy :: Ord tyname => Traversal' (Type tyname uni ann) tyname

Get all the free type variables in a type.

ftvTyCtx :: Ord tyname => Set tyname -> Traversal' (Type tyname uni ann) tyname

vTerm :: Fold (Term tyname name uni fun ann) name

Get all the term variables in a term.

tvTerm :: Fold (Term tyname name uni fun ann) tyname

Get all the type variables in a term.

tvTy :: Fold (Type tyname uni ann) tyname

Get all the type variables in a type.