PlutusIR

AST

data Term tyname name uni fun a

Constructors

Let a Recursivity (NonEmpty (Binding tyname name uni fun a)) (Term tyname name uni fun a)

Var a name

TyAbs a tyname (Kind a) (Term tyname name uni fun a)

LamAbs a name (Type tyname uni a) (Term tyname name uni fun a)

Apply a (Term tyname name uni fun a) (Term tyname name uni fun a)

Constant a (Some (ValueOf uni))

Builtin a fun

TyInst a (Term tyname name uni fun a) (Type tyname uni a)

Error a (Type tyname uni a)

IWrap a (Type tyname uni a) (Type tyname uni a) (Term tyname name uni fun a)

Unwrap a (Term tyname name uni fun a)

termSubterms :: Traversal' (Term tyname name uni fun a) (Term tyname name uni fun a)

Get all the direct child Terms of the given Term, including those within Bindings.

termSubtypes :: Traversal' (Term tyname name uni fun a) (Type tyname uni a)

Get all the direct child Kinds of the given Term, including those within Bindings.

termBindings :: Traversal' (Term tyname name uni fun a) (Binding tyname name uni fun a)

Get all the direct child Bindings of the given Term.

termAnn :: Term tyname name uni fun a -> a

data Type tyname uni ann

A Kind assigned to expressions.

Constructors

TyVar ann tyname

TyFun ann (Type tyname uni ann) (Type tyname uni ann)

TyIFix ann (Type tyname uni ann) (Type tyname uni ann)
Fix-point type, for constructing self-recursive types

TyForall ann tyname (Kind ann) (Type tyname uni ann)

TyBuiltin ann (SomeTypeIn uni)
Builtin type

TyLam ann tyname (Kind ann) (Type tyname uni ann)

TyApp ann (Type tyname uni ann) (Type tyname uni ann)

typeSubtypes :: Traversal' (Type tyname uni ann) (Type tyname uni ann)

Get all the direct child Kinds of the given Kind.

data Datatype tyname name uni a

Constructors

Datatype a (TyVarDecl tyname a) [TyVarDecl tyname a] name [VarDecl tyname name uni a]

datatypeNameString :: Datatype TyName name uni a -> String

datatypeSubtypes :: Traversal' (Datatype tyname name uni a) (Type tyname uni a)

Get all the direct child Kinds of the given Datatype.

data Kind ann

Constructors

Type ann

KindArrow ann (Kind ann) (Kind ann)

data Recursivity

Each multi-let-group has to be marked with its scoping: * NonRec: the identifiers introduced by this multi-let are only linearly-scoped, i.e. an identifier cannot refer to itself or later-introduced identifiers of the group. * Rec: an identifiers introduced by this multi-let group can use all other multi-lets of the same group (including itself), thus permitting (mutual) recursion.

Constructors

NonRec

Rec

data Strictness

Constructors

NonStrict

Strict

data Binding tyname name uni fun a

Constructors

TermBind a Strictness (VarDecl tyname name uni a) (Term tyname name uni fun a)

TypeBind a (TyVarDecl tyname a) (Type tyname uni a)

DatatypeBind a (Datatype tyname name uni a)

bindingSubterms :: Traversal' (Binding tyname name uni fun a) (Term tyname name uni fun a)

Get all the direct child Terms of the given Binding.

bindingSubtypes :: Traversal' (Binding tyname name uni fun a) (Type tyname uni a)

Get all the direct child Kinds of the given Binding.

bindingIds :: (HasUnique tyname TypeUnique, HasUnique name TermUnique) => Traversal1' (Binding tyname name uni fun a) Unique

All the identifiers/names introduced by this binding In case of a datatype-binding it has multiple identifiers: the type, constructors, match function

data Program tyname name uni fun ann

Constructors

Program

Fields

_progAnn :: ann

_progTerm :: Term tyname name uni fun ann

applyProgram :: Monoid a => Program tyname name uni fun a -> Program tyname name uni fun a -> Program tyname name uni fun a

newtype TyName

We use a newtype to enforce separation between names used for types and those used for terms.

Constructors

TyName

Fields

unTyName :: Name

data Name

A Name represents variables/names in Plutus Core.

Constructors

Name

Fields

_nameText :: Text

The identifier name, for use in error messages.

_nameUnique :: Unique

A Unique assigned to the name, allowing for cheap comparisons in the compiler.

data VarDecl tyname name uni ann

A "variable declaration", i.e. a name and a type for a variable.

Constructors

VarDecl

Fields

_varDeclAnn :: ann

_varDeclName :: name

_varDeclType :: Type tyname uni ann

data TyVarDecl tyname ann

A "type variable declaration", i.e. a name and a kind for a type variable.

Constructors

TyVarDecl

Fields

_tyVarDeclAnn :: ann

_tyVarDeclName :: tyname

_tyVarDeclKind :: Kind ann

varDeclNameString :: VarDecl tyname Name uni a -> String

tyVarDeclNameString :: TyVarDecl TyName a -> String