PlutusIR.Analysis.Dependencies

Description

Functions for computing the dependency graph of variables within a term or type. A "dependency" between two nodes "A depends on B" means that B cannot be removed from the program without also removing A.

Constructors

Variable Unique

Root

runTermDeps :: (DepGraph g, ToBuiltinMeaning uni fun) => BuiltinsInfo uni fun -> VarsInfo tyname name uni a -> Term tyname name uni fun a -> g

Compute the dependency graph of a Term. The Root node will correspond to the term itself.

For example, the graph of [(let (nonrec) (vardecl x t) y) [x z]] is ROOT -> x ROOT -> z x -> y x -> t