PlutusIR.Compiler

Documentation

compileTerm :: Compiling m e uni fun a => Term TyName Name uni fun a -> m (PLCTerm uni fun a)

compileToReadable :: (Compiling m e uni fun a, b ~ Provenance a) => Term TyName Name uni fun a -> m (Term TyName Name uni fun b)

The 1st half of the PIR compiler pipeline up to floating/merging the lets. We stop momentarily here to give a chance to the tx-plugin to dump a "readable" version of pir (i.e. floated).

compileReadableToPlc :: (Compiling m e uni fun a, b ~ Provenance a) => Term TyName Name uni fun b -> m (PLCTerm uni fun a)

The 2nd half of the PIR compiler pipeline. Compiles a Term into a PLC Term, by removing/translating step-by-step the PIR's language constructs to PLC. Note: the result *does* have globally unique names.

type Compiling m e uni fun a = (Monad m, MonadReader (CompilationCtx uni fun a) m, AsTypeError e (Term TyName Name uni fun ()) uni fun (Provenance a), AsTypeErrorExt e uni (Provenance a), AsError e uni fun (Provenance a), MonadError e m, MonadQuote m, Ord a, Typecheckable uni fun, GEq uni, Pretty fun, Closed uni, Pretty (SomeTypeIn uni), uni `Everywhere` PrettyConst, Pretty a)

data Error uni fun a

Constructors

CompilationError a Text
A generic compilation error.

UnsupportedError a Text
An error relating specifically to an unsupported feature.

PLCError (Error uni fun a)
An error from running some PLC function, lifted into this error type for convenience.

PLCTypeError (TypeError (Term TyName Name uni fun ()) uni fun a)

PIRTypeError (TypeErrorExt uni a)

class AsError r uni fun a | r -> uni fun a where

Minimal complete definition

_Error

Methods

_Error :: Prism' r (Error uni fun a)

_CompilationError :: Prism' r (a, Text)

_UnsupportedError :: Prism' r (a, Text)

_PLCError :: Prism' r (Error uni fun a)

_PLCTypeError :: Prism' r (TypeError (Term TyName Name uni fun ()) uni fun a)

_PIRTypeError :: Prism' r (TypeErrorExt uni a)

class AsTypeError r term uni fun ann | r -> term uni fun ann where

Minimal complete definition

_TypeError

Methods

_TypeError :: Prism' r (TypeError term uni fun ann)

_KindMismatch :: Prism' r (ann, Type TyName uni (), Kind (), Kind ())

_TypeMismatch :: Prism' r (ann, term, Type TyName uni (), Normalized (Type TyName uni ()))

_TyNameMismatch :: Prism' r (ann, TyName, TyName)

_NameMismatch :: Prism' r (ann, Name, Name)

_FreeTypeVariableE :: Prism' r (ann, TyName)

_FreeVariableE :: Prism' r (ann, Name)

_UnknownBuiltinFunctionE :: Prism' r (ann, fun)

class AsTypeErrorExt r uni ann | r -> uni ann where

Minimal complete definition

_TypeErrorExt

Methods

_TypeErrorExt :: Prism' r (TypeErrorExt uni ann)

_MalformedDataConstrResType :: Prism' r (ann, Type TyName uni ann)

data Provenance a

Indicates where a value comes from.

This is either an original annotation or a pieces of context explaining how the term relates to a previous Provenance. We also provide NoProvenance for convenience.

The provenance should always be just the original annotation, if we have one. It should only be another kind of provenance if we're in the process of generating some term that doesn't correspond directly to a term in the original AST.

Constructors

Original a

LetBinding Recursivity (Provenance a)

TermBinding String (Provenance a)

TypeBinding String (Provenance a)

DatatypeComponent DatatypeComponent (Provenance a)

MultipleSources (Set (Provenance a))
Added for accumulating difference provenances when floating lets

data DatatypeComponent

Constructors

Constructor

ConstructorType

Destructor

DestructorType

DatatypeType

PatternFunctor

noProvenance :: Provenance a

data CompilationOpts a

coOptimize :: forall a. Lens' (CompilationOpts a) Bool

coPedantic :: forall a. Lens' (CompilationOpts a) Bool

coVerbose :: forall a. Lens' (CompilationOpts a) Bool

coDebug :: forall a. Lens' (CompilationOpts a) Bool

coMaxSimplifierIterations :: forall a. Lens' (CompilationOpts a) Int

coDoSimplifierUnwrapCancel :: forall a. Lens' (CompilationOpts a) Bool

coDoSimplifierBeta :: forall a. Lens' (CompilationOpts a) Bool

coDoSimplifierInline :: forall a. Lens' (CompilationOpts a) Bool

coInlineHints :: forall a a. Lens (CompilationOpts a) (CompilationOpts a) (InlineHints Name (Provenance a)) (InlineHints Name (Provenance a))

coProfile :: forall a. Lens' (CompilationOpts a) Bool

defaultCompilationOpts :: CompilationOpts a

data CompilationCtx uni fun a

ccOpts :: forall uni fun a. Lens' (CompilationCtx uni fun a) (CompilationOpts a)

ccEnclosing :: forall uni fun a. Lens' (CompilationCtx uni fun a) (Provenance a)

ccTypeCheckConfig :: forall uni fun a uni. Lens (CompilationCtx uni fun a) (CompilationCtx uni fun a) (Maybe (PirTCConfig uni fun)) (Maybe (PirTCConfig uni fun))

data PirTCConfig uni fun

extending the plc typecheck config with AllowEscape

Constructors

PirTCConfig

Fields

_pirConfigTCConfig :: TypeCheckConfig uni fun

_pirConfigAllowEscape :: AllowEscape

data AllowEscape

Extra flag to be passed in the TypeCheckM Reader context, to signal if the PIR expression currently being typechecked is at the top-level and thus its type can escape, or nested and thus not allowed to escape.

Constructors

YesEscape

NoEscape

toDefaultCompilationCtx :: Default (BuiltinVersion fun) => TypeCheckConfig uni fun -> CompilationCtx uni fun a