PlutusIR.Compiler.Datatype

Description

Functions for compiling let-bound PIR datatypes into PLC.

Documentation

compileDatatype :: Compiling m e uni fun a => Recursivity -> PIRTerm uni fun a -> Datatype TyName Name uni (Provenance a) -> m (PIRTerm uni fun a)

Compile a Datatype bound with the given body.

compileDatatypeDefs :: MonadQuote m => Recursivity -> Datatype TyName Name uni (Provenance a) -> m (Def (TyVarDecl TyName (Provenance a)) (PLCRecType uni fun a), [Def (VarDecl TyName Name uni (Provenance a)) (PIRTerm uni fun a)], Def (VarDecl TyName Name uni (Provenance a)) (PIRTerm uni fun a))

Compile a Datatype to a triple of type-constructor, data-constructors, destructor definitions.

compileRecDatatypes :: Compiling m e uni fun a => PIRTerm uni fun a -> NonEmpty (Datatype TyName Name uni (Provenance a)) -> m (PIRTerm uni fun a)

funTyArgs :: Type tyname uni a -> [Type tyname uni a]

Get the argument types of a function type. funTyArgs (A->B->C) = [A, B]

funResultType :: Type tyname uni a -> Type tyname uni a

Get the result type of a function. If not a function, then is the same as id funResultType (A->B->C) = C funResultType (X) = X

mkDatatypeValueType :: a -> Datatype TyName Name uni a -> Type TyName uni a

The type of a datatype-value is of the form `[TyCon tyarg1 tyarg2 ... tyargn]`

mkDestructorTy :: Type TyName uni a -> Datatype TyName Name uni a -> Type TyName uni a

Make the type of a destructor for a Datatype. mkDestructorTy functor of List List = forall (a :: *) . (List a) -> (functor of List) = forall (a :: *) . (List a) -> (forall (out_List :: *) . (out_List -> (a -> List a -> out_List) -> out_List))

mkScottTy :: ann -> Datatype TyName Name uni ann -> TyName -> Type TyName uni ann

Make the "Scott-encoded" type for a Datatype, with type variables free. mkScottTy Maybe = forall out_Maybe. out_Maybe -> (a -> out_Maybe) -> out_Maybe

replaceFunTyTarget :: Type tyname uni a -> Type tyname uni a -> Type tyname uni a

replaceFunTyTarget X (A->..->Z) = (A->..->X)

resultTypeName :: MonadQuote m => Datatype TyName Name uni a -> m TyName