PlutusIR.Compiler.Types

Documentation

data AllowEscape

Extra flag to be passed in the TypeCheckM Reader context, to signal if the PIR expression currently being typechecked is at the top-level and thus its type can escape, or nested and thus not allowed to escape.

Constructors

YesEscape

NoEscape

data PirTCConfig uni fun

extending the plc typecheck config with AllowEscape

Constructors

PirTCConfig

Fields

_pirConfigTCConfig :: TypeCheckConfig uni fun

_pirConfigAllowEscape :: AllowEscape

data DatatypeStyle

What style to use when encoding datatypes. Generally, SumsOfProducts is superior, unless you are targeting an old Plutus Core language version.

See Note [Encoding of datatypes]

Constructors

ScottEncoding

SumsOfProducts

BuiltinCasing
A temporary data type style used to make a couple of V3 ledger-api-test tests pass before we can support casing on values of built-in types in newer protocol versions and merge this into SumsOfProducts (which is what controls whether Case is available or not).

data SharedName

We generate some shared definitions compilation, this datatype defines the "keys" for those definitions.

Constructors

FixpointCombinator Quote Name