PlutusIR.Core.Instance.Flat

Orphan instances

Flat Strictness

Instance details

Methods

encode :: Strictness -> Encoding

decode :: Get Strictness

size :: Strictness -> NumBits -> NumBits

Flat Recursivity

Instance details

Methods

encode :: Recursivity -> Encoding

decode :: Get Recursivity

size :: Recursivity -> NumBits -> NumBits

(Closed uni, Everywhere uni Flat, Flat fun, Flat a, Flat tyname, Flat name) => Flat (Program tyname name uni fun a)

Instance details

Methods

encode :: Program tyname name uni fun a -> Encoding

decode :: Get (Program tyname name uni fun a)

size :: Program tyname name uni fun a -> NumBits -> NumBits

(Closed uni, Everywhere uni Flat, Flat fun, Flat a, Flat tyname, Flat name) => Flat (Term tyname name uni fun a)

Instance details

Methods

encode :: Term tyname name uni fun a -> Encoding

decode :: Get (Term tyname name uni fun a)

size :: Term tyname name uni fun a -> NumBits -> NumBits

(Closed uni, Everywhere uni Flat, Flat fun, Flat a, Flat tyname, Flat name) => Flat (Binding tyname name uni fun a)

Instance details

Methods

encode :: Binding tyname name uni fun a -> Encoding

decode :: Get (Binding tyname name uni fun a)

size :: Binding tyname name uni fun a -> NumBits -> NumBits

(Closed uni, Everywhere uni Flat, Flat a, Flat tyname, Flat name, Flat fun) => Flat (Datatype tyname name uni fun a)

Instance details

Methods

encode :: Datatype tyname name uni fun a -> Encoding

decode :: Get (Datatype tyname name uni fun a)

size :: Datatype tyname name uni fun a -> NumBits -> NumBits