PlutusIR.Core.Plated

Documentation

termSubterms :: Traversal' (Term tyname name uni fun a) (Term tyname name uni fun a)

Get all the direct child Terms of the given Term, including those within Bindings.

termSubtermsDeep :: Fold (Term tyname name uni fun ann) (Term tyname name uni fun ann)

Get all the transitive child Terms of the given Term.

termSubtypes :: Traversal' (Term tyname name uni fun a) (Type tyname uni a)

Get all the direct child Kinds of the given Term, including those within Bindings.

termSubtypesDeep :: Fold (Term tyname name uni fun ann) (Type tyname uni ann)

Get all the transitive child Kinds of the given Term.

termSubkinds :: Traversal' (Term tyname name uni fun ann) (Kind ann)

Get all the direct child Kinds of the given Term.

termBindings :: Traversal' (Term tyname name uni fun a) (Binding tyname name uni fun a)

Get all the direct child Bindings of the given Term.

termVars :: Traversal' (Term tyname name uni fun ann) name

Get all the direct child 'name a's of the given Term from Vars.

typeSubtypes :: Traversal' (Type tyname uni ann) (Type tyname uni ann)

Get all the direct child Kinds of the given Kind.

typeSubtypesDeep :: Fold (Type tyname uni ann) (Type tyname uni ann)

Get all the transitive child Kinds of the given Kind.

typeSubkinds :: Traversal' (Type tyname uni ann) (Kind ann)

Get all the direct child Kinds of the given Kind.

typeUniques :: HasUniques (Type tyname uni ann) => Traversal' (Type tyname uni ann) Unique

Get all the direct child Uniques of the given Kind from binders TyVars.

typeUniquesDeep :: HasUniques (Type tyname uni ann) => Fold (Type tyname uni ann) Unique

Get all the transitive child Uniques of the given Kind.

datatypeSubtypes :: Traversal' (Datatype tyname name uni a) (Type tyname uni a)

Get all the direct child Kinds of the given Datatype.

datatypeSubkinds :: Traversal' (Datatype tyname name uni a) (Kind a)

Get all the direct child Kinds of the given Datatype.

datatypeTyNames :: Traversal' (Datatype tyname name uni a) tyname

Get all the type-names introduces by a datatype

bindingSubterms :: Traversal' (Binding tyname name uni fun a) (Term tyname name uni fun a)

Get all the direct child Terms of the given Binding.

bindingSubtypes :: Traversal' (Binding tyname name uni fun a) (Type tyname uni a)

Get all the direct child Kinds of the given Binding.

bindingSubkinds :: Traversal' (Binding tyname name uni fun a) (Kind a)

Get all the direct child Kinds of the given Binding.

bindingNames :: Traversal' (Binding tyname name uni fun a) name

Get all the names introduces by a binding

bindingTyNames :: Traversal' (Binding tyname name uni fun a) tyname

Get all the type-names introduces by a binding

bindingIds :: (HasUnique tyname TypeUnique, HasUnique name TermUnique) => Traversal1' (Binding tyname name uni fun a) Unique

All the identifiers/names introduced by this binding In case of a datatype-binding it has multiple identifiers: the type, constructors, match function

termUniques :: HasUniques (Term tyname name uni fun ann) => Traversal' (Term tyname name uni fun ann) Unique

Get all the direct child Uniques of the given Term (including the type-level ones).

termUniquesDeep :: HasUniques (Term tyname name uni fun ann) => Fold (Term tyname name uni fun ann) Unique

Get all the transitive child Uniques of the given Term (including the type-level ones).