PlutusIR.Generators.QuickCheck.ShrinkTerms

fixupTerm :: TypeCtx -> DefaultFun ())]

Shrink a typed term in a type and term context. NOTE: if you want to understand what's going on in this function it's a good idea to look at how we do this for types first (it's a lot simpler).

shrinkClosedTypedTerm :: (DefaultFun ())]