PlutusIR.Parser

Description

Parsers for PIR terms in DefaultUni.

Documentation

parse :: Parser a -> String -> Text -> Either (ParseErrorBundle Text ParseError) a

parseQuoted :: Parser a -> String -> Text -> Quote (Either (ParseErrorBundle Text ParseError) a)

program :: Parser (Program TyName Name DefaultUni DefaultFun SourcePos)

pType :: Parser PType

Parser for PType.

pTerm :: Parser PTerm

type Parser = ParsecT ParseError Text (StateT ParserState Quote)

data SourcePos

The data type SourcePos represents source positions. It contains the name of the source file, a line number, and a column number. Source line and column positions change intensively during parsing, so we need to make them strict to avoid memory leaks.