PlutusIR.Pass.Test

numTestsForPassProp :: Int

An appropriate number of tests for a compiler pass property, so that we get some decent exploration of the program space. If you also take other arguments, then consider multiplying this up in order to account for the larger space.

runTestPass :: (Pass on generated Termss, setting up the typechecking config and throwing errors.

testPassProp' :: forall m tyname name a prop. (DefaultFun a) -> Property

A version of testPassProp with more control, allowing some pre-processing of the term, and a more specific "exit" function.

Orphan instances

Arbitrary (DefaultFun]