PlutusIR.Purity

Documentation

isPure :: ToBuiltinMeaning uni fun => BuiltinVersion fun -> (name -> Strictness) -> Term tyname name uni fun a -> Bool

Will evaluating this term have side effects (looping or error)? This is slightly wider than the definition of a value, as it includes applications that are known to be pure, as well as things that can't be returned from the machine (as they'd be ill-scoped).

firstEffectfulTerm :: Term tyname name uni fun a -> Maybe (Term tyname name uni fun a)

Try to identify the first sub term which will be evaluated in the given term and which could have an effect. Nothing indicates that we don't know, this function is conservative.

asBuiltinApp :: Term tyname name uni fun a -> Maybe (BuiltinApp tyname name uni fun a)

View a Term as a BuiltinApp if possible.

isSaturated :: forall tyname name uni fun a. ToBuiltinMeaning uni fun => BuiltinVersion fun -> BuiltinApp tyname name uni fun a -> Maybe Bool

Is the given BuiltinApp saturated? Returns Nothing if something is badly wrong and we can't tell.

data BuiltinApp tyname name uni fun a

A (not necessarily saturated) builtin application, consisting of the builtin and the arguments it has been applied to.

Constructors

BuiltinApp fun [Arg tyname name uni fun a]

data Arg tyname name uni fun a

An argument taken by a builtin: could be a term of a type.

Constructors

TypeArg (Type tyname uni a)

TermArg (Term tyname name uni fun a)