PlutusIR.Subst

Documentation

uniquesTerm :: HasUniques (Term tyname name uni fun ann) => Term tyname name uni fun ann -> Set Unique

uniquesType :: HasUniques (Type tyname uni ann) => Type tyname uni ann -> Set Unique

Get all the uniques in a type.

fvTerm :: Ord name => Term tyname name uni fun ann -> Set name

Get all the free term variables in a PIR term.

ftvTerm :: Ord tyname => Term tyname name uni fun ann -> Set tyname

Get all the free type variables in a PIR term.

fvBinding :: Ord name => Binding tyname name uni fun ann -> Set name

Get all the free variables in a PIR single let-binding.

ftvBinding :: forall tyname name uni fun ann. Ord tyname => Recursivity -> Binding tyname name uni fun ann -> Set tyname

Get all the free type variables in a PIR single let-binding.

ftvTy :: Ord tyname => Type tyname uni ann -> Set tyname

Get all the free type variables in a type.