PlutusIR.Transform.DeadCode

Description

Optimization passes for removing dead code, mainly dead let bindings.

Documentation

removeDeadBindings :: (HasUnique name TermUnique, ToBuiltinMeaning uni fun, MonadQuote m) => BuiltinVersion fun -> Term TyName name uni fun a -> m (Term TyName name uni fun a)

Remove all the dead let bindings in a term.