PlutusIR.Transform.Inline

Description

An inlining pass.

Note that there is (essentially) a copy of this in the UPLC inliner, and the two should be KEPT IN SYNC, so if you make changes here please either make them in the other one too or add to the comment that summarises the differences.

Documentation

inline :: forall tyname name uni fun a m. ExternalConstraints tyname name uni fun m => InlineHints name a -> BuiltinVersion fun -> Term tyname name uni fun a -> m (Term tyname name uni fun a)

Inline simple bindings. Relies on global uniqueness, and preserves it. See Note [Inlining and global uniqueness]

newtype InlineHints name a

Constructors

InlineHints

Fields

shouldInline :: a -> name -> Bool