PlutusIR.Transform.Inline.Utils

Description

Types and their functions, and general utility (including heuristics) for inlining.

Documentation

newtype InlineTerm tyname name uni fun a

Substitution range, SubstRng in the paper but no Susp case. See Note [Inlining approach and 'Secrets of the GHC Inliner']

Constructors

Done (Dupable (Term tyname name uni fun a))

newtype TermEnv tyname name uni fun a

Term substitution, Subst in the paper. A map of unprocessed variable and its substitution range.

Constructors

TermEnv

Fields

_unTermEnv :: UniqueMap TermUnique (InlineTerm tyname name uni fun a)

newtype TypeEnv tyname uni a

Type substitution, similar to TermEnv but for types. A map of unprocessed type variable and its substitution range.

Constructors

TypeEnv

Fields

_unTypeEnv :: UniqueMap TypeUnique (Dupable (Type tyname uni a))

data Subst tyname name uni fun a

Substitution for both terms and types. Similar to Subst in the paper, but the paper only has terms.

Constructors

Subst

Fields

_termEnv :: TermEnv tyname name uni fun a

_typeEnv :: TypeEnv tyname uni a

unTermEnv :: forall tyname name uni fun a tyname name uni fun a. Iso (TermEnv tyname name uni fun a) (TermEnv tyname name uni fun a) (UniqueMap TermUnique (InlineTerm tyname name uni fun a)) (UniqueMap TermUnique (InlineTerm tyname name uni fun a))

unTypeEnv :: forall tyname uni a tyname uni a. Iso (TypeEnv tyname uni a) (TypeEnv tyname uni a) (UniqueMap TypeUnique (Dupable (Type tyname uni a))) (UniqueMap TypeUnique (Dupable (Type tyname uni a)))

typeEnv :: forall tyname name uni fun a. Lens' (Subst tyname name uni fun a) (TypeEnv tyname uni a)

termEnv :: forall tyname name uni fun a name fun. Lens (Subst tyname name uni fun a) (Subst tyname name uni fun a) (TermEnv tyname name uni fun a) (TermEnv tyname name uni fun a)

lookupTerm

Arguments

:: HasUnique name TermUnique

=> name
The name of the variable.

-> Subst tyname name uni fun a
The substitution.

-> Maybe (InlineTerm tyname name uni fun a)

Look up the unprocessed variable in the substitution.

extendTerm

Arguments

:: HasUnique name TermUnique

=> name
The name of the variable.

-> InlineTerm tyname name uni fun a
The substitution range.

-> Subst tyname name uni fun a
The substitution.

-> Subst tyname name uni fun a

Insert the unprocessed variable into the substitution.

lookupType :: HasUnique tyname TypeUnique => tyname -> Subst tyname name uni fun a -> Maybe (Dupable (Type tyname uni a))

Look up the unprocessed type variable in the substitution.

isTypeSubstEmpty :: Subst tyname name uni fun a -> Bool

Check if the type substitution is empty.

extendType

Arguments

:: HasUnique tyname TypeUnique

=> tyname
The name of the type variable.

-> Type tyname uni a
Its type.

-> Subst tyname name uni fun a
The substitution.

-> Subst tyname name uni fun a

Insert the unprocessed type variable into the substitution.

type ExternalConstraints tyname name uni fun m = (HasUnique name TermUnique, HasUnique tyname TypeUnique, Eq name, ToBuiltinMeaning uni fun, MonadQuote m)

type InliningConstraints tyname name uni fun = (HasUnique name TermUnique, HasUnique tyname TypeUnique, Eq name, ToBuiltinMeaning uni fun)

data InlineInfo name fun a

Constructors

InlineInfo

Fields

_iiStrictnessMap :: StrictnessMap

Is it strict? Only needed for PIR, not UPLC

_iiUsages :: Usages

how many times is it used?

_iiHints :: InlineHints name a

have we explicitly been told to inline.

_iiBuiltinVer :: BuiltinVersion fun

the builtin version.

iiUsages :: forall name fun a. Lens' (InlineInfo name fun a) Usages

iiStrictnessMap :: forall name fun a. Lens' (InlineInfo name fun a) StrictnessMap

iiHints :: forall name fun a name a. Lens (InlineInfo name fun a) (InlineInfo name fun a) (InlineHints name a) (InlineHints name a)

iiBuiltinVer :: forall name fun a fun. Lens (InlineInfo name fun a) (InlineInfo name fun a) (BuiltinVersion fun) (BuiltinVersion fun)

type InlineM tyname name uni fun a = ReaderT (InlineInfo name fun a) (StateT (Subst tyname name uni fun a) Quote)

The monad the inliner runs in.

checkPurity :: forall tyname name uni fun a. InliningConstraints tyname name uni fun => Term tyname name uni fun a -> InlineM tyname name uni fun a Bool

Check if term is pure. See Note [Inlining and purity]

nameUsedAtMostOnce :: forall tyname name uni fun a. InliningConstraints tyname name uni fun => name -> InlineM tyname name uni fun a Bool

effectSafe

Arguments

:: forall tyname name uni fun a. InliningConstraints tyname name uni fun

=> Term tyname name uni fun a

-> Strictness

-> name

-> Bool
is it pure?

-> InlineM tyname name uni fun a Bool

acceptable :: Term tyname name uni fun a -> InlineM tyname name uni fun a Bool

Should we inline? Should only inline things that won't duplicate work or code. See Note [Inlining approach and 'Secrets of the GHC Inliner']

costIsAcceptable :: Term tyname name uni fun a -> Bool

Is the cost increase (in terms of evaluation work) of inlining a variable whose RHS is the given term acceptable?

sizeIsAcceptable :: Term tyname name uni fun a -> Bool

Is the size increase (in the AST) of inlining a variable whose RHS is the given term acceptable?

trivialType :: Type tyname uni a -> Bool

Is this a an utterly trivial type which might as well be inlined?