PlutusIR.Transform.Substitute

Description

Implements naive substitution functions for replacing type and term variables.

Documentation

substVar :: (name -> Maybe (Term tyname name uni fun a)) -> Term tyname name uni fun a -> Maybe (Term tyname name uni fun a)

Replace a variable using the given function.

substTyVar :: (tyname -> Maybe (Type tyname uni ann)) -> Type tyname uni ann -> Type tyname uni ann

Replace a type variable using the given function.

typeSubstTyNames :: (tyname -> Maybe (Type tyname uni ann)) -> Type tyname uni ann -> Type tyname uni ann

Naively substitute type names (i.e. do not substitute binders).

termSubstNames :: (name -> Maybe (Term tyname name uni fun a)) -> Term tyname name uni fun a -> Term tyname name uni fun a

Naively substitute names using the given functions (i.e. do not substitute binders).

termSubstTyNames :: (tyname -> Maybe (Type tyname uni a)) -> Term tyname name uni fun a -> Term tyname name uni fun a

Naively substitute type names using the given functions (i.e. do not substitute binders).

bindingSubstNames :: (name -> Maybe (Term tyname name uni fun a)) -> Binding tyname name uni fun a -> Binding tyname name uni fun a

Naively substitute names using the given functions (i.e. do not substitute binders).

bindingSubstTyNames :: (tyname -> Maybe (Type tyname uni a)) -> Binding tyname name uni fun a -> Binding tyname name uni fun a

Naively substitute type names using the given functions (i.e. do not substitute binders).