Prettyprinter.Custom

Documentation

brackets' :: Doc a -> Doc a

This prints a document enclosed by brackets, possibly indenting the output on a new line if it does not fit.

braces' :: Doc a -> Doc a

This prints a document enclosed by braces, possibly indenting the output on a new line if it does not fit.

parens' :: Doc a -> Doc a

This prints a document enclosed by parentheses, aligning the opening and closing parentheses.

sexp :: Doc a -> [Doc a] -> Doc a

Print a "sexp", i.e. something like "(keyword arg1 ... argN)".