UntypedPlutusCore.Core

Documentation

module UntypedPlutusCore.Core.Type

module UntypedPlutusCore.Core.Plated

termTagWidth :: NumBits

Using 4 bits to encode term tags.

encodeTermTag :: Word8 -> Encoding

decodeTermTag :: Get Word8

encodeTerm :: forall name uni fun ann. (Closed uni, uni `Everywhere` Flat, Flat fun, Flat ann, Flat name, Flat (Binder name)) => Term name uni fun ann -> Encoding

decodeTerm :: forall name uni fun ann. (Closed uni, uni `Everywhere` Flat, Flat fun, Flat ann, Flat name, Flat (Binder name)) => Version -> (fun -> Maybe String) -> Get (Term name uni fun ann)

sizeTerm :: forall name uni fun ann. (Closed uni, uni `Everywhere` Flat, Flat fun, Flat ann, Flat name, Flat (Binder name)) => Term name uni fun ann -> NumBits -> NumBits

encodeProgram :: forall name uni fun ann. (Closed uni, uni `Everywhere` Flat, Flat fun, Flat ann, Flat name, Flat (Binder name)) => Program name uni fun ann -> Encoding

An encoder for programs.

It is not easy to use this correctly with flat. The simplest thing is to go via the instance for UnrestrictedProgram, which uses this

decodeProgram :: forall name uni fun ann. (Closed uni, uni `Everywhere` Flat, Flat fun, Flat ann, Flat name, Flat (Binder name)) => (fun -> Maybe String) -> Get (Program name uni fun ann)

sizeProgram :: forall name uni fun ann. (Closed uni, uni `Everywhere` Flat, Flat fun, Flat ann, Flat name, Flat (Binder name)) => Program name uni fun ann -> NumBits -> NumBits

newtype UnrestrictedProgram name uni fun ann

A program that can be serialized without any restrictions, e.g. on the set of allowable builtins or term constructs. It is generally safe to use this newtype for serializing, but it should only be used for deserializing in tests.

Constructors

UnrestrictedProgram

Fields

unUnrestrictedProgram :: Program name uni fun ann

splitParams :: Term name uni fun a -> ([name], Term name uni fun a)

Strips off lambda binders.

splitApplication :: Term name uni fun a -> (Term name uni fun a, [(a, Term name uni fun a)])

Strip off arguments