UntypedPlutusCore.Core.Plated

Documentation

termConstants :: Traversal' (Term name uni fun ann) (Some (ValueOf uni))

Get all the direct constants of the given Term from Constants.

termBinds :: Traversal' (Term name uni fun ann) name

Get all the direct child 'name a's of the given Term from LamAbses.

termVars :: Traversal' (Term name uni fun ann) name

Get all the direct child 'name a's of the given Term from Vars.

termUniques :: HasUniques (Term name uni fun ann) => Traversal' (Term name uni fun ann) Unique

Get all the direct child Uniques of the given Term.

termSubterms :: Traversal' (Term name uni fun ann) (Term name uni fun ann)

Get all the direct child Terms of the given Term.

termConstantsDeep :: Fold (Term name uni fun ann) (Some (ValueOf uni))

Get all the transitive child Constants of the given Term.

termSubtermsDeep :: Fold (Term name uni fun ann) (Term name uni fun ann)

Get all the transitive child Terms of the given Term.

termUniquesDeep :: HasUniques (Term name uni fun ann) => Fold (Term name uni fun ann) Unique

Get all the transitive child Uniques of the given Term.