UntypedPlutusCore.Core.Type

Documentation

type family UniOf a :: Type -> Type

Extract the universe from a type.

data Version ann

Version of Plutus Core to be used for the program.

Constructors

Version ann Natural Natural Natural

newtype Binder name

This is a wrapper to mark the place where the binder is introduced (i.e. LamAbs/TyAbs) and not where it is actually used (TyVar/Var..). This marking allows us to skip the (de)serialization of binders at LamAbs/TyAbs positions iff name is DeBruijn-encoded (level or index). See for example the instance of Flat

Constructors

Binder

Fields

unBinder :: name

data Term name uni fun ann

The type of Untyped Plutus Core terms. Mirrors the type of Typed Plutus Core terms except

all types are removed

IWrap and Unwrap are removed

type abstractions are replaced with Delay

type instantiations are replaced with Force

The latter two are due to the fact that we don't have value restriction in Typed Plutus Core and hence a computation can be stuck expecting only a single type argument for the computation to become unstuck. Therefore we can't just silently remove type abstractions and instantiations and need to replace them with something else that also blocks evaluation (in order for the semantics of an erased program to match with the semantics of the original typed one). Delay and Force serve exactly this purpose.

Constructors

Var !ann !name

LamAbs !ann !name !(Term name uni fun ann)

Apply !ann !(Term name uni fun ann) !(Term name uni fun ann)

Force !ann !(Term name uni fun ann)

Delay !ann !(Term name uni fun ann)

Constant !ann !(Some (ValueOf uni))

Builtin !ann !fun

Error !ann

data Program name uni fun ann

A Program is simply a Term coupled with a Version of the core language.

Constructors

Program

Fields

_progAnn :: ann

_progVer :: Version ann

_progTerm :: Term name uni fun ann

progAnn :: forall name uni fun ann. Lens' (Program name uni fun ann) ann

progVer :: forall name uni fun ann. Lens' (Program name uni fun ann) (Version ann)

progTerm :: forall name uni fun ann name uni fun. Lens (Program name uni fun ann) (Program name uni fun ann) (Term name uni fun ann) (Term name uni fun ann)

bindFunM :: Monad m => (ann -> fun -> m (Term name uni fun' ann)) -> Term name uni fun ann -> m (Term name uni fun' ann)

bindFun :: (ann -> fun -> Term name uni fun' ann) -> Term name uni fun ann -> Term name uni fun' ann

mapFun :: (ann -> fun -> fun') -> Term name uni fun ann -> Term name uni fun' ann

termAnn :: Term name uni fun ann -> ann

Return the outermost annotation of a Term.

erase :: Term tyname name uni fun ann -> Term name uni fun ann

Erase a Typed Plutus Core term to its untyped counterpart.

eraseProgram :: Program tyname name uni fun ann -> Program name uni fun ann

Erase a Typed Plutus Core Program to its untyped counterpart.

data UVarDecl name ann

An untyped "variable declaration", i.e. a name for a variable.

Constructors

UVarDecl

Fields

_uvarDeclAnn :: ann

_uvarDeclName :: name

uvarDeclName :: forall name ann name. Lens (UVarDecl name ann) (UVarDecl name ann) name name

uvarDeclAnn :: forall name ann ann. Lens (UVarDecl name ann) (UVarDecl name ann) ann ann