UntypedPlutusCore.DeBruijn

Description

Support for using de Bruijn indices for term names.

Documentation

newtype Index

A relative index used for de Bruijn identifiers.

Constructors

Index Word64

class HasIndex a where

Methods

index :: Lens' a Index

newtype DeBruijn

A term name as a de Bruijn index, without the name string.

Constructors

DeBruijn

Fields

dbnIndex :: Index

data NamedDeBruijn

A term name as a de Bruijn index.

Constructors

NamedDeBruijn

Fields

ndbnString :: !Text

ndbnIndex :: !Index

newtype FakeNamedDeBruijn

A wrapper around nameddebruijn that must hold the invariant of name=fakeName.

Constructors

FakeNamedDeBruijn NamedDeBruijn

data FreeVariableError

We cannot do a correct translation to or from de Bruijn indices if the program is not well-scoped. So we throw an error in such a case.

Constructors

FreeUnique Unique

FreeIndex Index

class AsFreeVariableError r where

Minimal complete definition

_FreeVariableError

Methods

_FreeVariableError :: Prism' r FreeVariableError

_FreeUnique :: Prism' r Unique

_FreeIndex :: Prism' r Index

deBruijnTerm :: (AsFreeVariableError e, MonadError e m) => Term Name uni fun ann -> m (Term NamedDeBruijn uni fun ann)

Convert a Term with Names into a Term with DeBruijns. Will throw an error if a free variable is encountered.

unDeBruijnTerm :: (MonadQuote m, AsFreeVariableError e, MonadError e m) => Term NamedDeBruijn uni fun ann -> m (Term Name uni fun ann)

Convert a Term with DeBruijns into a Term with Names. Will throw an error if a free variable is encountered.

unNameDeBruijn :: NamedDeBruijn -> DeBruijn

fakeNameDeBruijn :: DeBruijn -> NamedDeBruijn

unsafe api, use with care

deBruijnTermWith :: Monad m => (Unique -> ReaderT Levels m Index) -> Term Name uni fun ann -> m (Term NamedDeBruijn uni fun ann)

Takes a "handler" function to execute when encountering free variables.

unDeBruijnTermWith :: MonadQuote m => (Index -> ReaderT Levels m Unique) -> Term NamedDeBruijn uni fun ann -> m (Term Name uni fun ann)

Takes a "handler" function to execute when encountering free variables.

freeIndexAsConsistentLevel :: (MonadReader Levels m, MonadState (Map Level Unique) m, MonadQuote m) => Index -> m Unique

A different implementation of a handler, where "free" debruijn indices do not throw an error but are instead gracefully converted to fresh uniques. These generated uniques remain free; i.e. if the original term was open, it will remain open after applying this handler. These generated free uniques are consistent across the open term (by using a state cache).

deBruijnInitIndex :: Index

The LamAbs index (for debruijn indices) and the starting level of DeBruijn monad