UntypedPlutusCore.Evaluation.Machine.Cek

Description

The API to the CEK machine.

Running the machine

runCek :: (uni `Everywhere` ExMemoryUsage, Ix fun, PrettyUni uni fun) => MachineParameters CekMachineCosts CekValue uni fun -> ExBudgetMode cost uni fun -> EmitterMode uni fun -> Term Name uni fun () -> (Either (CekEvaluationException Name uni fun) (Term Name uni fun ()), cost, [Text])

Evaluate a term using the CEK machine with logging enabled and keep track of costing. A wrapper around the internal runCek to debruijn input and undebruijn output. *THIS FUNCTION IS PARTIAL if the input term contains free variables*

runCekDeBruijn :: (uni `Everywhere` ExMemoryUsage, Ix fun, PrettyUni uni fun) => MachineParameters CekMachineCosts CekValue uni fun -> ExBudgetMode cost uni fun -> EmitterMode uni fun -> Term NamedDeBruijn uni fun () -> (Either (CekEvaluationException NamedDeBruijn uni fun) (Term NamedDeBruijn uni fun ()), cost, [Text])

Evaluate a term using the CEK machine and keep track of costing, logging is optional.

runCekNoEmit :: (uni `Everywhere` ExMemoryUsage, Ix fun, PrettyUni uni fun) => MachineParameters CekMachineCosts CekValue uni fun -> ExBudgetMode cost uni fun -> Term Name uni fun () -> (Either (CekEvaluationException Name uni fun) (Term Name uni fun ()), cost)

Evaluate a term using the CEK machine with logging disabled and keep track of costing. *THIS FUNCTION IS PARTIAL if the input term contains free variables*

unsafeRunCekNoEmit :: (GShow uni, Typeable uni, Closed uni, uni `EverywhereAll` '[ExMemoryUsage, PrettyConst], Ix fun, Pretty fun, Typeable fun) => MachineParameters CekMachineCosts CekValue uni fun -> ExBudgetMode cost uni fun -> Term Name uni fun () -> (EvaluationResult (Term Name uni fun ()), cost)

Unsafely evaluate a term using the CEK machine with logging disabled and keep track of costing. May throw a CekMachineException. *THIS FUNCTION IS PARTIAL if the input term contains free variables*

evaluateCek :: (uni `Everywhere` ExMemoryUsage, Ix fun, PrettyUni uni fun) => EmitterMode uni fun -> MachineParameters CekMachineCosts CekValue uni fun -> Term Name uni fun () -> (Either (CekEvaluationException Name uni fun) (Term Name uni fun ()), [Text])

Evaluate a term using the CEK machine with logging enabled. *THIS FUNCTION IS PARTIAL if the input term contains free variables*

evaluateCekNoEmit :: (uni `Everywhere` ExMemoryUsage, Ix fun, PrettyUni uni fun) => MachineParameters CekMachineCosts CekValue uni fun -> Term Name uni fun () -> Either (CekEvaluationException Name uni fun) (Term Name uni fun ())

Evaluate a term using the CEK machine with logging disabled. *THIS FUNCTION IS PARTIAL if the input term contains free variables*

unsafeEvaluateCek :: (GShow uni, Typeable uni, Closed uni, uni `EverywhereAll` '[ExMemoryUsage, PrettyConst], Ix fun, Pretty fun, Typeable fun) => EmitterMode uni fun -> MachineParameters CekMachineCosts CekValue uni fun -> Term Name uni fun () -> (EvaluationResult (Term Name uni fun ()), [Text])

Evaluate a term using the CEK machine with logging enabled. May throw a CekMachineException. *THIS FUNCTION IS PARTIAL if the input term contains free variables*

unsafeEvaluateCekNoEmit :: (GShow uni, Typeable uni, Closed uni, uni `EverywhereAll` '[ExMemoryUsage, PrettyConst], Ix fun, Pretty fun, Typeable fun) => MachineParameters CekMachineCosts CekValue uni fun -> Term Name uni fun () -> EvaluationResult (Term Name uni fun ())

Evaluate a term using the CEK machine with logging disabled. May throw a CekMachineException. *THIS FUNCTION IS PARTIAL if the input term contains free variables*

data EvaluationResult a

The parameterized type of results various evaluation engines return. On the PLC side this becomes (via makeKnown) either a call to Error or a value of the PLC counterpart of type a.

Constructors

EvaluationSuccess a

EvaluationFailure

extractEvaluationResult :: Either (EvaluationException user internal term) a -> Either (ErrorWithCause internal term) (EvaluationResult a)

Turn any UserEvaluationError into an EvaluationFailure.

unsafeExtractEvaluationResult :: (PrettyPlc internal, PrettyPlc term, Typeable internal, Typeable term) => Either (EvaluationException user internal term) a -> EvaluationResult a

Errors

data CekUserError

Constructors

CekOutOfExError ExRestrictingBudget
The final overspent (i.e. negative) budget.

CekEvaluationFailure
Error has been called or a builtin application has failed

data ErrorWithCause err cause

An error and (optionally) what caused it.

Constructors

ErrorWithCause

Fields

_ewcError :: err

_ewcCause :: Maybe cause

type CekEvaluationException name uni fun = EvaluationException CekUserError (MachineError fun) (Term name uni fun ())

The CEK machine-specific EvaluationException.

data EvaluationError user internal

The type of errors (all of them) which can occur during evaluation (some are used-caused, some are internal).

Constructors

InternalEvaluationError internal
Indicates bugs.

UserEvaluationError user
Indicates user errors.

Costing

data ExBudgetCategory fun

Constructors

BStep StepKind

BBuiltinApp fun

BStartup

newtype CekBudgetSpender uni fun s

The CEK machine is parameterized over a spendBudget function. This makes the budgeting machinery extensible and allows us to separate budgeting logic from evaluation logic and avoid branching on the union of all possible budgeting state types during evaluation.

Constructors

CekBudgetSpender

Fields

unCekBudgetSpender :: ExBudgetCategory fun -> ExBudget -> CekM uni fun s ()

newtype ExBudgetMode cost uni fun

A budgeting mode to execute the CEK machine in.

Constructors

ExBudgetMode

Fields

unExBudgetMode :: forall s. ST s (ExBudgetInfo cost uni fun s)

data StepKind

Constructors

BConst

BVar

BLamAbs

BApply

BDelay

BForce

BBuiltin

newtype CekExTally fun

For a detailed report on what costs how much + the same overall budget that Counting gives. The (derived) Monoid instance of CekExTally is the main piece of the machinery.

Constructors

CekExTally (MonoidalHashMap (ExBudgetCategory fun) ExBudget)

newtype CountingSt

For calculating the cost of execution by counting up using the Monoid instance of ExBudget.

Constructors

CountingSt ExBudget

data TallyingSt fun

Constructors

TallyingSt (CekExTally fun) ExBudget

newtype RestrictingSt

Constructors

RestrictingSt ExRestrictingBudget

data CekMachineCosts

Costs for evaluating AST nodes. Times should be specified in picoseconds, memory sizes in bytes.

Costing modes

counting :: ExBudgetMode CountingSt uni fun

For calculating the cost of execution.

tallying :: (Eq fun, Hashable fun) => ExBudgetMode (TallyingSt fun) uni fun

For a detailed report on what costs how much + the same overall budget that Counting gives.

restricting :: forall uni fun. PrettyUni uni fun => ExRestrictingBudget -> ExBudgetMode RestrictingSt uni fun

For execution, to avoid overruns.

restrictingEnormous :: PrettyUni uni fun => ExBudgetMode RestrictingSt uni fun

restricting instantiated at enormousBudget.

enormousBudget :: ExRestrictingBudget

When we want to just evaluate the program we use the Restricting mode with an enormous budget, so that evaluation costs of on-chain budgeting are reflected accurately in benchmarks.

Emitter modes

noEmitter :: EmitterMode uni fun

No emitter.

logEmitter :: EmitterMode uni fun

Emits log only.

logWithTimeEmitter :: EmitterMode uni fun

Emits log with timestamp.

logWithBudgetEmitter :: EmitterMode uni fun

Emits log with the budget.

Misc

data CekValue uni fun

Constructors

VCon !(Some (ValueOf uni))

VDelay (Term NamedDeBruijn uni fun ()) !(CekValEnv uni fun)

VLamAbs NamedDeBruijn (Term NamedDeBruijn uni fun ()) !(CekValEnv uni fun)

VBuiltin !fun (Term NamedDeBruijn uni fun ()) (CekValEnv uni fun) !(BuiltinRuntime (CekValue uni fun))

readKnownCek :: (uni `Everywhere` ExMemoryUsage, ReadKnown (Term Name uni fun ()) a, Ix fun, PrettyUni uni fun) => MachineParameters CekMachineCosts CekValue uni fun -> Term Name uni fun () -> Either (CekEvaluationException Name uni fun) a

Unlift a value using the CEK machine. *THIS FUNCTION IS PARTIAL if the input term contains free variables*

class Hashable a

The class of types that can be converted to a hash value.

Minimal implementation: hashWithSalt.

Note: the hash is not guaranteed to be stable across library versions, operating systems or architectures. For stable hashing use named hashes: SHA256, CRC32 etc.

If you are looking for Hashable instance in time package, check time-compat

type PrettyUni uni fun = (GShow uni, Closed uni, Pretty fun, Typeable uni, Typeable fun, Everywhere uni PrettyConst)

The set of constraints we need to be able to print things in universes, which we need in order to throw exceptions.