UntypedPlutusCore.Evaluation.Machine.Cek.CekMachineCosts

Documentation

data CekMachineCosts

Costs for evaluating AST nodes. Times should be specified in picoseconds, memory sizes in bytes.

Constructors

CekMachineCosts

Fields

cekStartupCost :: ExBudget

cekVarCost :: ExBudget

cekConstCost :: ExBudget

cekLamCost :: ExBudget

cekDelayCost :: ExBudget

cekForceCost :: ExBudget

cekApplyCost :: ExBudget

cekBuiltinCost :: ExBudget

Just the cost of evaluating a Builtin node, not the builtin itself. There's no entry for Error since we'll be exiting anyway; also, what would happen if calling Error caused the budget to be exceeded?

cekMachineCostsPrefix :: Text

The prefix of the field names in the CekMachineCosts type, used for extracting the CekMachineCosts component of the ledger's cost model parameters. See Note [Cost model parameters] in CostModelInterface.

unitCekMachineCosts :: CekMachineCosts