UntypedPlutusCore.Evaluation.Machine.Cek.Debug.Driver

Documentation

class Breakpointable ann bps | ann -> bps where

Leave abstract the types of annotation and breakpoints. The only thing the driver requires is an inclusion relation of breakpoints into the Annotation

Methods

hasBreakpoints :: ann -> bps -> Bool

data CekState uni fun ann

data Cmd bps

The commands that the driver may receive from the client (tui,cli,test,etc)

Constructors

Step
Instruct the driver to a *SINGLE* step. Note: No need to pass breakpoints here because the stepping granularity is *minimal*.

Continue bps
Instruct to multi-step until end-of-program or until breakpoint reached

Next bps
Instruct to multi-step over the function call at point or until breakpoint reached

Finish bps
Instruct to multi-step to end of current function or until breakpoint reached

runDriver :: forall uni fun ann bps m. (Breakpointable ann bps, MonadFree (DebugF uni fun ann bps) m) => NTerm uni fun ann -> m ()

Entrypoint of the driver

data DebugF uni fun ann bps a

The drivers's suspension functor

Constructors

InputF (Cmd bps -> a)
Await for the client (e.g. TUI) to tell what to do next (Cmd).

LogF String a
A generator of logging messages of the driver

StepF
An enumeratee of Driver State (generator+iteratee): Yield a state before doing a step, then await for a state to resume after the step. See Note [Stepping the driver].

Fields

(CekState uni fun ann)

yield with the current driver's state before running a step

(CekState uni fun ann -> a)

resume back with a state after the step interpretation | A generator of CekState to yield to client (e.g. TUI)

UpdateClientF (CekState uni fun ann) a

Reexport some functions for convenience

handleStep :: forall uni fun ann s c. (PrettyUni uni fun, GivenCekReqs uni fun ann s, c ~ CekState uni fun ann) => c -> CekM uni fun s c

The state transition function of the machine.

tryHandleStep :: forall uni fun ann s c. (PrettyUni uni fun, GivenCekReqs uni fun ann s, c ~ CekState uni fun ann) => c -> CekM uni fun s (Either (CekEvaluationException NamedDeBruijn uni fun) c)

As handleStep but captures any exception during a single-state transition

Note that we use the "pure" MonadError.tryError and not the MonadCatch.tryError. See Note [Throwing exceptions in ST]

class Monad m => MonadFree (f :: Type -> TYPE LiftedRep) (m :: Type -> Type) | m -> f

Monads provide substitution (fmap) and renormalization (join):

m >>= f = join (fmap f m)

A free Monad is one that does no work during the normalization step beyond simply grafting the two monadic values together.

[] is not a free Monad (in this sense) because join [[a]] smashes the lists flat.

On the other hand, consider:

data Tree a = Bin (Tree a) (Tree a) | Tip a

instance Monad Tree where
  return = Tip
  Tip a >>= f = f a
  Bin l r >>= f = Bin (l >>= f) (r >>= f)

This Monad is the free Monad of Pair:

data Pair a = Pair a a

And we could make an instance of MonadFree for it directly:

instance MonadFree Pair Tree where
   wrap (Pair l r) = Bin l r

Or we could choose to program with Free Pair instead of Tree and thereby avoid having to define our own Monad instance.

Moreover, Control.Monad.Free.Church provides a MonadFree instance that can improve the asymptotic complexity of code that constructs free monads by effectively reassociating the use of (>>=). You may also want to take a look at the kan-extensions package (http://hackage.haskell.org/package/kan-extensions).

See Free for a more formal definition of the free Monad for a Functor.

iterM :: (Functor f, Monad m) => (f (m a) -> m a) -> Free f a -> m a

Like iter for monadic values.

iterTM :: forall f (m :: Type -> Type) t a. (Functor f, Monad m, MonadTrans t, Monad (t m)) => (f (t m a) -> t m a) -> FreeT f m a -> t m a

Tear down a free monad transformer using iteration over a transformer.

partialIterT :: Monad m => Integer -> (forall a. f a -> m a) -> FreeT f m b -> FreeT f m b

partialIterT n phi m interprets first n layers of m using phi. This is sort of the opposite for cutoff.

Some examples (n ≥ 0):

partialIterT 0 _ m              ≡ m
partialIterT (n+1) phi . return ≡ return
partialIterT (n+1) phi . lift   ≡ lift
partialIterT (n+1) phi . wrap   ≡ join . lift . phi

cutoff :: forall (f :: Type -> Type) (m :: Type -> Type) a. (Functor f, Monad m) => Integer -> FreeT f m a -> FreeT f m (Maybe a)

Cuts off a tree of computations at a given depth. If the depth is 0 or less, no computation nor monadic effects will take place.

Some examples (n ≥ 0):

cutoff 0     _        ≡ return Nothing
cutoff (n+1) . return ≡ return . Just
cutoff (n+1) . lift   ≡ lift . liftM Just
cutoff (n+1) . wrap   ≡ wrap . fmap (cutoff n)

Calling retract . cutoff n is always terminating, provided each of the steps in the iteration is terminating.

data FreeT (f :: Type -> Type) (m :: Type -> TYPE LiftedRep) a

The "free monad transformer" for a functor f