UntypedPlutusCore.Evaluation.Machine.SteppableCek

Description

The API to the Steppable CEK machine. Provides the same interface to original CEK machine.

Running the machine

runCek :: ThrowableBuiltins uni fun => MachineParameters CekMachineCosts fun (CekValue uni fun ann) -> ExBudgetMode cost uni fun -> EmitterMode uni fun -> Term Name uni fun ann -> (Either (CekEvaluationException Name uni fun) (Term Name uni fun ()), cost, [Text])

Evaluate a term using the Steppable CEK machine with logging enabled and keep track of costing. A wrapper around the internal runCek to debruijn input and undebruijn output. *THIS FUNCTION IS PARTIAL if the input term contains free variables*

runCekDeBruijn :: ThrowableBuiltins uni fun => MachineParameters CekMachineCosts fun (CekValue uni fun ann) -> ExBudgetMode cost uni fun -> EmitterMode uni fun -> NTerm uni fun ann -> (Either (CekEvaluationException NamedDeBruijn uni fun) (NTerm uni fun ()), cost, [Text])

runCekNoEmit :: ThrowableBuiltins uni fun => MachineParameters CekMachineCosts fun (CekValue uni fun ann) -> ExBudgetMode cost uni fun -> Term Name uni fun ann -> (Either (CekEvaluationException Name uni fun) (Term Name uni fun ()), cost)

Evaluate a term using the Steppable CEK machine with logging disabled and keep track of costing. *THIS FUNCTION IS PARTIAL if the input term contains free variables*

unsafeRunCekNoEmit :: ThrowableBuiltins uni fun => MachineParameters CekMachineCosts fun (CekValue uni fun ann) -> ExBudgetMode cost uni fun -> Term Name uni fun ann -> (EvaluationResult (Term Name uni fun ()), cost)

Unsafely evaluate a term using the Steppable CEK machine with logging disabled -- and keep track of costing. May throw a CekMachineException. *THIS FUNCTION IS PARTIAL if the input term contains free variables*

evaluateCek :: ThrowableBuiltins uni fun => EmitterMode uni fun -> MachineParameters CekMachineCosts fun (CekValue uni fun ann) -> Term Name uni fun ann -> (Either (CekEvaluationException Name uni fun) (Term Name uni fun ()), [Text])

Evaluate a term using the Steppable CEK machine with logging enabled. *THIS FUNCTION IS PARTIAL if the input term contains free variables*

evaluateCekNoEmit :: ThrowableBuiltins uni fun => MachineParameters CekMachineCosts fun (CekValue uni fun ann) -> Term Name uni fun ann -> Either (CekEvaluationException Name uni fun) (Term Name uni fun ())

Evaluate a term using the Steppable CEK machine with logging disabled. *THIS FUNCTION IS PARTIAL if the input term contains free variables*

unsafeEvaluateCek :: ThrowableBuiltins uni fun => EmitterMode uni fun -> MachineParameters CekMachineCosts fun (CekValue uni fun ann) -> Term Name uni fun ann -> (EvaluationResult (Term Name uni fun ()), [Text])

Evaluate a term using the Steppable CEK machine with logging enabled. May throw a CekMachineException. *THIS FUNCTION IS PARTIAL if the input term contains free variables*

unsafeEvaluateCekNoEmit :: ThrowableBuiltins uni fun => MachineParameters CekMachineCosts fun (CekValue uni fun ann) -> Term Name uni fun ann -> EvaluationResult (Term Name uni fun ())

Evaluate a term using the Steppable CEK machine with logging disabled. May throw a CekMachineException. *THIS FUNCTION IS PARTIAL if the input term contains free variables*

data EvaluationResult a

The parameterized type of results various evaluation engines return. On the PLC side this becomes (via makeKnown) either a call to Error or a value of the PLC counterpart of type a.

Constructors

EvaluationSuccess !a

EvaluationFailure

extractEvaluationResult :: Either (EvaluationException user internal term) a -> Either (ErrorWithCause internal term) (EvaluationResult a)

Turn any UserEvaluationError into an EvaluationFailure.

unsafeExtractEvaluationResult :: (PrettyPlc internal, PrettyPlc term, Typeable internal, Typeable term) => Either (EvaluationException user internal term) a -> EvaluationResult a

Errors

data CekUserError

Constructors

CekOutOfExError !ExRestrictingBudget
The final overspent (i.e. negative) budget.

CekEvaluationFailure
Error has been called or a builtin application has failed

data ErrorWithCause err cause

An error and (optionally) what caused it.

Constructors

ErrorWithCause

Fields

_ewcError :: !err

_ewcCause :: !(Maybe cause)

type CekEvaluationException name uni fun = EvaluationException CekUserError (MachineError fun) (Term name uni fun ())

The CEK machine-specific EvaluationException.

data EvaluationError user internal

The type of errors (all of them) which can occur during evaluation (some are used-caused, some are internal).

Constructors

InternalEvaluationError !internal
Indicates bugs.

UserEvaluationError !user
Indicates user errors.

Costing

data ExBudgetCategory fun

Constructors

BStep StepKind

BBuiltinApp fun

BStartup

newtype CekBudgetSpender uni fun s

The CEK machine is parameterized over a spendBudget function. This makes the budgeting machinery extensible and allows us to separate budgeting logic from evaluation logic and avoid branching on the union of all possible budgeting state types during evaluation.

Constructors

CekBudgetSpender

Fields

unCekBudgetSpender :: ExBudgetCategory fun -> ExBudget -> CekM uni fun s ()

newtype ExBudgetMode cost uni fun

A budgeting mode to execute the CEK machine in.

Constructors

ExBudgetMode

Fields

unExBudgetMode :: forall s. ST s (ExBudgetInfo cost uni fun s)

data StepKind

Constructors

BConst

BVar

BLamAbs

BApply

BDelay

BForce

BBuiltin

BConstr

BCase

newtype CekExTally fun

For a detailed report on what costs how much + the same overall budget that Counting gives. The (derived) Monoid instance of CekExTally is the main piece of the machinery.

Constructors

CekExTally (MonoidalHashMap (ExBudgetCategory fun) ExBudget)

newtype CountingSt

For calculating the cost of execution by counting up using the Monoid instance of ExBudget.

Constructors

CountingSt ExBudget

data TallyingSt fun

Constructors

TallyingSt (CekExTally fun) ExBudget

newtype RestrictingSt

Constructors

RestrictingSt ExRestrictingBudget

type CekMachineCosts = CekMachineCostsBase Identity

Costs for evaluating AST nodes. Times should be specified in picoseconds, memory sizes in bytes.

Costing modes

counting :: ExBudgetMode CountingSt uni fun

For calculating the cost of execution.

tallying :: Hashable fun => ExBudgetMode (TallyingSt fun) uni fun

For a detailed report on what costs how much + the same overall budget that Counting gives.

restricting :: ThrowableBuiltins uni fun => ExRestrictingBudget -> ExBudgetMode RestrictingSt uni fun

For execution, to avoid overruns.

restrictingEnormous :: ThrowableBuiltins uni fun => ExBudgetMode RestrictingSt uni fun

restricting instantiated at enormousBudget.

enormousBudget :: ExRestrictingBudget

When we want to just evaluate the program we use the Restricting mode with an enormous budget, so that evaluation costs of on-chain budgeting are reflected accurately in benchmarks.

Emitter modes

noEmitter :: EmitterMode uni fun

No emitter.

logEmitter :: EmitterMode uni fun

Emits log only.

logWithTimeEmitter :: EmitterMode uni fun

Emits log with timestamp.

logWithBudgetEmitter :: EmitterMode uni fun

Emits log with the budget.

Misc

data CekValue uni fun ann

Constructors

VCon !(Some (ValueOf uni))

VDelay !(NTerm uni fun ann) !(CekValEnv uni fun ann)

VLamAbs !NamedDeBruijn !(NTerm uni fun ann) !(CekValEnv uni fun ann)

VBuiltin
A partial builtin application, accumulating arguments for eventual full application. We don't need a CekValEnv here unlike in the other constructors, because VBuiltin values always store their corresponding Terms fully discharged, see the comments at the call sites (search for VBuiltin).

Fields

!fun

So that we know, for what builtin we're calculating the cost. We can sneak this into BuiltinRuntime, so that we don't need to store it here, but somehow doing so was consistently slowing evaluation down by half a percent. Might be noise, might be not, but at least we know that removing this fun is not helpful anyway. See this commit reversing the change: https://github.com/IntersectMBO/plutus/pull/4778/commits/86a3e24ca3c671cc27c6f4344da2bcd14f961706

(NTerm uni fun ())

This must be lazy. It represents the fully discharged partial application of the builtin function that we're going to run when it's fully saturated. We need the Term to be able to return it in case full saturation is never achieved and a partial application needs to be returned in the result. The laziness is important, because the arguments are discharged values and discharging is expensive, so we don't want to do it unless we really have to. Making this field strict resulted in a 3-4.5% slowdown at the time of writing.

!(BuiltinRuntime (CekValue uni fun ann))

The partial application and its costing function. Check the docs of BuiltinRuntime for details. | A constructor value, including fully computed arguments and the tag.

VConstr !Word64 !(ArgStack uni fun ann)

readKnownCek :: (ThrowableBuiltins uni fun, ReadKnown (Term Name uni fun ()) a) => MachineParameters CekMachineCosts fun (CekValue uni fun ann) -> Term Name uni fun ann -> Either (CekEvaluationException Name uni fun) a

Unlift a value using the Steppable CEK machine. *THIS FUNCTION IS PARTIAL if the input term contains free variables*

class Eq a => Hashable a

type ThrowableBuiltins uni fun = (PrettyUni uni, Pretty fun, Typeable uni, Typeable fun)

The set of constraints we need to be able to throw exceptions with things with built-in types and functions in them.