UntypedPlutusCore.Evaluation.Machine.SteppableCek.Internal

Description

The CEK machine. The CEK machine relies on variables having non-equal Uniques whenever they have non-equal string names. I.e. Uniques are used instead of string names. This is for efficiency reasons. The CEK machines handles name capture by design.

lenContext :: Context uni fun ann -> Word

cekStateContext :: Traversal' (CekState uni fun ann) (Context uni fun ann)

cekStateAnn :: CekState uni fun ann -> Maybe ann

runCekDeBruijn :: ThrowableBuiltins uni fun => MachineParameters CekMachineCosts fun (CekValue uni fun ann) -> ExBudgetMode cost uni fun -> EmitterMode uni fun -> NTerm uni fun ann -> CekReport cost NamedDeBruijn uni fun

computeCek :: forall uni fun ann s. (ThrowableBuiltins uni fun, GivenCekReqs uni fun ann s) => Context uni fun ann -> CekValEnv uni fun ann -> NTerm uni fun ann -> CekM uni fun s (CekState uni fun ann)

returnCek :: forall uni fun ann s. (ThrowableBuiltins uni fun, GivenCekReqs uni fun ann s) => Context uni fun ann -> CekValue uni fun ann -> CekM uni fun s (CekState uni fun ann)

mkCekTrans :: forall cost uni fun ann m s. (ThrowableBuiltins uni fun, PrimMonad m, s ~ PrimState m) => MachineParameters CekMachineCosts fun (CekValue uni fun ann) -> ExBudgetMode cost uni fun -> EmitterMode uni fun -> Slippage -> m (CekTrans uni fun ann s, ExBudgetInfo cost uni fun s)

Based on the supplied arguments, initialize the CEK environment and construct a state transition function. Returns the constructed transition function paired with the methods to live access the running budget.

type CekTrans uni fun ann s = Trans (CekM uni fun s) (CekState uni fun ann)

nilSlippage :: Slippage

A CEK parameter that turns the slippage optimization *off*.

This is needed in the case of the debugger, where the accounting/budgeting costs must be *live* updated.

data EvaluationResult a

The parameterized type of results various evaluation engines return.

Constructors

EvaluationSuccess !a

EvaluationFailure

Methods

typeAst :: Type0 tyname uni ()

(TypeError ('Text "Use \8216BuiltinResult\8217 instead of \8216EvaluationResult\8217") :: Constraint, uni ~ UniOf val) => MakeKnownIn uni val (EvaluationResult a)

Instance details

Defined in PlutusCore.Builtin.KnownType

Methods

makeKnown :: EvaluationResult a -> BuiltinResult val

(TypeError ('Text "Use \8216BuiltinResult\8217 instead of \8216EvaluationResult\8217") :: Constraint, uni ~ UniOf val) => ReadKnownIn uni val (EvaluationResult a)

Instance details

Defined in PlutusCore.Builtin.KnownType

Methods

readKnown :: val -> ReadKnownM (EvaluationResult a)

PrettyBy config a => PrettyBy config (EvaluationResult a)

Instance details

Defined in PlutusCore.Evaluation.Result

Methods

prettyBy :: config -> EvaluationResult a -> Doc ann

prettyListBy :: config -> [EvaluationResult a] -> Doc ann

Generic (EvaluationResult a)

Instance details

Defined in PlutusCore.Evaluation.Result

Associated Types

type Rep (EvaluationResult a) :: Type -> Type

Methods

from :: EvaluationResult a -> Rep (EvaluationResult a) x

to :: Rep (EvaluationResult a) x -> EvaluationResult a

Show a => Show (EvaluationResult a)

Instance details

Defined in PlutusCore.Evaluation.Result

Methods

showsPrec :: Int -> EvaluationResult a -> ShowS

show :: EvaluationResult a -> String

showList :: [EvaluationResult a] -> ShowS

NFData a => NFData (EvaluationResult a)

Instance details

Defined in PlutusCore.Evaluation.Result

Methods

rnf :: EvaluationResult a -> ()

Eq a => Eq (EvaluationResult a)

Instance details

Defined in PlutusCore.Evaluation.Result

Methods

(==) :: EvaluationResult a -> EvaluationResult a -> Bool

(/=) :: EvaluationResult a -> EvaluationResult a -> Bool

PrettyClassic a => Pretty (EvaluationResult a)

Instance details

Defined in PlutusCore.Evaluation.Result

Methods

pretty :: EvaluationResult a -> Doc ann

prettyList :: [EvaluationResult a] -> Doc ann

type ToBinds uni acc (EvaluationResult a :: Type)

Instance details

Defined in PlutusCore.Builtin.KnownTypeAst

type ToBinds uni acc (EvaluationResult a :: Type) = ToBinds uni acc a

type ToHoles uni _1 (EvaluationResult a :: Type)

Instance details

Defined in PlutusCore.Builtin.KnownTypeAst

type ToHoles uni _1 (EvaluationResult a :: Type) = '[TypeHole a :: Hole]

type IsBuiltin uni (EvaluationResult a :: Type)

Instance details

Defined in PlutusCore.Builtin.KnownTypeAst

type IsBuiltin uni (EvaluationResult a :: Type) = 'False

type Rep (EvaluationResult a)

Instance details

Defined in PlutusCore.Evaluation.Result

type Rep (EvaluationResult a) = D1 ('MetaData "EvaluationResult" "PlutusCore.Evaluation.Result" "plutus-core-1.60.0.0-inplace" 'False) (C1 ('MetaCons "EvaluationSuccess" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'SourceStrict 'DecidedStrict) (Rec0 a)) :+: C1 ('MetaCons "EvaluationFailure" 'PrefixI 'False) (U1 :: Type -> Type))

data EvaluationError structural operational

The type of errors that can occur during evaluation. There are two kinds of errors:

Structural ones -- these are errors that are indicative of the _structure_ of the program being wrong. For example, a free variable was encountered during evaluation, a non-function was applied to an argument or tailList was applied to a non-list.

Operational ones -- these are errors that are indicative of the _logic_ of the program being wrong. For example, error was executed, tailList was applied to an empty list or evaluation ran out of gas.

On the chain both of these are just regular failures and we don't distinguish between them there: if a script fails, it fails, it doesn't matter what the reason was. However in the tests it does matter why the failure occurred: a structural error may indicate that the test was written incorrectly while an operational error may be entirely expected.

In other words, structural errors are "runtime type errors" and operational errors are regular runtime errors. Which means that evaluating an (erased) well-typed program should never produce a structural error, only an operational one. This creates a sort of "runtime type system" for UPLC and it would be great to stick to it and enforce in tests etc, but we currently don't.

Constructors

StructuralError !structural

OperationalError !operational

data ErrorWithCause err cause

An error and (optionally) what caused it.

Constructors

ErrorWithCause

Fields

_ewcError :: !err

_ewcCause :: !(Maybe cause)

type ThrowableBuiltins uni fun = (PrettyUni uni, Pretty fun, Typeable uni, Typeable fun)

The set of constraints we need to be able to throw exceptions with things with built-in types and functions in them.

data CekUserError

Constructors

CekCaseBuiltinError Text)) :+: (C1 ('MetaCons "CekOutOfExError" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'SourceStrict 'DecidedStrict) (Rec0 ExRestrictingBudget)) :+: C1 ('MetaCons "CekEvaluationFailure" 'PrefixI 'False) (U1 :: Type -> Type)))

type CekEvaluationException name uni fun = EvaluationException (MachineError fun) CekUserError (Term name uni fun ())

The CEK machine-specific EvaluationException.

data ExBudgetCategory fun

Constructors

BStep StepKind

BBuiltinApp fun

BStartup

newtype CekBudgetSpender uni fun s

The CEK machine is parameterized over a spendBudget function. This makes the budgeting machinery extensible and allows us to separate budgeting logic from evaluation logic and avoid branching on the union of all possible budgeting state types during evaluation.

Constructors

CekBudgetSpender

Fields

unCekBudgetSpender :: ExBudgetCategory fun -> ExBudget -> CekM uni fun s ()

newtype ExBudgetMode cost uni fun

A budgeting mode to execute the CEK machine in.

Constructors

ExBudgetMode

Fields

unExBudgetMode :: forall s. Text]

Logs emitted during evaluation.

data CekValue uni fun ann

Constructors

VCon !(Some (ValueOf uni))

VDelay !(NTerm uni fun ann) !(CekValEnv uni fun ann)

VLamAbs !NamedDeBruijn !(NTerm uni fun ann) !(CekValEnv uni fun ann)

VBuiltin
A partial builtin application, accumulating arguments for eventual full application. We don't need a CekValEnv here unlike in the other constructors, because VBuiltin values always store their corresponding Terms fully discharged, see the comments at the call sites (search for VBuiltin).

Fields

!fun

So that we know, for what builtin we're calculating the cost. We can sneak this into BuiltinRuntime, so that we don't need to store it here, but somehow doing so was consistently slowing evaluation down by half a percent. Might be noise, might be not, but at least we know that removing this fun is not helpful anyway. See this commit reversing the change: https://github.com/IntersectMBO/plutus/pull/4778/commits/86a3e24ca3c671cc27c6f4344da2bcd14f961706

(NTerm uni fun ())

This must be lazy. It represents the fully discharged partial application of the builtin function that we're going to run when it's fully saturated. We need the Term to be able to return it in case full saturation is never achieved and a partial application needs to be returned in the result. The laziness is important, because the arguments are discharged values and discharging is expensive, so we don't want to do it unless we really have to. Making this field strict resulted in a 3-4.5% slowdown at the time of writing.

!(BuiltinRuntime (CekValue uni fun ann))

The partial application and its costing function. Check the docs of BuiltinRuntime for details. | A constructor value, including fully computed arguments and the tag.

VConstr !Word64 !(EmptyOrMultiStack uni fun ann)

data DischargeResult uni fun

The result of dischargeCekValue.

Constructors

DischargeConstant (Some (ValueOf uni))

DischargeNonConstant (NTerm uni fun ())

newtype EmitterMode uni fun

An emitting mode to execute the CEK machine in, similar to ExBudgetMode.

Constructors

EmitterMode

Fields

unEmitterMode :: forall s. ST s ExBudget -> ST s (CekEmitterInfo uni fun s)

data ArgStack uni fun ann

A LIFO stack of CekValues, used to record multiple arguments that need to be pushed onto the context in reverse order. Currently used by FrameConstr for collecting the elements of a Constr as it is cheap to prepend new elements in ArgStack.

Constructors

NilStack

ConsStack !(CekValue uni fun ann) !(ArgStack uni fun ann)

data EmptyOrMultiStack uni fun ann

An alternative version of ArgStack that uses ArgNonEmptyStack when non-empty. Used in VConstr. Once all evaluated elements of Constr is collecting to ArgStack in FrameConstr, the collected elements gets reversed and put into VConstr as EmptyOrMultiStack. VConstr using EmptyOrMultiStack is more efficient than ArgStack when casing, since FrameAwaitFunValueN can be dispatched with a single pattern match.

Constructors

EmptyStack

MultiStack !(ArgStackNonEmpty uni fun ann)

data ArgStackNonEmpty uni fun ann

A non-empty variant of ArgStack, used in FrameAwaitFunValueN to store arguments that will be applied to a term. More efficient than ArgStack, since this saves one evaluation cycle by ensuring there is no NilStack.

Constructors

LastStackNonEmpty !(CekValue uni fun ann)

ConsStackNonEmpty !(CekValue uni fun ann) !(ArgStackNonEmpty uni fun ann)

type GivenCekReqs uni fun ann s = (GivenCekRuntime uni fun ann, GivenCekCaserBuiltin uni, GivenCekEmitter uni fun s, GivenCekSpender uni fun s, GivenCekSlippage, GivenCekStepCounter s, GivenCekCosts)

Constraint requiring all of the machine's implicit parameters.

type GivenCekSpender uni fun s = ?cekBudgetSpender :: CekBudgetSpender uni fun s

Implicit parameter for budget spender.

data StepCounter (n :: Nat) s

A set of Word8 counters that is used in the CEK machine to count steps.

type NumberOfStepCounters = CountConstructorsEnum (Rep StepKind)

The number of step counters that we need, should be the same as the number of constructors of StepKind.

type CounterSize = NumberOfStepCounters + 1

The total number of counters that we need, one extra for the total counter. See Note [Structure of the step counter]

type TotalCountIndex = NumberOfStepCounters

The index at which the total step counter is kept. See Note [Structure of the step counter]

type Slippage = Word8

type NTerm uni fun = Term NamedDeBruijn uni fun

The Terms that CEK can execute must have DeBruijn binders Name is not necessary but we leave it here for simplicity and debuggability.

splitStructuralOperational :: Either (EvaluationException structural operational term) a -> Either (ErrorWithCause structural term) (EvaluationResult a)

Preserve the contents of an StructuralError as a Left and turn an OperationalError into a Right EvaluationFailure (thus erasing the content of the error in the latter case).

unsafeSplitStructuralOperational :: (PrettyPlc structural, PrettyPlc term, Typeable structural, Typeable term) => Either (EvaluationException structural operational term) a -> EvaluationResult a

Throw on a StructuralError and turn an OperationalError into an EvaluationFailure (thus erasing the content of the error in the latter case).

cekResultToEither :: CekResult name uni fun -> Either (CekEvaluationException name uni fun) (Term name uni fun ())

Convert the given CekResult into an Either. This is useful, because in the ledger API we care whether the result is a constant or not, but in tests, executables etc we don't and so handling an either-error-or-term is more natural.

dischargeResultToTerm :: DischargeResult uni fun -> NTerm uni fun ()

mapTermCekResult :: (Term name uni fun () -> Term name' uni fun ()) -> CekResult name uni fun -> CekResult name' uni fun

Apply the given function to the Term (if any) stored in the given CekResult.

dischargeCekValue :: forall uni fun ann. CekValue uni fun ann -> DischargeResult uni fun

Convert a CekValue into a Term by replacing all bound variables with the terms they're bound to (which themselves have to be obtained by recursively discharging values).

defaultSlippage :: Slippage

The default number of slippage (in machine steps) to allow.

runCekM :: forall cost uni fun ann. ThrowableBuiltins uni fun => MachineParameters CekMachineCosts fun (CekValue uni fun ann) -> ExBudgetMode cost uni fun -> EmitterMode uni fun -> (forall s. GivenCekReqs uni fun ann s => CekM uni fun s (DischargeResult uni fun)) -> CekReport cost NamedDeBruijn uni fun