UntypedPlutusCore.Mark

Documentation

markNonFreshTerm :: (HasUniques (Term name uni fun ann), MonadQuote m) => Term name uni fun ann -> m ()

Marks all the Uniques in a term as used, so they will not be generated in future. Useful if you have a term which was not generated in Quote.

markNonFreshProgram :: (HasUnique name TermUnique, MonadQuote m) => Program name uni fun ann -> m ()

Marks all the Uniques in a program as used, so they will not be generated in future. Useful if you have a program which was not generated in Quote.