UntypedPlutusCore.Parser

parseProgram :: (Text -> m (Program Name DefaultUni DefaultFun SrcSpan)

Parse a UPLC program. The resulting program will have fresh names. The underlying monad must be capable of handling any parse errors. This passes "test" to the parser as the name of the input stream; to supply a name explicity, use `parse program name inputText -> m (Program Name DefaultUni DefaultFun SrcSpan)

Parse and rewrite so that names are globally unique, not just unique within their scope.

type Parser = Text (StateT ParserState (ReaderT (Maybe Version) Quote))

data SourcePos

The data type SourcePos represents source positions. It contains the name of the source file, a line number, and a column number. Source line and column positions change intensively during parsing, so we need to make them strict to avoid memory leaks.