PlutusCore

Parser

parseProgram :: ByteString -> Either (ParseErrorBundle Text ParseError) (Program TyName Name DefaultUni DefaultFun SourcePos)

Parse a PLC program. The resulting program will have fresh names. The underlying monad must be capable of handling any parse errors.

parseTerm :: ByteString -> Either (ParseErrorBundle Text ParseError) (Term TyName Name DefaultUni DefaultFun SourcePos)

Parse a PLC term. The resulting program will have fresh names. The underlying monad must be capable of handling any parse errors.

parseType :: ByteString -> Either (ParseErrorBundle Text ParseError) (Type TyName DefaultUni SourcePos)

Parse a PLC type. The resulting program will have fresh names. The underlying monad must be capable of handling any parse errors.

parseScoped :: MonadQuote f => ByteString -> f (Program TyName Name DefaultUni DefaultFun SourcePos)

Parse and rewrite so that names are globally unique, not just unique within their scope. don't require there to be no free variables at this point, we might be parsing an open term

topSourcePos :: SourcePos

Builtins

data Some (tag :: k -> Type) where

Existential. This is type is useful to hide GADTs' parameters.

>>> data Tag :: * -> * where TagInt :: Tag Int; TagBool :: Tag Bool
>>> instance GShow Tag where gshowsPrec _ TagInt = showString "TagInt"; gshowsPrec _ TagBool = showString "TagBool"
>>> classify s = case s of "TagInt" -> [mkGReadResult TagInt]; "TagBool" -> [mkGReadResult TagBool]; _ -> []
>>> instance GRead Tag where greadsPrec _ s = [ (r, rest) | (con, rest) <-  lex s, r <- classify con ]

You can either use PatternSynonyms (available with GHC >= 8.0)

>>> let x = Some TagInt
>>> x
Some TagInt

>>> case x of { Some TagInt -> "I"; Some TagBool -> "B" } :: String
"I"

or you can use functions

>>> let y = mkSome TagBool
>>> y
Some TagBool

>>> withSome y $ \y' -> case y' of { TagInt -> "I"; TagBool -> "B" } :: String
"B"

The implementation of mapSome is safe.

>>> let f :: Tag a -> Tag a; f TagInt = TagInt; f TagBool = TagBool
>>> mapSome f y
Some TagBool

but you can also use:

>>> withSome y (mkSome . f)
Some TagBool

>>> read "Some TagBool" :: Some Tag
Some TagBool

>>> read "mkSome TagInt" :: Some Tag
Some TagInt

Bundled Patterns

pattern Some :: forall k tag (a :: k). () => tag a -> Some tag

data SomeTypeIn uni

A particular type from a universe.

Constructors

forall k (a :: k). SomeTypeIn !(uni (Esc a))

data Kinded uni ta where

Constructors

Kinded :: Typeable k => !(uni (Esc a)) -> Kinded uni (Esc (a :: k))

data ValueOf uni a

A value of a particular type from a universe.

Constructors

ValueOf !(uni (Esc a)) !a

someValueOf :: forall a uni. uni (Esc a) -> a -> Some (ValueOf uni)

Wrap a value into Some (ValueOf uni), given its explicit type tag.

someValue :: forall a uni. uni `Includes` a => a -> Some (ValueOf uni)

Wrap a value into Some (ValueOf uni), provided its type is in the universe.

someValueType :: Some (ValueOf uni) -> SomeTypeIn uni

data Esc a

"Escapes" a type of an arbitrary kind to fit into Type.

class Contains uni a where

A class for enumerating types and fully instantiated type formers that uni contains. For example, a particular ExampleUni may have monomorphic types in it:

instance ExampleUni Contains Integer where ... instance ExampleUni Contains Bool where ...

as well as polymorphic ones:

instance ExampleUni Contains [] where ... instance ExampleUni Contains (,) where ...

as well as their instantiations:

instance ExampleUni Contains a => ExampleUni Contains [a] where ... instance (ExampleUni Contains a, ExampleUni Contains b) => ExampleUni Contains (a, b) where ...

(a universe can have any subset of the mentioned sorts of types, for example it's fine to have instantiated polymorphic types and not have uninstantiated ones and vice versa)

Note that when used as a constraint of a function Contains does not allow you to directly express things like "uni has the Integer, Bool and [] types and type formers", because [] is not fully instantiated. So you can only say "uni has Integer, Bool, [Integer], [Bool], [[Integer]], [[Bool]] etc" and such manual enumeration is annoying, so we'd really like to be able to say that uni has lists of arbitrary built-in types (including lists of lists etc). Contains does not allow that, but Includes does. For example, in the body of the following definition:

foo :: (uni Includes Integer, uni Includes Bool, uni Includes []) => ... foo = ...

you can make use of the fact that uni has lists of arbitrary included types (integers, booleans and lists).

Hence most of the time opt for using the more flexible Includes.

Includes is defined in terms of Contains, so you only need to provide a Contains instance per type from the universe and you'll get Includes for free.

Methods

knownUni :: uni (Esc a)

type Includes uni = Permits (Contains uni)

uni Includes a reads as "a is in the uni". a can be of a higher-kind, see the docs of Contains on why you might want that.

class Closed uni where

A universe is Closed, if it's known how to constrain every type from the universe and every type can be encoded to / decoded from a sequence of integer tags. The universe doesn't have to be finite and providing support for infinite universes is the reason why we encode a type as a sequence of integer tags as opposed to a single integer tag. For example, given

  data U a where
      UList :: !(U a) -> U [a]
      UInt  :: U Int

UList (UList UInt) can be encoded to [0,0,1] where 0 and 1 are the integer tags of the UList and UInt constructors, respectively.

Associated Types

type Everywhere uni (constr :: Type -> Constraint) :: Constraint

A constrant for "constr a holds for any a from uni".

Methods

encodeUni :: uni a -> [Int]

Encode a type as a sequence of Int tags. The opposite of decodeUni.

withDecodedUni :: (forall k (a :: k). Typeable k => uni (Esc a) -> DecodeUniM r) -> DecodeUniM r

Decode a type and feed it to the continuation.

bring :: uni `Everywhere` constr => proxy constr -> uni (Esc a) -> (constr a => r) -> r

Bring a constr a instance in scope, provided a is a type from the universe and constr holds for any type from the universe.

type family EverywhereAll uni constrs where ...

Equations

EverywhereAll uni '[] = ()

EverywhereAll uni (constr ': constrs) = (uni `Everywhere` constr, uni `EverywhereAll` constrs)

knownUniOf :: uni `Contains` a => proxy a -> uni (Esc a)

Same as knownUni, but receives a proxy.

class GShow (t :: k -> Type) where

Show-like class for 1-type-parameter GADTs. GShow t => ... is equivalent to something like (forall a. Show (t a)) => .... The easiest way to create instances would probably be to write (or derive) an instance Show (T a), and then simply say:

instance GShow t where gshowsPrec = showsPrec

Methods

gshowsPrec :: forall (a :: k). Int -> t a -> ShowS

show :: Show a => a -> String

A specialised variant of showsPrec, using precedence context zero, and returning an ordinary String.

class GEq (f :: k -> Type) where

A class for type-contexts which contain enough information to (at least in some cases) decide the equality of types occurring within them.

Methods

geq :: forall (a :: k) (b :: k). f a -> f b -> Maybe (a :~: b)

Produce a witness of type-equality, if one exists.

A handy idiom for using this would be to pattern-bind in the Maybe monad, eg.:

extract :: GEq tag => tag a -> DSum tag -> Maybe a
extract t1 (t2 :=> x) = do
    Refl <- geq t1 t2
    return x

Or in a list comprehension:

extractMany :: GEq tag => tag a -> [DSum tag] -> [a]
extractMany t1 things = [ x | (t2 :=> x) <- things, Refl <- maybeToList (geq t1 t2)]

(Making use of the DSum type from Data.Dependent.Sum in both examples)

deriveGEq :: DeriveGEQ t => t -> Q [Dec]

class HasUniApply (uni :: Type -> Type) where

A class for "uni has general type application".

Methods

matchUniApply

Arguments

:: uni tb
The type.

-> r
What to return if the type is not an application.

-> (forall k l (f :: k -> l) a. tb ~ Esc (f a) => uni (Esc f) -> uni (Esc a) -> r)
The continuation taking a function and an argument.

-> r

Deconstruct a type application into the function and the argument and feed them to the continuation. If the type is not an application, then return the default value.

checkStar :: forall uni a (x :: a). Typeable a => uni (Esc x) -> Maybe (a :~: Type)

Check if the kind of the given type from the universe is Type.

withApplicable :: forall (a :: Type) (ab :: Type) f x uni m r. (Typeable ab, Typeable a, MonadPlus m) => uni (Esc (f :: ab)) -> uni (Esc (x :: a)) -> (forall (b :: Type). (Typeable b, ab ~ (a -> b)) => m r) -> m r

Check if one type from the universe can be applied to another (i.e. check that the expected kind of the argument matches the actual one) and call the continuation in the refined context. Fail with mzero otherwise.

data (a :: k) :~: (b :: k) where infix 4

Propositional equality. If a :~: b is inhabited by some terminating value, then the type a is the same as the type b. To use this equality in practice, pattern-match on the a :~: b to get out the Refl constructor; in the body of the pattern-match, the compiler knows that a ~ b.

Since: base-4.7.0.0

Constructors

Refl :: forall k (a :: k). a :~: a

type (<:) uni1 uni2 = uni1 `Everywhere` Includes uni2

A constraint for "uni1 is a subuniverse of uni2".

data DefaultUni a where

The universe used by default.

Constructors

DefaultUniInteger :: DefaultUni (Esc Integer)

DefaultUniByteString :: DefaultUni (Esc ByteString)

DefaultUniString :: DefaultUni (Esc Text)

DefaultUniUnit :: DefaultUni (Esc ())

DefaultUniBool :: DefaultUni (Esc Bool)

DefaultUniProtoList :: DefaultUni (Esc [])

DefaultUniProtoPair :: DefaultUni (Esc (,))

DefaultUniApply :: !(DefaultUni (Esc f)) -> !(DefaultUni (Esc a)) -> DefaultUni (Esc (f a))

DefaultUniData :: DefaultUni (Esc Data)

pattern DefaultUniList :: forall a k1 k2 (f :: k1 -> k2) (a1 :: k1). () => forall. (a ~ Esc (f a1), Esc f ~ Esc []) => DefaultUni (Esc a1) -> DefaultUni a

pattern DefaultUniPair :: forall a k1 k2 (f1 :: k1 -> k2) (a1 :: k1) k3 k4 (f2 :: k3 -> k4) (a2 :: k3). () => forall. (a ~ Esc (f1 a1), Esc f1 ~ Esc (f2 a2), Esc f2 ~ Esc (,)) => DefaultUni (Esc a2) -> DefaultUni (Esc a1) -> DefaultUni a

data DefaultFun

Default built-in functions.

When updating these, make sure to add them to the protocol version listing! See Note [New builtins and protocol versions]

Constructors

AddInteger

SubtractInteger

MultiplyInteger

DivideInteger

QuotientInteger

RemainderInteger

ModInteger

EqualsInteger

LessThanInteger

LessThanEqualsInteger

AppendByteString

ConsByteString

SliceByteString

LengthOfByteString

IndexByteString

EqualsByteString

LessThanByteString

LessThanEqualsByteString

Sha2_256

Sha3_256

Blake2b_256

VerifyEd25519Signature

VerifyEcdsaSecp256k1Signature

VerifySchnorrSecp256k1Signature

AppendString

EqualsString

EncodeUtf8

DecodeUtf8

IfThenElse

ChooseUnit

Trace

FstPair

SndPair

ChooseList

MkCons

HeadList

TailList

NullList

ChooseData

ConstrData

MapData

ListData

IData

BData

UnConstrData

UnMapData

UnListData

UnIData

UnBData

EqualsData

SerialiseData

MkPairData

MkNilData

MkNilPairData

AST

data Term tyname name uni fun ann

Constructors

Var ann name
a named variable

TyAbs ann tyname (Kind ann) (Term tyname name uni fun ann)

LamAbs ann name (Type tyname uni ann) (Term tyname name uni fun ann)

Apply ann (Term tyname name uni fun ann) (Term tyname name uni fun ann)

Constant ann (Some (ValueOf uni))
a constant term

Builtin ann fun

TyInst ann (Term tyname name uni fun ann) (Type tyname uni ann)

Unwrap ann (Term tyname name uni fun ann)

IWrap ann (Type tyname uni ann) (Type tyname uni ann) (Term tyname name uni fun ann)

Error ann (Type tyname uni ann)

termSubterms :: Traversal' (Term tyname name uni fun ann) (Term tyname name uni fun ann)

Get all the direct child Terms of the given Term.

termSubtypes :: Traversal' (Term tyname name uni fun ann) (Type tyname uni ann)

Get all the direct child Types of the given Term.

type family UniOf a :: Type -> Type

Extract the universe from a type.

data Type tyname uni ann

A Type assigned to expressions.

Constructors

TyVar ann tyname

TyFun ann (Type tyname uni ann) (Type tyname uni ann)

TyIFix ann (Type tyname uni ann) (Type tyname uni ann)
Fix-point type, for constructing self-recursive types

TyForall ann tyname (Kind ann) (Type tyname uni ann)

TyBuiltin ann (SomeTypeIn uni)
Builtin type

TyLam ann tyname (Kind ann) (Type tyname uni ann)

TyApp ann (Type tyname uni ann) (Type tyname uni ann)

typeSubtypes :: Traversal' (Type tyname uni ann) (Type tyname uni ann)

Get all the direct child Types of the given Type.

data Kind ann

Constructors

Type ann

KindArrow ann (Kind ann) (Kind ann)

data ParseError

An error encountered during parsing.

Constructors

UnknownBuiltinType Text SourcePos

BuiltinTypeNotAStar Text SourcePos

UnknownBuiltinFunction Text SourcePos

InvalidBuiltinConstant Text Text SourcePos

data Version ann

Version of Plutus Core to be used for the program.

Constructors

Version ann Natural Natural Natural

data Program tyname name uni fun ann

A Program is simply a Term coupled with a Version of the core language.

Constructors

Program

Fields

_progAnn :: ann

_progVer :: Version ann

_progTerm :: Term tyname name uni fun ann

data Name

A Name represents variables/names in Plutus Core.

Constructors

Name

Fields

nameString :: Text

The identifier name, for use in error messages.

nameUnique :: Unique

A Unique assigned to the name, allowing for cheap comparisons in the compiler.

newtype TyName

We use a newtype to enforce separation between names used for types and those used for terms.

Constructors

TyName

Fields

unTyName :: Name

newtype Unique

A unique identifier

Constructors

Unique

Fields

unUnique :: Int

newtype UniqueMap unique a

A mapping from uniques to values of type a.

Constructors

UniqueMap

Fields

unUniqueMap :: IntMap a

newtype Normalized a

Constructors

Normalized

Fields

unNormalized :: a

defaultVersion :: ann -> Version ann

The default version of Plutus Core supported by this library.

termAnn :: Term tyname name uni fun ann -> ann

typeAnn :: Type tyname uni ann -> ann

tyVarDeclAnn :: forall tyname ann. Lens' (TyVarDecl tyname ann) ann

tyVarDeclName :: forall tyname ann tyname. Lens (TyVarDecl tyname ann) (TyVarDecl tyname ann) tyname tyname

tyVarDeclKind :: forall tyname ann. Lens' (TyVarDecl tyname ann) (Kind ann)

varDeclAnn :: forall tyname name uni fun ann fun. Lens (VarDecl tyname name uni fun ann) (VarDecl tyname name uni fun ann) ann ann

varDeclName :: forall tyname name uni fun ann name fun. Lens (VarDecl tyname name uni fun ann) (VarDecl tyname name uni fun ann) name name

varDeclType :: forall tyname name uni fun ann tyname uni fun. Lens (VarDecl tyname name uni fun ann) (VarDecl tyname name uni fun ann) (Type tyname uni ann) (Type tyname uni ann)

tyDeclAnn :: forall tyname uni ann. Lens' (TyDecl tyname uni ann) ann

tyDeclType :: forall tyname uni ann tyname uni. Lens (TyDecl tyname uni ann) (TyDecl tyname uni ann) (Type tyname uni ann) (Type tyname uni ann)

tyDeclKind :: forall tyname uni ann. Lens' (TyDecl tyname uni ann) (Kind ann)

progAnn :: forall tyname name uni fun ann. Lens' (Program tyname name uni fun ann) ann

progVer :: forall tyname name uni fun ann. Lens' (Program tyname name uni fun ann) (Version ann)

progTerm :: forall tyname name uni fun ann tyname name uni fun. Lens (Program tyname name uni fun ann) (Program tyname name uni fun ann) (Term tyname name uni fun ann) (Term tyname name uni fun ann)

mapFun :: (fun -> fun') -> Term tyname name uni fun ann -> Term tyname name uni fun' ann

Map a function over the set of built-in functions.

DeBruijn representation

newtype DeBruijn

A term name as a de Bruijn index, without the name string.

Constructors

DeBruijn

Fields

dbnIndex :: Index

data NamedDeBruijn

A term name as a de Bruijn index.

Constructors

NamedDeBruijn

Fields

ndbnString :: Text

ndbnIndex :: Index

deBruijnTerm :: (AsFreeVariableError e, MonadError e m) => Term TyName Name uni fun ann -> m (Term NamedTyDeBruijn NamedDeBruijn uni fun ann)

Convert a Term with TyNames and Names into a Term with NamedTyDeBruijns and NamedDeBruijns. Will throw an error if a free variable is encountered.

unDeBruijnTerm :: (MonadQuote m, AsFreeVariableError e, MonadError e m) => Term NamedTyDeBruijn NamedDeBruijn uni fun ann -> m (Term TyName Name uni fun ann)

Convert a Term with NamedTyDeBruijns and NamedDeBruijns into a Term with TyNames and Names. Will throw an error if a free variable is encountered.

Lexer

data SourcePos

The data type SourcePos represents source positions. It contains the name of the source file, a line number, and a column number. Source line and column positions change intensively during parsing, so we need to make them strict to avoid memory leaks.

Formatting

format :: (Monad m, PrettyBy config (Program TyName Name DefaultUni DefaultFun SourcePos)) => config -> ByteString -> m Text

Processing

type family HasUniques a :: Constraint

All kinds of uniques an entity contains.

class Rename a where

The class of things that can be renamed. I.e. things that are capable of satisfying the global uniqueness condition.

Methods

rename :: MonadQuote m => a -> m a

Rename Uniques so that they're globally unique. In case there are any free variables, they must be left untouched and bound variables must not get renamed to free ones. Must always assign new names to bound variables, so that rename can be used for alpha-renaming as well.

Type checking

class ToKind (uni :: Type -> Type)

For computing the Plutus kind of a built-in type. See kindOfBuiltinType.

Minimal complete definition

toSingKind

type Typecheckable uni fun = (ToKind uni, HasUniApply uni, ToBuiltinMeaning uni fun)

Configuration.

newtype BuiltinTypes uni fun

Mapping from Builtins to their Types.

Constructors

BuiltinTypes

Fields

unBuiltinTypes :: Maybe (Array fun (Dupable (Normalized (Type TyName uni ()))))

newtype TypeCheckConfig uni fun

Configuration of the type checker.

Constructors

TypeCheckConfig

Fields

_tccBuiltinTypes :: BuiltinTypes uni fun

tccBuiltinTypes :: HasTypeCheckConfig c uni fun => Lens' c (BuiltinTypes uni fun)

builtinMeaningsToTypes :: (MonadError err m, AsTypeError err term uni fun ann, Typecheckable uni fun) => ann -> m (BuiltinTypes uni fun)

Extract the TypeScheme from a BuiltinMeaning and convert it to the corresponding Type for each built-in function.

getDefTypeCheckConfig :: (MonadError err m, AsTypeError err term uni fun ann, Typecheckable uni fun) => ann -> m (TypeCheckConfig uni fun)

Get the default type checking config.

Kindtype inferencechecking.

inferKind :: (MonadQuote m, MonadError err m, AsTypeError err term uni fun ann, ToKind uni) => TypeCheckConfig uni fun -> Type TyName uni ann -> m (Kind ())

Infer the kind of a type.

checkKind :: (MonadQuote m, MonadError err m, AsTypeError err term uni fun ann, ToKind uni) => TypeCheckConfig uni fun -> ann -> Type TyName uni ann -> Kind () -> m ()

Check a type against a kind. Infers the kind of the type and checks that it's equal to the given kind throwing a TypeError (annotated with the value of the ann argument) otherwise.

inferType :: (MonadError err m, MonadQuote m, AsTypeError err (Term TyName Name uni fun ()) uni fun ann, ToKind uni, HasUniApply uni, GEq uni, Ix fun) => TypeCheckConfig uni fun -> Term TyName Name uni fun ann -> m (Normalized (Type TyName uni ()))

Infer the type of a term.

checkType :: (MonadError err m, MonadQuote m, AsTypeError err (Term TyName Name uni fun ()) uni fun ann, ToKind uni, HasUniApply uni, GEq uni, Ix fun) => TypeCheckConfig uni fun -> ann -> Term TyName Name uni fun ann -> Normalized (Type TyName uni ()) -> m ()

Check a term against a type. Infers the type of the term and checks that it's equal to the given type throwing a TypeError (annotated with the value of the ann argument) otherwise.

inferTypeOfProgram :: (MonadError err m, MonadQuote m, AsTypeError err (Term TyName Name uni fun ()) uni fun ann, ToKind uni, HasUniApply uni, GEq uni, Ix fun) => TypeCheckConfig uni fun -> Program TyName Name uni fun ann -> m (Normalized (Type TyName uni ()))

Infer the type of a program.

checkTypeOfProgram :: (MonadError err m, MonadQuote m, AsTypeError err (Term TyName Name uni fun ()) uni fun ann, ToKind uni, HasUniApply uni, GEq uni, Ix fun) => TypeCheckConfig uni fun -> ann -> Program TyName Name uni fun ann -> Normalized (Type TyName uni ()) -> m ()

Check a program against a type. Infers the type of the program and checks that it's equal to the given type throwing a TypeError (annotated with the value of the ann argument) otherwise.

printType :: (AsTypeError e (Term TyName Name DefaultUni DefaultFun ()) DefaultUni DefaultFun SourcePos, MonadError e m) => ByteString -> m Text

normalizeTypesIn :: (HasUnique tyname TypeUnique, HasUnique name TermUnique, MonadQuote m, HasUniApply uni) => Term tyname name uni fun ann -> m (Term tyname name uni fun ann)

Normalize every Type in a Term.

normalizeTypesInProgram :: (HasUnique tyname TypeUnique, HasUnique name TermUnique, MonadQuote m, HasUniApply uni) => Program tyname name uni fun ann -> m (Program tyname name uni fun ann)

Normalize every Type in a Program.

class AsTypeError r term uni fun ann | r -> term uni fun ann where

Minimal complete definition

_TypeError

Methods

_TypeError :: Prism' r (TypeError term uni fun ann)

_KindMismatch :: Prism' r (ann, Type TyName uni (), Kind (), Kind ())

_TypeMismatch :: Prism' r (ann, term, Type TyName uni (), Normalized (Type TyName uni ()))

_FreeTypeVariableE :: Prism' r (ann, TyName)

_FreeVariableE :: Prism' r (ann, Name)

_UnknownBuiltinFunctionE :: Prism' r (ann, fun)

data TypeError term uni fun ann

parseTypecheck :: (AsTypeError e (Term TyName Name DefaultUni DefaultFun ()) DefaultUni DefaultFun SourcePos, MonadError e m, MonadQuote m) => TypeCheckConfig DefaultUni DefaultFun -> ByteString -> m (Normalized (Type TyName DefaultUni ()))

Parse a program and typecheck it.

typecheckPipeline :: (AsTypeError e (Term TyName Name DefaultUni DefaultFun ()) DefaultUni DefaultFun a, MonadError e m, MonadQuote m) => TypeCheckConfig DefaultUni DefaultFun -> Program TyName Name DefaultUni DefaultFun a -> m (Normalized (Type TyName DefaultUni ()))

Typecheck a program.

Errors

data Error uni fun ann

Constructors

ParseErrorE ParseError

UniqueCoherencyErrorE (UniqueError ann)

TypeErrorE (TypeError (Term TyName Name uni fun ()) uni fun ann)

NormCheckErrorE (NormCheckError TyName Name uni fun ann)

FreeVariableErrorE FreeVariableError

class AsError r uni fun ann | r -> uni fun ann where

Minimal complete definition

_Error

Methods

_Error :: Prism' r (Error uni fun ann)

_ParseErrorE :: Prism' r ParseError

_UniqueCoherencyErrorE :: Prism' r (UniqueError ann)

_TypeErrorE :: Prism' r (TypeError (Term TyName Name uni fun ()) uni fun ann)

_NormCheckErrorE :: Prism' r (NormCheckError TyName Name uni fun ann)

_FreeVariableErrorE :: Prism' r FreeVariableError

data NormCheckError tyname name uni fun ann

Constructors

BadType ann (Type tyname uni ann) Text

BadTerm ann (Term tyname name uni fun ann) Text

class AsNormCheckError r tyname name uni fun ann | r -> tyname name uni fun ann where

Minimal complete definition

_NormCheckError

Methods

_NormCheckError :: Prism' r (NormCheckError tyname name uni fun ann)

_BadType :: Prism' r (ann, Type tyname uni ann, Text)

_BadTerm :: Prism' r (ann, Term tyname name uni fun ann, Text)

data UniqueError ann

Constructors

MultiplyDefined Unique ann ann

IncoherentUsage Unique ann ann

FreeVariable Unique ann

class AsUniqueError r ann | r -> ann where

Minimal complete definition

_UniqueError

Methods

_UniqueError :: Prism' r (UniqueError ann)

_MultiplyDefined :: Prism' r (Unique, ann, ann)

_IncoherentUsage :: Prism' r (Unique, ann, ann)

_FreeVariable :: Prism' r (Unique, ann)

data FreeVariableError

We cannot do a correct translation to or from de Bruijn indices if the program is not well-scoped. So we throw an error in such a case.

Constructors

FreeUnique Unique

FreeIndex Index

class AsFreeVariableError r where

Minimal complete definition

_FreeVariableError

Methods

_FreeVariableError :: Prism' r FreeVariableError

_FreeUnique :: Prism' r Unique

_FreeIndex :: Prism' r Index

Base functors

data TermF (tyname :: Type) (name :: Type) (uni :: Type -> Type) (fun :: Type) (ann :: Type) r

Constructors

VarF ann name

TyAbsF ann tyname (Kind ann) r

LamAbsF ann name (Type tyname uni ann) r

ApplyF ann r r

ConstantF ann (Some (ValueOf uni))

BuiltinF ann fun

TyInstF ann r (Type tyname uni ann)

UnwrapF ann r

IWrapF ann (Type tyname uni ann) (Type tyname uni ann) r

ErrorF ann (Type tyname uni ann)

data TypeF (tyname :: Type) (uni :: Type -> Type) (ann :: Type) r

Constructors

TyVarF ann tyname

TyFunF ann r r

TyIFixF ann r r

TyForallF ann tyname (Kind ann) r

TyBuiltinF ann (SomeTypeIn uni)

TyLamF ann tyname (Kind ann) r

TyAppF ann r r

Quotation and term construction

type Quote = QuoteT Identity

A non-transformer version of QuoteT.

runQuote :: Quote a -> a

See runQuoteT.

data QuoteT m a

The "quotation" monad transformer. Within this monad you can do safe construction of PLC terms using quasiquotation, fresh-name generation, and parsing.

runQuoteT :: Monad m => QuoteT m a -> m a

Run a quote from an empty identifier state. Note that the resulting term cannot necessarily be safely combined with other terms - that should happen inside QuoteT.

class Monad m => MonadQuote m

A monad that allows lifting of quoted expressions.

liftQuote :: MonadQuote m => Quote a -> m a

Name generation

freshUnique :: MonadQuote m => m Unique

Get a fresh Unique.

freshName :: MonadQuote m => Text -> m Name

Get a fresh Name, given the annotation and the Text name.

freshTyName :: MonadQuote m => Text -> m TyName

Get a fresh TyName, given the annotation and the Text name.

Evaluation

data EvaluationResult a

The parameterized type of results various evaluation engines return. On the PLC side this becomes (via makeKnown) either a call to Error or a value of the PLC counterpart of type a.

Constructors

EvaluationSuccess a

EvaluationFailure

data UnliftingMode

Determines how to unlift arguments. The difference is that with UnliftingImmediate unlifting is performed immediately after a builtin gets the argument and so can fail immediately too, while with deferred unlifting all arguments are unlifted upon full saturation, hence no failure can occur until that. The former makes it much harder to specify the behaviour of builtins and so UnliftingDeferred is the preferred mode.

Constructors

UnliftingImmediate

UnliftingDeferred

Combining programs

applyProgram :: Monoid a => Program tyname name uni fun a -> Program tyname name uni fun a -> Program tyname name uni fun a

Take one PLC program and apply it to another.

Benchmarking

termSize :: Term tyname name uni fun ann -> Size

Count the number of AST nodes in a term.

typeSize :: Type tyname uni ann -> Size

Count the number of AST nodes in a type.

kindSize :: Kind a -> Size

Count the number of AST nodes in a kind.

>>> kindSize $ Type ()
Size {unSize = 1}
>>> kindSize $ KindArrow () (KindArrow () (Type ()) (Type ())) (Type ())
Size {unSize = 5}

programSize :: Program tyname name uni fun ann -> Size

Count the number of AST nodes in a program.

serialisedSize :: Flat a => a -> Integer

Compute the size of the serializabled form of a value.

Budgeting defaults

defaultBuiltinCostModel :: BuiltinCostModel

The default cost model for built-in functions.

defaultBuiltinsRuntime :: HasConstantIn DefaultUni term => BuiltinsRuntime DefaultFun term

defaultCekCostModel :: CostModel CekMachineCosts BuiltinCostModel

defaultCekMachineCosts :: CekMachineCosts

Default costs for CEK machine instructions.

defaultCekParameters :: MachineParameters CekMachineCosts CekValue DefaultUni DefaultFun

defaultCostModelParams :: Maybe CostModelParams

The default cost model data. This is exposed to the ledger, so let's not confuse anybody by mentioning the CEK machine

defaultUnliftingMode :: UnliftingMode

unitCekParameters :: MachineParameters CekMachineCosts CekValue DefaultUni DefaultFun

CEK machine costs

cekMachineCostsPrefix :: Text

The prefix of the field names in the CekMachineCosts type, used for extracting the CekMachineCosts component of the ledger's cost model parameters. See Note [Cost model parameters] in CostModelInterface.

data CekMachineCosts

Costs for evaluating AST nodes. Times should be specified in picoseconds, memory sizes in bytes.

Constructors

CekMachineCosts

Fields

cekStartupCost :: ExBudget

cekVarCost :: ExBudget

cekConstCost :: ExBudget

cekLamCost :: ExBudget

cekDelayCost :: ExBudget

cekForceCost :: ExBudget

cekApplyCost :: ExBudget

cekBuiltinCost :: ExBudget

Just the cost of evaluating a Builtin node, not the builtin itself. There's no entry for Error since we'll be exiting anyway; also, what would happen if calling Error caused the budget to be exceeded?