PlutusCore

Parser

parseProgram :: (AsParserErrorBundle e, MonadError e m, MonadQuote m) => Text -> m (Program TyName Name DefaultUni DefaultFun SrcSpan)

Parse a PLC program. The resulting program will have fresh names. The underlying monad must be capable of handling any parse errors. This passes "test" to the parser as the name of the input stream; to supply a name explicity, use `parse program name input`.

parseTerm :: (AsParserErrorBundle e, MonadError e m, MonadQuote m) => Text -> m (Term TyName Name DefaultUni DefaultFun SrcSpan)

Parse a PLC term. The resulting program will have fresh names. The underlying monad must be capable of handling any parse errors.

parseType :: (AsParserErrorBundle e, MonadError e m, MonadQuote m) => Text -> m (Type TyName DefaultUni SrcSpan)

Parse a PLC type. The resulting program will have fresh names. The underlying monad must be capable of handling any parse errors.

data SourcePos

data SrcSpan

The span between two source locations.

This corresponds roughly to the SrcSpan used by GHC, but we define our own version so we don't have to depend on ghc to use it.

The line and column numbers are 1-based, and the unit is Unicode code point (or Char).

Constructors

SrcSpan

Fields

srcSpanFile :: FilePath

srcSpanSLine :: Int

srcSpanSCol :: Int

srcSpanELine :: Int

srcSpanECol :: Int

Same as GHC's SrcSpan, srcSpanECol is usually one more than the column of the last character of the thing this SrcSpan is for (unless the last character is the line break).

data SrcSpans

Builtins

data Some (tag :: k -> Type) where

Bundled Patterns

pattern Some :: forall {k} tag (a :: k). () => tag a -> Some tag

data SomeTypeIn uni

A particular type from a universe.

Constructors

forall k (a :: k). SomeTypeIn !(uni (Esc a))

data Kinded uni ta where

Constructors

Kinded :: Typeable k => !(uni (Esc a)) -> Kinded uni (Esc (a :: k))

data ValueOf uni a

A value of a particular type from a universe.

Constructors

ValueOf !(uni (Esc a)) !a

someValueOf :: forall a uni. uni (Esc a) -> a -> Some (ValueOf uni)

Wrap a value into Some (ValueOf uni), given its explicit type tag.

someValue :: forall a uni. uni `Contains` a => a -> Some (ValueOf uni)

Wrap a value into Some (ValueOf uni), provided its type is in the universe.

someValueType :: Some (ValueOf uni) -> SomeTypeIn uni

data Esc a

"Escapes" a type of an arbitrary kind to fit into Type.

class Contains uni a where

A class for enumerating types and fully instantiated type formers that uni contains. For example, a particular ExampleUni may have monomorphic types in it:

instance ExampleUni Contains Integer where ... instance ExampleUni Contains Bool where ...

as well as polymorphic ones:

instance ExampleUni Contains [] where ... instance ExampleUni Contains (,) where ...

as well as their instantiations:

instance ExampleUni Contains a => ExampleUni Contains [a] where ... instance (ExampleUni Contains a, ExampleUni Contains b) => ExampleUni Contains (a, b) where ...

(a universe can have any subset of the mentioned sorts of types, for example it's fine to have instantiated polymorphic types and not have uninstantiated ones and vice versa)

Note that when used as a constraint of a function Contains does not allow you to directly express things like "uni has the Integer, Bool and [] types and type formers", because [] is not fully instantiated. So you can only say "uni has Integer, Bool, [Integer], [Bool], [[Integer]], [[Bool]] etc" and such manual enumeration is annoying, so we'd really like to be able to say that uni has lists of arbitrary built-in types (including lists of lists etc). Contains does not allow that, but Includes does. For example, in the body of the following definition:

foo :: (uni Includes Integer, uni Includes Bool, uni Includes []) => ... foo = ...

you can make use of the fact that uni has lists of arbitrary included types (integers, booleans and lists).

Hence most of the time opt for using the more flexible Includes.

Includes is defined in terms of Contains, so you only need to provide a Contains instance per type from the universe and you'll get Includes for free.

Methods

knownUni :: uni (Esc a)

class Closed uni where

A universe is Closed, if it's known how to constrain every type from the universe and every type can be encoded to / decoded from a sequence of integer tags. The universe doesn't have to be finite and providing support for infinite universes is the reason why we encode a type as a sequence of integer tags as opposed to a single integer tag. For example, given

  data U a where
      UList :: !(U a) -> U [a]
      UInt  :: U Int

UList (UList UInt) can be encoded to [0,0,1] where 0 and 1 are the integer tags of the UList and UInt constructors, respectively.

Associated Types

type Everywhere uni (constr :: Type -> Constraint) :: Constraint

A constrant for "constr a holds for any a from uni".

Methods

encodeUni :: uni a -> [Int]

Encode a type as a sequence of Int tags. The opposite of decodeUni.

withDecodedUni :: (forall k (a :: k). Typeable k => uni (Esc a) -> DecodeUniM r) -> DecodeUniM r

Decode a type and feed it to the continuation.

bring :: uni `Everywhere` constr => proxy constr -> uni (Esc a) -> (constr a => r) -> r

Bring a constr a instance in scope, provided a is a type from the universe and constr holds for any type from the universe.

type family EverywhereAll uni constrs where ...

Equations

EverywhereAll uni '[] = ()

EverywhereAll uni (constr ': constrs) = (uni `Everywhere` constr, uni `EverywhereAll` constrs)

knownUniOf :: uni `Contains` a => proxy a -> uni (Esc a)

Same as knownUni, but receives a proxy.

class GShow (t :: k -> Type) where

Methods

gshowsPrec :: forall (a :: k). Int -> t a -> ShowS

show :: Show a => a -> String

A specialised variant of showsPrec, using precedence context zero, and returning an ordinary String.

class GEq (f :: k -> Type) where

Methods

geq :: forall (a :: k) (b :: k). f a -> f b -> Maybe (a :~: b)

class HasUniApply (uni :: Type -> Type) where

A class for "uni has general type application".

Methods

uniApply :: forall k l (f :: k -> l) a. uni (Esc f) -> uni (Esc a) -> uni (Esc (f a))

Apply a type constructor to an argument.

matchUniApply

Arguments

:: uni tb
The type.

-> r
What to return if the type is not an application.

-> (forall k l (f :: k -> l) a. tb ~ Esc (f a) => uni (Esc f) -> uni (Esc a) -> r)
The continuation taking a function and an argument.

-> r

Deconstruct a type application into the function and the argument and feed them to the continuation. If the type is not an application, then return the default value.

checkStar :: forall uni a (x :: a). Typeable a => uni (Esc x) -> Maybe (a :~: Type)

Check if the kind of the given type from the universe is Type.

withApplicable :: forall (a :: Type) (ab :: Type) f x uni m r. (Typeable ab, Typeable a, MonadPlus m) => uni (Esc (f :: ab)) -> uni (Esc (x :: a)) -> (forall (b :: Type). (Typeable b, ab ~ (a -> b)) => m r) -> m r

Check if one type from the universe can be applied to another (i.e. check that the expected kind of the argument matches the actual one) and call the continuation in the refined context. Fail with mzero otherwise.

data (a :: k) :~: (b :: k) where infix 4

Propositional equality. If a :~: b is inhabited by some terminating value, then the type a is the same as the type b. To use this equality in practice, pattern-match on the a :~: b to get out the Refl constructor; in the body of the pattern-match, the compiler knows that a ~ b.

Since: base-4.7.0.0

Constructors

Refl :: forall {k} (a :: k). a :~: a

type (<:) uni1 uni2 = uni1 `Everywhere` Includes uni2

A constraint for "uni1 is a subuniverse of uni2".

type HasTypeLevel uni x = KnownTypeAst Void uni (ElaborateBuiltin uni x)

Specifies that the given type is a built-in one and can be embedded into a Kind.

type HasTermLevel uni = Includes uni

Specifies that the given type is a built-in one and its values can be embedded into a Term.

type HasTypeAndTermLevel uni x = (uni `HasTypeLevel` x, uni `HasTermLevel` x)

The product of HasTypeLevel and HasTermLevel.

data DefaultUni a where

The universe used by default.

Constructors

DefaultUniInteger :: DefaultUni (Esc Integer)

DefaultUniByteString :: DefaultUni (Esc ByteString)

DefaultUniString :: DefaultUni (Esc Text)

DefaultUniUnit :: DefaultUni (Esc ())

DefaultUniBool :: DefaultUni (Esc Bool)

DefaultUniProtoList :: DefaultUni (Esc List)

DefaultUniProtoPair :: DefaultUni (Esc (,))

DefaultUniApply :: !(DefaultUni (Esc f)) -> !(DefaultUni (Esc a)) -> DefaultUni (Esc (f a))

DefaultUniData :: DefaultUni (Esc Data)

DefaultUniBLS12_381_G1_Element :: DefaultUni (Esc Element)

DefaultUniBLS12_381_G2_Element :: DefaultUni (Esc Element)

DefaultUniBLS12_381_MlResult :: DefaultUni (Esc MlResult)

pattern DefaultUniList :: forall {a} {k1} {k2} {f :: k1 -> k2} {a1 :: k1}. () => forall. (a ~ Esc (f a1), Esc f ~ Esc List) => DefaultUni (Esc a1) -> DefaultUni a

pattern DefaultUniPair :: forall {a} {k1} {k2} {f1 :: k1 -> k2} {a1 :: k1} {k3} {k4} {f2 :: k3 -> k4} {a2 :: k3}. () => forall. (a ~ Esc (f1 a1), Esc f1 ~ Esc (f2 a2), Esc f2 ~ Esc (,)) => DefaultUni (Esc a2) -> DefaultUni (Esc a1) -> DefaultUni a

data DefaultFun

Default built-in functions.

When updating these, make sure to add them to the protocol version listing! See Note [New builtins/language versions and protocol versions]

Constructors

AddInteger

SubtractInteger

MultiplyInteger

DivideInteger

QuotientInteger

RemainderInteger

ModInteger

EqualsInteger

LessThanInteger

LessThanEqualsInteger

AppendByteString

ConsByteString

SliceByteString

LengthOfByteString

IndexByteString

EqualsByteString

LessThanByteString

LessThanEqualsByteString

Sha2_256

Sha3_256

Blake2b_256

VerifyEd25519Signature

VerifyEcdsaSecp256k1Signature

VerifySchnorrSecp256k1Signature

AppendString

EqualsString

EncodeUtf8

DecodeUtf8

IfThenElse

ChooseUnit

Trace

FstPair

SndPair

ChooseList

MkCons

HeadList

TailList

NullList

ChooseData

ConstrData

MapData

ListData

IData

BData

UnConstrData

UnMapData

UnListData

UnIData

UnBData

EqualsData

SerialiseData

MkPairData

MkNilData

MkNilPairData

Bls12_381_G1_add

Bls12_381_G1_neg

Bls12_381_G1_scalarMul

Bls12_381_G1_equal

Bls12_381_G1_hashToGroup

Bls12_381_G1_compress

Bls12_381_G1_uncompress

Bls12_381_G2_add

Bls12_381_G2_neg

Bls12_381_G2_scalarMul

Bls12_381_G2_equal

Bls12_381_G2_hashToGroup

Bls12_381_G2_compress

Bls12_381_G2_uncompress

Bls12_381_millerLoop

Bls12_381_mulMlResult

Bls12_381_finalVerify

Keccak_256

Blake2b_224

IntegerToByteString

ByteStringToInteger

AST

data Term tyname name uni fun ann

Constructors

Var ann name
a named variable

LamAbs ann name (Type tyname uni ann) (Term tyname name uni fun ann)
lambda abstraction

Apply ann (Term tyname name uni fun ann) (Term tyname name uni fun ann)
application

TyAbs ann tyname (Kind ann) (Term tyname name uni fun ann)
type abstraction

TyInst ann (Term tyname name uni fun ann) (Type tyname uni ann)
instantiation

IWrap ann (Type tyname uni ann) (Type tyname uni ann) (Term tyname name uni fun ann)
wrapping

Unwrap ann (Term tyname name uni fun ann)
unwrapping See Note [Constr tag type]

Constr ann (Type tyname uni ann) Word64 [Term tyname name uni fun ann]
constructor

Case ann (Type tyname uni ann) (Term tyname name uni fun ann) [Term tyname name uni fun ann]
case

Constant ann (Some (ValueOf uni))
constants

Builtin ann fun
builtin functions

Error ann (Type tyname uni ann)
fail with error

termSubterms :: Traversal' (Term tyname name uni fun ann) (Term tyname name uni fun ann)

Get all the direct child Terms of the given Term.

termSubtypes :: Traversal' (Term tyname name uni fun ann) (Type tyname uni ann)

Get all the direct child Kinds of the given Term.

termMapNames :: forall tyname tyname' name name' uni fun ann. (tyname -> tyname') -> (name -> name') -> Term tyname name uni fun ann -> Term tyname' name' uni fun ann

programMapNames :: forall tyname tyname' name name' uni fun ann. (tyname -> tyname') -> (name -> name') -> Program tyname name uni fun ann -> Program tyname' name' uni fun ann

type family UniOf a :: Type -> Type

Extract the universe from a type.

data Type tyname uni ann

A Kind assigned to expressions.

Constructors

TyVar ann tyname
Type variable

TyFun ann (Type tyname uni ann) (Type tyname uni ann)
Function type

TyIFix ann (Type tyname uni ann) (Type tyname uni ann)
Fix-point type, for constructing self-recursive types

TyForall ann tyname (Kind ann) (Type tyname uni ann)
Polymorphic type

TyBuiltin ann (SomeTypeIn uni)
Builtin type

TyLam ann tyname (Kind ann) (Type tyname uni ann)
Type lambda

TyApp ann (Type tyname uni ann) (Type tyname uni ann)
Type application

TySOP ann [[Type tyname uni ann]]
Sum-of-products type

typeSubtypes :: Traversal' (Type tyname uni ann) (Type tyname uni ann)

Get all the direct child Kinds of the given Kind.

typeMapNames :: forall tyname tyname' uni ann. (tyname -> tyname') -> Type tyname uni ann -> Type tyname' uni ann

data Kind ann

Constructors

Type ann

KindArrow ann (Kind ann) (Kind ann)

toPatFuncKind :: Kind () -> Kind ()

The kind of a pattern functor (the first Kind argument of TyIFix) at a given kind (of the second Kind argument of TyIFix):

toPatFuncKind k = (k -> *) -> k -> *

fromPatFuncKind :: Kind () -> Maybe (Kind ())

argsFunKind :: Kind ann -> [Kind ann]

Extract all a_i from a_0 -> a_1 -> ... -> r.

data ParserError

An error encountered during parsing.

Constructors

UnknownBuiltinType !Text !SourcePos

BuiltinTypeNotAStar !Text !SourcePos

UnknownBuiltinFunction !Text !SourcePos ![Text]

InvalidBuiltinConstant !Text !Text !SourcePos

data Version

The version of Plutus Core used by this program.

The intention is to convey different levels of backwards compatibility for existing scripts: - Major version changes are backwards-incompatible - Minor version changes are backwards-compatible - Patch version changes should be entirely invisible (and we will likely not use this level)

The version used should be changed only when the language itself changes. For example, adding a new kind of term to the language would require a minor version bump; removing a kind of term would require a major version bump.

Similarly, changing the semantics of the language will require a version bump, typically a major one. This is the main reason why the version is actually tracked in the AST: we can have two language versions with identical ASTs but different semantics, so we need to track the version explicitly.

Compatibility is about compatibility for specific scripts, not about e.g. tools which consume scripts. Adding a new kind of term does not change how existing scripts behave, but does change what tools would need to do to process scripts.

Constructors

Version

Fields

_versionMajor :: Natural

_versionMinor :: Natural

_versionPatch :: Natural

data Program tyname name uni fun ann

A Program is simply a Term coupled with a Version of the core language.

Constructors

Program

Fields

_progAnn :: ann

_progVer :: Version

_progTerm :: Term tyname name uni fun ann

data Name

A Name represents variables/names in Plutus Core.

Constructors

Name

Fields

_nameText :: Text

The identifier name, for use in error messages.

_nameUnique :: Unique

A Unique assigned to the name, allowing for cheap comparisons in the compiler.

newtype TyName

We use a newtype to enforce separation between names used for types and those used for terms.

Constructors

TyName

Fields

unTyName :: Name

newtype Unique

A unique identifier

Constructors

Unique

Fields

unUnique :: Int

newtype UniqueMap unique a

A mapping from Uniques to arbitrary values of type a. Since Unique is equivalent to Int (see PlutusCore.Name.Unique), we can use an IntMap representation for this type.

Constructors

UniqueMap

Fields

unUniqueMap :: IntMap a

newtype UniqueSet unique

A set containing Uniques. Since Unique is equivalent to Int (see PlutusCore.Name.Unique), we can use an IntSet representation for this type.

Constructors

UniqueSet

Fields

unUniqueSet :: IntSet

newtype Normalized a

Constructors

Normalized

Fields

unNormalized :: a

latestVersion :: Version

The latest version of Plutus Core supported by this library.

termAnn :: Term tyname name uni fun ann -> ann

typeAnn :: Type tyname uni ann -> ann

tyVarDeclAnn :: forall tyname ann. Lens' (TyVarDecl tyname ann) ann

tyVarDeclName :: forall tyname ann tyname. Lens (TyVarDecl tyname ann) (TyVarDecl tyname ann) tyname tyname

tyVarDeclKind :: forall tyname ann. Lens' (TyVarDecl tyname ann) (Kind ann)

varDeclAnn :: forall tyname name uni ann. Lens' (VarDecl tyname name uni ann) ann

varDeclName :: forall tyname name uni ann name. Lens (VarDecl tyname name uni ann) (VarDecl tyname name uni ann) name name

varDeclType :: forall tyname name uni ann tyname uni. Lens (VarDecl tyname name uni ann) (VarDecl tyname name uni ann) (Type tyname uni ann) (Type tyname uni ann)

tyDeclAnn :: forall tyname uni ann. Lens' (TyDecl tyname uni ann) ann

tyDeclType :: forall tyname uni ann tyname uni. Lens (TyDecl tyname uni ann) (TyDecl tyname uni ann) (Type tyname uni ann) (Type tyname uni ann)

tyDeclKind :: forall tyname uni ann. Lens' (TyDecl tyname uni ann) (Kind ann)

progAnn :: forall tyname name uni fun ann. Lens' (Program tyname name uni fun ann) ann

progVer :: forall tyname name uni fun ann. Lens' (Program tyname name uni fun ann) Version

progTerm :: forall tyname name uni fun ann tyname name uni fun. Lens (Program tyname name uni fun ann) (Program tyname name uni fun ann) (Term tyname name uni fun ann) (Term tyname name uni fun ann)

mapFun :: (fun -> fun') -> Term tyname name uni fun ann -> Term tyname name uni fun' ann

Map a function over the set of built-in functions.

DeBruijn representation

newtype DeBruijn

A term name as a de Bruijn index, without the name string.

Constructors

DeBruijn

Fields

dbnIndex :: Index

newtype TyDeBruijn

A type name as a de Bruijn index, without the name string.

Constructors

TyDeBruijn DeBruijn

data NamedDeBruijn

A term name as a de Bruijn index.

Constructors

NamedDeBruijn

Fields

ndbnString :: !Text

ndbnIndex :: !Index

newtype NamedTyDeBruijn

A type name as a de Bruijn index.

Constructors

NamedTyDeBruijn NamedDeBruijn

deBruijnTerm :: (AsFreeVariableError e, MonadError e m) => Term TyName Name uni fun ann -> m (Term NamedTyDeBruijn NamedDeBruijn uni fun ann)

Convert a Term with TyNames and Names into a Term with NamedTyDeBruijns and NamedDeBruijns. Will throw an error if a free variable is encountered.

unDeBruijnTerm :: (MonadQuote m, AsFreeVariableError e, MonadError e m) => Term NamedTyDeBruijn NamedDeBruijn uni fun ann -> m (Term TyName Name uni fun ann)

Convert a Term with NamedTyDeBruijns and NamedDeBruijns into a Term with TyNames and Names. Will throw an error if a free variable is encountered.

Processing

type family HasUniques a :: Constraint

All kinds of uniques an entity contains.

class Rename a where

The class of things that can be renamed. I.e. things that are capable of satisfying the global uniqueness condition.

Methods

rename :: MonadQuote m => a -> m a

Rename Uniques so that they're globally unique. In case there are any free variables, they must be left untouched and bound variables must not get renamed to free ones. Must always assign new names to bound variables, so that rename can be used for alpha-renaming as well.

Type checking

module PlutusCore.TypeCheck

normalizeTypesIn :: (HasUnique tyname TypeUnique, HasUnique name TermUnique, MonadNormalizeType uni m) => Term tyname name uni fun ann -> m (Term tyname name uni fun ann)

Normalize every Kind in a Term.

normalizeTypesInProgram :: (HasUnique tyname TypeUnique, HasUnique name TermUnique, MonadNormalizeType uni m) => Program tyname name uni fun ann -> m (Program tyname name uni fun ann)

Normalize every Kind in a Program.

class AsTypeError r term uni fun ann | r -> term uni fun ann where

Minimal complete definition

_TypeError

Methods

_TypeError :: Prism' r (TypeError term uni fun ann)

_KindMismatch :: Prism' r (ann, Type TyName uni (), ExpectedShapeOr (Kind ()), Kind ())

_TypeMismatch :: Prism' r (ann, term, ExpectedShapeOr (Type TyName uni ()), Normalized (Type TyName uni ()))

_TyNameMismatch :: Prism' r (ann, TyName, TyName)

_NameMismatch :: Prism' r (ann, Name, Name)

_FreeTypeVariableE :: Prism' r (ann, TyName)

_FreeVariableE :: Prism' r (ann, Name)

_UnknownBuiltinFunctionE :: Prism' r (ann, fun)

data TypeError term uni fun ann

Errors

data Error uni fun ann

Constructors

ParseErrorE !ParserErrorBundle

UniqueCoherencyErrorE !(UniqueError ann)

TypeErrorE !(TypeError (Term TyName Name uni fun ()) uni fun ann)

NormCheckErrorE !(NormCheckError TyName Name uni fun ann)

FreeVariableErrorE !FreeVariableError

class AsError r uni fun ann | r -> uni fun ann where

Minimal complete definition

_Error

Methods

_Error :: Prism' r (Error uni fun ann)

_ParseErrorE :: Prism' r ParserErrorBundle

_UniqueCoherencyErrorE :: Prism' r (UniqueError ann)

_TypeErrorE :: Prism' r (TypeError (Term TyName Name uni fun ()) uni fun ann)

_NormCheckErrorE :: Prism' r (NormCheckError TyName Name uni fun ann)

_FreeVariableErrorE :: Prism' r FreeVariableError

data NormCheckError tyname name uni fun ann

Constructors

BadType !ann !(Type tyname uni ann) !Text

BadTerm !ann !(Term tyname name uni fun ann) !Text

class AsNormCheckError r tyname name uni fun ann | r -> tyname name uni fun ann where

Minimal complete definition

_NormCheckError

Methods

_NormCheckError :: Prism' r (NormCheckError tyname name uni fun ann)

_BadType :: Prism' r (ann, Type tyname uni ann, Text)

_BadTerm :: Prism' r (ann, Term tyname name uni fun ann, Text)

data UniqueError ann

Constructors

MultiplyDefined !Unique !ann !ann

IncoherentUsage !Unique !ann !ann

FreeVariable !Unique !ann

class AsUniqueError r ann | r -> ann where

Minimal complete definition

_UniqueError

Methods

_UniqueError :: Prism' r (UniqueError ann)

_MultiplyDefined :: Prism' r (Unique, ann, ann)

_IncoherentUsage :: Prism' r (Unique, ann, ann)

_FreeVariable :: Prism' r (Unique, ann)

data FreeVariableError

We cannot do a correct translation to or from de Bruijn indices if the program is not well-scoped. So we throw an error in such a case.

Constructors

FreeUnique !Unique

FreeIndex !Index

class AsFreeVariableError r where

Minimal complete definition

_FreeVariableError

Methods

_FreeVariableError :: Prism' r FreeVariableError

_FreeUnique :: Prism' r Unique

_FreeIndex :: Prism' r Index

Quotation and term construction

type Quote = QuoteT Identity

A non-transformer version of QuoteT.

runQuote :: Quote a -> a

See runQuoteT.

data QuoteT m a

The "quotation" monad transformer. Within this monad you can do safe construction of PLC terms using quasiquotation, fresh-name generation, and parsing.

runQuoteT :: Monad m => QuoteT m a -> m a

Run a quote from an empty identifier state. Note that the resulting term cannot necessarily be safely combined with other terms - that should happen inside QuoteT.

class Monad m => MonadQuote m

A monad that allows lifting of quoted expressions.

liftQuote :: MonadQuote m => Quote a -> m a

Name generation

freshUnique :: MonadQuote m => m Unique

Get a fresh Unique.

freshName :: MonadQuote m => Text -> m Name

Get a fresh Name, given the annotation and the Text name.

freshTyName :: MonadQuote m => Text -> m TyName

Get a fresh TyName, given the annotation and the Text name.

Evaluation

data EvaluationResult a

The parameterized type of results various evaluation engines return. On the PLC side this becomes (via makeKnown) either a call to Error or a value of the PLC counterpart of type a.

Constructors

EvaluationSuccess !a

EvaluationFailure

Combining programs

applyProgram :: Semigroup a => Program tyname name uni fun a -> Program tyname name uni fun a -> Either ApplyProgramError (Program tyname name uni fun a)

Applies one program to another. Fails if the versions do not match and tries to merge annotations.

Benchmarking

termSize :: Term tyname name uni fun ann -> Size

Count the number of AST nodes in a term.

typeSize :: Type tyname uni ann -> Size

Count the number of AST nodes in a type.

kindSize :: Kind a -> Size

Count the number of AST nodes in a kind.

>>> kindSize $ Type ()
Size {unSize = 1}
>>> kindSize $ KindArrow () (KindArrow () (Type ()) (Type ())) (Type ())
Size {unSize = 5}

programSize :: Program tyname name uni fun ann -> Size

Count the number of AST nodes in a program.

serialisedSize :: Flat a => a -> Integer

Compute the size of the serializabled form of a value.