PlutusCore

parseTerm :: (Text -> m (Term TyName Name DefaultUni DefaultFun SrcSpan)

Parse a PLC term. The resulting program will have fresh names. The underlying monad must be capable of handling any parse errors.

parseType :: (Text -> m (Type TyName DefaultUni SrcSpan)

Parse a PLC type. The resulting program will have fresh names. The underlying monad must be capable of handling any parse errors.

data SourcePos

The data type SourcePos represents source positions. It contains the name of the source file, a line number, and a column number. Source line and column positions change intensively during parsing, so we need to make them strict to avoid memory leaks.

data SrcSpan

The span between two source locations.

This corresponds roughly to the SrcSpan used by GHC, but we define our own version so we don't have to depend on ghc to use it.

The line and column numbers are 1-based, and the unit is Unicode code point (or Char).

Constructors

SrcSpan

Fields

srcSpanFile :: FilePath

srcSpanSLine :: Int

srcSpanSCol :: Int

srcSpanELine :: Int

srcSpanECol :: Int

Same as GHC's SrcSpan, srcSpanECol is usually one more than the column of the last character of the thing this SrcSpan is for (unless the last character is the line break).

Builtins

data Some (tag :: k -> Type) where

Existential. This is type is useful to hide GADTs' parameters.

>>> data Tag :: Type -> Type where TagInt :: Tag Int; TagBool :: Tag Bool
>>> instance GShow Tag where gshowsPrec _ TagInt = showString "TagInt"; gshowsPrec _ TagBool = showString "TagBool"
>>> classify s = case s of "TagInt" -> [mkGReadResult TagInt]; "TagBool" -> [mkGReadResult TagBool]; _ -> []
>>> instance GRead Tag where greadsPrec _ s = [ (r, rest) | (con, rest) <-  lex s, r <- classify con ]

You can either use PatternSynonyms (available with GHC >= 8.0)

>>> let x = Some TagInt
>>> x
Some TagInt

>>> case x of { Some TagInt -> "I"; Some TagBool -> "B" } :: String
"I"

or you can use functions

>>> let y = mkSome TagBool
>>> y
Some TagBool

>>> withSome y $ \y' -> case y' of { TagInt -> "I"; TagBool -> "B" } :: String
"B"

The implementation of mapSome is safe.

>>> let f :: Tag a -> Tag a; f TagInt = TagInt; f TagBool = TagBool
>>> mapSome f y
Some TagBool

but you can also use:

>>> withSome y (mkSome . f)
Some TagBool

>>> read "Some TagBool" :: Some Tag
Some TagBool

>>> read "mkSome TagInt" :: Some Tag
Some TagInt

Bundled Patterns

pattern Some :: forall {k} tag (a :: k). () => tag a -> Some tag

(Closed uni, Everywhere uni PrettyConst) => Pretty (ValueOf uni a)

Instance details

Defined in PlutusCore.Pretty.PrettyConst

Methods

pretty :: ValueOf uni a -> Doc ann

prettyList :: [ValueOf uni a] -> Doc ann

(Closed uni, Everywhere uni PrettyConst) => Pretty (Some (ValueOf uni))

Instance details

Defined in PlutusCore.Pretty.PrettyConst

Methods

pretty :: Some (ValueOf uni) -> Doc ann

prettyList :: [Some (ValueOf uni)] -> Doc ann

someValueOf :: forall a uni. uni (Esc a) -> a -> Some (ValueOf uni)

Wrap a value into Some (ValueOf uni), given its explicit type tag.

someValue :: forall a uni. uni `Contains` a => a -> Some (ValueOf uni)

Wrap a value into Some (ValueOf uni), provided its type is in the universe.

someValueType :: Some (ValueOf uni) -> SomeTypeIn uni

data Esc a

"Escapes" a type of an arbitrary kind to fit into Type.

class Contains uni a where

A class for enumerating types and fully instantiated type formers that uni contains. For example, a particular ExampleUni may have monomorphic types in it:

instance ExampleUni Contains Integer where ... instance ExampleUni Contains Bool where ...

as well as polymorphic ones:

instance ExampleUni Contains [] where ... instance ExampleUni Contains (,) where ...

as well as their instantiations:

instance ExampleUni Contains a => ExampleUni Contains [a] where ... instance (ExampleUni Contains a, ExampleUni Contains b) => ExampleUni Contains (a, b) where ...

(a universe can have any subset of the mentioned sorts of types, for example it's fine to have instantiated polymorphic types and not have uninstantiated ones and vice versa)

Note that when used as a constraint of a function Contains does not allow you to directly express things like "uni has the Integer, Bool and [] types and type formers", because [] is not fully instantiated. So you can only say "uni has Integer, Bool, [Integer], [Bool], [[Integer]], [[Bool]] etc" and such manual enumeration is annoying, so we'd really like to be able to say that uni has lists of arbitrary built-in types (including lists of lists etc). Contains does not allow that, but Includes does. For example, in the body of the following definition:

foo :: (uni Includes Integer, uni Includes Bool, uni Includes []) => ... foo = ...

you can make use of the fact that uni has lists of arbitrary included types (integers, booleans and lists).

Hence most of the time opt for using the more flexible Includes.

Includes is defined in terms of Contains, so you only need to provide a Contains instance per type from the universe and you'll get Includes for free.

Methods

knownUni :: uni (Esc a)

Contains DefaultUni Vector)

Contains DefaultUni List

Instance details

Defined in PlutusCore.Default.Universe

Methods

knownUni :: DefaultUni (Esc List)

(TypeError NoStandalonePolymorphicDataErrMsg :: Constraint) => Contains uni (TyVarRep :: TyNameRep kind -> kind)

Instance details

Defined in PlutusCore.Builtin.Polymorphism

Methods

knownUni :: uni (Esc TyVarRep)

class Closed uni where

A universe is Closed, if it's known how to constrain every type from the universe and every type can be encoded to / decoded from a sequence of integer tags. The universe doesn't have to be finite and providing support for infinite universes is the reason why we encode a type as a sequence of integer tags as opposed to a single integer tag. For example, given

  data U a where
      UList :: !(U a) -> U [a]
      UInt  :: U Int

UList (UList UInt) can be encoded to [0,0,1] where 0 and 1 are the integer tags of the UList and UInt constructors, respectively.

Associated Types

type Everywhere uni (constr :: Type -> Constraint) :: Constraint

A constrant for "constr a holds for any a from uni".

Methods

encodeUni :: uni a -> [Int]

Encode a type as a sequence of Int tags. The opposite of decodeUni.

withDecodedUni :: (forall k (a :: k). Typeable k => uni (Esc a) -> DecodeUniM r) -> DecodeUniM r

Decode a type and feed it to the continuation.

bring :: uni `Everywhere` constr => proxy constr -> uni (Esc a) -> (constr a => r) -> r

Bring a constr a instance in scope, provided a is a type from the universe and constr holds for any type from the universe.

type family EverywhereAll uni constrs where ...

Equations

EverywhereAll uni '[] = ()

EverywhereAll uni (constr ': constrs) = (uni `Everywhere` constr, uni `EverywhereAll` constrs)

knownUniOf :: uni `Contains` a => proxy a -> uni (Esc a)

Same as knownUni, but receives a proxy.

class GShow (t :: k -> Type) where

Show-like class for 1-type-parameter GADTs. GShow t => ... is equivalent to something like (forall a. Show (t a)) => .... The easiest way to create instances would probably be to write (or derive) an instance Show (T a), and then simply say:

instance GShow t where gshowsPrec = defaultGshowsPrec

Methods

gshowsPrec :: forall (a :: k). GOrdering a a0 -> ShowS

(GShow a, GShow b) => GShow (Product a b :: k -> Type)
>>> gshow (Pair Refl Refl :: Product ((:~:) Int) ((:~:) Int) Int)
"Pair Refl Refl"

Instance details

Defined in Data.GADT.Internal

Methods

gshowsPrec :: forall (a0 :: k0). Int -> Product a b a0 -> ShowS

(GShow a, GShow b) => GShow (Sum a b :: k -> Type)
>>> gshow (InL Refl :: Sum ((:~:) Int) ((:~:) Bool) Int)
"InL Refl"

Instance details

Defined in Data.GADT.Internal

Methods

gshowsPrec :: forall (a0 :: k0). Int -> (a :~~: a0) -> ShowS

(GShow a, GShow b) => GShow (a :*: b :: k -> Type)
>>> gshow (Pair Refl Refl :: Product ((:~:) Int) ((:~:) Int) Int)
"Refl :*: Refl"

Since: some-1.0.4

Instance details

Defined in Data.GADT.Internal

Methods

gshowsPrec :: forall (a0 :: k0). Int -> (a :*: b) a0 -> ShowS

(GShow a, GShow b) => GShow (a :+: b :: k -> Type)
>>> gshow (L1 Refl :: ((:~:) Int :+: (:~:) Bool) Int)
"L1 Refl"

Since: some-1.0.4

Instance details

Defined in Data.GADT.Internal

Methods

gshowsPrec :: forall (a0 :: k0). Int -> (a :+: b) a0 -> ShowS

show :: Show a => a -> String

A specialised variant of showsPrec, using precedence context zero, and returning an ordinary String.

class GEq (f :: k -> Type) where

A class for type-contexts which contain enough information to (at least in some cases) decide the equality of types occurring within them.

This class is sometimes confused with TestEquality from base. TestEquality only checks type equality.

Consider

>>> data Tag a where TagInt1 :: Tag Int; TagInt2 :: Tag Int

The correct TestEquality Tag instance is

>>> :{
instance TestEquality Tag where
    testEquality TagInt1 TagInt1 = Just Refl
    testEquality TagInt1 TagInt2 = Just Refl
    testEquality TagInt2 TagInt1 = Just Refl
    testEquality TagInt2 TagInt2 = Just Refl
:}

While we can define

instance GEq Tag where
   geq = testEquality

this will mean we probably want to have

instance Eq Tag where
   _ == _ = True

Note: In the future version of some package (to be released around GHC-9.6 / 9.8) the forall a. Eq (f a) constraint will be added as a constraint to GEq, with a law relating GEq and Eq:

geq x y = Just Refl   ⇒  x == y = True        ∀ (x :: f a) (y :: f b)
x == y                ≡  isJust (geq x y)     ∀ (x, y :: f a)

So, the more useful GEq Tag instance would differentiate between different constructors:

>>> :{
instance GEq Tag where
    geq TagInt1 TagInt1 = Just Refl
    geq TagInt1 TagInt2 = Nothing
    geq TagInt2 TagInt1 = Nothing
    geq TagInt2 TagInt2 = Just Refl
:}

which is consistent with a derived Eq instance for Tag

>>> deriving instance Eq (Tag a)

Note that even if a ~ b, the geq (x :: f a) (y :: f b) may be Nothing (when value terms are inequal).

The consistency of GEq and Eq is easy to check by exhaustion:

>>> let checkFwdGEq :: (forall a. Eq (f a), GEq f) => f a -> f b -> Bool; checkFwdGEq x y = case geq x y of Just Refl -> x == y; Nothing -> True
>>> (checkFwdGEq TagInt1 TagInt1, checkFwdGEq TagInt1 TagInt2, checkFwdGEq TagInt2 TagInt1, checkFwdGEq TagInt2 TagInt2)
(True,True,True,True)

>>> let checkBwdGEq :: (Eq (f a), GEq f) => f a -> f a -> Bool; checkBwdGEq x y = if x == y then isJust (geq x y) else isNothing (geq x y)
>>> (checkBwdGEq TagInt1 TagInt1, checkBwdGEq TagInt1 TagInt2, checkBwdGEq TagInt2 TagInt1, checkBwdGEq TagInt2 TagInt2)
(True,True,True,True)

Methods

geq :: forall (a :: k) (b :: k). f a -> f b -> Maybe (a :~: b)

Produce a witness of type-equality, if one exists.

A handy idiom for using this would be to pattern-bind in the Maybe monad, eg.:

extract :: GEq tag => tag a -> DSum tag -> Maybe a
extract t1 (t2 :=> x) = do
    Refl <- geq t1 t2
    return x

Or in a list comprehension:

extractMany :: GEq tag => tag a -> [DSum tag] -> [a]
extractMany t1 things = [ x | (t2 :=> x) <- things, Refl <- maybeToList (geq t1 t2)]

(Making use of the DSum type from Data.Dependent.Sum in both examples)

class HasUniApply (uni :: Type -> Type) where

A class for "uni has general type application".

Methods

uniApply :: forall k l (f :: k -> l) a. uni (Esc f) -> uni (Esc a) -> uni (Esc (f a))

Apply a type constructor to an argument.

matchUniApply

Arguments

:: uni tb
The type.

-> r
What to return if the type is not an application.

-> (forall k l (f :: k -> l) a. tb ~ Esc (f a) => uni (Esc f) -> uni (Esc a) -> r)
The continuation taking a function and an argument.

-> r

Deconstruct a type application into the function and the argument and feed them to the continuation. If the type is not an application, then return the default value.

checkStar :: forall uni a (x :: a). Typeable a => uni (Esc x) -> Maybe (a :~: Type)

Check if the kind of the given type from the universe is Type.

withApplicable :: forall (a :: Type) (ab :: Type) f x uni m r. (Typeable ab, Typeable a, MonadPlus m) => uni (Esc (f :: ab)) -> uni (Esc (x :: a)) -> (forall (b :: Type). (Typeable b, ab ~ (a -> b)) => m r) -> m r

Check if one type from the universe can be applied to another (i.e. check that the expected kind of the argument matches the actual one) and call the continuation in the refined context. Fail with mzero otherwise.

data (a :: k) :~: (b :: k) where infix 4

Propositional equality. If a :~: b is inhabited by some terminating value, then the type a is the same as the type b. To use this equality in practice, pattern-match on the a :~: b to get out the Refl constructor; in the body of the pattern-match, the compiler knows that a ~ b.

Since: base-4.7.0.0

Constructors

Refl :: forall {k} (a :: k). a :~: a

Show (DefaultUni a)

Instance details

Defined in PlutusCore.Default.Universe

Methods

showsPrec :: Int -> DefaultUni a -> ShowS

show :: DefaultUni a -> String

showList :: [DefaultUni a] -> ShowS

Pretty (DefaultUni a)
This always pretty-prints parens around type applications (e.g. (list bool)) and doesn't pretty-print them otherwise (e.g. integer).

Instance details

Defined in PlutusCore.Default.Universe

Methods

pretty :: DefaultUni a -> Doc ann

prettyList :: [DefaultUni a] -> Doc ann

Pretty (SomeTypeIn DefaultUni)

Instance details

Defined in PlutusCore.Default.Universe

Methods

pretty :: SomeTypeIn DefaultUni -> Doc ann

prettyList :: [SomeTypeIn DefaultUni] -> Doc ann

Contains DefaultUni Text)

type ToHoles DefaultUni hole Vector) => DefaultUni (Esc a1) -> DefaultUni a

data DefaultFun

Default built-in functions.

When updating these, make sure to add them to the protocol version listing! See Note [New builtins/language versions and protocol versions]

Constructors

AddInteger

SubtractInteger

MultiplyInteger

DivideInteger

QuotientInteger

RemainderInteger

ModInteger

EqualsInteger

LessThanInteger

LessThanEqualsInteger

AppendByteString

ConsByteString

SliceByteString

LengthOfByteString

IndexByteString

EqualsByteString

LessThanByteString

LessThanEqualsByteString

Sha2_256

Sha3_256

Blake2b_256

VerifyEd25519Signature

VerifyEcdsaSecp256k1Signature

VerifySchnorrSecp256k1Signature

AppendString

EqualsString

EncodeUtf8

DecodeUtf8

IfThenElse

ChooseUnit

Trace

FstPair

SndPair

ChooseList

MkCons

HeadList

TailList

NullList

ChooseData

ConstrData

MapData

ListData

IData

BData

UnConstrData

UnMapData

UnListData

UnIData

UnBData

EqualsData

SerialiseData

MkPairData

MkNilData

MkNilPairData

Bls12_381_G1_add

Bls12_381_G1_neg

Bls12_381_G1_scalarMul

Bls12_381_G1_equal

Bls12_381_G1_hashToGroup

Bls12_381_G1_compress

Bls12_381_G1_uncompress

Bls12_381_G2_add

Bls12_381_G2_neg

Bls12_381_G2_scalarMul

Bls12_381_G2_equal

Bls12_381_G2_hashToGroup

Bls12_381_G2_compress

Bls12_381_G2_uncompress

Bls12_381_millerLoop

Bls12_381_mulMlResult

Bls12_381_finalVerify

Keccak_256

Blake2b_224

IntegerToByteString

ByteStringToInteger

AndByteString

OrByteString

XorByteString

ComplementByteString

ReadBit

WriteBits

ReplicateByte

ShiftByteString

RotateByteString

CountSetBits

FindFirstSetBit

Ripemd_160

ExpModInteger

DropList

LengthOfArray

ListToArray

IndexArray

Bls12_381_G1_multiScalarMul

Bls12_381_G2_multiScalarMul

InsertCoin

LookupCoin

UnionValue

ValueContains

ValueData

UnValueData

ScaleValue

Pretty DefaultFun

Instance details

Defined in PlutusCore.Default.Builtins

Methods

pretty :: DefaultFun -> Doc ann

prettyList :: [DefaultFun] -> Doc ann

uni ThunkInfo)
showTypeOf :: Proxy (BuiltinSemanticsVariant DefaultFun) -> String

Pretty (BuiltinSemanticsVariant DefaultFun)

Instance details

Defined in PlutusCore.Default.Builtins

Methods

pretty :: BuiltinSemanticsVariant DefaultFun -> Doc ann

prettyList :: [BuiltinSemanticsVariant DefaultFun] -> Doc ann

type Rep DefaultFun

Instance details

Defined in PlutusCore.Default.Builtins

type Rep DefaultFun = D1 ('MetaData "DefaultFun" "PlutusCore.Default.Builtins" "plutus-core-1.60.0.0-inplace" 'False) ((((((C1 ('MetaCons "AddInteger" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "SubtractInteger" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "MultiplyInteger" 'PrefixI 'False) (U1 :: Type -> Type))) :+: (C1 ('MetaCons "DivideInteger" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "QuotientInteger" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "RemainderInteger" 'PrefixI 'False) (U1 :: Type -> Type)))) :+: ((C1 ('MetaCons "ModInteger" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "EqualsInteger" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "LessThanInteger" 'PrefixI 'False) (U1 :: Type -> Type))) :+: (C1 ('MetaCons "LessThanEqualsInteger" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "AppendByteString" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "ConsByteString" 'PrefixI 'False) (U1 :: Type -> Type))))) :+: (((C1 ('MetaCons "SliceByteString" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "LengthOfByteString" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "IndexByteString" 'PrefixI 'False) (U1 :: Type -> Type))) :+: (C1 ('MetaCons "EqualsByteString" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "LessThanByteString" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "LessThanEqualsByteString" 'PrefixI 'False) (U1 :: Type -> Type)))) :+: ((C1 ('MetaCons "Sha2_256" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "Sha3_256" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "Blake2b_256" 'PrefixI 'False) (U1 :: Type -> Type))) :+: ((C1 ('MetaCons "VerifyEd25519Signature" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "VerifyEcdsaSecp256k1Signature" 'PrefixI 'False) (U1 :: Type -> Type)) :+: (C1 ('MetaCons "VerifySchnorrSecp256k1Signature" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "AppendString" 'PrefixI 'False) (U1 :: Type -> Type)))))) :+: ((((C1 ('MetaCons "EqualsString" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "EncodeUtf8" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "DecodeUtf8" 'PrefixI 'False) (U1 :: Type -> Type))) :+: (C1 ('MetaCons "IfThenElse" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "ChooseUnit" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "Trace" 'PrefixI 'False) (U1 :: Type -> Type)))) :+: ((C1 ('MetaCons "FstPair" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "SndPair" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "ChooseList" 'PrefixI 'False) (U1 :: Type -> Type))) :+: (C1 ('MetaCons "MkCons" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "HeadList" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "TailList" 'PrefixI 'False) (U1 :: Type -> Type))))) :+: (((C1 ('MetaCons "NullList" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "ChooseData" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "ConstrData" 'PrefixI 'False) (U1 :: Type -> Type))) :+: (C1 ('MetaCons "MapData" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "ListData" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "IData" 'PrefixI 'False) (U1 :: Type -> Type)))) :+: ((C1 ('MetaCons "BData" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "UnConstrData" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "UnMapData" 'PrefixI 'False) (U1 :: Type -> Type))) :+: ((C1 ('MetaCons "UnListData" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "UnIData" 'PrefixI 'False) (U1 :: Type -> Type)) :+: (C1 ('MetaCons "UnBData" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "EqualsData" 'PrefixI 'False) (U1 :: Type -> Type))))))) :+: (((((C1 ('MetaCons "SerialiseData" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "MkPairData" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "MkNilData" 'PrefixI 'False) (U1 :: Type -> Type))) :+: (C1 ('MetaCons "MkNilPairData" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "Bls12_381_G1_add" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "Bls12_381_G1_neg" 'PrefixI 'False) (U1 :: Type -> Type)))) :+: ((C1 ('MetaCons "Bls12_381_G1_scalarMul" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "Bls12_381_G1_equal" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "Bls12_381_G1_hashToGroup" 'PrefixI 'False) (U1 :: Type -> Type))) :+: (C1 ('MetaCons "Bls12_381_G1_compress" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "Bls12_381_G1_uncompress" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "Bls12_381_G2_add" 'PrefixI 'False) (U1 :: Type -> Type))))) :+: (((C1 ('MetaCons "Bls12_381_G2_neg" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "Bls12_381_G2_scalarMul" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "Bls12_381_G2_equal" 'PrefixI 'False) (U1 :: Type -> Type))) :+: (C1 ('MetaCons "Bls12_381_G2_hashToGroup" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "Bls12_381_G2_compress" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "Bls12_381_G2_uncompress" 'PrefixI 'False) (U1 :: Type -> Type)))) :+: ((C1 ('MetaCons "Bls12_381_millerLoop" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "Bls12_381_mulMlResult" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "Bls12_381_finalVerify" 'PrefixI 'False) (U1 :: Type -> Type))) :+: ((C1 ('MetaCons "Keccak_256" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "Blake2b_224" 'PrefixI 'False) (U1 :: Type -> Type)) :+: (C1 ('MetaCons "IntegerToByteString" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "ByteStringToInteger" 'PrefixI 'False) (U1 :: Type -> Type)))))) :+: ((((C1 ('MetaCons "AndByteString" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "OrByteString" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "XorByteString" 'PrefixI 'False) (U1 :: Type -> Type))) :+: (C1 ('MetaCons "ComplementByteString" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "ReadBit" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "WriteBits" 'PrefixI 'False) (U1 :: Type -> Type)))) :+: ((C1 ('MetaCons "ReplicateByte" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "ShiftByteString" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "RotateByteString" 'PrefixI 'False) (U1 :: Type -> Type))) :+: ((C1 ('MetaCons "CountSetBits" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "FindFirstSetBit" 'PrefixI 'False) (U1 :: Type -> Type)) :+: (C1 ('MetaCons "Ripemd_160" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "ExpModInteger" 'PrefixI 'False) (U1 :: Type -> Type))))) :+: (((C1 ('MetaCons "DropList" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "LengthOfArray" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "ListToArray" 'PrefixI 'False) (U1 :: Type -> Type))) :+: (C1 ('MetaCons "IndexArray" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "Bls12_381_G1_multiScalarMul" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "Bls12_381_G2_multiScalarMul" 'PrefixI 'False) (U1 :: Type -> Type)))) :+: ((C1 ('MetaCons "InsertCoin" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "LookupCoin" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "UnionValue" 'PrefixI 'False) (U1 :: Type -> Type))) :+: ((C1 ('MetaCons "ValueContains" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "ValueData" 'PrefixI 'False) (U1 :: Type -> Type)) :+: (C1 ('MetaCons "UnValueData" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "ScaleValue" 'PrefixI 'False) (U1 :: Type -> Type))))))))

data BuiltinSemanticsVariant DefaultFun

Instance details

Defined in PlutusCore.Default.Builtins

data BuiltinSemanticsVariant DefaultFun

= DefaultFunSemanticsVariantA

| DefaultFunSemanticsVariantB

| DefaultFunSemanticsVariantC

| DefaultFunSemanticsVariantD

| DefaultFunSemanticsVariantE

type CostingPart uni DefaultFun

Instance details

Defined in PlutusCore.Default.Builtins

type CostingPart uni DefaultFun = BuiltinCostModel

type Rep (BuiltinSemanticsVariant DefaultFun)

Instance details

Defined in PlutusCore.Default.Builtins

type Rep (BuiltinSemanticsVariant DefaultFun) = D1 ('MetaData "BuiltinSemanticsVariant" "PlutusCore.Default.Builtins" "plutus-core-1.60.0.0-inplace" 'False) ((C1 ('MetaCons "DefaultFunSemanticsVariantA" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "DefaultFunSemanticsVariantB" 'PrefixI 'False) (U1 :: Type -> Type)) :+: (C1 ('MetaCons "DefaultFunSemanticsVariantC" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "DefaultFunSemanticsVariantD" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "DefaultFunSemanticsVariantE" 'PrefixI 'False) (U1 :: Type -> Type))))

AST

data Term tyname name uni fun ann

Constructors

Var ann name
a named variable

LamAbs ann name (Type tyname uni ann) (Term tyname name uni fun ann)
lambda abstraction

Apply ann (Term tyname name uni fun ann) (Term tyname name uni fun ann)
application

TyAbs ann tyname (Kind ann) (Term tyname name uni fun ann)
type abstraction

TyInst ann (Term tyname name uni fun ann) (Type tyname uni ann)
instantiation

IWrap ann (Type tyname uni ann) (Type tyname uni ann) (Term tyname name uni fun ann)
wrapping

Unwrap ann (Term tyname name uni fun ann)
unwrapping See Note [Constr tag type]

Constr ann (Type tyname uni ann) Word64 [Term tyname name uni fun ann]
constructor

Case ann (Type tyname uni ann) (Term tyname name uni fun ann) [Term tyname name uni fun ann]
case

Constant ann (Some (ValueOf uni))
constants

Builtin ann fun
builtin functions

Error ann (Type tyname uni ann)
fail with error

Pretty ann => Pretty (Kind ann)

Instance details

Defined in PlutusCore.Core.Instance.Pretty.Default

Methods

pretty :: Kind ann -> Doc ann0

prettyList :: [Kind ann] -> Doc ann0

Pretty ann => PrettyBy (PrettyConfigClassic configName) (Kind ann)

Instance details

Defined in PlutusCore.Core.Instance.Pretty.Classic

Methods

prettyBy :: PrettyConfigClassic configName -> Kind ann -> Doc ann0

prettyListBy :: PrettyConfigClassic configName -> [Kind ann] -> Doc ann0

PrettyBy (PrettyConfigReadable configName) (Kind a)

Instance details

Defined in PlutusCore.Core.Instance.Pretty.Readable

Methods

prettyBy :: PrettyConfigReadable configName -> Kind a -> Doc ann

prettyListBy :: PrettyConfigReadable configName -> [Kind a] -> Doc ann

type Rep (Kind ann)

Instance details

Defined in PlutusCore.Core.Type

type Rep (Kind ann) = D1 ('MetaData "Kind" "PlutusCore.Core.Type" "plutus-core-1.60.0.0-inplace" 'False) (C1 ('MetaCons "Type" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 ann)) :+: C1 ('MetaCons "KindArrow" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 ann) :*: (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 (Kind ann)) :*: S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 (Kind ann)))))

type HasUniques (Kind ann)

Instance details

Defined in PlutusCore.Core.Type

type HasUniques (Kind ann) = ()

toPatFuncKind :: Kind () -> Kind ()

The kind of a pattern functor (the first Kind argument of TyIFix) at a given kind (of the second Kind argument of TyIFix):

toPatFuncKind k = (k -> *) -> k -> * 

fromPatFuncKind :: Kind () -> Text) :*: S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'SourceStrict 'DecidedStrict) (Rec0 SourcePos)))))

data Version

The version of Plutus Core used by this program.

The intention is to convey different levels of backwards compatibility for existing scripts: - Major version changes are backwards-incompatible - Minor version changes are backwards-compatible - Patch version changes should be entirely invisible (and we will likely not use this level)

The version used should be changed only when the language itself changes. For example, adding a new kind of term to the language would require a minor version bump; removing a kind of term would require a major version bump.

Similarly, changing the semantics of the language will require a version bump, typically a major one. This is the main reason why the version is actually tracked in the AST: we can have two language versions with identical ASTs but different semantics, so we need to track the version explicitly.

Compatibility is about compatibility for specific scripts, not about e.g. tools which consume scripts. Adding a new kind of term does not change how existing scripts behave, but does change what tools would need to do to process scripts.

Constructors

Version

Fields

_versionMajor :: Natural

_versionMinor :: Natural

_versionPatch :: Natural

newtype TyName

We use a newtype to enforce separation between names used for types and those used for terms.

Constructors

TyName

Fields

unTyName :: Name

newtype Unique

A unique identifier This is normally a positive integral number, except in topUnique where we make use of a negative unique to signify top-level.

Constructors

Unique

Fields

unUnique :: NumBits

Pretty Unique

Instance details

Defined in PlutusCore.Name.Unique

Methods

pretty :: Unique -> Doc ann

prettyList :: [Unique] -> Doc ann

HasUnique Unique Unique

Instance details

Defined in PlutusCore.Name.Unique

Methods

unique :: Lens' Unique Unique

Lift Unique

Instance details

Defined in PlutusCore.Name.Unique

Methods

lift :: Quote m => Unique -> m Exp

liftTyped :: forall (m :: Type -> Type). Quote m => Unique -> Code m Unique

newtype UniqueMap unique a

A mapping from Uniques to arbitrary values of type a. Since Unique is equivalent to Int (see PlutusCore.Name.Unique), we can use an IntMap representation for this type.

Constructors

UniqueMap

Fields

unUniqueMap :: IntMap a

newtype UniqueSet unique

A set containing Uniques. Since Unique is equivalent to Int (see PlutusCore.Name.Unique), we can use an IntSet representation for this type.

Constructors

UniqueSet

Fields

unUniqueSet :: NumBits

(GEq uni, Closed uni, Everywhere uni Eq, Eq ann) => Eq (Type NamedTyDeBruijn uni ann)

Instance details

Defined in PlutusCore.Core.Instance.Eq

Methods

(==) :: Type NamedTyDeBruijn uni ann -> Type NamedTyDeBruijn uni ann -> Bool

(/=) :: Type NamedTyDeBruijn uni ann -> Type NamedTyDeBruijn uni ann -> Bool

(GEq uni, Closed uni, Everywhere uni Eq, Eq fun, Eq ann) => Eq (Term NamedTyDeBruijn NamedDeBruijn uni fun ann)

Instance details

Defined in PlutusCore.Core.Instance.Eq

Methods

(==) :: Term NamedTyDeBruijn NamedDeBruijn uni fun ann -> Term NamedTyDeBruijn NamedDeBruijn uni fun ann -> Bool

(/=) :: Term NamedTyDeBruijn NamedDeBruijn uni fun ann -> Term NamedTyDeBruijn NamedDeBruijn uni fun ann -> Bool

type Rep NamedTyDeBruijn

Instance details

Defined in PlutusCore.DeBruijn.Internal

type Rep NamedTyDeBruijn = D1 ('MetaData "NamedTyDeBruijn" "PlutusCore.DeBruijn.Internal" "plutus-core-1.60.0.0-inplace" 'True) (C1 ('MetaCons "NamedTyDeBruijn" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 NamedDeBruijn)))

type Unwrapped NamedTyDeBruijn

Instance details

Defined in PlutusCore.DeBruijn.Internal

type Unwrapped NamedTyDeBruijn = GUnwrapped (Rep NamedTyDeBruijn)

deBruijnTerm :: MonadError FreeVariableError m => Term TyName Name uni fun ann -> m (Term NamedTyDeBruijn NamedDeBruijn uni fun ann)

Convert a Term with TyNames and Names into a Term with NamedTyDeBruijns and NamedDeBruijns. Will throw an error if a free variable is encountered.

unDeBruijnTerm :: (MonadQuote m, MonadError FreeVariableError m) => Term NamedTyDeBruijn NamedDeBruijn uni fun ann -> m (Term TyName Name uni fun ann)

Convert a Term with NamedTyDeBruijns and NamedDeBruijns into a Term with TyNames and Names. Will throw an error if a free variable is encountered.

Processing

type family HasUniques a :: Constraint

All kinds of uniques an entity contains.

class Rename a where

The class of things that can be renamed. I.e. things that are capable of satisfying the global uniqueness condition.

Methods

rename :: MonadQuote m => a -> m a

Rename Uniques so that they're globally unique. In case there are any free variables, they must be left untouched and bound variables must not get renamed to free ones. Must always assign new names to bound variables, so that rename can be used for alpha-renaming as well.

Type checking

module PlutusCore.TypeCheck

normalizeTypesIn :: (HasUnique tyname TypeUnique, HasUnique name TermUnique, MonadNormalizeType uni m) => Term tyname name uni fun ann -> m (Term tyname name uni fun ann)

Normalize every Kind in a Term.

normalizeTypesInProgram :: (HasUnique tyname TypeUnique, HasUnique name TermUnique, MonadNormalizeType uni m) => Program tyname name uni fun ann -> m (Program tyname name uni fun ann)

Normalize every Kind in a Program.

data TypeError term uni fun ann

data UniqueError ann

Constructors

MultiplyDefined !Unique !ann !ann

IncoherentUsage !Unique !ann !ann

FreeVariable !Unique !ann

data FreeVariableError

We cannot do a correct translation to or from de Bruijn indices if the program is not well-scoped. So we throw an error in such a case.

Constructors

FreeUnique !Unique

FreeIndex !Index

Quotation and term construction

type Quote = QuoteT Identity

A non-transformer version of QuoteT.

runQuote :: Quote a -> a

See runQuoteT.

data QuoteT m a

The "quotation" monad transformer. Within this monad you can do safe construction of PLC terms using quasiquotation, fresh-name generation, and parsing.

runQuoteT :: Monad m => QuoteT m a -> m a

Run a quote from an empty identifier state. Note that the resulting term cannot necessarily be safely combined with other terms - that should happen inside QuoteT.

class Text name.

Evaluation

data EvaluationResult a

The parameterized type of results various evaluation engines return.

Constructors

EvaluationSuccess !a

EvaluationFailure

Methods

typeAst :: Type0 tyname uni ()

(TypeError ('Text "Use \8216BuiltinResult\8217 instead of \8216EvaluationResult\8217") :: Constraint, uni ~ UniOf val) => MakeKnownIn uni val (EvaluationResult a)

Instance details

Defined in PlutusCore.Builtin.KnownType

Methods

makeKnown :: EvaluationResult a -> BuiltinResult val

(TypeError ('Text "Use \8216BuiltinResult\8217 instead of \8216EvaluationResult\8217") :: Constraint, uni ~ UniOf val) => ReadKnownIn uni val (EvaluationResult a)

Instance details

Defined in PlutusCore.Builtin.KnownType

Methods

readKnown :: val -> ReadKnownM (EvaluationResult a)

PrettyBy config a => PrettyBy config (EvaluationResult a)

Instance details

Defined in PlutusCore.Evaluation.Result

Methods

prettyBy :: config -> EvaluationResult a -> Doc ann

prettyListBy :: config -> [EvaluationResult a] -> Doc ann

Generic (EvaluationResult a)

Instance details

Defined in PlutusCore.Evaluation.Result

Associated Types

type Rep (EvaluationResult a) :: Type -> Type

Methods

from :: EvaluationResult a -> Rep (EvaluationResult a) x

to :: Rep (EvaluationResult a) x -> EvaluationResult a

Show a => Show (EvaluationResult a)

Instance details

Defined in PlutusCore.Evaluation.Result

Methods

showsPrec :: Int -> EvaluationResult a -> ShowS

show :: EvaluationResult a -> String

showList :: [EvaluationResult a] -> ShowS

NFData a => NFData (EvaluationResult a)

Instance details

Defined in PlutusCore.Evaluation.Result

Methods

rnf :: EvaluationResult a -> ()

Eq a => Eq (EvaluationResult a)

Instance details

Defined in PlutusCore.Evaluation.Result

Methods

(==) :: EvaluationResult a -> EvaluationResult a -> Bool

(/=) :: EvaluationResult a -> EvaluationResult a -> Bool

PrettyClassic a => Pretty (EvaluationResult a)

Instance details

Defined in PlutusCore.Evaluation.Result

Methods

pretty :: EvaluationResult a -> Doc ann

prettyList :: [EvaluationResult a] -> Doc ann

type ToBinds uni acc (EvaluationResult a :: Type)

Instance details

Defined in PlutusCore.Builtin.KnownTypeAst

type ToBinds uni acc (EvaluationResult a :: Type) = ToBinds uni acc a

type ToHoles uni _1 (EvaluationResult a :: Type)

Instance details

Defined in PlutusCore.Builtin.KnownTypeAst

type ToHoles uni _1 (EvaluationResult a :: Type) = '[TypeHole a :: Hole]

type IsBuiltin uni (EvaluationResult a :: Type)

Instance details

Defined in PlutusCore.Builtin.KnownTypeAst

type IsBuiltin uni (EvaluationResult a :: Type) = 'False

type Rep (EvaluationResult a)

Instance details

Defined in PlutusCore.Evaluation.Result

type Rep (EvaluationResult a) = D1 ('MetaData "EvaluationResult" "PlutusCore.Evaluation.Result" "plutus-core-1.60.0.0-inplace" 'False) (C1 ('MetaCons "EvaluationSuccess" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'SourceStrict 'DecidedStrict) (Rec0 a)) :+: C1 ('MetaCons "EvaluationFailure" 'PrefixI 'False) (U1 :: Type -> Type))

Combining programs

applyProgram :: (MonadError ApplyProgramError m, Semigroup a) => Program tyname name uni fun a -> Program tyname name uni fun a -> m (Program tyname name uni fun a)

Applies one program to another. Fails if the versions do not match and tries to merge annotations.

Benchmarking

termAstSize :: Term tyname name uni fun ann -> AstSize

Count the number of AST nodes in a term.

typeAstSize :: Type tyname uni ann -> AstSize

Count the number of AST nodes in a type.

kindAstSize :: Kind a -> AstSize

Count the number of AST nodes in a kind.

>>> kindAstSize $ Type ()
AstSize {unAstSize = 1}
>>> kindAstSize $ KindArrow () (KindArrow () (Type ()) (Type ())) (Type ())
AstSize {unAstSize = 5} 

programAstSize :: Program tyname name uni fun ann -> AstSize

Count the number of AST nodes in a program.