PlutusCore.Annotation

Documentation

data Ann

An annotation type used during the compilation.

Constructors

Ann

Fields

annInline :: Inline

annSrcSpans :: SrcSpans

annIsAsDataMatcher :: Bool

data SrcSpan

The span between two source locations.

This corresponds roughly to the SrcSpan used by GHC, but we define our own version so we don't have to depend on ghc to use it.

The line and column numbers are 1-based, and the unit is Unicode code point (or Char).

Constructors

SrcSpan

Fields

srcSpanFile :: FilePath

srcSpanSLine :: Int

srcSpanSCol :: Int

srcSpanELine :: Int

srcSpanECol :: Int

Same as GHC's SrcSpan, srcSpanECol is usually one more than the column of the last character of the thing this SrcSpan is for (unless the last character is the line break).

newtype InlineHints name a

Constructors

InlineHints

Fields

shouldInline :: a -> name -> Inline

data Inline

Constructors

AlwaysInline
When calling PlutusIR.Compiler.Definitions.defineTerm to add a new term definition, if we annotation the var on the LHS of the definition with AlwaysInline, the inliner will always inline that var.

This is currently used to ensure builtin functions such as trace (when the remove-trace flag is on and trace is rewritten to const) are inlined, because the inliner would otherwise not inline them. To achieve that, we annotate the definition with AlwaysInline when defining trace, i.e., trace AlwaysInline = _ a -> a.

SafeToInline
Signaling to the compiler that a binding is safe to inline. This is useful for annotating strict bindings that aren't obviously safe to inline.

MayInline

class AnnInline a where

Methods

annAlwaysInline :: a

An annotation instructing the inliner to always inline a binding.

annSafeToInline :: a

An annotation signaling to the inliner that a binding is safe to inline. The inlining decision is left to the inliner. This is useful for annotating strict bindings that aren't obviously safe to inline.

annMayInline :: a

An annotation that leaves the inlining decision to the inliner.

data SourcePos

The data type SourcePos represents source positions. It contains the name of the source file, a line number, and a column number. Source line and column positions change intensively during parsing, so we need to make them strict to avoid memory leaks.

Constructors

SourcePos

Fields

sourceName :: Text.Megaparsec.Pos
type Rep SourcePos = D1 ('MetaData "SourcePos" "Text.Megaparsec.Pos" "megaparsec-9.7.0-9IxnAL1Z5n8HX9bf04jh9I" 'False) (C1 ('MetaCons "SourcePos" 'PrefixI 'True) (S1 ('MetaSel ('Just "sourceName") 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 FilePath) :*: (S1 ('MetaSel ('Just "sourceLine") 'NoSourceUnpackedness 'SourceStrict 'DecidedUnpack) (Rec0 Pos) :*: S1 ('MetaSel ('Just "sourceColumn") 'NoSourceUnpackedness 'SourceStrict 'DecidedUnpack) (Rec0 Pos))))

data Pos

Pos is the type for positive integers. This is used to represent line number, column number, and similar things like indentation level. Semigroup instance can be used to safely and efficiently add Poses together.

Since: megaparsec-5.0.0

addSrcSpan :: SrcSpan -> Ann -> Ann

Add an extra SrcSpan to existing SrcSpans of Ann

lineInSrcSpan :: Pos -> SrcSpan -> Bool

Tells if a line (positive integer) falls inside a SrcSpan.