PlutusCore.Builtin

Description

Reexports from modules from the Builtin folder.

Documentation

class Typeable (a :: k)

The class Typeable allows a concrete representation of a type to be calculated.

Minimal complete definition

typeRep#

data TypeScheme val (args :: [Type]) res where

Type schemes of primitive operations. as is a list of types of arguments, r is the resulting type. E.g. Text -> Bool -> Integer is encoded as TypeScheme val [Text, Bool] Integer.

Constructors

TypeSchemeResult :: (Typeable res, KnownTypeAst (UniOf val) res, MakeKnown val res) => TypeScheme val '[] res

TypeSchemeArrow :: (Typeable arg, KnownTypeAst (UniOf val) arg, MakeKnown val arg, ReadKnown val arg) => TypeScheme val args res -> TypeScheme val (arg ': args) res infixr 9

TypeSchemeAll :: (KnownSymbol text, KnownNat uniq, KnownKind kind) => Proxy '(text, uniq, kind) -> TypeScheme val args res -> TypeScheme val args res

argProxy :: TypeScheme val (arg ': args) res -> Proxy arg

typeSchemeToType :: TypeScheme val args res -> Type TyName (UniOf val) ()

Convert a TypeScheme to the corresponding Kind. Basically, a map from the PHOAS representation to the FOAS one.

class (uni `Everywhere` ImplementedKnownTypeAst uni, uni `Everywhere` ImplementedReadKnownIn uni, uni `Everywhere` ImplementedMakeKnownIn uni) => TestTypesFromTheUniverseAreAllKnown uni

An instance of this class not having any constraints ensures that every type (according to Everywhere) from the universe has 'KnownTypeAst, ReadKnownIn and MakeKnownIn instances.

data BuiltinRuntime val

A BuiltinRuntime represents a possibly partial builtin application. We get an initial BuiltinRuntime representing an empty builtin application (i.e. just the builtin with no arguments) by instantiating (via fromBuiltinRuntimeOptions) a BuiltinRuntimeOptions.

Applying or type-instantiating a builtin peels off the corresponding constructor from its BuiltinRuntime.

BuiltinResult contains the cost (an ExBudget) and the result (a MakeKnownM val) of the builtin application. The cost is stored strictly, since the evaluator is going to look at it and the result is stored lazily, since it's not supposed to be forced before accounting for the cost of the application. If the cost exceeds the available budget, the evaluator discards the the result of the builtin application without ever forcing it and terminates with evaluation failure. Allowing the user to compute something that they don't have the budget for would be a major bug.

Evaluators that ignore the entire concept of costing (e.g. the CK machine) may of course force the result of the builtin application unconditionally.

Constructors

BuiltinResult ExBudget ~(MakeKnownM val)

BuiltinExpectArgument (val -> ReadKnownM (BuiltinRuntime val))

BuiltinExpectForce (BuiltinRuntime val)

data UnliftingMode

Determines how to unlift arguments. The difference is that with UnliftingImmediate unlifting is performed immediately after a builtin gets the argument and so can fail immediately too, while with deferred unlifting all arguments are unlifted upon full saturation, hence no failure can occur until that. The former makes it much harder to specify the behaviour of builtins and so UnliftingDeferred is the preferred mode.

Constructors

UnliftingImmediate

UnliftingDeferred

data BuiltinRuntimeOptions val cost

A BuiltinRuntimeOptions is a precursor to BuiltinRuntime. One gets the latter from the former by applying a function returning the runtime denotation of the builtin (either _broImmediateF for immediate unlifting or _broDeferredF for deferred unlifting, see UnliftingMode for details) to a cost model.

Constructors

BuiltinRuntimeOptions

Fields

_broImmediateF :: cost -> BuiltinRuntime val

_broDeferredF :: cost -> BuiltinRuntime val

fromBuiltinRuntimeOptions :: UnliftingMode -> cost -> BuiltinRuntimeOptions val cost -> BuiltinRuntime val

Convert a BuiltinRuntimeOptions to a BuiltinRuntime given an UnliftingMode and a cost model.

newtype BuiltinsRuntime fun val

A BuiltinRuntime for each builtin from a set of builtins. Just an Array of cached BuiltinRuntimes. We could instantiate BuiltinMeaning on the fly and avoid caching BuiltinRuntime in an array, but we've tried it and it was much slower as we do rely on caching (especially for costing).

Constructors

BuiltinsRuntime

Fields

unBuiltinRuntime :: Array fun (BuiltinRuntime val)

lookupBuiltin :: (MonadError (ErrorWithCause err cause) m, AsMachineError err fun, Ix fun) => fun -> BuiltinsRuntime fun val -> m (BuiltinRuntime val)

Look up the runtime info of a built-in function during evaluation.

newtype Opaque val (rep :: Type)

The AST of a value with a Plutus type attached to it. The type is for the Plutus type checker to look at. Opaque can appear in the type of the denotation of a builtin.

Constructors

Opaque

Fields

unOpaque :: val

newtype SomeConstant uni (rep :: Type)

For unlifting from the Constant constructor when the stored value is of a monomorphic built-in type

The rep parameter specifies how the type looks on the PLC side (i.e. just like with Opaque val rep).

Constructors

SomeConstant

Fields

unSomeConstant :: Some (ValueOf uni)

data TyNameRep (kind :: Type)

Representation of a type variable: its name and unique and an implicit kind.

Constructors

TyNameRep Symbol Nat

data family TyVarRep (name :: TyNameRep kind) :: kind

Representation of an intrinsically-kinded type variable: a name.

data family TyAppRep (fun :: dom -> cod) (arg :: dom) :: cod

Representation of an intrinsically-kinded type application: a function and an argument.

data family TyForallRep (name :: TyNameRep kind) (a :: Type) :: Type

Representation of of an intrinsically-kinded universal quantifier: a bound name and a body.

type family FoldArgs args res where ...

Turn a list of Haskell types args into a functional type ending in res.

>>> :set -XDataKinds
>>> :kind! FoldArgs [(), Bool] Integer
FoldArgs [(), Bool] Integer :: *
= () -> Bool -> Integer

Equations

FoldArgs '[] res = res

FoldArgs (arg ': args) res = arg -> FoldArgs args res

data BuiltinMeaning val cost

The meaning of a built-in function consists of its type represented as a TypeScheme, its Haskell denotation and a BuiltinRuntimeOptions.

The TypeScheme of a built-in function is used for example for

computing the PLC type of the function to be used during type checking

getting arity information

generating arbitrary values to apply the function to in tests

The denotation is lazy, so that we don't need to worry about a builtin being bottom (happens in tests). The production path is not affected by that, since only runtime denotations are used for evaluation.

Constructors

forall args res. BuiltinMeaning (TypeScheme val args res) ~(FoldArgs args res) (BuiltinRuntimeOptions val cost)

type HasMeaningIn uni val = (Typeable val, ExMemoryUsage val, HasConstantIn uni val)

Constraints available when defining a built-in function.

class (Typeable uni, Typeable fun, Bounded fun, Enum fun, Ix fun, Default (BuiltinVersion fun)) => ToBuiltinMeaning uni fun where

A type class for "each function from a set of built-in functions has a BuiltinMeaning".

Associated Types

type CostingPart uni fun

The cost part of BuiltinMeaning.

data BuiltinVersion fun

See Note [Versioned builtins]

Methods

toBuiltinMeaning :: HasMeaningIn uni val => BuiltinVersion fun -> fun -> BuiltinMeaning val (CostingPart uni fun)

Get the BuiltinMeaning of a built-in function.

typeOfBuiltinFunction :: forall uni fun. ToBuiltinMeaning uni fun => BuiltinVersion fun -> fun -> Type TyName uni ()

Get the type of a built-in function.

type family GetArgs a where ...

Chop a function type to get a list of its argument types.

Equations

GetArgs (a -> b) = a ': GetArgs b

GetArgs _ = '[]

class KnownMonotype val args res where

A class that allows us to derive a monotype for a builtin. We could've computed the runtime denotations (toImmediateF and toDeferredF) from the TypeScheme and the denotation of the builtin, but not statically (due to unfolding not working for recursive functions and TypeScheme being recursive, i.e. requiring the conversion function to be recursive), and so it would cause us to retain a lot of evaluation-irrelevant stuff in the constructors of BuiltinRuntime, which has to make evaluation slower (we didn't check) and certainly makes the generated Core much harder to read. Technically speaking, we could get a RuntimeScheme from the TypeScheme and the denotation statically if we changed the definition of TypeScheme and made it a singleton, but then the conversion function would have to become a class anyway and we'd just replicate what we have here, except in a much more complicated way.

Methods

knownMonotype :: TypeScheme val args res

toMonoImmediateF :: (FoldArgs args res, FoldArgs args ExBudget) -> BuiltinRuntime val

Convert the denotation of a builtin to its runtime counterpart with immediate unlifting. We use a tuple rather than two arguments for symmetry with toPolyDeferredF. It all gets inlined anyway.

toMonoDeferredF :: ReadKnownM (FoldArgs args res, FoldArgs args ExBudget) -> BuiltinRuntime val

Convert the denotation of a builtin to its runtime counterpart with deferred unlifting. The argument is in ReadKnownM, because that's what deferred unlifting amounts to: passing the action returning the builtin application around until full saturation, which is when the action actually gets run.

class KnownMonotype val args res => KnownPolytype (binds :: [Some TyNameRep]) val args res where

A class that allows us to derive a polytype for a builtin.

Methods

knownPolytype :: TypeScheme val args res

toPolyImmediateF :: (FoldArgs args res, FoldArgs args ExBudget) -> BuiltinRuntime val

Convert the denotation of a builtin to its runtime counterpart with immediate unlifting. We use a tuple rather than two arguments for symmetry with toPolyDeferredF. It all gets inlined anyway.

toPolyDeferredF :: ReadKnownM (FoldArgs args res, FoldArgs args ExBudget) -> BuiltinRuntime val

Convert the denotation of a builtin to its runtime counterpart with deferred unlifting. The argument is in ReadKnownM, because that's what deferred unlifting amounts to: passing the action returning the builtin application around until full saturation, which is when the action actually gets run.

type family ThrowOnBothEmpty binds args isBuiltin a where ...

Ensure a built-in function is not nullary and throw a nice error otherwise.

Equations

ThrowOnBothEmpty '[] '[] 'True a = TypeError (('Text "A built-in function must take at least one type or term argument" :$$: ((('Text "\8216" :<>: 'ShowType a) :<>: 'Text "\8217 is a built-in type") :<>: 'Text " so you can embed any of its values as a constant")) :$$: 'Text "If you still want a built-in function, add a dummy \8216()\8217 argument")

ThrowOnBothEmpty '[] '[] 'False a = TypeError ('Text "A built-in function must take at least one type or term argument" :$$: 'Text "To fix this error add a dummy \8216()\8217 argument")

ThrowOnBothEmpty _ _ _ _ = ()

class MakeBuiltinMeaning a val where

A function turned into a type class with exactly one fully general instance. We can't package up the constraints of makeBuiltinMeaning (see the instance) into a type or class synonym, because they contain a bunch of variables defined by ~ or determined via functional dependencies and neither class nor type definitions can handle that (see https://gitlab.haskell.org/ghc/ghc/-/issues/7100). Inlining three lines of constraints whenever we need to call makeBuiltinMeaning over a non-concrete type is a bad option and this abstraction is free anyway, hence its existence.

The a type variable goes first, because makeBuiltinMeaning @A is a common pattern.

Methods

makeBuiltinMeaning :: a -> (cost -> FoldArgs (GetArgs a) ExBudget) -> BuiltinMeaning val cost

Construct the meaning for a built-in function by automatically deriving its TypeScheme, given

the denotation of the builtin

an uninstantiated costing function

toBuiltinRuntime :: UnliftingMode -> cost -> BuiltinMeaning val cost -> BuiltinRuntime val

Convert a BuiltinMeaning to a BuiltinRuntime given an UnliftingMode and a cost model.

toBuiltinsRuntime :: (cost ~ CostingPart uni fun, ToBuiltinMeaning uni fun, HasMeaningIn uni val) => BuiltinVersion fun -> UnliftingMode -> cost -> BuiltinsRuntime fun val

Calculate runtime info for all built-in functions given denotations of builtins, an UnliftingMode and a cost model.

data TyNameRep (kind :: Type)

Representation of a type variable: its name and unique and an implicit kind.

Constructors

TyNameRep Symbol Nat

data family TyVarRep (name :: TyNameRep kind) :: kind

Representation of an intrinsically-kinded type variable: a name.

data family TyAppRep (fun :: dom -> cod) (arg :: dom) :: cod

Representation of an intrinsically-kinded type application: a function and an argument.

data family TyForallRep (name :: TyNameRep kind) (a :: Type) :: Type

Representation of of an intrinsically-kinded universal quantifier: a bound name and a body.

data Hole

The kind of holes.

data family RepHole x

A hole in the Rep context.

data family TypeHole a

A hole in the Type context.

type KnownBuiltinTypeAst uni a = KnownTypeAst uni (ElaborateBuiltin a)

A constraint for "a is a KnownTypeAst by means of being included in uni".

class KnownTypeAst uni x where

Minimal complete definition

Nothing

Associated Types

type IsBuiltin x :: Bool

Whether x is a built-in type.

type IsBuiltin x = IsBuiltin (ElaborateBuiltin x)

type ToHoles x :: [Hole]

Return every part of the type that can be a to-be-instantiated type variable. For example, in Integer there's no such types and in (a, b) it's the two arguments (a and b) and the same applies to a -> b (to mention a type that is not built-in).

type ToHoles x = ToHoles (ElaborateBuiltin x)

type ToBinds x :: [Some TyNameRep]

Collect all unique variables (a variable consists of a textual name, a unique and a kind) in an a.

type ToBinds x = ToBinds (ElaborateBuiltin x)

Methods

toTypeAst :: proxy x -> Type TyName uni ()

The type representing a used on the PLC side.

default toTypeAst :: KnownBuiltinTypeAst uni x => proxy x -> Type TyName uni ()

type family Merge xs ys :: [a] where ...

Delete all elements appearing in the first list from the second one and concatenate the lists.

Equations

Merge '[] ys = ys

Merge (x ': xs) ys = x ': Delete x (Merge xs ys)

data KnownTypeError

The type of errors that readKnown and makeKnown can return.

throwKnownTypeErrorWithCause :: (MonadError (ErrorWithCause err cause) m, AsUnliftingError err, AsEvaluationFailure err) => cause -> KnownTypeError -> m void

Attach a cause to a KnownTypeError and throw that. Note that an evaluator might require the cause to be computed lazily for best performance on the happy path, hence this function must not force its first argument. TODO: wrap cause in Lazy once we have it.

type KnownBuiltinTypeIn uni val a = (HasConstantIn uni val, Pretty (SomeTypeIn uni), GEq uni, uni `Contains` a)

A constraint for "a is a ReadKnownIn and MakeKnownIn by means of being included in uni".

type KnownBuiltinType val a = KnownBuiltinTypeIn (UniOf val) val a

A constraint for "a is a ReadKnownIn and MakeKnownIn by means of being included in UniOf term".

data MakeKnownM a

The monad that makeKnown runs in. Equivalent to ExceptT KnownTypeError Emitter, except optimized in two ways:

everything is strict

has the MakeKnownSuccess constructor that is used for returning a value with no logs attached, which is the most common case for us, so it helps a lot not to construct and deconstruct a redundant tuple

Moving from ExceptT KnownTypeError Emitter to this data type gave us a speedup of 8% of total evaluation time.

Logs are represented as a DList, because we don't particularly care about the efficiency of logging, since there's no logging on the chain and builtins don't emit much anyway. Otherwise we'd have to use text-builder or text-builder-linear or something of this sort.

Constructors

MakeKnownFailure (DList Text) KnownTypeError

MakeKnownSuccess a

MakeKnownSuccessWithLogs (DList Text) a

type ReadKnownM = Either KnownTypeError

The monad that readKnown runs in.

liftReadKnownM :: ReadKnownM a -> MakeKnownM a

Lift a ReadKnownM computation into MakeKnownM.

readKnownConstant :: forall val a. KnownBuiltinType val a => val -> ReadKnownM a

Convert a constant embedded into a PLC term to the corresponding Haskell value.

class uni ~ UniOf val => MakeKnownIn uni val a where

Minimal complete definition

Nothing

Methods

makeKnown :: a -> MakeKnownM val

Convert a Haskell value to the corresponding PLC val. The inverse of readKnown.

default makeKnown :: KnownBuiltinType val a => a -> MakeKnownM val

type MakeKnown val = MakeKnownIn (UniOf val) val

class uni ~ UniOf val => ReadKnownIn uni val a where

Minimal complete definition

Nothing

Methods

readKnown :: val -> ReadKnownM a

Convert a PLC val to the corresponding Haskell value. The inverse of makeKnown.

default readKnown :: KnownBuiltinType val a => val -> ReadKnownM a

type ReadKnown val = ReadKnownIn (UniOf val) val

makeKnownOrFail :: MakeKnownIn uni val a => a -> EvaluationResult val

Same as makeKnown, but allows for neither emitting nor storing the cause of a failure.

readKnownSelf :: (ReadKnown val a, AsUnliftingError err, AsEvaluationFailure err) => val -> Either (ErrorWithCause err val) a

Same as readKnown, but the cause of a potential failure is the provided term itself.

data SingKind k where

The type of singletonized Haskell kinds representing Plutus kinds. Indexing by a Haskell kind allows us to avoid an error call in the ToKind instance of DefaultUni and not worry about proxies in the type of knownKind.

Constructors

SingType :: SingKind Type

SingKindArrow :: SingKind k -> SingKind l -> SingKind (k -> l) infixr 5

class KnownKind k where

For reifying Haskell kinds representing Plutus kinds at the term level.

Methods

knownKind :: SingKind k

withKnownKind :: Kind ann -> (forall k. KnownKind k => Proxy k -> r) -> r

class ToKind (uni :: Type -> Type) where

For computing the Plutus kind of a built-in type. See kindOfBuiltinType.

Methods

toSingKind :: uni (Esc (a :: k)) -> SingKind k

Reify the kind of a type from the universe at the term level.

demoteKind :: SingKind k -> Kind ()

Convert a reified Haskell kind to a Plutus kind.

kindOfBuiltinType :: ToKind uni => uni (Esc a) -> Kind ()

Compute the kind of a type from a universe.

data KnownTypeError

The type of errors that readKnown and makeKnown can return.

Constructors

KnownTypeUnliftingError UnliftingError

KnownTypeEvaluationFailure

throwNotAConstant :: MonadError KnownTypeError m => m void

class HasConstant term where

Ensures that term has a Constant-like constructor to lift values to and unlift values from.

Methods

asConstant :: term -> Either KnownTypeError (Some (ValueOf (UniOf term)))

Unwrap from a Constant-like constructor throwing an UnliftingError if the provided term is not a wrapped Haskell value.

fromConstant :: Some (ValueOf (UniOf term)) -> term

Wrap a Haskell value as a term.

type HasConstantIn uni term = (UniOf term ~ uni, HasConstant term)

Ensures that term has a Constant-like constructor to lift values to and unlift values from and connects term and its uni.

fromValueOf :: HasConstant term => UniOf term (Esc a) -> a -> term

Wrap a Haskell value (given its explicit type tag) as a term.

fromValue :: (HasConstant term, UniOf term `Includes` a) => a -> term

Wrap a Haskell value (provided its type is in the universe) as a term.

module PlutusCore.Builtin.Emitter