PlutusCore.Builtin

Description

Reexports from modules from the Builtin folder.

Documentation

data TypeScheme val (args :: [Type]) res where

Type schemes of primitive operations. as is a list of types of arguments, r is the resulting type. E.g. Text -> Bool -> Integer is encoded as TypeScheme val [Text, Bool] Integer.

Constructors

TypeSchemeResult :: (KnownTypeAst (UniOf val) res, MakeKnown val res) => TypeScheme val '[] res

TypeSchemeArrow :: (KnownTypeAst (UniOf val) arg, MakeKnown val arg, ReadKnown val arg) => TypeScheme val args res -> TypeScheme val (arg ': args) res infixr 9

TypeSchemeAll :: (KnownSymbol text, KnownNat uniq, KnownKind kind) => Proxy '(text, uniq, kind) -> TypeScheme val args res -> TypeScheme val args res

argProxy :: TypeScheme val (arg ': args) res -> Proxy arg

typeSchemeToType :: TypeScheme val args res -> Type TyName (UniOf val) ()

Convert a TypeScheme to the corresponding Type. Basically, a map from the PHOAS representation to the FOAS one.

class (uni `Everywhere` ImplementedKnownTypeAst uni, uni `Everywhere` ImplementedReadKnownIn uni, uni `Everywhere` ImplementedMakeKnownIn uni) => TestTypesFromTheUniverseAreAllKnown uni

An instance of this class not having any constraints ensures that every type (according to Everywhere) from the universe has 'KnownTypeAst, ReadKnownIn and MakeKnownIn instances.

data Peano

Peano numbers. Normally called Nat, but that is already reserved by base.

Constructors

Z

S Peano

data RuntimeScheme n where

Same as TypeScheme except this one doesn't contain any evaluation-irrelevant types stuff. n represents the number of term-level arguments that a builtin takes.

Constructors

RuntimeSchemeResult :: RuntimeScheme 'Z

RuntimeSchemeArrow :: RuntimeScheme n -> RuntimeScheme ('S n)

RuntimeSchemeAll :: RuntimeScheme n -> RuntimeScheme n

type family ToRuntimeDenotationType val n where ...

Compute the runtime denotation type of a builtin given the type of values and the number of arguments that the builtin takes. A "runtime denotation type" is different from a regular denotation type in that a regular one can have any ReadKnownIn type as an argument and can return any MakeKnownIn type, while in a runtime one only val can appear as the type of an argument, as well as in the result type. This is what we get by calling readKnown over each argument of the denotation and calling makeKnown over its result.

Equations

ToRuntimeDenotationType val 'Z = MakeKnownM () val

ToRuntimeDenotationType val ('S n) = val -> ReadKnownM () (ToRuntimeDenotationType val n)

type family ToCostingType n where ...

Compute the costing type for a builtin given the number of arguments that the builtin takes.

Equations

ToCostingType 'Z = ExBudget

ToCostingType ('S n) = ExMemory -> ToCostingType n

data BuiltinRuntime val

A BuiltinRuntime represents a possibly partial builtin application. We get an initial BuiltinRuntime representing an empty builtin application (i.e. just the builtin with no arguments) by instantiating (via fromBuiltinRuntimeOptions) a BuiltinRuntimeOptions.

A BuiltinRuntime contains info that is used during evaluation:

the RuntimeScheme of the uninstantiated part of the builtin. I.e. initially it's the runtime scheme of the whole builtin, but applying or type-instantiating the builtin peels off the corresponding constructor from the runtime scheme

the (possibly partially instantiated) runtime denotation

the (possibly partially instantiated) costing function

All the three are in sync in terms of partial instantiatedness due to RuntimeScheme being a GADT and ToRuntimeDenotationType and ToCostingType operating on the index of that GADT.

Constructors

forall n. BuiltinRuntime (RuntimeScheme n) ~(ToRuntimeDenotationType val n) (ToCostingType n)

noThunksInCosting :: Context -> ToCostingType n -> RuntimeScheme n -> IO (Maybe ThunkInfo)

Check whether the given `ToCostingType n` contains thunks. The `RuntimeScheme n` is used to refine n.

data UnliftingMode

Determines how to unlift arguments. The difference is that with UnliftingImmediate unlifting is performed immediately after a builtin gets the argument and so can fail immediately too, while with deferred unlifting all arguments are unlifted upon full saturation, hence no failure can occur until that. The former makes it much harder to specify the behaviour of builtins and so UnliftingDeferred is the preferred mode.

Constructors

UnliftingImmediate

UnliftingDeferred

data BuiltinRuntimeOptions n val cost

A BuiltinRuntimeOptions is a precursor to BuiltinRuntime. One gets the latter from the former by choosing the runtime denotation of the builtin (either _broImmediateF for immediate unlifting or _broDeferredF for deferred unlifting, see UnliftingMode for details) and by instantiating _broToExF with a cost model to get the costing function for the builtin.

Constructors

BuiltinRuntimeOptions

Fields

_broRuntimeScheme :: RuntimeScheme n

_broImmediateF :: ToRuntimeDenotationType val n

_broDeferredF :: ToRuntimeDenotationType val n

_broToExF :: cost -> ToCostingType n

fromBuiltinRuntimeOptions :: UnliftingMode -> cost -> BuiltinRuntimeOptions n val cost -> BuiltinRuntime val

Convert a BuiltinRuntimeOptions to a BuiltinRuntime given an UnliftingMode and a cost model.

newtype BuiltinsRuntime fun val

A BuiltinRuntime for each builtin from a set of builtins.

Constructors

BuiltinsRuntime

Fields

unBuiltinRuntime :: Array fun (BuiltinRuntime val)

lookupBuiltin :: (MonadError (ErrorWithCause err cause) m, AsMachineError err fun, Ix fun) => fun -> BuiltinsRuntime fun val -> m (BuiltinRuntime val)

Look up the runtime info of a built-in function during evaluation.

newtype Opaque val (rep :: Type)

The denotation of a term whose PLC type is encoded in rep (for example a type variable or an application of a type variable). I.e. the denotation of such a term is the term itself. This is because we have parametricity in Haskell, so we can't inspect a value whose type is, say, a bound variable, so we never need to convert such a term from Plutus Core to Haskell and back and instead can keep it intact.

Constructors

Opaque

Fields

unOpaque :: val

newtype SomeConstant uni (rep :: Type)

For unlifting from the Constant constructor when the stored value is of a monomorphic built-in type

The rep parameter specifies how the type looks on the PLC side (i.e. just like with Opaque val rep).

Constructors

SomeConstant

Fields

unSomeConstant :: Some (ValueOf uni)

data TyNameRep (kind :: Type)

Representation of a type variable: its name and unique and an implicit kind.

Constructors

TyNameRep Symbol Nat

data family TyVarRep (name :: TyNameRep kind) :: kind

Representation of an intrinsically-kinded type variable: a name.

data family TyAppRep (fun :: dom -> cod) (arg :: dom) :: cod

Representation of an intrinsically-kinded type application: a function and an argument.

data family TyForallRep (name :: TyNameRep kind) (a :: Type) :: Type

Representation of of an intrinsically-kinded universal quantifier: a bound name and a body.

type family FoldArgs args res where ...

Turn a list of Haskell types args into a functional type ending in res.

>>> :set -XDataKinds
>>> :kind! FoldArgs [Text, Bool] Integer
FoldArgs [Text, Bool] Integer :: *
= Text -> Bool -> Integer

Equations

FoldArgs '[] res = res

FoldArgs (arg ': args) res = arg -> FoldArgs args res

data BuiltinMeaning val cost

The meaning of a built-in function consists of its type represented as a TypeScheme, its Haskell denotation and a costing function (both in uninstantiated form).

The TypeScheme of a built-in function is used for

computing the PLC type of the function to be used during type checking

checking equality of the expected type of an argument of a builtin and the actual one during evaluation, so that we can avoid unsafe-coercing

A costing function for a built-in function is computed from a cost model for all built-in functions from a certain set, hence the cost parameter.

Constructors

forall args res. BuiltinMeaning (TypeScheme val args res) (FoldArgs args res) (BuiltinRuntimeOptions (Length args) val cost)

class (Bounded fun, Enum fun, Ix fun) => ToBuiltinMeaning uni fun where

A type class for "each function from a set of built-in functions has a BuiltinMeaning".

Associated Types

type CostingPart uni fun

The cost part of BuiltinMeaning.

Methods

toBuiltinMeaning :: HasConstantIn uni val => fun -> BuiltinMeaning val (CostingPart uni fun)

Get the BuiltinMeaning of a built-in function.

typeOfBuiltinFunction :: ToBuiltinMeaning uni fun => fun -> Type TyName uni ()

Get the type of a built-in function.

type family Length xs where ...

Compute the length of a type-level list.

Equations

Length '[] = 'Z

Length (_ ': xs) = 'S (Length xs)

type family GetArgs a where ...

Chop a function type to get a list of its argument types.

Equations

GetArgs (a -> b) = a ': GetArgs b

GetArgs _ = '[]

class KnownMonotype val args res a | args res -> a, a -> res where

A class that allows us to derive a monotype for a builtin. We could've easily computed a RuntimeScheme from a TypeScheme but not statically (due to unfolding not working for recursive functions and TypeScheme is recursive, i.e. the conversion can only be a recursive function), and so it would cause us to retain a lot of evaluation-irrelevant stuff in the constructors of TypeScheme, which makes it much harder to read the Core. Technically speaking, we could convert a TypeScheme to a RuntimeScheme statically if we changed the definition of TypeScheme and made it a singleton, but then the conversion function would have to become a class anyway and we'd just replicate what we have here, except in a much more complicated way. It's also more efficient to generate RuntimeSchemes directly, but that shouldn't really matter, -- given that they computed only once and cached afterwards.

Similarly, we could've computed toImmediateF and toDeferredF from a TypeScheme but not statically again, and that would break inlining and basically all the optimization.

Methods

knownMonotype :: TypeScheme val args res

knownMonoruntime :: RuntimeScheme (Length args)

toImmediateF :: FoldArgs args res -> ToRuntimeDenotationType val (Length args)

Convert the denotation of a builtin to its runtime counterpart with immediate unlifting.

toDeferredF :: ReadKnownM () (FoldArgs args res) -> ToRuntimeDenotationType val (Length args)

Convert the denotation of a builtin to its runtime counterpart with deferred unlifting. The argument is in ReadKnownM, because that's what deferred unlifting amounts to: passing the action returning the builtin application around until full saturation, which is when the action actually gets run.

class KnownMonotype val args res a => KnownPolytype (binds :: [Some TyNameRep]) val args res a | args res -> a, a -> res where

A class that allows us to derive a polytype for a builtin.

Methods

knownPolytype :: TypeScheme val args res

knownPolyruntime :: RuntimeScheme (Length args)

makeBuiltinMeaning :: forall a val cost binds args res j. (binds ~ ToBinds a, args ~ GetArgs a, a ~ FoldArgs args res, ElaborateFromTo 0 j val a, KnownPolytype binds val args res a) => a -> (cost -> ToCostingType (Length args)) -> BuiltinMeaning val cost

Construct the meaning for a built-in function by automatically deriving its TypeScheme, given

the denotation of the builtin

an uninstantiated costing function

toBuiltinRuntime :: UnliftingMode -> cost -> BuiltinMeaning val cost -> BuiltinRuntime val

Convert a BuiltinMeaning to a BuiltinRuntime given an UnliftingMode and a cost model.

toBuiltinsRuntime :: (cost ~ CostingPart uni fun, HasConstantIn uni val, ToBuiltinMeaning uni fun) => UnliftingMode -> cost -> BuiltinsRuntime fun val

Calculate runtime info for all built-in functions given denotations of builtins and a cost model.

data TyNameRep (kind :: Type)

Representation of a type variable: its name and unique and an implicit kind.

Constructors

TyNameRep Symbol Nat

data family TyVarRep (name :: TyNameRep kind) :: kind

Representation of an intrinsically-kinded type variable: a name.

data family TyAppRep (fun :: dom -> cod) (arg :: dom) :: cod

Representation of an intrinsically-kinded type application: a function and an argument.

data family TyForallRep (name :: TyNameRep kind) (a :: Type) :: Type

Representation of of an intrinsically-kinded universal quantifier: a bound name and a body.

data Hole

The kind of holes.

data family RepHole x

A hole in the Rep context.

data family TypeHole a

A hole in the Type context.

type KnownBuiltinTypeAst uni a = KnownTypeAst uni (ElaborateBuiltin a)

A constraint for "a is a KnownTypeAst by means of being included in uni".

class KnownTypeAst uni x where

Minimal complete definition

Nothing

Associated Types

type ToHoles x :: [Hole]

Return every part of the type that can be a to-be-instantiated type variable. For example, in Integer there's no such types and in (a, b) it's the two arguments (a and b) and the same applies to a -> b (to mention a type that is not built-in).

type ToHoles x = ToHoles (ElaborateBuiltin x)

type ToBinds x :: [Some TyNameRep]

Collect all unique variables (a variable consists of a textual name, a unique and a kind) in an a.

type ToBinds x = ToBinds (ElaborateBuiltin x)

Methods

toTypeAst :: proxy x -> Type TyName uni ()

The type representing a used on the PLC side.

default toTypeAst :: KnownBuiltinTypeAst uni x => proxy x -> Type TyName uni ()

type family Merge xs ys :: [a] where ...

Delete all elements appearing in the first list from the second one and concatenate the lists.

Equations

Merge '[] ys = ys

Merge (x ': xs) ys = x ': Delete x (Merge xs ys)

data KnownTypeError

The type of errors that readKnown and makeKnown can return.

throwKnownTypeErrorWithCause :: (MonadError (ErrorWithCause err cause) m, AsUnliftingError err, AsEvaluationFailure err) => ErrorWithCause KnownTypeError cause -> m void

Throw a ErrorWithCause KnownTypeError cause.

type KnownBuiltinTypeIn uni val a = (HasConstantIn uni val, GShow uni, GEq uni, uni `Contains` a)

A constraint for "a is a ReadKnownIn and MakeKnownIn by means of being included in uni".

type KnownBuiltinType val a = KnownBuiltinTypeIn (UniOf val) val a

A constraint for "a is a ReadKnownIn and MakeKnownIn by means of being included in UniOf term".

type MakeKnownM cause = ExceptT (ErrorWithCause KnownTypeError cause) Emitter

The monad that makeKnown runs in.

type ReadKnownM cause = Either (ErrorWithCause KnownTypeError cause)

The monad that readKnown runs in.

readKnownConstant :: forall val a cause. KnownBuiltinType val a => Maybe cause -> val -> ReadKnownM cause a

Convert a constant embedded into a PLC term to the corresponding Haskell value.

class uni ~ UniOf val => MakeKnownIn uni val a where

Minimal complete definition

Nothing

Methods

makeKnown :: Maybe cause -> a -> MakeKnownM cause val

Convert a Haskell value to the corresponding PLC val. The inverse of readKnown.

default makeKnown :: KnownBuiltinType val a => Maybe cause -> a -> MakeKnownM cause val

type MakeKnown val = MakeKnownIn (UniOf val) val

class uni ~ UniOf val => ReadKnownIn uni val a where

Minimal complete definition

Nothing

Methods

readKnown :: Maybe cause -> val -> ReadKnownM cause a

Convert a PLC val to the corresponding Haskell value. The inverse of makeKnown.

default readKnown :: KnownBuiltinType val a => Maybe cause -> val -> ReadKnownM cause a

type ReadKnown val = ReadKnownIn (UniOf val) val

makeKnownRun :: MakeKnownIn uni val a => Maybe cause -> a -> (ReadKnownM cause val, DList Text)

makeKnownOrFail :: MakeKnownIn uni val a => a -> EvaluationResult val

Same as makeKnown, but allows for neither emitting nor storing the cause of a failure.

readKnownSelf :: (ReadKnown val a, AsUnliftingError err, AsEvaluationFailure err) => val -> Either (ErrorWithCause err val) a

Same as readKnown, but the cause of a potential failure is the provided term itself.

data SingKind k where

The type of singletonized Haskell kinds representing Plutus kinds. Indexing by a Haskell kind allows us to avoid an error call in the ToKind instance of DefaultUni and not worry about proxies in the type of knownKind.

Constructors

SingType :: SingKind Type

SingKindArrow :: SingKind k -> SingKind l -> SingKind (k -> l)

class KnownKind k where

For reifying Haskell kinds representing Plutus kinds at the term level.

Methods

knownKind :: SingKind k

class ToKind (uni :: Type -> Type) where

For computing the Plutus kind of a built-in type. See kindOfBuiltinType.

Methods

toSingKind :: uni (Esc (a :: k)) -> SingKind k

Reify the kind of a type from the universe at the term level.

demoteKind :: SingKind k -> Kind ()

Convert a reified Haskell kind to a Plutus kind.

kindOfBuiltinType :: ToKind uni => uni (Esc a) -> Kind ()

Compute the kind of a type from a universe.

throwNotAConstant :: (MonadError (ErrorWithCause err cause) m, AsUnliftingError err) => Maybe cause -> m r

Throw an UnliftingError saying that the received argument is not a constant.

class HasConstant term where

Ensures that term has a Constant-like constructor to lift values to and unlift values from.

Methods

asConstant :: AsUnliftingError err => Maybe cause -> term -> Either (ErrorWithCause err cause) (Some (ValueOf (UniOf term)))

Unlift from the Constant constructor throwing an UnliftingError if the provided term is not a Constant.

fromConstant :: Some (ValueOf (UniOf term)) -> term

Wrap a Haskell value as a term.

type HasConstantIn uni term = (UniOf term ~ uni, HasConstant term)

Ensures that term has a Constant-like constructor to lift values to and unlift values from and connects term and its uni.

module PlutusCore.Builtin.Emitter