PlutusCore.Builtin

Description

Reexports from modules from the Builtin folder.

Documentation

class Typeable (a :: k)

The class Typeable allows a concrete representation of a type to be calculated.

Minimal complete definition

typeRep#

data TypeScheme val (args :: [Type]) res where

The type of type schemes of built-in functions. args is a list of types of arguments, res is the resulting type. E.g. Text -> Bool -> Integer is encoded as TypeScheme val [Text, Bool] Integer.

Constructors

TypeSchemeResult :: (Typeable res, KnownTypeAst TyName (UniOf val) res, MakeKnown val res) => TypeScheme val '[] res

TypeSchemeArrow :: (Typeable arg, KnownTypeAst TyName (UniOf val) arg, MakeKnown val arg, ReadKnown val arg) => TypeScheme val args res -> TypeScheme val (arg ': args) res infixr 9

TypeSchemeAll :: (KnownSymbol text, KnownNat uniq, KnownKind kind) => Proxy '(text, uniq, kind) -> TypeScheme val args res -> TypeScheme val args res

argProxy :: TypeScheme val (arg ': args) res -> Proxy arg

typeSchemeToType :: TypeScheme val args res -> Type TyName (UniOf val) ()

Convert a TypeScheme to the corresponding Kind.

class (uni `Everywhere` ImplementedKnownTypeAst uni, uni `Everywhere` ImplementedReadKnownIn uni, uni `Everywhere` ImplementedMakeKnownIn uni) => TestTypesFromTheUniverseAreAllKnown uni

An instance of this class not having any constraints ensures that every type (according to Everywhere) from the universe has 'KnownTypeAst, ReadKnownIn and MakeKnownIn instances.

data BuiltinRuntime val

A BuiltinRuntime represents a possibly partial builtin application, including an empty builtin application (i.e. just the builtin with no arguments).

Applying or type-instantiating a builtin peels off the corresponding constructor from its BuiltinRuntime.

BuiltinCostedResult contains the cost (an ExBudgetStream) and the result (a BuiltinResult (HeadSpine val)) of the builtin application. The cost is stored strictly, since the evaluator is going to look at it and the result is stored lazily, since it's not supposed to be forced before accounting for the cost of the application. If the cost exceeds the available budget, the evaluator discards the result of the builtin application without ever forcing it and terminates with evaluation failure. Allowing the user to compute something that they don't have the budget for would be a major bug.

Evaluators that ignore the entire concept of costing (e.g. the CK machine) may of course force the result of the builtin application unconditionally.

Constructors

BuiltinCostedResult ExBudgetStream ~(BuiltinResult val)

BuiltinExpectArgument (val -> BuiltinRuntime val)

BuiltinExpectForce (BuiltinRuntime val)

data BuiltinsRuntime fun val

A data wrapper around a function returning the BuiltinRuntime of a built-in function.

In order for lookups to be efficient the BuiltinRuntimes need to be cached, i.e. pulled out of the function statically. See makeBuiltinMeaning for how we achieve that.

Constructors

BuiltinsRuntime

Fields

unBuiltinsRuntime :: fun -> BuiltinRuntime val

builtinRuntimeFailure :: BuiltinError -> BuiltinRuntime val

lookupBuiltin :: fun -> BuiltinsRuntime fun val -> BuiltinRuntime val

Look up the runtime info of a built-in function during evaluation.

data EvaluationError structural operational

The type of errors that can occur during evaluation. There are two kinds of errors:

Structural ones -- these are errors that are indicative of the _structure_ of the program being wrong. For example, a free variable was encountered during evaluation, a non-function was applied to an argument or tailList was applied to a non-list.

Operational ones -- these are errors that are indicative of the _logic_ of the program being wrong. For example, error was executed, tailList was applied to an empty list or evaluation ran out of gas.

On the chain both of these are just regular failures and we don't distinguish between them there: if a script fails, it fails, it doesn't matter what the reason was. However in the tests it does matter why the failure occurred: a structural error may indicate that the test was written incorrectly while an operational error may be entirely expected.

In other words, structural errors are "runtime type errors" and operational errors are regular runtime errors. Which means that evaluating an (erased) well-typed program should never produce a structural error, only an operational one. This creates a sort of "runtime type system" for UPLC and it would be great to stick to it and enforce in tests etc, but we currently don't.

Constructors

StructuralError !structural

OperationalError !operational

newtype UnliftingEvaluationError

When unlifting of a PLC term into a Haskell value fails, this error is thrown.

Constructors

MkUnliftingEvaluationError

Fields

unUnliftingEvaluationError :: EvaluationError UnliftingError UnliftingError

data BuiltinError

The type of errors that readKnown and makeKnown can return.

Constructors

BuiltinUnliftingEvaluationError UnliftingEvaluationError

BuiltinEvaluationFailure

data BuiltinResult a

The monad that makeKnown runs in. Equivalent to ExceptT BuiltinError (Writer (DList Text)), except optimized in two ways:

everything is strict

has the BuiltinSuccess constructor that is used for returning a value with no logs attached, which is the most common case for us, so it helps a lot not to construct and deconstruct a redundant tuple

Moving from ExceptT BuiltinError (Writer (DList Text)) to this data type gave us a speedup of 8% of total evaluation time.

Logs are represented as a DList, because we don't particularly care about the efficiency of logging, since there's no logging on the chain and builtins don't emit much anyway. Otherwise we'd have to use text-builder or text-builder-linear or something of this sort.

Constructors

BuiltinSuccess a

BuiltinSuccessWithLogs (Text) BuiltinError

structuralUnliftingError :: Text -> BuiltinResult a -> BuiltinResult a

Prepend logs to a BuiltinResult computation.

throwing :: MonadError e m => AReview e t -> t -> m x

throwing_ :: MonadError e m => AReview e () -> m x

builtinResultFailure :: BuiltinResult a

newtype Opaque val (rep :: Type)

The AST of a value with a Plutus type attached to it. The type is for the Plutus type checker to look at. Opaque can appear in the type of the denotation of a builtin.

Constructors

Opaque

Fields

unOpaque :: val

Methods

typeAst :: Type0 tyname uni ()

uni ~ UniOf val => MakeKnownIn uni val (Opaque val rep)

Instance details

Defined in PlutusCore.Builtin.KnownType

Methods

makeKnown :: Opaque val rep -> BuiltinResult val

uni ~ UniOf val => ReadKnownIn uni val (Opaque val rep)

Instance details

Defined in PlutusCore.Builtin.KnownType

Methods

readKnown :: val -> ReadKnownM (Opaque val rep)

(TypeError NoConstraintsErrMsg :: Constraint) => Monoid (Opaque val rep)

Instance details

Defined in PlutusCore.Builtin.Polymorphism

Methods

mempty :: Opaque val rep

mappend :: Opaque val rep -> Opaque val rep -> Opaque val rep

mconcat :: [Opaque val rep] -> Opaque val rep

(TypeError NoConstraintsErrMsg :: Constraint) => Semigroup (Opaque val rep)

Instance details

Defined in PlutusCore.Builtin.Polymorphism

Methods

(<>) :: Opaque val rep -> Opaque val rep -> Opaque val rep

sconcat :: NonEmpty (Opaque val rep) -> Opaque val rep

stimes :: Integral b => b -> Opaque val rep -> Opaque val rep

(TypeError NoConstraintsErrMsg :: Constraint) => Bounded (Opaque val rep)

Instance details

Defined in PlutusCore.Builtin.Polymorphism

Methods

minBound :: Opaque val rep

maxBound :: Opaque val rep

(TypeError NoConstraintsErrMsg :: Constraint) => Enum (Opaque val rep)

Instance details

Defined in PlutusCore.Builtin.Polymorphism

Methods

succ :: Opaque val rep -> Opaque val rep

pred :: Opaque val rep -> Opaque val rep

toEnum :: Int -> Opaque val rep

fromEnum :: Opaque val rep -> Int

enumFrom :: Opaque val rep -> [Opaque val rep]

enumFromThen :: Opaque val rep -> Opaque val rep -> [Opaque val rep]

enumFromTo :: Opaque val rep -> Opaque val rep -> [Opaque val rep]

enumFromThenTo :: Opaque val rep -> Opaque val rep -> Opaque val rep -> [Opaque val rep]

(TypeError NoConstraintsErrMsg :: Constraint) => Ix (Opaque val rep)

Instance details

Defined in PlutusCore.Builtin.Polymorphism

Methods

range :: (Opaque val rep, Opaque val rep) -> [Opaque val rep]

index :: (Opaque val rep, Opaque val rep) -> Opaque val rep -> Int

unsafeIndex :: (Opaque val rep, Opaque val rep) -> Opaque val rep -> Int

inRange :: (Opaque val rep, Opaque val rep) -> Opaque val rep -> Bool

rangeSize :: (Opaque val rep, Opaque val rep) -> Int

unsafeRangeSize :: (Opaque val rep, Opaque val rep) -> Int

(TypeError NoConstraintsErrMsg :: Constraint) => Num (Opaque val rep)

Instance details

Defined in PlutusCore.Builtin.Polymorphism

Methods

(+) :: Opaque val rep -> Opaque val rep -> Opaque val rep

(-) :: Opaque val rep -> Opaque val rep -> Opaque val rep

(*) :: Opaque val rep -> Opaque val rep -> Opaque val rep

negate :: Opaque val rep -> Opaque val rep

abs :: Opaque val rep -> Opaque val rep

signum :: Opaque val rep -> Opaque val rep

fromInteger :: Integer -> Opaque val rep

(TypeError NoConstraintsErrMsg :: Constraint) => Integral (Opaque val rep)

Instance details

Defined in PlutusCore.Builtin.Polymorphism

Methods

quot :: Opaque val rep -> Opaque val rep -> Opaque val rep

rem :: Opaque val rep -> Opaque val rep -> Opaque val rep

div :: Opaque val rep -> Opaque val rep -> Opaque val rep

mod :: Opaque val rep -> Opaque val rep -> Opaque val rep

quotRem :: Opaque val rep -> Opaque val rep -> (Opaque val rep, Opaque val rep)

divMod :: Opaque val rep -> Opaque val rep -> (Opaque val rep, Opaque val rep)

toInteger :: Opaque val rep -> Integer

(TypeError NoConstraintsErrMsg :: Constraint) => Real (Opaque val rep)

Instance details

Defined in PlutusCore.Builtin.Polymorphism

Methods

toRational :: Opaque val rep -> Rational

(TypeError NoConstraintsErrMsg :: Constraint) => Eq (Opaque val rep)

Instance details

Defined in PlutusCore.Builtin.Polymorphism

Methods

(==) :: Opaque val rep -> Opaque val rep -> Bool

(/=) :: Opaque val rep -> Opaque val rep -> Bool

(TypeError NoConstraintsErrMsg :: Constraint) => Ord (Opaque val rep)

Instance details

Defined in PlutusCore.Builtin.Polymorphism

Methods

compare :: Opaque val rep -> Opaque val rep -> Ordering

(<) :: Opaque val rep -> Opaque val rep -> Bool

(<=) :: Opaque val rep -> Opaque val rep -> Bool

(>) :: Opaque val rep -> Opaque val rep -> Bool

(>=) :: Opaque val rep -> Opaque val rep -> Bool

max :: Opaque val rep -> Opaque val rep -> Opaque val rep

min :: Opaque val rep -> Opaque val rep -> Opaque val rep

HasConstant val => HasConstant (Opaque val rep)

Instance details

Defined in PlutusCore.Builtin.Polymorphism

Methods

asConstant :: Opaque val rep -> Either BuiltinError (Some (ValueOf (UniOf (Opaque val rep))))

fromConstant :: Some (ValueOf (UniOf (Opaque val rep))) -> Opaque val rep

ExMemoryUsage val => ExMemoryUsage (Opaque val rep)

Instance details

Defined in PlutusCore.Builtin.Polymorphism

Methods

memoryUsage :: Opaque val rep -> CostRose

type ToBinds uni acc (Opaque val rep :: Type)

Instance details

Defined in PlutusCore.Builtin.KnownTypeAst

type ToBinds uni acc (Opaque val rep :: Type) = ToBinds uni acc rep

type ToHoles uni _1 (Opaque val rep :: Type)

Instance details

Defined in PlutusCore.Builtin.KnownTypeAst

type ToHoles uni _1 (Opaque val rep :: Type) = '[RepHole rep :: Hole]

type IsBuiltin uni (Opaque val rep :: Type)

Instance details

Defined in PlutusCore.Builtin.KnownTypeAst

type IsBuiltin uni (Opaque val rep :: Type) = 'False

type UniOf (Opaque val rep)

Instance details

Defined in PlutusCore.Builtin.Polymorphism

type UniOf (Opaque val rep) = UniOf val

newtype SomeConstant uni (rep :: Type)

For unlifting from the Constant constructor when the stored value is of a monomorphic built-in type

The rep parameter specifies how the type looks on the PLC side (i.e. just like with Opaque val rep).

Constructors

SomeConstant

Fields

unSomeConstant :: Some (ValueOf uni)

type ToHoles uni _1 (TyForallRep name a :: Type)

Instance details

Defined in PlutusCore.Builtin.KnownTypeAst

type ToHoles uni _1 (TyForallRep name a :: Type) = '[RepHole a :: Hole]

type IsBuiltin uni (TyForallRep name a :: Type)

Instance details

Defined in PlutusCore.Builtin.KnownTypeAst

type IsBuiltin uni (TyForallRep name a :: Type) = 'False

data family BuiltinHead x

For annotating an uninstantiated built-in type, so that it gets handled by the right instance or type family.

data family LastArg x y

LastArg x y is the same thing as y in the signature of the denotation of a built-in functions and this type is only used for referencing x before y, so that the elaboration machinery generates x before y in the all part of the Plutus signature of the builtin. This is a very hacky and indirect way of specifying the ordering of type variables in a Plutus signature, in future we'll do it explicitly by introducing a Forall binder for use in type signatures of denotations of builtins.

Methods

typeAst :: Type0 tyname uni ()

type ToBinds uni acc (LastArg x y :: Type)

Instance details

Defined in PlutusCore.Builtin.KnownTypeAst

type ToBinds uni acc (LastArg x y :: Type) = ToBinds uni (ToBinds uni acc x) y

type ToHoles uni hole (LastArg x y :: Type)

Instance details

Defined in PlutusCore.Builtin.KnownTypeAst

type ToHoles uni hole (LastArg x y :: Type) = '[RunHole hole x, RunHole hole y]

type IsBuiltin uni (LastArg x y :: Type)

Instance details

Defined in PlutusCore.Builtin.KnownTypeAst

type IsBuiltin uni (LastArg x y :: Type) = IsBuiltin uni y

type family ElaborateBuiltin uni x

Take an iterated application of a built-in type and elaborate every function application inside of it to TyAppRep and annotate the head with BuiltinHead.

The idea is that we don't need to process built-in types manually if we simply add some annotations for instance resolution to look for. Think what we'd have to do manually for, say, ToHoles: traverse the spine of the application and collect all the holes into a list, which is troubling, because type applications are left-nested and lists are right-nested, so we'd have to use accumulators or an explicit Reverse type family. And then we also have KnownTypeAst and ToBinds, so handling built-in types in a special way for each of those would be a hassle, especially given the fact that type-level Haskell is not exactly good at computing things. With the ElaborateBuiltin approach we get KnownTypeAst, ToHoles and ToBinds for free.

We make this an open type family, so that elaboration is customizable for each universe.

type family AllElaboratedArgs constr x where ...

Take a constraint and use it to constrain every argument of a possibly 0-ary elaborated application of a built-in type.

Equations

AllElaboratedArgs constr (f `TyAppRep` x) = (constr x, AllElaboratedArgs constr f)

AllElaboratedArgs _ (BuiltinHead _) = ()

class AllElaboratedArgs constr (ElaborateBuiltin uni x) => AllBuiltinArgs uni constr x

Take a constraint and use it to constrain every argument of a possibly 0-ary application of a built-in type.

type family FoldArgs args res where ...

Turn a list of Haskell types args into a functional type ending in res.

>>> :set -XDataKinds
>>> :kind! FoldArgs [(), Bool] Integer
FoldArgs [(), Bool] Integer :: *
= () -> Bool -> Integer 

Equations

FoldArgs '[] res = res

FoldArgs (arg ': args) res = arg -> FoldArgs args res

data BuiltinMeaning val cost

The meaning of a built-in function consists of its type represented as a TypeScheme, its Haskell denotation and its uninstantiated runtime denotation.

The TypeScheme of a built-in function is used for example for

computing the PLC type of the function to be used during type checking

getting arity information

generating arbitrary values to apply the function to in tests

The denotation is lazy, so that we don't need to worry about a builtin being bottom (happens in tests). The production path is not affected by that, since only runtime denotations are used for evaluation.

Constructors

forall args res. BuiltinMeaning (TypeScheme val args res) ~(FoldArgs args res) (cost -> BuiltinRuntime val)

type HasMeaningIn uni val = (Typeable val, ExMemoryUsage val, HasConstantIn uni val)

Constraints available when defining a built-in function.

class (Typeable uni, Typeable fun, Bounded fun, Enum fun, Ix fun, Default (BuiltinSemanticsVariant fun)) => ToBuiltinMeaning uni fun where

A type class for "each function from a set of built-in functions has a BuiltinMeaning".

Associated Types

type CostingPart uni fun

The cost part of BuiltinMeaning.

data BuiltinSemanticsVariant fun

See Note [Builtin semantics variants]

Methods

toBuiltinMeaning :: HasMeaningIn uni val => BuiltinSemanticsVariant fun -> fun -> BuiltinMeaning val (CostingPart uni fun)

Get the BuiltinMeaning of a built-in function.

withTypeSchemeOfBuiltinFunction :: forall val fun r. (ToBuiltinMeaning (UniOf val) fun, ExMemoryUsage val, Typeable val, HasConstant val) => BuiltinSemanticsVariant fun -> fun -> (forall args res. TypeScheme val args res -> r) -> r

Feed the TypeScheme of the given built-in function to the continuation.

typeOfBuiltinFunction :: forall uni fun. ToBuiltinMeaning uni fun => BuiltinSemanticsVariant fun -> fun -> Type TyName uni ()

Get the type of a built-in function.

type family GetArgs a where ...

Chop a function type to get a list of its argument types.

Equations

GetArgs (a -> b) = a ': GetArgs b

GetArgs _ = '[]

class KnownMonotype val args res where

A class that allows us to derive a monotype for a builtin. We could've computed the runtime denotation from the TypeScheme and the denotation of the builtin, but not statically (due to unfolding not working for recursive functions and TypeScheme being recursive, i.e. requiring the conversion function to be recursive), and so it would cause us to retain a lot of evaluation-irrelevant stuff in the constructors of BuiltinRuntime, which has to make evaluation slower (we didn't check) and certainly makes the generated Core much harder to read. Technically speaking, we could get a RuntimeScheme from the TypeScheme and the denotation statically if we changed the definition of TypeScheme and made it a singleton, but then the conversion function would have to become a class anyway and we'd just replicate what we have here, except in a much more complicated way.

Methods

knownMonotype :: TypeScheme val args res

toMonoF :: ReadKnownM (FoldArgs args res, FoldArgs args ExBudgetStream) -> BuiltinRuntime val

Convert the denotation of a builtin to its runtime counterpart. The argument is in ReadKnownM, because that's what deferred unlifting amounts to: passing the action returning the builtin application around until full saturation, which is when the action actually gets run.

class KnownMonotype val args res => KnownPolytype (binds :: [Some TyNameRep]) val args res where

A class that allows us to derive a polytype for a builtin.

Methods

knownPolytype :: TypeScheme val args res

toPolyF :: ReadKnownM (FoldArgs args res, FoldArgs args ExBudgetStream) -> BuiltinRuntime val

Convert the denotation of a builtin to its runtime counterpart. The argument is in ReadKnownM, because that's what we need to do: passing the action returning the builtin application around until full saturation, which is when the action actually gets run.

type family ThrowOnBothEmpty binds args isBuiltin a where ...

Ensure a built-in function is not nullary and throw a nice error otherwise.

Equations

ThrowOnBothEmpty '[] '[] 'True a = TypeError (('Text "A built-in function must take at least one type or term argument" ':$$: ((('Text "\8216" ':<>: 'ShowType a) ':<>: 'Text "\8217 is a built-in type") ':<>: 'Text " so you can embed any of its values as a constant")) ':$$: 'Text "If you still want a built-in function, add a dummy \8216()\8217 argument")

ThrowOnBothEmpty '[] '[] 'False a = TypeError ('Text "A built-in function must take at least one type or term argument" ':$$: 'Text "To fix this error add a dummy \8216()\8217 argument")

ThrowOnBothEmpty _ _ _ _ = ()

class MakeBuiltinMeaning a val where

A function turned into a type class with exactly one fully general instance. We can't package up the constraints of makeBuiltinMeaning (see the instance) into a type or class synonym, because they contain a bunch of variables defined by ~ or determined via functional dependencies and neither class nor type definitions can handle that (see https://gitlab.haskell.org/ghc/ghc/-/issues/7100). Inlining three lines of constraints whenever we need to call makeBuiltinMeaning over a non-concrete type is a bad option and this abstraction is free anyway, hence its existence.

The a type variable goes first, because makeBuiltinMeaning @A is a common pattern.

Methods

makeBuiltinMeaning :: a -> (cost -> FoldArgs (GetArgs a) ExBudgetStream) -> BuiltinMeaning val cost

Construct the meaning for a built-in function by automatically deriving its TypeScheme, given

the denotation of the builtin

an uninstantiated costing function

toBuiltinRuntime :: cost -> BuiltinMeaning val cost -> BuiltinRuntime val

Convert a BuiltinMeaning to a BuiltinRuntime given a cost model.

toBuiltinsRuntime :: (cost ~ CostingPart uni fun, ToBuiltinMeaning uni fun, HasMeaningIn uni val) => BuiltinSemanticsVariant fun -> cost -> BuiltinsRuntime fun val

Calculate runtime info for all built-in functions given meanings of builtins (as a constraint), the semantics variant of the set of builtins and a cost model.

data TyNameRep (kind :: Some name) (ToBinds uni acc a)

type ToHoles uni _1 (TyForallRep name a :: Type)

Instance details

Defined in PlutusCore.Builtin.KnownTypeAst

type ToHoles uni _1 (TyForallRep name a :: Type) = '[RepHole a :: Hole]

type IsBuiltin uni (TyForallRep name a :: Type)

Instance details

Defined in PlutusCore.Builtin.KnownTypeAst

type IsBuiltin uni (TyForallRep name a :: Type) = 'False

data Hole

The kind of holes.

data family RepHole x

A hole in the Rep context.

data family TypeHole a

A hole in the Type context.

type family RunHole hole where ...

Turn a hole in the GHC.Type -> GHC.Type form into one of the Hole form. This only changes the kind of the given argument. This is a way of encoding forall a. a -> Hole at the kind level, which we don't attempt to use, because GHC apparently hates polymorphism at the kind level and chokes upon encountering it.

Equations

RunHole RepHole = RepHole

RunHole TypeHole = TypeHole

type HasTermLevel uni = Includes uni

Specifies that the given type is a built-in one and its values can be embedded into a Term.

type HasTypeLevel uni x = KnownTypeAst Void uni (ElaborateBuiltin uni x)

Specifies that the given type is a built-in one and can be embedded into a Kind.

type HasTypeAndTermLevel uni x = (uni `HasTypeLevel` x, uni `HasTermLevel` x)

The product of HasTypeLevel and HasTermLevel.

mkTyBuiltin :: forall a (x :: a) uni ann tyname. uni `HasTypeLevel` x => ann -> Type tyname uni ann

Convert a Haskell representation of a possibly 0-ary application of a built-in type to arbitrary types implementing KnownTypeAst.

type KnownBuiltinTypeAst tyname uni x = AllBuiltinArgs uni (KnownTypeAst tyname uni) x

A constraint for "a is a KnownTypeAst by means of being included in uni".

class KnownTypeAst tyname uni x where

This class allows one to convert the type-level Haskell representation of a Plutus type into the corresponding Plutus type. Associated type families are needed to help elaboration.

Depending on the universe converting a Haskell type to a Plutus team can give different results (e.g. Int can be a built-in type instead of being encoded via built-in Integer), hence this class takes a uni argument. Plus, elaboration is universe-specific too.

Minimal complete definition

Nothing

Associated Types

type IsBuiltin uni x :: Bool

Whether x is a built-in type.

type IsBuiltin uni x = IsBuiltin uni (ElaborateBuiltin uni x)

type ToHoles uni (hole :: Type -> Type) x :: [Hole]

Return every part of the type that can be a to-be-instantiated type variable. For example, in Integer there's no such types and in (a, b) it's the two arguments (a and b) and the same applies to a -> b (to mention a type that is not built-in).

Takes a hole in the GHC.Type -> GHC.Type form (a convention originally adopted in the elaborator, perhaps not a very helpful one), which can be turned into an actual Hole via RunHole.

type ToHoles uni hole x = ToHoles uni hole (ElaborateBuiltin uni x)

type ToBinds uni (acc :: [Some TyNameRep]

Collect all unique variables (a variable consists of a textual name, a unique and a kind) in an accumulator and return the accumulator once a leaf is reached.

type ToBinds uni acc x = ToBinds uni acc (ElaborateBuiltin uni x)

Methods

typeAst :: Type tyname uni ()

The Plutus counterpart of the x type.

default typeAst :: KnownTypeAst tyname uni (ElaborateBuiltin uni x) => Type tyname uni ()

Methods

typeAst :: Type tyname uni ()

toTypeAst :: forall a tyname uni (x :: a) proxy. KnownTypeAst tyname uni x => proxy x -> Type tyname uni ()

Return the Plutus counterpart of the x type.

type family Insert x xs where ...

Insert x into xs unless it's already there.

Equations

Insert x '[] = '[x]

Insert x (x : xs) = x ': xs

Insert x (y : xs) = y ': Insert x xs

type family Delete x xs where ...

Delete the first x from a list. Which is okay since we only ever put things in once.

Equations

Delete _ '[] = '[]

Delete x (x ': xs) = xs

Delete x (y ': xs) = y ': Delete x xs

data BuiltinError

The type of errors that readKnown and makeKnown can return.

class GEqL f a where

A version of GEq that fixes a in place, which allows us to create an inlinable recursive implementation of geqL.

The way it works is that whenever there's recursion, we look up the recursive case in the current context (i.e. the dictionary) instead of actually calling geqL recursively (even though it's gonna look like we do exactly that, because there's no way to distinguish between a recursive call and a dictionary lookup as the two share the same name, although to help GHC choose a lookup we sprinkle the perhaps unreliable LoopBreaker in the DefaultUni instance of this class).

Alligning things this way allows us to inline arbitrarily deep recursion for as long as types keep being monomorphic.

For example, the MapData builtin accepts a [(Data, Data)] and with geqL matching on all of DefaultUniProtoList, DefaultUniProtoPair and DefaultUniData gets inlined in the denotation of the builtin. For the Constr builtin that resulted in a 4.3% speedup at the time this comment was written.

Methods

geqL :: f (Esc a) -> f (Esc b) -> EvaluationResult (a :~: b)

newtype LoopBreaker uni a

In f = ... f ... where f is a class method, how do you know if f is going to be a recursive call or a type class method call? If both type check, then you don't really know how GHC is going to play it. So we add this data type to make sure that the RHS f will have to become a type class method call.

Can GHC turn that method call into a recursive one once type classes are resolved? Dunno, but at least we've introduced an obstacle preventing GHC from immediately creating a non-inlinable recursive definition.

Constructors

LoopBreaker (uni a)

type KnownBuiltinTypeIn uni val a = (HasConstantIn uni val, PrettyParens (SomeTypeIn uni), GEqL uni a, uni `HasTermLevel` a)

A constraint for "a is a ReadKnownIn and MakeKnownIn by means of being included in uni".

type KnownBuiltinType val a = KnownBuiltinTypeIn (UniOf val) val a

A constraint for "a is a ReadKnownIn and MakeKnownIn by means of being included in UniOf term".

data BuiltinResult a

The monad that makeKnown runs in. Equivalent to ExceptT BuiltinError (Writer (DList Text)), except optimized in two ways:

everything is strict

has the BuiltinSuccess constructor that is used for returning a value with no logs attached, which is the most common case for us, so it helps a lot not to construct and deconstruct a redundant tuple

Moving from ExceptT BuiltinError (Writer (DList Text)) to this data type gave us a speedup of 8% of total evaluation time.

Logs are represented as a DList, because we don't particularly care about the efficiency of logging, since there's no logging on the chain and builtins don't emit much anyway. Otherwise we'd have to use text-builder or text-builder-linear or something of this sort.

Constructors

BuiltinSuccess a

BuiltinSuccessWithLogs (Text) BuiltinError

Methods

typeAst :: Type0 tyname uni ()

MakeKnownIn uni val a => MakeKnownIn uni val (BuiltinResult a)

Instance details

Defined in PlutusCore.Builtin.KnownType

Methods

makeKnown :: BuiltinResult a -> BuiltinResult val

(TypeError ('Text "\8216BuiltinResult\8217 cannot appear in the type of an argument") :: Constraint, uni ~ UniOf val) => ReadKnownIn uni val (BuiltinResult a)

Instance details

Defined in PlutusCore.Builtin.KnownType

Methods

readKnown :: val -> ReadKnownM (BuiltinResult a)

Show a => Show (BuiltinResult a)

Instance details

Defined in PlutusCore.Builtin.Result

Methods

showsPrec :: Int -> BuiltinResult a -> ShowS

show :: BuiltinResult a -> String

showList :: [BuiltinResult a] -> ShowS

type ToBinds uni acc (BuiltinResult a :: Type)

Instance details

Defined in PlutusCore.Builtin.KnownTypeAst

type ToBinds uni acc (BuiltinResult a :: Type) = ToBinds uni acc a

type ToHoles uni _1 (BuiltinResult a :: Type)

Instance details

Defined in PlutusCore.Builtin.KnownTypeAst

type ToHoles uni _1 (BuiltinResult a :: Type) = '[TypeHole a :: Hole]

type IsBuiltin uni (BuiltinResult a :: Type)

Instance details

Defined in PlutusCore.Builtin.KnownTypeAst

type IsBuiltin uni (BuiltinResult a :: Type) = 'False

type ReadKnownM = Either BuiltinError

The monad that readKnown runs in.

data Spine a

A non-empty spine. Isomorphic to NonEmpty, except is strict and is defined as a single recursive data type.

Constructors

SpineLast a

SpineCons a (Spine a)

data HeadSpine err a b

The head-spine form of an iterated application. Provides O(1) access to the head of the application. NonEmpty a ~ HeadSpine a a, except is strict and the no-spine case is made a separate constructor for performance reasons (it only takes a single pattern match to access the head when there's no spine this way, while otherwise we'd also need to match on the spine to ensure that it's empty -- and the no-spine case is by far the most common one, hence we want to optimize it).

Used in built-in functions returning function applications such as CaseList.

Constructors

HeadOnly a

HeadSpine a (Spine b)

HeadError ~err

headSpine :: a -> [b] -> HeadSpine err a b

Construct HeadSpine from head and list.

type MonoHeadSpine err a = HeadSpine err a a

HeadSpine but the type of head and spine is same

class uni ~ UniOf val => MakeKnownIn uni val a where

Minimal complete definition

Nothing

Methods

makeKnown :: a -> BuiltinResult val

Convert a Haskell value to the corresponding PLC value. The inverse of readKnown.

default makeKnown :: KnownBuiltinType val a => a -> BuiltinResult val

KnownBuiltinTypeIn DefaultUni term Vector a -> BuiltinResult term

KnownBuiltinTypeIn DefaultUni term [a] => MakeKnownIn DefaultUni term [a]

Instance details

Defined in PlutusCore.Default.Universe

Methods

makeKnown :: [a] -> BuiltinResult term

MakeKnownIn uni val a => MakeKnownIn uni val (BuiltinResult a)

Instance details

Defined in PlutusCore.Builtin.KnownType

Methods

makeKnown :: BuiltinResult a -> BuiltinResult val

(TypeError ('Text "Use \8216BuiltinResult\8217 instead of \8216EvaluationResult\8217") :: Constraint, uni ~ UniOf val) => MakeKnownIn uni val (EvaluationResult a)

Instance details

Defined in PlutusCore.Builtin.KnownType

Methods

makeKnown :: EvaluationResult a -> BuiltinResult val

KnownBuiltinTypeIn DefaultUni term (a, b) => MakeKnownIn DefaultUni term (a, b)

Instance details

Defined in PlutusCore.Default.Universe

Methods

makeKnown :: (a, b) -> BuiltinResult term

uni ~ UniOf val => MakeKnownIn uni val (Opaque val rep)

Instance details

Defined in PlutusCore.Builtin.KnownType

Methods

makeKnown :: Opaque val rep -> BuiltinResult val

HasConstantIn uni val => MakeKnownIn uni val (SomeConstant uni rep)

Instance details

Defined in PlutusCore.Builtin.KnownType

Methods

makeKnown :: SomeConstant uni rep -> BuiltinResult val

MakeKnownIn DefaultUni term a => MakeKnownIn DefaultUni term (MetaForall name a)

Instance details

Defined in PlutusCore.Examples.Builtins

Methods

makeKnown :: MetaForall name a -> BuiltinResult term

readKnownConstant :: forall val a. KnownBuiltinType val a => val -> ReadKnownM a

Convert a constant embedded into a PLC term to the corresponding Haskell value.

type MakeKnown val = MakeKnownIn (UniOf val) val

class uni ~ UniOf val => ReadKnownIn uni val a where

Minimal complete definition

Nothing

Methods

readKnown :: val -> ReadKnownM a

Convert a PLC value to the corresponding Haskell value. The inverse of makeKnown.

default readKnown :: KnownBuiltinType val a => val -> ReadKnownM a

KnownBuiltinTypeIn DefaultUni term Vector a)

KnownBuiltinTypeIn DefaultUni term [a] => ReadKnownIn DefaultUni term [a]

Instance details

Defined in PlutusCore.Default.Universe

Methods

readKnown :: term -> ReadKnownM [a]

(TypeError ('Text "\8216BuiltinResult\8217 cannot appear in the type of an argument") :: Constraint, uni ~ UniOf val) => ReadKnownIn uni val (BuiltinResult a)

Instance details

Defined in PlutusCore.Builtin.KnownType

Methods

readKnown :: val -> ReadKnownM (BuiltinResult a)

(TypeError ('Text "Use \8216BuiltinResult\8217 instead of \8216EvaluationResult\8217") :: Constraint, uni ~ UniOf val) => ReadKnownIn uni val (EvaluationResult a)

Instance details

Defined in PlutusCore.Builtin.KnownType

Methods

readKnown :: val -> ReadKnownM (EvaluationResult a)

KnownBuiltinTypeIn DefaultUni term (a, b) => ReadKnownIn DefaultUni term (a, b)

Instance details

Defined in PlutusCore.Default.Universe

Methods

readKnown :: term -> ReadKnownM (a, b)

uni ~ UniOf val => ReadKnownIn uni val (Opaque val rep)

Instance details

Defined in PlutusCore.Builtin.KnownType

Methods

readKnown :: val -> ReadKnownM (Opaque val rep)

HasConstantIn uni val => ReadKnownIn uni val (SomeConstant uni rep)

Instance details

Defined in PlutusCore.Builtin.KnownType

Methods

readKnown :: val -> ReadKnownM (SomeConstant uni rep)

type ReadKnown val = ReadKnownIn (UniOf val) val

makeKnownOrFail :: MakeKnownIn uni val a => a -> EvaluationResult val

Same as makeKnown, but allows for neither emitting nor storing the cause of a failure.

readKnownSelf :: (ReadKnown val a, BuiltinErrorToEvaluationError structural operational) => val -> Either (ErrorWithCause (EvaluationError structural operational) val) a

Same as readKnown, but the cause of a potential failure is the provided term itself.

data SingKind k where

The type of singletonized Haskell kinds representing Plutus kinds. Indexing by a Haskell kind allows us to avoid an error call in the ToKind instance of DefaultUni and not worry about proxies in the type of knownKind.

Constructors

SingType :: SingKind Type

SingKindArrow :: SingKind k -> SingKind l -> SingKind (k -> l) infixr 5

withSingKind :: Kind ann -> (forall k. SingKind k -> r) -> r

Feed the SingKind version of the given Kind to the given continuation.

class KnownKind k where

For reifying Haskell kinds representing Plutus kinds at the term level.

Methods

knownKind :: SingKind k

bringKnownKind :: SingKind k -> (KnownKind k => r) -> r

Satisfy the KnownKind constraint of a continuation using the given SingKind.

withKnownKind :: Kind ann -> (forall k. KnownKind k => Proxy k -> r) -> r

class ToKind (uni :: Type -> Type) where

For computing the Plutus kind of a built-in type. See kindOfBuiltinType.

Methods

toSingKind :: uni (Esc (a :: k)) -> SingKind k

Reify the kind of a type from the universe at the term level.

demoteKind :: SingKind k -> Kind ()

Convert a reified Haskell kind to a Plutus kind.

kindOfBuiltinType :: ToKind uni => uni (Esc a) -> Kind ()

Compute the kind of a type from a universe.

data BuiltinError

The type of errors that readKnown and makeKnown can return.

Constructors

BuiltinUnliftingEvaluationError UnliftingEvaluationError

BuiltinEvaluationFailure

notAConstant :: BuiltinError

class HasConstant term where

Ensures that term has a Constant-like constructor to lift values to and unlift values from.

Methods

asConstant :: term -> Either BuiltinError (Some (ValueOf (UniOf term)))

Unwrap from a Constant-like constructor throwing an UnliftingError if the provided term is not a wrapped Haskell value.

fromConstant :: Some (ValueOf (UniOf term)) -> term

Wrap a Haskell value as a term.

type HasConstantIn uni term = (UniOf term ~ uni, HasConstant term)

Ensures that term has a Constant-like constructor to lift values to and unlift values from and connects term and its uni.

fromValueOf :: forall a term. HasConstant term => UniOf term (Esc a) -> a -> term

Wrap a Haskell value (given its explicit type tag) as a term.

fromValue :: forall a term. (HasConstant term, UniOf term `HasTermLevel` a) => a -> term

Wrap a Haskell value (provided its type is in the universe) as a term.

class AnnotateCaseBuiltin uni where

Methods

annotateCaseBuiltin :: UniOf term Text [(term, [Type TyName uni ann])]

Given a tag for a built-in type and a list of branches, annotate each of the branches with its expected argument types or fail if casing on values of the built-in type isn't supported. Note: you don't need to include the resulting type of the whole case matching in the returning list here.

data CaserBuiltin uni

A data version of CaseBuiltin. we parameterize the evaluator by a CaserBuiltin so that the caller can choose whether to use the caseBuiltin method or the always failing caser (the latter is required for earlier protocol versions when we didn't support casing on builtins).

Constructors

CaserBuiltin

Fields

unCaserBuiltin :: !(forall term. UniOf term Text term (Some (ValueOf uni)))

unavailableCaserBuiltin :: Int -> CaserBuiltin uni