PlutusCore.Core

Documentation

data Kind ann

Constructors

Type ann

KindArrow ann (Kind ann) (Kind ann)

toPatFuncKind :: Kind () -> Kind ()

The kind of a pattern functor (the first Kind argument of TyIFix) at a given kind (of the second Kind argument of TyIFix):

toPatFuncKind k = (k -> *) -> k -> *

fromPatFuncKind :: Kind () -> Maybe (Kind ())

argsFunKind :: Kind ann -> [Kind ann]

Extract all a_i from a_0 -> a_1 -> ... -> r.

data Type tyname uni ann

A Kind assigned to expressions.

Constructors

TyVar ann tyname

TyFun ann (Type tyname uni ann) (Type tyname uni ann)

TyIFix ann (Type tyname uni ann) (Type tyname uni ann)
Fix-point type, for constructing self-recursive types

TyForall ann tyname (Kind ann) (Type tyname uni ann)

TyBuiltin ann (SomeTypeIn uni)
Builtin type

TyLam ann tyname (Kind ann) (Type tyname uni ann)

TyApp ann (Type tyname uni ann) (Type tyname uni ann)

data Term tyname name uni fun ann

Constructors

Var ann name
a named variable

TyAbs ann tyname (Kind ann) (Term tyname name uni fun ann)

LamAbs ann name (Type tyname uni ann) (Term tyname name uni fun ann)

Apply ann (Term tyname name uni fun ann) (Term tyname name uni fun ann)

Constant ann (Some (ValueOf uni))
a constant term

Builtin ann fun

TyInst ann (Term tyname name uni fun ann) (Type tyname uni ann)

Unwrap ann (Term tyname name uni fun ann)

IWrap ann (Type tyname uni ann) (Type tyname uni ann) (Term tyname name uni fun ann)

Error ann (Type tyname uni ann)

data Version ann

The version of Plutus Core used by this program.

The intention is to convey different levels of backwards compatibility for existing scripts: - Major version changes are backwards-incompatible - Minor version changes are backwards-compatible - Patch version changes should be entirely invisible (and we will likely not use this level)

The version used should be changed only when the language itself changes. For example, adding a new kind of term to the language would require a minor version bump; removing a kind of term would require a major version bump.

Similarly, changing the semantics of the language will require a version bump, typically a major one. This is the main reason why the version is actually tracked in the AST: we can have two language versions with identical ASTs but different semantics, so we need to track the version explicitly.

Compatibility is about compatibility for specific scripts, not about e.g. tools which consume scripts. Adding a new kind of term does not change how existing scripts behave, but does change what tools would need to do to process scripts.

Constructors

Version ann Natural Natural Natural

data Program tyname name uni fun ann

A Program is simply a Term coupled with a Version of the core language.

Constructors

Program

Fields

_progAnn :: ann

_progVer :: Version ann

_progTerm :: Term tyname name uni fun ann

type family UniOf a :: Type -> Type

Extract the universe from a type.

newtype Normalized a

Constructors

Normalized

Fields

unNormalized :: a

data TyVarDecl tyname ann

A "type variable declaration", i.e. a name and a kind for a type variable.

Constructors

TyVarDecl

Fields

_tyVarDeclAnn :: ann

_tyVarDeclName :: tyname

_tyVarDeclKind :: Kind ann

data VarDecl tyname name uni ann

A "variable declaration", i.e. a name and a type for a variable.

Constructors

VarDecl

Fields

_varDeclAnn :: ann

_varDeclName :: name

_varDeclType :: Type tyname uni ann

data TyDecl tyname uni ann

A "type declaration", i.e. a kind for a type.

Constructors

TyDecl

Fields

_tyDeclAnn :: ann

_tyDeclType :: Type tyname uni ann

_tyDeclKind :: Kind ann

tyDeclVar :: TyVarDecl tyname ann -> TyDecl tyname uni ann

type family HasUniques a :: Constraint

All kinds of uniques an entity contains.

newtype Binder name

This is a wrapper to mark the place where the binder is introduced (i.e. LamAbs/TyAbs) and not where it is actually used (TyVar/Var..). This marking allows us to skip the (de)serialization of binders at LamAbs/TyAbs positions iff name is DeBruijn-encoded (level or index). See for example the instance of Flat

Constructors

Binder

Fields

unBinder :: name

defaultVersion :: ann -> Version ann

The default version of Plutus Core supported by this library.

Helper functions

termAnn :: Term tyname name uni fun ann -> ann

typeAnn :: Type tyname uni ann -> ann

mapFun :: (fun -> fun') -> Term tyname name uni fun ann -> Term tyname name uni fun' ann

Map a function over the set of built-in functions.

tyVarDeclAnn :: forall tyname ann. Lens' (TyVarDecl tyname ann) ann

tyVarDeclName :: forall tyname ann tyname. Lens (TyVarDecl tyname ann) (TyVarDecl tyname ann) tyname tyname

tyVarDeclKind :: forall tyname ann. Lens' (TyVarDecl tyname ann) (Kind ann)

varDeclAnn :: forall tyname name uni ann. Lens' (VarDecl tyname name uni ann) ann

varDeclName :: forall tyname name uni ann name. Lens (VarDecl tyname name uni ann) (VarDecl tyname name uni ann) name name

varDeclType :: forall tyname name uni ann tyname uni. Lens (VarDecl tyname name uni ann) (VarDecl tyname name uni ann) (Type tyname uni ann) (Type tyname uni ann)

tyDeclAnn :: forall tyname uni ann. Lens' (TyDecl tyname uni ann) ann

tyDeclType :: forall tyname uni ann tyname uni. Lens (TyDecl tyname uni ann) (TyDecl tyname uni ann) (Type tyname uni ann) (Type tyname uni ann)

tyDeclKind :: forall tyname uni ann. Lens' (TyDecl tyname uni ann) (Kind ann)

progAnn :: forall tyname name uni fun ann. Lens' (Program tyname name uni fun ann) ann

progVer :: forall tyname name uni fun ann. Lens' (Program tyname name uni fun ann) (Version ann)

progTerm :: forall tyname name uni fun ann tyname name uni fun. Lens (Program tyname name uni fun ann) (Program tyname name uni fun ann) (Term tyname name uni fun ann) (Term tyname name uni fun ann)

kindSubkinds :: Traversal' (Kind ann) (Kind ann)

kindSubkindsDeep :: Fold (Kind ann) (Kind ann)

tyVarDeclSubkinds :: Traversal' (TyVarDecl tyname a) (Kind a)

Get all the direct child Kinds of the given TyVarDecl.

typeTyBinds :: Traversal' (Type tyname uni ann) tyname

Get all the direct child 'tyname a's of the given Kind from binders.

typeTyVars :: Traversal' (Type tyname uni ann) tyname

Get all the direct child 'tyname a's of the given Kind from TyVars.

typeUniques :: HasUniques (Type tyname uni ann) => Traversal' (Type tyname uni ann) Unique

Get all the direct child Uniques of the given Kind from binders TyVars.

typeSubkinds :: Traversal' (Type tyname uni ann) (Kind ann)

Get all the direct child Kinds of the given Kind.

typeSubtypes :: Traversal' (Type tyname uni ann) (Type tyname uni ann)

Get all the direct child Kinds of the given Kind.

typeSubtypesDeep :: Fold (Type tyname uni ann) (Type tyname uni ann)

Get all the transitive child Kinds of the given Kind.

varDeclSubtypes :: Traversal' (VarDecl tyname name uni a) (Type tyname uni a)

Get all the direct child Kinds of the given VarDecl.

termTyBinds :: Traversal' (Term tyname name uni fun ann) tyname

Get all the direct child 'tyname a's of the given Term from TyAbses.

termBinds :: Traversal' (Term tyname name uni fun ann) name

Get all the direct child 'name a's of the given Term from LamAbses.

termVars :: Traversal' (Term tyname name uni fun ann) name

Get all the direct child 'name a's of the given Term from Vars.

termUniques :: HasUniques (Term tyname name uni fun ann) => Traversal' (Term tyname name uni fun ann) Unique

Get all the direct child Uniques of the given Term (including the type-level ones).

termSubkinds :: Traversal' (Term tyname name uni fun ann) (Kind ann)

Get all the direct child Kinds of the given Term.

termSubtypes :: Traversal' (Term tyname name uni fun ann) (Type tyname uni ann)

Get all the direct child Kinds of the given Term.

termSubtypesDeep :: Fold (Term tyname name uni fun ann) (Type tyname uni ann)

Get all the transitive child Kinds of the given Term.

termSubterms :: Traversal' (Term tyname name uni fun ann) (Term tyname name uni fun ann)

Get all the direct child Terms of the given Term.

termSubtermsDeep :: Fold (Term tyname name uni fun ann) (Term tyname name uni fun ann)

Get all the transitive child Terms of the given Term.

typeUniquesDeep :: HasUniques (Type tyname uni ann) => Fold (Type tyname uni ann) Unique

Get all the transitive child Uniques of the given Kind.

termUniquesDeep :: HasUniques (Term tyname name uni fun ann) => Fold (Term tyname name uni fun ann) Unique

Get all the transitive child Uniques of the given Term (including the type-level ones).

Base functors

data TermF (tyname :: Type) (name :: Type) (uni :: Type -> Type) (fun :: Type) (ann :: Type) r

Constructors

VarF ann name

TyAbsF ann tyname (Kind ann) r

LamAbsF ann name (Type tyname uni ann) r

ApplyF ann r r

ConstantF ann (Some (ValueOf uni))

BuiltinF ann fun

TyInstF ann r (Type tyname uni ann)

UnwrapF ann r

IWrapF ann (Type tyname uni ann) (Type tyname uni ann) r

ErrorF ann (Type tyname uni ann)

data TypeF (tyname :: Type) (uni :: Type -> Type) (ann :: Type) r

Constructors

TyVarF ann tyname

TyFunF ann r r

TyIFixF ann r r

TyForallF ann tyname (Kind ann) r

TyBuiltinF ann (SomeTypeIn uni)

TyLamF ann tyname (Kind ann) r

TyAppF ann r r

data KindF (ann :: Type) r

Constructors

TypeF ann

KindArrowF ann r r