PlutusCore.Data

Documentation

data Data

A generic "data" type.

The main constructor Constr represents a datatype value in sum-of-products form: Constr i args represents a use of the ith constructor along with its arguments.

The other constructors are various primitives.

Constructors

Constr NumBits

Pretty Data

Instance details

Defined in PlutusCore.Data

Methods

pretty :: Data -> Doc ann

prettyList :: [Data] -> Doc ann

Some TyNameRep]

Methods

typeAst :: Type tyname DefaultUni ()

type Rep Data

Instance details

Defined in PlutusCore.Data

type Rep Data = D1 ('MetaData "Data" "PlutusCore.Data" "plutus-core-1.60.0.0-inplace" 'False) ((C1 ('MetaCons "Constr" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 Integer) :*: S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 [Data])) :+: C1 ('MetaCons "Map" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 [(Data, Data)]))) :+: (C1 ('MetaCons "List" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 [Data])) :+: (C1 ('MetaCons "I" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 Integer)) :+: C1 ('MetaCons "B" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 ByteString)))))

type IsBuiltin DefaultUni Data

Instance details

Defined in PlutusCore.Default.Universe

type IsBuiltin DefaultUni Data = IsBuiltin DefaultUni (ElaborateBuiltin DefaultUni Data)

type ToBinds DefaultUni acc Data

Instance details

Defined in PlutusCore.Default.Universe

type ToBinds DefaultUni acc Data = ToBinds DefaultUni acc (ElaborateBuiltin DefaultUni Data)

type ToHoles DefaultUni hole Data

Instance details

Defined in PlutusCore.Default.Universe

type ToHoles DefaultUni hole Data = ToHoles DefaultUni hole (ElaborateBuiltin DefaultUni Data)