PlutusCore.Default

Documentation

data DefaultUni a where

The universe used by default.

Constructors

DefaultUniInteger :: DefaultUni (Esc Integer)

DefaultUniByteString :: DefaultUni (Esc ByteString)

DefaultUniString :: DefaultUni (Esc Text)

DefaultUniUnit :: DefaultUni (Esc ())

DefaultUniBool :: DefaultUni (Esc Bool)

DefaultUniProtoList :: DefaultUni (Esc [])

DefaultUniProtoPair :: DefaultUni (Esc (,))

DefaultUniApply :: !(DefaultUni (Esc f)) -> !(DefaultUni (Esc a)) -> DefaultUni (Esc (f a))

DefaultUniData :: DefaultUni (Esc Data)

DefaultUniBLS12_381_G1_Element :: DefaultUni (Esc Element)

DefaultUniBLS12_381_G2_Element :: DefaultUni (Esc Element)

DefaultUniBLS12_381_MlResult :: DefaultUni (Esc MlResult)

pattern DefaultUniList :: forall {a} {k1} {k2} {f :: k1 -> k2} {a1 :: k1}. () => forall. (a ~ Esc (f a1), Esc f ~ Esc []) => DefaultUni (Esc a1) -> DefaultUni a

pattern DefaultUniPair :: forall {a} {k1} {k2} {f1 :: k1 -> k2} {a1 :: k1} {k3} {k4} {f2 :: k3 -> k4} {a2 :: k3}. () => forall. (a ~ Esc (f1 a1), Esc f1 ~ Esc (f2 a2), Esc f2 ~ Esc (,)) => DefaultUni (Esc a2) -> DefaultUni (Esc a1) -> DefaultUni a

module Universe

noMoreTypeFunctions :: DefaultUni (Esc (f :: a -> b -> c -> d)) -> any

For pleasing the coverage checker.

data DefaultFun

Default built-in functions.

When updating these, make sure to add them to the protocol version listing! See Note [New builtins and protocol versions]

Constructors

AddInteger

SubtractInteger

MultiplyInteger

DivideInteger

QuotientInteger

RemainderInteger

ModInteger

EqualsInteger

LessThanInteger

LessThanEqualsInteger

AppendByteString

ConsByteString

SliceByteString

LengthOfByteString

IndexByteString

EqualsByteString

LessThanByteString

LessThanEqualsByteString

Sha2_256

Sha3_256

Blake2b_256

VerifyEd25519Signature

VerifyEcdsaSecp256k1Signature

VerifySchnorrSecp256k1Signature

AppendString

EqualsString

EncodeUtf8

DecodeUtf8

IfThenElse

ChooseUnit

Trace

FstPair

SndPair

ChooseList

MkCons

HeadList

TailList

NullList

ChooseData

ConstrData

MapData

ListData

IData

BData

UnConstrData

UnMapData

UnListData

UnIData

UnBData

EqualsData

SerialiseData

MkPairData

MkNilData

MkNilPairData

Bls12_381_G1_add

Bls12_381_G1_neg

Bls12_381_G1_scalarMul

Bls12_381_G1_equal

Bls12_381_G1_hashToGroup

Bls12_381_G1_compress

Bls12_381_G1_uncompress

Bls12_381_G2_add

Bls12_381_G2_neg

Bls12_381_G2_scalarMul

Bls12_381_G2_equal

Bls12_381_G2_hashToGroup

Bls12_381_G2_compress

Bls12_381_G2_uncompress

Bls12_381_millerLoop

Bls12_381_mulMlResult

Bls12_381_finalVerify

nonZeroArg :: (Integer -> Integer -> Integer) -> Integer -> Integer -> EvaluationResult Integer

Turn a function into another function that returns EvaluationFailure when its second argument is 0 or calls the original function otherwise and wraps the result in EvaluationSuccess. Useful for correctly handling div, mod, etc.

eitherToEmitter :: Show e => Either e r -> Emitter (EvaluationResult r)

Turn a function returning Either into another function that emits an error message and returns EvaluationFailure in the Left case and wraps the result in EvaluationSuccess in the Right case.

builtinTagWidth :: NumBits

Using 7 bits to encode builtin tags.

encodeBuiltin :: Word8 -> Encoding

decodeBuiltin :: Get Word8