PlutusCore.Evaluation.Machine.BuiltinCostModel

Documentation

type BuiltinCostModel = BuiltinCostModelBase CostingFun

data BuiltinCostModelBase f

Constructors

BuiltinCostModelBase

Fields

paramAddInteger :: f ModelTwoArguments

paramSubtractInteger :: f ModelTwoArguments

paramMultiplyInteger :: f ModelTwoArguments

paramDivideInteger :: f ModelTwoArguments

paramQuotientInteger :: f ModelTwoArguments

paramRemainderInteger :: f ModelTwoArguments

paramModInteger :: f ModelTwoArguments

paramEqualsInteger :: f ModelTwoArguments

paramLessThanInteger :: f ModelTwoArguments

paramLessThanEqualsInteger :: f ModelTwoArguments

paramAppendByteString :: f ModelTwoArguments

paramConsByteString :: f ModelTwoArguments

paramSliceByteString :: f ModelThreeArguments

paramLengthOfByteString :: f ModelOneArgument

paramIndexByteString :: f ModelTwoArguments

paramEqualsByteString :: f ModelTwoArguments

paramLessThanByteString :: f ModelTwoArguments

paramLessThanEqualsByteString :: f ModelTwoArguments

paramSha2_256 :: f ModelOneArgument

paramSha3_256 :: f ModelOneArgument

paramBlake2b_256 :: f ModelOneArgument

paramVerifyEd25519Signature :: f ModelThreeArguments

paramVerifyEcdsaSecp256k1Signature :: f ModelThreeArguments

paramVerifySchnorrSecp256k1Signature :: f ModelThreeArguments

paramAppendString :: f ModelTwoArguments

paramEqualsString :: f ModelTwoArguments

paramEncodeUtf8 :: f ModelOneArgument

paramDecodeUtf8 :: f ModelOneArgument

paramIfThenElse :: f ModelThreeArguments

paramChooseUnit :: f ModelTwoArguments

paramTrace :: f ModelTwoArguments

paramFstPair :: f ModelOneArgument

paramSndPair :: f ModelOneArgument

paramChooseList :: f ModelThreeArguments

paramMkCons :: f ModelTwoArguments

paramHeadList :: f ModelOneArgument

paramTailList :: f ModelOneArgument

paramNullList :: f ModelOneArgument

paramChooseData :: f ModelSixArguments

paramConstrData :: f ModelTwoArguments

paramMapData :: f ModelOneArgument

paramListData :: f ModelOneArgument

paramIData :: f ModelOneArgument

paramBData :: f ModelOneArgument

paramUnConstrData :: f ModelOneArgument

paramUnMapData :: f ModelOneArgument

paramUnListData :: f ModelOneArgument

paramUnIData :: f ModelOneArgument

paramUnBData :: f ModelOneArgument

paramEqualsData :: f ModelTwoArguments

paramMkPairData :: f ModelTwoArguments

paramMkNilData :: f ModelOneArgument

paramMkNilPairData :: f ModelOneArgument

paramSerialiseData :: f ModelOneArgument

data CostingFun model

Constructors

CostingFun

Fields

costingFunCpu :: model

costingFunMemory :: model

data ModelAddedSizes

s * (x + y) + I

Constructors

ModelAddedSizes

Fields

modelAddedSizesIntercept :: CostingInteger

modelAddedSizesSlope :: CostingInteger

data ModelSubtractedSizes

s * (x - y) + I

Constructors

ModelSubtractedSizes

Fields

modelSubtractedSizesIntercept :: CostingInteger

modelSubtractedSizesSlope :: CostingInteger

modelSubtractedSizesMinimum :: CostingInteger

data ModelConstantOrLinear

if p then s*x else c; p depends on usage

Constructors

ModelConstantOrLinear

Fields

modelConstantOrLinearConstant :: CostingInteger

modelConstantOrLinearIntercept :: CostingInteger

modelConstantOrLinearSlope :: CostingInteger

data ModelConstantOrTwoArguments

if p then f(x,y) else c; p depends on usage

Constructors

ModelConstantOrTwoArguments

Fields

modelConstantOrTwoArgumentsConstant :: CostingInteger

modelConstantOrTwoArgumentsModel :: ModelTwoArguments

data ModelLinearSize

Constructors

ModelLinearSize

Fields

modelLinearSizeIntercept :: CostingInteger

modelLinearSizeSlope :: CostingInteger

data ModelMultipliedSizes

s * (x * y) + I

Constructors

ModelMultipliedSizes

Fields

modelMultipliedSizesIntercept :: CostingInteger

modelMultipliedSizesSlope :: CostingInteger

data ModelMinSize

s * min(x, y) + I

Constructors

ModelMinSize

Fields

modelMinSizeIntercept :: CostingInteger

modelMinSizeSlope :: CostingInteger

data ModelMaxSize

s * max(x, y) + I

Constructors

ModelMaxSize

Fields

modelMaxSizeIntercept :: CostingInteger

modelMaxSizeSlope :: CostingInteger

data ModelOneArgument

Constructors

ModelOneArgumentConstantCost CostingInteger

ModelOneArgumentLinearCost ModelLinearSize

data ModelTwoArguments

Constructors

ModelTwoArgumentsConstantCost CostingInteger

ModelTwoArgumentsLinearInX ModelLinearSize

ModelTwoArgumentsLinearInY ModelLinearSize

ModelTwoArgumentsAddedSizes ModelAddedSizes

ModelTwoArgumentsSubtractedSizes ModelSubtractedSizes

ModelTwoArgumentsMultipliedSizes ModelMultipliedSizes

ModelTwoArgumentsMinSize ModelMinSize

ModelTwoArgumentsMaxSize ModelMaxSize

ModelTwoArgumentsLinearOnDiagonal ModelConstantOrLinear

ModelTwoArgumentsConstAboveDiagonal ModelConstantOrTwoArguments

ModelTwoArgumentsConstBelowDiagonal ModelConstantOrTwoArguments

data ModelThreeArguments

Constructors

ModelThreeArgumentsConstantCost CostingInteger

ModelThreeArgumentsAddedSizes ModelAddedSizes

ModelThreeArgumentsLinearInX ModelLinearSize

ModelThreeArgumentsLinearInY ModelLinearSize

ModelThreeArgumentsLinearInZ ModelLinearSize

data ModelFourArguments

Constructors

ModelFourArgumentsConstantCost CostingInteger

data ModelFiveArguments

Constructors

ModelFiveArgumentsConstantCost CostingInteger

data ModelSixArguments

Constructors

ModelSixArgumentsConstantCost CostingInteger

runCostingFunOneArgument :: CostingFun ModelOneArgument -> ExMemory -> ExBudget

runCostingFunTwoArguments :: CostingFun ModelTwoArguments -> ExMemory -> ExMemory -> ExBudget

runCostingFunThreeArguments :: CostingFun ModelThreeArguments -> ExMemory -> ExMemory -> ExMemory -> ExBudget

runCostingFunFourArguments :: CostingFun ModelFourArguments -> ExMemory -> ExMemory -> ExMemory -> ExMemory -> ExBudget

runCostingFunFiveArguments :: CostingFun ModelFiveArguments -> ExMemory -> ExMemory -> ExMemory -> ExMemory -> ExMemory -> ExBudget

runCostingFunSixArguments :: CostingFun ModelSixArguments -> ExMemory -> ExMemory -> ExMemory -> ExMemory -> ExMemory -> ExMemory -> ExBudget

class Hashable a

The class of types that can be converted to a hash value.

Minimal implementation: hashWithSalt.

Note: the hash is not guaranteed to be stable across library versions, operating systems or architectures. For stable hashing use named hashes: SHA256, CRC32 etc.

If you are looking for Hashable instance in time package, check time-compat

newtype MCostingFun a

Same as CostingFun but maybe missing. We could use 'Compose Maybe CostinFun' instead but we would then need an orphan ToJSON instance.

Constructors

MCostingFun (Maybe (CostingFun a))