PlutusCore.Evaluation.Machine.BuiltinCostModel

Documentation

type BuiltinCostModel = BuiltinCostModelBase CostingFun

data BuiltinCostModelBase f

The main model which contains all data required to predict the cost of builtin functions. See md for how this is generated. Calibrated for the CEK machine.

Constructors

BuiltinCostModelBase

Fields

paramAddInteger :: f ModelTwoArguments

paramSubtractInteger :: f ModelTwoArguments

paramMultiplyInteger :: f ModelTwoArguments

paramDivideInteger :: f ModelTwoArguments

paramQuotientInteger :: f ModelTwoArguments

paramRemainderInteger :: f ModelTwoArguments

paramModInteger :: f ModelTwoArguments

paramEqualsInteger :: f ModelTwoArguments

paramLessThanInteger :: f ModelTwoArguments

paramLessThanEqualsInteger :: f ModelTwoArguments

paramAppendByteString :: f ModelTwoArguments

paramConsByteString :: f ModelTwoArguments

paramSliceByteString :: f ModelThreeArguments

paramLengthOfByteString :: f ModelOneArgument

paramIndexByteString :: f ModelTwoArguments

paramEqualsByteString :: f ModelTwoArguments

paramLessThanByteString :: f ModelTwoArguments

paramLessThanEqualsByteString :: f ModelTwoArguments

paramSha2_256 :: f ModelOneArgument

paramSha3_256 :: f ModelOneArgument

paramBlake2b_256 :: f ModelOneArgument

paramVerifyEd25519Signature :: f ModelThreeArguments

paramVerifyEcdsaSecp256k1Signature :: f ModelThreeArguments

paramVerifySchnorrSecp256k1Signature :: f ModelThreeArguments

paramAppendString :: f ModelTwoArguments

paramEqualsString :: f ModelTwoArguments

paramEncodeUtf8 :: f ModelOneArgument

paramDecodeUtf8 :: f ModelOneArgument

paramIfThenElse :: f ModelThreeArguments

paramChooseUnit :: f ModelTwoArguments

paramTrace :: f ModelTwoArguments

paramFstPair :: f ModelOneArgument

paramSndPair :: f ModelOneArgument

paramChooseList :: f ModelThreeArguments

paramMkCons :: f ModelTwoArguments

paramHeadList :: f ModelOneArgument

paramTailList :: f ModelOneArgument

paramNullList :: f ModelOneArgument

paramChooseData :: f ModelSixArguments

paramConstrData :: f ModelTwoArguments

paramMapData :: f ModelOneArgument

paramListData :: f ModelOneArgument

paramIData :: f ModelOneArgument

paramBData :: f ModelOneArgument

paramUnConstrData :: f ModelOneArgument

paramUnMapData :: f ModelOneArgument

paramUnListData :: f ModelOneArgument

paramUnIData :: f ModelOneArgument

paramUnBData :: f ModelOneArgument

paramEqualsData :: f ModelTwoArguments

paramMkPairData :: f ModelTwoArguments

paramMkNilData :: f ModelOneArgument

paramMkNilPairData :: f ModelOneArgument

paramSerialiseData :: f ModelOneArgument

paramBls12_381_G1_add :: f ModelTwoArguments

paramBls12_381_G1_neg :: f ModelOneArgument

paramBls12_381_G1_scalarMul :: f ModelTwoArguments

paramBls12_381_G1_equal :: f ModelTwoArguments

paramBls12_381_G1_compress :: f ModelOneArgument

paramBls12_381_G1_uncompress :: f ModelOneArgument

paramBls12_381_G1_hashToGroup :: f ModelTwoArguments

paramBls12_381_G2_add :: f ModelTwoArguments

paramBls12_381_G2_neg :: f ModelOneArgument

paramBls12_381_G2_scalarMul :: f ModelTwoArguments

paramBls12_381_G2_equal :: f ModelTwoArguments

paramBls12_381_G2_compress :: f ModelOneArgument

paramBls12_381_G2_uncompress :: f ModelOneArgument

paramBls12_381_G2_hashToGroup :: f ModelTwoArguments

paramBls12_381_millerLoop :: f ModelTwoArguments

paramBls12_381_mulMlResult :: f ModelTwoArguments

paramBls12_381_finalVerify :: f ModelTwoArguments

paramKeccak_256 :: f ModelOneArgument

paramBlake2b_224 :: f ModelOneArgument

paramIntegerToByteString :: f ModelThreeArguments

paramByteStringToInteger :: f ModelTwoArguments

data CostingFun model

Constructors

CostingFun

Fields

costingFunCpu :: model

costingFunMemory :: model

newtype Intercept

A wrapped CostingInteger that is supposed to be used as an intercept.

Constructors

Intercept

Fields

unIntercept :: CostingInteger

newtype Slope

A wrapped CostingInteger that is supposed to be used as a slope.

Constructors

Slope

Fields

unSlope :: CostingInteger

newtype Coefficient0

A wrapped CostingInteger that is supposed to be used as the degree 0 coefficient of a polynomial.

Constructors

Coefficient0

Fields

unCoefficient0 :: CostingInteger

newtype Coefficient1

A wrapped CostingInteger that is supposed to be used as the degree 1 coefficient of a polynomial.

Constructors

Coefficient1

Fields

unCoefficient1 :: CostingInteger

newtype Coefficient2

A wrapped CostingInteger that is supposed to be used as the degree 2 coefficient of a polynomial.

Constructors

Coefficient2

Fields

unCoefficient2 :: CostingInteger

data OneVariableLinearFunction

s * x + I

Constructors

OneVariableLinearFunction

Fields

oneVariableLinearFunctionIntercept :: Intercept

oneVariableLinearFunctionSlope :: Slope

data OneVariableQuadraticFunction

c2*x^2 + c1*x + c0

Constructors

OneVariableQuadraticFunction

Fields

oneVariableQuadraticFunctionC0 :: Coefficient0

oneVariableQuadraticFunctionC1 :: Coefficient1

oneVariableQuadraticFunctionC2 :: Coefficient2

data TwoVariableLinearFunction

s1 * x + s2 * y + I

Constructors

TwoVariableLinearFunction

Fields

twoVariableLinearFunctionIntercept :: Intercept

twoVariableLinearFunctionSlopeX :: Slope

twoVariableLinearFunctionSlopeY :: Slope

data ModelAddedSizes

s * (x + y) + I

Constructors

ModelAddedSizes

Fields

modelAddedSizesIntercept :: Intercept

modelAddedSizesSlope :: Slope

data ModelSubtractedSizes

s * (x - y) + I

Constructors

ModelSubtractedSizes

Fields

modelSubtractedSizesIntercept :: Intercept

modelSubtractedSizesSlope :: Slope

modelSubtractedSizesMinimum :: CostingInteger

data ModelConstantOrLinear

if p then s*x else c; p depends on usage

Constructors

ModelConstantOrLinear

Fields

modelConstantOrLinearConstant :: CostingInteger

modelConstantOrLinearIntercept :: Intercept

modelConstantOrLinearSlope :: Slope

data ModelConstantOrTwoArguments

if p then f(x,y) else c; p depends on usage

Constructors

ModelConstantOrTwoArguments

Fields

modelConstantOrTwoArgumentsConstant :: CostingInteger

modelConstantOrTwoArgumentsModel :: ModelTwoArguments

data ModelMultipliedSizes

s * (x * y) + I

Constructors

ModelMultipliedSizes

Fields

modelMultipliedSizesIntercept :: Intercept

modelMultipliedSizesSlope :: Slope

data ModelMinSize

s * min(x, y) + I

Constructors

ModelMinSize

Fields

modelMinSizeIntercept :: Intercept

modelMinSizeSlope :: Slope

data ModelMaxSize

s * max(x, y) + I

Constructors

ModelMaxSize

Fields

modelMaxSizeIntercept :: Intercept

modelMaxSizeSlope :: Slope

data ModelOneArgument

Constructors

ModelOneArgumentConstantCost CostingInteger

ModelOneArgumentLinearCost OneVariableLinearFunction

data ModelTwoArguments

Constructors

ModelTwoArgumentsConstantCost CostingInteger

ModelTwoArgumentsLinearInX OneVariableLinearFunction

ModelTwoArgumentsLinearInY OneVariableLinearFunction

ModelTwoArgumentsLinearInXAndY TwoVariableLinearFunction

ModelTwoArgumentsAddedSizes ModelAddedSizes

ModelTwoArgumentsSubtractedSizes ModelSubtractedSizes

ModelTwoArgumentsMultipliedSizes ModelMultipliedSizes

ModelTwoArgumentsMinSize ModelMinSize

ModelTwoArgumentsMaxSize ModelMaxSize

ModelTwoArgumentsLinearOnDiagonal ModelConstantOrLinear

ModelTwoArgumentsConstAboveDiagonal ModelConstantOrTwoArguments

ModelTwoArgumentsConstBelowDiagonal ModelConstantOrTwoArguments

ModelTwoArgumentsQuadraticInY OneVariableQuadraticFunction

data ModelThreeArguments

Constructors

ModelThreeArgumentsConstantCost CostingInteger

ModelThreeArgumentsAddedSizes ModelAddedSizes

ModelThreeArgumentsLinearInX OneVariableLinearFunction

ModelThreeArgumentsLinearInY OneVariableLinearFunction

ModelThreeArgumentsLinearInZ OneVariableLinearFunction

ModelThreeArgumentsQuadraticInZ OneVariableQuadraticFunction

ModelThreeArgumentsLiteralInYOrLinearInZ OneVariableLinearFunction

data ModelFourArguments

Constructors

ModelFourArgumentsConstantCost CostingInteger

data ModelFiveArguments

Constructors

ModelFiveArgumentsConstantCost CostingInteger

data ModelSixArguments

Constructors

ModelSixArgumentsConstantCost CostingInteger

runCostingFunOneArgument :: ExMemoryUsage a1 => CostingFun ModelOneArgument -> a1 -> ExBudgetStream

runCostingFunTwoArguments :: (ExMemoryUsage a1, ExMemoryUsage a2) => CostingFun ModelTwoArguments -> a1 -> a2 -> ExBudgetStream

runCostingFunThreeArguments :: (ExMemoryUsage a1, ExMemoryUsage a2, ExMemoryUsage a3) => CostingFun ModelThreeArguments -> a1 -> a2 -> a3 -> ExBudgetStream

runCostingFunFourArguments :: (ExMemoryUsage a1, ExMemoryUsage a2, ExMemoryUsage a3, ExMemoryUsage a4) => CostingFun ModelFourArguments -> a1 -> a2 -> a3 -> a4 -> ExBudgetStream

runCostingFunFiveArguments :: (ExMemoryUsage a1, ExMemoryUsage a2, ExMemoryUsage a3, ExMemoryUsage a4, ExMemoryUsage a5) => CostingFun ModelFiveArguments -> a1 -> a2 -> a3 -> a4 -> a5 -> ExBudgetStream

runCostingFunSixArguments :: (ExMemoryUsage a1, ExMemoryUsage a2, ExMemoryUsage a3, ExMemoryUsage a4, ExMemoryUsage a5, ExMemoryUsage a6) => CostingFun ModelSixArguments -> a1 -> a2 -> a3 -> a4 -> a5 -> a6 -> ExBudgetStream

class Eq a => Hashable a

newtype MCostingFun a

Same as CostingFun but maybe missing. We could use 'Compose Maybe CostinFun' instead but we would then need an orphan ToJSON instance.

Constructors

MCostingFun (Maybe (CostingFun a))