PlutusCore.Evaluation.Machine.BuiltinCostModel

Documentation

type BuiltinCostModel = BuiltinCostModelBase CostingFun

data BuiltinCostModelBase f

The main model which contains all data required to predict the cost of builtin functions. See md for how this is generated. Calibrated for the CEK machine.

Constructors

BuiltinCostModelBase

Fields

paramAddInteger :: f ModelTwoArguments

paramSubtractInteger :: f ModelTwoArguments

paramMultiplyInteger :: f ModelTwoArguments

paramDivideInteger :: f ModelTwoArguments

paramQuotientInteger :: f ModelTwoArguments

paramRemainderInteger :: f ModelTwoArguments

paramModInteger :: f ModelTwoArguments

paramEqualsInteger :: f ModelTwoArguments

paramLessThanInteger :: f ModelTwoArguments

paramLessThanEqualsInteger :: f ModelTwoArguments

paramAppendByteString :: f ModelTwoArguments

paramConsByteString :: f ModelTwoArguments

paramSliceByteString :: f ModelThreeArguments

paramLengthOfByteString :: f ModelOneArgument

paramIndexByteString :: f ModelTwoArguments

paramEqualsByteString :: f ModelTwoArguments

paramLessThanByteString :: f ModelTwoArguments

paramLessThanEqualsByteString :: f ModelTwoArguments

paramSha2_256 :: f ModelOneArgument

paramSha3_256 :: f ModelOneArgument

paramBlake2b_256 :: f ModelOneArgument

paramVerifyEd25519Signature :: f ModelThreeArguments

paramVerifyEcdsaSecp256k1Signature :: f ModelThreeArguments

paramVerifySchnorrSecp256k1Signature :: f ModelThreeArguments

paramAppendString :: f ModelTwoArguments

paramEqualsString :: f ModelTwoArguments

paramEncodeUtf8 :: f ModelOneArgument

paramDecodeUtf8 :: f ModelOneArgument

paramIfThenElse :: f ModelThreeArguments

paramChooseUnit :: f ModelTwoArguments

paramTrace :: f ModelTwoArguments

paramFstPair :: f ModelOneArgument

paramSndPair :: f ModelOneArgument

paramChooseList :: f ModelThreeArguments

paramMkCons :: f ModelTwoArguments

paramHeadList :: f ModelOneArgument

paramTailList :: f ModelOneArgument

paramNullList :: f ModelOneArgument

paramChooseData :: f ModelSixArguments

paramConstrData :: f ModelTwoArguments

paramMapData :: f ModelOneArgument

paramListData :: f ModelOneArgument

paramIData :: f ModelOneArgument

paramBData :: f ModelOneArgument

paramUnConstrData :: f ModelOneArgument

paramUnMapData :: f ModelOneArgument

paramUnListData :: f ModelOneArgument

paramUnIData :: f ModelOneArgument

paramUnBData :: f ModelOneArgument

paramEqualsData :: f ModelTwoArguments

paramMkPairData :: f ModelTwoArguments

paramMkNilData :: f ModelOneArgument

paramMkNilPairData :: f ModelOneArgument

paramSerialiseData :: f ModelOneArgument

paramBls12_381_G1_add :: f ModelTwoArguments

paramBls12_381_G1_neg :: f ModelOneArgument

paramBls12_381_G1_scalarMul :: f ModelTwoArguments

paramBls12_381_G1_multiScalarMul :: f ModelTwoArguments

paramBls12_381_G1_equal :: f ModelTwoArguments

paramBls12_381_G1_compress :: f ModelOneArgument

paramBls12_381_G1_uncompress :: f ModelOneArgument

paramBls12_381_G1_hashToGroup :: f ModelTwoArguments

paramBls12_381_G2_add :: f ModelTwoArguments

paramBls12_381_G2_neg :: f ModelOneArgument

paramBls12_381_G2_scalarMul :: f ModelTwoArguments

paramBls12_381_G2_equal :: f ModelTwoArguments

paramBls12_381_G2_multiScalarMul :: f ModelTwoArguments

paramBls12_381_G2_compress :: f ModelOneArgument

paramBls12_381_G2_uncompress :: f ModelOneArgument

paramBls12_381_G2_hashToGroup :: f ModelTwoArguments

paramBls12_381_millerLoop :: f ModelTwoArguments

paramBls12_381_mulMlResult :: f ModelTwoArguments

paramBls12_381_finalVerify :: f ModelTwoArguments

paramKeccak_256 :: f ModelOneArgument

paramBlake2b_224 :: f ModelOneArgument

paramIntegerToByteString :: f ModelThreeArguments

paramByteStringToInteger :: f ModelTwoArguments

paramAndByteString :: f ModelThreeArguments

paramOrByteString :: f ModelThreeArguments

paramXorByteString :: f ModelThreeArguments

paramComplementByteString :: f ModelOneArgument

paramReadBit :: f ModelTwoArguments

paramWriteBits :: f ModelThreeArguments

paramReplicateByte :: f ModelTwoArguments

paramShiftByteString :: f ModelTwoArguments

paramRotateByteString :: f ModelTwoArguments

paramCountSetBits :: f ModelOneArgument

paramFindFirstSetBit :: f ModelOneArgument

paramRipemd_160 :: f ModelOneArgument

paramExpModInteger :: f ModelThreeArguments

paramDropList :: f ModelTwoArguments

paramLengthOfArray :: f ModelOneArgument

paramListToArray :: f ModelOneArgument

paramIndexArray :: f ModelTwoArguments

paramLookupCoin :: f ModelThreeArguments

paramValueContains :: f ModelTwoArguments

paramValueData :: f ModelOneArgument

paramUnValueData :: f ModelOneArgument

paramInsertCoin :: f ModelFourArguments

paramUnionValue :: f ModelTwoArguments

paramScaleValue :: f ModelTwoArguments

data CostingFun model

A type of costing functions parametric over a model type. In practice the we have one model type `ModelNArguments` for every N, where N is the arity of the builtin whose costs we want to model. Each model type has a number of constructors defining different "shapes" of N-parameter functions which calculate a cost given the sizes of the builtin's arguments.

Constructors

CostingFun

Fields

costingFunCpu :: model

costingFunMemory :: model

class UnimplementedCostingFun a where

In the initial stages of implementing a new builtin it is necessary to provide a temporary costing function which is used until the builtin has been properly costed: `see CostModelGeneration.md`. Each `ModelNArguments` type defines an instance of this class where unimplementedCostingFun is a constant costing function which returns a very high cost for all inputs. This prevents new functions from being used in situations where costs are important until a sensible costing function has been implemented.

Methods

unimplementedCostingFun :: b -> CostingFun a

newtype Intercept

A wrapped CostingInteger that is supposed to be used as an intercept.

Constructors

Intercept

Fields

unIntercept :: CostingInteger

newtype Coefficient0

A wrapped CostingInteger that is supposed to be used as the degree 0 coefficient of a polynomial.

Constructors

Coefficient0

Fields

unCoefficient0 :: CostingInteger

newtype Coefficient1

A wrapped CostingInteger that is supposed to be used as the degree 1 coefficient of a polynomial.

Constructors

Coefficient1

Fields

unCoefficient1 :: CostingInteger

newtype Coefficient2

A wrapped CostingInteger that is supposed to be used as the degree 2 coefficient of a polynomial.

Constructors

Coefficient2

Fields

unCoefficient2 :: CostingInteger

newtype Coefficient00

A wrapped CostingInteger that is supposed to be used as the degree (0,0) coefficient of a two-variable polynomial.

Constructors

Coefficient00

Fields

unCoefficient00 :: CostingInteger

newtype Coefficient10

A wrapped CostingInteger that is supposed to be used as the degree (1,0) coefficient of a two-variable polynomial.

Constructors

Coefficient10

Fields

unCoefficient10 :: CostingInteger

newtype Coefficient01

A wrapped CostingInteger that is supposed to be used as the degree (0,1) coefficient of a two-variable polynomial.

Constructors

Coefficient01

Fields

unCoefficient01 :: CostingInteger

newtype Coefficient20

A wrapped CostingInteger that is supposed to be used as the degree (2,0) coefficient of a two-variable polynomial.

Constructors

Coefficient20

Fields

unCoefficient20 :: CostingInteger

newtype Coefficient11

A wrapped CostingInteger that is supposed to be used as the degree (1,1) coefficient of a two-variable polynomial.

Constructors

Coefficient11

Fields

unCoefficient11 :: CostingInteger

newtype Coefficient02

A wrapped CostingInteger that is supposed to be used as the degree (0,2) coefficient of a two-variable polynomial.

Constructors

Coefficient02

Fields

unCoefficient02 :: CostingInteger

newtype Coefficient12

A wrapped CostingInteger that is supposed to be used as the degree (1,2) coefficient of a two-variable polynomial.

Constructors

Coefficient12

Fields

unCoefficient12 :: CostingInteger

data ExpModCostingFunction

c00 + c01x*y + c12x*y^2 This is used only for expModInteger, whose costing is quite complex.

Constructors

ExpModCostingFunction

Fields

coefficient00 :: Coefficient00

coefficient11 :: Coefficient11

coefficient12 :: Coefficient12

newtype MCostingFun a

Same as CostingFun but maybe missing. We could use 'Compose Maybe CostinFun' instead but we would then need an orphan ToJSON instance.

Constructors

MCostingFun (ToJSON (MCostingFun a)

Instance details

Defined in PlutusCore.Evaluation.Machine.BuiltinCostModel

Methods

toJSON :: MCostingFun a -> Value

toEncoding :: MCostingFun a -> Encoding

toJSONList :: [MCostingFun a] -> Value

toEncodingList :: [MCostingFun a] -> Encoding

omitField :: MCostingFun a -> Bool

Monoid (MCostingFun a)

Instance details

Defined in PlutusCore.Evaluation.Machine.BuiltinCostModel

Methods

mempty :: MCostingFun a

mappend :: MCostingFun a -> MCostingFun a -> MCostingFun a

mconcat :: [MCostingFun a] -> MCostingFun a

Semigroup (MCostingFun a)

Instance details

Defined in PlutusCore.Evaluation.Machine.BuiltinCostModel

Methods

(<>) :: MCostingFun a -> MCostingFun a -> MCostingFun a

sconcat :: NonEmpty (MCostingFun a) -> MCostingFun a

stimes :: Integral b => b -> MCostingFun a -> MCostingFun a