PlutusCore.Evaluation.Machine.CostModelInterface

data CostModelApplyError

A fatal error when trying to create a cost given some plain costmodel parameters.

Constructors

CMUnknownParamError !Text)) :+: (C1 ('MetaCons "CMInternalReadError" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "CMInternalWriteError" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'SourceStrict 'DecidedStrict) (Rec0 String))))

data CostModelApplyWarn

A non-fatal warning when trying to create a cost given some plain costmodel parameters.

Constructors

CMTooManyParamsWarn
See Note [Cost model parameters from the ledger's point of view]

Fields

cmExpected :: !Int

cmActual :: !Int

CMTooFewParamsWarn
See Note [Cost model parameters from the ledger's point of view]

Fields

cmExpected :: !Int

cmActual :: !Int