PlutusCore.Evaluation.Machine.CostingFun.Core

Documentation

data CostingFun model

A type of costing functions parametric over a model type. In practice the we have one model type `ModelNArguments` for every N, where N is the arity of the builtin whose costs we want to model. Each model type has a number of constructors defining different "shapes" of N-parameter functions which calculate a cost given the sizes of the builtin's arguments.

Constructors

CostingFun

Fields

costingFunCpu :: model

costingFunMemory :: model

class UnimplementedCostingFun a where

In the initial stages of implementing a new builtin it is necessary to provide a temporary costing function which is used until the builtin has been properly costed: `see CostModelGeneration.md`. Each `ModelNArguments` type defines an instance of this class where unimplementedCostingFun is a constant costing function which returns a very high cost for all inputs. This prevents new functions from being used in situations where costs are important until a sensible costing function has been implemented.

Methods

unimplementedCostingFun :: b -> CostingFun a

newtype Intercept

A wrapped CostingInteger that is supposed to be used as an intercept.

Constructors

Intercept

Fields

unIntercept :: CostingInteger

newtype Coefficient0

A wrapped CostingInteger that is supposed to be used as the degree 0 coefficient of a polynomial.

Constructors

Coefficient0

Fields

unCoefficient0 :: CostingInteger

newtype Coefficient1

A wrapped CostingInteger that is supposed to be used as the degree 1 coefficient of a polynomial.

Constructors

Coefficient1

Fields

unCoefficient1 :: CostingInteger

newtype Coefficient2

A wrapped CostingInteger that is supposed to be used as the degree 2 coefficient of a polynomial.

Constructors

Coefficient2

Fields

unCoefficient2 :: CostingInteger

newtype Coefficient00

A wrapped CostingInteger that is supposed to be used as the degree (0,0) coefficient of a two-variable polynomial.

Constructors

Coefficient00

Fields

unCoefficient00 :: CostingInteger

newtype Coefficient10

A wrapped CostingInteger that is supposed to be used as the degree (1,0) coefficient of a two-variable polynomial.

Constructors

Coefficient10

Fields

unCoefficient10 :: CostingInteger

newtype Coefficient01

A wrapped CostingInteger that is supposed to be used as the degree (0,1) coefficient of a two-variable polynomial.

Constructors

Coefficient01

Fields

unCoefficient01 :: CostingInteger

newtype Coefficient20

A wrapped CostingInteger that is supposed to be used as the degree (2,0) coefficient of a two-variable polynomial.

Constructors

Coefficient20

Fields

unCoefficient20 :: CostingInteger

newtype Coefficient11

A wrapped CostingInteger that is supposed to be used as the degree (1,1) coefficient of a two-variable polynomial.

Constructors

Coefficient11

Fields

unCoefficient11 :: CostingInteger

newtype Coefficient02

A wrapped CostingInteger that is supposed to be used as the degree (0,2) coefficient of a two-variable polynomial.

Constructors

Coefficient02

Fields

unCoefficient02 :: CostingInteger

newtype Coefficient12

A wrapped CostingInteger that is supposed to be used as the degree (1,2) coefficient of a two-variable polynomial.

Constructors

Coefficient12

Fields

unCoefficient12 :: CostingInteger

data ExpModCostingFunction

c00 + c01x*y + c12x*y^2 This is used only for expModInteger, whose costing is quite complex.

Constructors

ExpModCostingFunction

Fields

coefficient00 :: Coefficient00

coefficient11 :: Coefficient11

coefficient12 :: Coefficient12