PlutusCore.Evaluation.Machine.CostingFun.SimpleJSON

Description

A JSON representation of costing functions for Plutus Core builtins which produces a simple cost model which can be used from Agda and other executables

data Model

This type reflects what is actually in the JSON. The stuff in CostingFun.Core and CostingFun.JSON is much more rigid, allowing parsing only for the model types applicable to the various ModelNArguments types; it also requires entries for everything in DefaultFun. Using the type defined here allows us to be more flexible and parse stuff that's not exactly what's expected in builtinCostModel.json.

Constructors

ConstantCost FromJSON Model

Instance details

Defined in PlutusCore.Evaluation.Machine.CostingFun.SimpleJSON

Methods

parseJSON :: Value -> Parser Model

parseJSONList :: Value -> Parser [Model]

omittedField :: Maybe Model

Show Model

Instance details

Defined in PlutusCore.Evaluation.Machine.CostingFun.SimpleJSON

Methods

showsPrec :: Int -> Model -> ShowS

show :: Model -> String

showList :: [Model] -> ShowS

Lift Model

Instance details

Defined in PlutusCore.Evaluation.Machine.CostingFun.SimpleJSON

Methods

lift :: Quote m => Model -> m Exp

liftTyped :: forall (m :: Type -> Type). Quote m => Model -> Code m Model

data CpuAndMemoryModel

A CPU usage modelling function and a memory usage modelling function bundled together

Constructors

CpuAndMemoryModel

Fields

cpuModel :: Model

memoryModel :: Model