PlutusCore.Evaluation.Machine.Exception

Description

The exceptions that an abstract machine can throw.

Documentation

newtype UnliftingError

When unlifting of a PLC term into a Haskell value fails, this error is thrown.

Constructors

UnliftingErrorE Text

class AsUnliftingError r where

Minimal complete definition

_UnliftingError

Methods

_UnliftingError :: Prism' r UnliftingError

_UnliftingErrorE :: Prism' r Text

data KnownTypeError

The type of errors that readKnown and makeKnown can return.

Constructors

KnownTypeUnliftingError UnliftingError

KnownTypeEvaluationFailure

data MachineError fun

Errors which can occur during a run of an abstract machine.

Constructors

NonPolymorphicInstantiationMachineError
An attempt to reduce a not immediately reducible type instantiation.

NonWrapUnwrappedMachineError
An attempt to unwrap a not wrapped term.

NonFunctionalApplicationMachineError
An attempt to reduce a not immediately reducible application.

OpenTermEvaluatedMachineError
An attempt to evaluate an open term.

UnliftingMachineError UnliftingError
An attempt to compute a constant application resulted in ConstAppError.

BuiltinTermArgumentExpectedMachineError
A builtin expected a term argument, but something else was received

UnexpectedBuiltinTermArgumentMachineError
A builtin received a term argument when something else was expected

UnknownBuiltin fun

class AsMachineError r fun | r -> fun where

Minimal complete definition

_MachineError

Methods

_MachineError :: Prism' r (MachineError fun)

_NonPolymorphicInstantiationMachineError :: Prism' r ()

_NonWrapUnwrappedMachineError :: Prism' r ()

_NonFunctionalApplicationMachineError :: Prism' r ()

_OpenTermEvaluatedMachineError :: Prism' r ()

_UnliftingMachineError :: Prism' r UnliftingError

_BuiltinTermArgumentExpectedMachineError :: Prism' r ()

_UnexpectedBuiltinTermArgumentMachineError :: Prism' r ()

_UnknownBuiltin :: Prism' r fun

data EvaluationError user internal

The type of errors (all of them) which can occur during evaluation (some are used-caused, some are internal).

Constructors

InternalEvaluationError internal
Indicates bugs.

UserEvaluationError user
Indicates user errors.

class AsEvaluationError r user internal | r -> user internal where

Minimal complete definition

_EvaluationError

Methods

_EvaluationError :: Prism' r (EvaluationError user internal)

_InternalEvaluationError :: Prism' r internal

_UserEvaluationError :: Prism' r user

data ErrorWithCause err cause

An error and (optionally) what caused it.

Constructors

ErrorWithCause

Fields

_ewcError :: err

_ewcCause :: Maybe cause

type EvaluationException user internal = ErrorWithCause (EvaluationError user internal)

throwNotAConstant :: MonadError KnownTypeError m => m void

mapCauseInMachineException :: (term1 -> term2) -> EvaluationException user (MachineError fun) term1 -> EvaluationException user (MachineError fun) term2

throwing :: MonadError e m => AReview e t -> t -> m x

Throw an Exception described by a ReifiedPrism.

throwing l ≡ reviews l throwError

throwing :: MonadError e m => Prism' e t -> t -> a
throwing :: MonadError e m => Iso' e t   -> t -> a

throwing_ :: MonadError e m => AReview e () -> m x

Similar to throwing but specialised for the common case of error constructors with no arguments.

data MyError = Foo | Bar
makePrisms ''MyError
throwing_ _Foo :: MonadError MyError m => m a

throwingWithCause :: forall exc e t term m x. (exc ~ ErrorWithCause e term, MonadError exc m) => AReview e t -> t -> Maybe term -> m x

Prismatically throw an error and its (optional) cause.

extractEvaluationResult :: Either (EvaluationException user internal term) a -> Either (ErrorWithCause internal term) (EvaluationResult a)

Turn any UserEvaluationError into an EvaluationFailure.

unsafeExtractEvaluationResult :: (PrettyPlc internal, PrettyPlc term, Typeable internal, Typeable term) => Either (EvaluationException user internal term) a -> EvaluationResult a