PlutusCore.Evaluation.Machine.Exception

Description

The exceptions that an abstract machine can throw.

data BuiltinError

The type of errors that readKnown and makeKnown can return.

Constructors

BuiltinUnliftingEvaluationError UnliftingEvaluationError

BuiltinEvaluationFailure

data MachineError fun

Errors which can occur during a run of an abstract machine.

Constructors

NonPolymorphicInstantiationMachineError
An attempt to reduce a not immediately reducible type instantiation.

NonWrapUnwrappedMachineError
An attempt to unwrap a not wrapped term.

NonFunctionalApplicationMachineError
An attempt to reduce a not immediately reducible application.

OpenTermEvaluatedMachineError
An attempt to evaluate an open term.

UnliftingMachineError UnliftingError
An attempt to compute a constant application resulted in UnliftingError.

BuiltinTermArgumentExpectedMachineError
A builtin expected a term argument, but something else was received.

UnexpectedBuiltinTermArgumentMachineError
A builtin received a term argument when something else was expected

NonConstrScrutinizedMachineError
An attempt to scrutinize a non-constr.

MissingCaseBranchMachineError Word64
An attempt to go into a non-existent case branch.

PanicMachineError String
A GHC exception was thrown.

data EvaluationError structural operational

The type of errors that can occur during evaluation. There are two kinds of errors:

Structural ones -- these are errors that are indicative of the _structure_ of the program being wrong. For example, a free variable was encountered during evaluation, a non-function was applied to an argument or tailList was applied to a non-list.

Operational ones -- these are errors that are indicative of the _logic_ of the program being wrong. For example, error was executed, tailList was applied to an empty list or evaluation ran out of gas.

On the chain both of these are just regular failures and we don't distinguish between them there: if a script fails, it fails, it doesn't matter what the reason was. However in the tests it does matter why the failure occurred: a structural error may indicate that the test was written incorrectly while an operational error may be entirely expected.

In other words, structural errors are "runtime type errors" and operational errors are regular runtime errors. Which means that evaluating an (erased) well-typed program should never produce a structural error, only an operational one. This creates a sort of "runtime type system" for UPLC and it would be great to stick to it and enforce in tests etc, but we currently don't.

Constructors

StructuralError !structural

OperationalError !operational

data ErrorWithCause err cause

An error and (optionally) what caused it.

Constructors

ErrorWithCause

Fields

_ewcError :: !err

_ewcCause :: !(Maybe cause)

type EvaluationException structural operational = ErrorWithCause (EvaluationError structural operational)

notAConstant :: BuiltinError

throwErrorWithCause :: MonadError (ErrorWithCause e cause) m => e -> cause -> m x

splitStructuralOperational :: Either (EvaluationException structural operational term) a -> Either (ErrorWithCause structural term) (EvaluationResult a)

Preserve the contents of an StructuralError as a Left and turn an OperationalError into a Right EvaluationFailure (thus erasing the content of the error in the latter case).

unsafeSplitStructuralOperational :: (PrettyPlc structural, PrettyPlc term, Typeable structural, Typeable term) => Either (EvaluationException structural operational term) a -> EvaluationResult a

Throw on a StructuralError and turn an OperationalError into an EvaluationFailure (thus erasing the content of the error in the latter case).

class BuiltinErrorToEvaluationError structural operational

Minimal complete definition

builtinErrorToEvaluationError

builtinErrorToEvaluationError :: BuiltinErrorToEvaluationError structural operational => BuiltinError -> EvaluationError structural operational

throwBuiltinErrorWithCause :: (MonadError (EvaluationException structural operational cause) m, BuiltinErrorToEvaluationError structural operational) => cause -> BuiltinError -> m void

Attach a cause to a BuiltinError and throw that. Note that an evaluator might require the cause to be computed lazily for best performance on the happy path, hence this function must not force its first argument. TODO: wrap cause in Lazy once we have it.