PlutusCore.Evaluation.Machine.MachineParameters

Documentation

data CostModel machinecosts builtincosts

We need to account for the costs of evaluator steps and also built-in function evaluation. The models for these have different structures and are used in different parts of the code, so inside the valuator we pass separate objects about most of the time . It's convenient for clients of the evaluator to only have to worry about a single object, so the CostModel type bundles the two together. We could conceivably have different evaluators with different internal costs, so we keep the machine costs abstract. The model for Cek machine steps is in UntypedPlutusCore.Evaluation.Machine.Cek.CekMachineCosts.

Constructors

CostModel

Fields

_machineCostModel :: machinecosts

_builtinCostModel :: builtincosts

machineCostModel :: forall machinecosts builtincosts machinecosts. Lens (CostModel machinecosts builtincosts) (CostModel machinecosts builtincosts) machinecosts machinecosts

builtinCostModel :: forall machinecosts builtincosts builtincosts. Lens (CostModel machinecosts builtincosts) (CostModel machinecosts builtincosts) builtincosts builtincosts

data MachineParameters machinecosts fun val

At execution time we need a BuiltinsRuntime object which includes both the cost model for builtins and their denotations. This bundles one of those together with the cost model for evaluator steps. The term type will be CekValue when we're using this with the CEK machine.

Constructors

MachineParameters

Fields

machineCosts :: machinecosts

builtinsRuntime :: BuiltinsRuntime fun val

mkMachineParameters :: (CostingPart uni fun ~ builtincosts, HasMeaningIn uni val, ToBuiltinMeaning uni fun) => BuiltinSemanticsVariant fun -> CostModel machinecosts builtincosts -> MachineParameters machinecosts fun val

This just uses toBuiltinsRuntime function to convert a BuiltinCostModel to a BuiltinsRuntime.