PlutusCore.Evaluation.Result

Description

This module defines a common type various evaluation machine use to return their results.

Documentation

class AsEvaluationFailure err where

A class for viewing errors as evaluation failures (in the sense of Plutus).

Methods

_EvaluationFailure :: Prism' err ()

evaluationFailure :: AsEvaluationFailure err => err

_EvaluationFailureVia :: Eq err => err -> Prism' err ()

Construct a Prism focusing on the *EvaluationFailure part of err by taking that *EvaluationFailure and

returning it for the setter part of the prism

checking the error for equality with *EvaluationFailure for the opposite direction.

data EvaluationResult a

The parameterized type of results various evaluation engines return. On the PLC side this becomes (via makeKnown) either a call to Error or a value of the PLC counterpart of type a.

Constructors

EvaluationSuccess !a

EvaluationFailure

isEvaluationSuccess :: EvaluationResult a -> Bool

Check whether an EvaluationResult is an EvaluationSuccess.

isEvaluationFailure :: EvaluationResult a -> Bool

Check whether an EvaluationResult is an EvaluationFailure.