PlutusCore.Parser

Description

Parsers for PLC terms in DefaultUni.

Documentation

type PType = Type TyName DefaultUni SourcePos

A PLC Type to be parsed. ATM the parser only works for types in the DefaultUni with DefaultFun.

varType :: Parser PType

funType :: Parser PType

allType :: Parser PType

lamType :: Parser PType

ifixType :: Parser PType

builtinType :: Parser PType

appType :: Parser PType

kind :: Parser (Kind SourcePos)

pType :: Parser PType

Parser for PType.

defaultUniApplication :: Parser (SomeTypeIn (Kinded DefaultUni))

Parser for built-in type applications.

defaultUni :: Parser (SomeTypeIn (Kinded DefaultUni))

Parser for built-in types (the ones from DefaultUni specifically).

Kinded is needed for checking that a type function can be applied to its argument. I.e. we do Plutus kind checking of builtin type applications during parsing, which is unfortunate, but there's no way we could construct a DefaultUni otherwise.

In case of kind error no sensible message is shown, only an overly general one:

>>> :set -XTypeApplications
>>> :set -XOverloadedStrings
>>> import PlutusCore.Error
>>> import PlutusCore.Quote
>>> let runP = putStrLn . either display display . runQuoteT . parseGen @ParserErrorBundle defaultUni
>>> runP "(list integer)"
(list integer)
>>> runP "(bool integer)"
test:1:14:
  |
1 | (bool integer)
  |              ^
expecting "bool", "bytestring", "data", "integer", "list", "pair", "string", "unit", or '('

This is to be fixed.

One thing we could do to avoid doing kind checking during parsing is to parse into

data TextualUni a where TextualUni :: TextualUni (Esc (Tree Text))

i.e. parse into Tree Text and do the kind checking afterwards, but given that we'll still need to do the kind checking of builtins regardless (even for UPLC), we don't win much by deferring doing it.

tyName :: Parser TyName

topSourcePos :: SourcePos

newtype ParserState

Constructors

ParserState

Fields

identifiers :: Map Text Unique

type Parser = ParsecT ParserError Text (StateT ParserState Quote)

initial :: ParserState

intern :: (MonadState ParserState m, MonadQuote m) => Text -> m Unique

Return the unique identifier of a name. If it's not in the current parser state, map the name to a fresh id and add it to the state. Used in the Name parser.

parse :: (AsParserErrorBundle e, MonadError e m, MonadQuote m) => Parser a -> String -> Text -> m a

toErrorB :: Either (ParseErrorBundle Text ParserError) a -> Either ParserErrorBundle a

parseGen :: (AsParserErrorBundle e, MonadError e m, MonadQuote m) => Parser a -> Text -> m a

Generic parser function in which the file path is just "test".

whitespace :: Parser ()

Space consumer.

lexeme :: Parser a -> Parser a

symbol :: Text -> Parser Text

inParens :: Parser a -> Parser a

inBrackets :: Parser a -> Parser a

inBraces :: Parser a -> Parser a

isIdentifierChar :: Char -> Bool

wordPos

Arguments

:: Text
The word to match

-> Parser SourcePos

Create a parser that matches the input word and returns its source position. This is for attaching source positions to parsed terms/programs.

version :: Parser (Version SourcePos)

name :: Parser Name

cachedBuiltin :: Map Text DefaultFun

builtinFunction :: Parser DefaultFun

Parser for builtin functions. Atm the parser can only parse DefaultFun.

conInteger :: Parser Integer

Parser for integer constants.

hexByte :: Parser Word8

Parser for a pair of hex digits to a Word8.

conBS :: Parser ByteString

Parser for bytestring constants. They start with "#".

conText :: Parser Text

Parser for string constants. They are wrapped in double quotes.

conUnit :: Parser ()

Parser for unit.

conBool :: Parser Bool

Parser for bool.

conList :: DefaultUni (Esc a) -> Parser [a]

Parser for lists.

conPair :: DefaultUni (Esc a) -> DefaultUni (Esc b) -> Parser (a, b)

Parser for pairs.

constantOf :: DefaultUni (Esc a) -> Parser a

Parser for constants of the given type.

constant :: Parser (Some (ValueOf DefaultUni))

Parser of constants whose type is in DefaultUni.

program :: Parser (Program TyName Name DefaultUni DefaultFun SourcePos)

Parser for PLC programs.

parseProgram :: (AsParserErrorBundle e, MonadError e m, MonadQuote m) => Text -> m (Program TyName Name DefaultUni DefaultFun SourcePos)

Parse a PLC program. The resulting program will have fresh names. The underlying monad must be capable of handling any parse errors. This passes "test" to the parser as the name of the input stream; to supply a name explicity, use `parse program name input`.

parseTerm :: (AsParserErrorBundle e, MonadError e m, MonadQuote m) => Text -> m (Term TyName Name DefaultUni DefaultFun SourcePos)

Parse a PLC term. The resulting program will have fresh names. The underlying monad must be capable of handling any parse errors.

parseType :: (AsParserErrorBundle e, MonadError e m, MonadQuote m) => Text -> m (Type TyName DefaultUni SourcePos)

Parse a PLC type. The resulting program will have fresh names. The underlying monad must be capable of handling any parse errors.

data SourcePos

The data type SourcePos represents source positions. It contains the name of the source file, a line number, and a column number. Source line and column positions change intensively during parsing, so we need to make them strict to avoid memory leaks.

data ParserError

An error encountered during parsing.

Constructors

UnknownBuiltinType Text SourcePos

BuiltinTypeNotAStar Text SourcePos

UnknownBuiltinFunction Text SourcePos [Text]

InvalidBuiltinConstant Text Text SourcePos