PlutusCore.Subst

Documentation

substTyVarA :: Applicative f => (tyname -> f (Maybe (Type tyname uni ann))) -> Type tyname uni ann -> f (Type tyname uni ann)

Applicatively replace a type variable using the given function.

substVarA :: Applicative f => (name -> f (Maybe (Term tyname name uni fun ann))) -> Term tyname name uni fun ann -> f (Term tyname name uni fun ann)

Applicatively replace a variable using the given function.

substTyVar :: (tyname -> Maybe (Type tyname uni ann)) -> Type tyname uni ann -> Type tyname uni ann

Replace a type variable using the given function.

substVar :: (name -> Maybe (Term tyname name uni fun ann)) -> Term tyname name uni fun ann -> Term tyname name uni fun ann

Replace a variable using the given function.

termSubstNamesM :: Monad m => (name -> m (Maybe (Term tyname name uni fun ann))) -> Term tyname name uni fun ann -> m (Term tyname name uni fun ann)

Naively monadically substitute names using the given function (i.e. do not substitute binders).

termSubstTyNamesM :: Monad m => (tyname -> m (Maybe (Type tyname uni ann))) -> Term tyname name uni fun ann -> m (Term tyname name uni fun ann)

Naively monadically substitute type names using the given function (i.e. do not substitute binders).

typeSubstTyNamesM :: Monad m => (tyname -> m (Maybe (Type tyname uni ann))) -> Type tyname uni ann -> m (Type tyname uni ann)

Naively monadically substitute type names (i.e. do not substitute binders). INLINE is important here because the function is too polymorphic (determined from profiling)

termSubstNames :: (name -> Maybe (Term tyname name uni fun ann)) -> Term tyname name uni fun ann -> Term tyname name uni fun ann

Naively substitute names using the given function (i.e. do not substitute binders).

termSubstTyNames :: (tyname -> Maybe (Type tyname uni ann)) -> Term tyname name uni fun ann -> Term tyname name uni fun ann

Naively substitute type names using the given function (i.e. do not substitute binders).

typeSubstTyNames :: (tyname -> Maybe (Type tyname uni ann)) -> Type tyname uni ann -> Type tyname uni ann

Naively substitute type names (i.e. do not substitute binders).

termSubstFreeNamesA :: (Applicative f, HasUnique name TermUnique) => (name -> f (Maybe (Term tyname name uni fun ann))) -> Term tyname name uni fun ann -> f (Term tyname name uni fun ann)

Applicatively substitute *free* names using the given function.

termSubstFreeNames :: HasUnique name TermUnique => (name -> Maybe (Term tyname name uni fun ann)) -> Term tyname name uni fun ann -> Term tyname name uni fun ann

Substitute *free* names using the given function.

fvTerm :: Ord name => Term tyname name uni fun ann -> Set name

Get all the free term variables in a term.

ftvTerm :: Ord tyname => Term tyname name uni fun ann -> Set tyname

Get all the free type variables in a term.

ftvTy :: Ord tyname => Type tyname uni ann -> Set tyname

Get all the free type variables in a type.

vTerm :: Ord name => Term tyname name uni fun ann -> Set name

Get all the term variables in a term.

tvTerm :: Ord tyname => Term tyname name uni fun ann -> Set tyname

Get all the type variables in a term.

tvTy :: Ord tyname => Type tyname uni ann -> Set tyname

Get all the type variables in a type.

uniquesType :: HasUniques (Type tyname uni ann) => Type tyname uni ann -> Set Unique

Get all the uniques in a type.

uniquesTerm :: HasUniques (Term tyname name uni fun ann) => Term tyname name uni fun ann -> Set Unique

Get all the uniques in a term (including the type-level ones).