PlutusIR.Parser

Description

Parsers for PIR terms in DefaultUni.

Documentation

parse :: (AsParserErrorBundle e, MonadError e m, MonadQuote m) => Parser a -> String -> Text -> m a

program :: Parser (Program TyName Name DefaultUni DefaultFun SourcePos)

pType :: Parser PType

Parser for PType.

pTerm :: Parser PTerm

parseProgram :: (AsParserErrorBundle e, MonadError e m, MonadQuote m) => Text -> m (Program TyName Name DefaultUni DefaultFun SourcePos)

Parse a PIR program. The resulting program will have fresh names. The underlying monad must be capable of handling any parse errors. This passes "test" to the parser as the name of the input stream; to supply a name explicity, use `parse program name input`.

type Parser = ParsecT ParserError Text (StateT ParserState Quote)

data SourcePos

The data type SourcePos represents source positions. It contains the name of the source file, a line number, and a column number. Source line and column positions change intensively during parsing, so we need to make them strict to avoid memory leaks.

topSourcePos :: SourcePos