PlutusIR.Parser

Description

Parsers for PIR terms in DefaultUni.

type Parser = Quote))

data SourcePos

The data type SourcePos represents source positions. It contains the name of the source file, a line number, and a column number. Source line and column positions change intensively during parsing, so we need to make them strict to avoid memory leaks.