PlutusIR.Purity

isSaturated :: forall tyname name uni fun a. TermUnique) => BuiltinsInfo uni fun -> VarsInfo tyname name uni a -> Term tyname name uni fun a -> Bool

Is the given term 'work-free'?

Note: The definition of 'work-free' is a little unclear, but the idea is that evaluating this term should do very a trivial amount of work.

data EvalOrder tyname name uni fun a

The order in which terms get evaluated, along with their purities. We use a DList here for efficient and lazy concatenation

Monoid (EvalOrder tyname name uni fun a)

Instance details

Defined in PlutusIR.Purity

Methods

mempty :: EvalOrder tyname name uni fun a

mappend :: EvalOrder tyname name uni fun a -> EvalOrder tyname name uni fun a -> EvalOrder tyname name uni fun a

mconcat :: [EvalOrder tyname name uni fun a] -> EvalOrder tyname name uni fun a

Semigroup (EvalOrder tyname name uni fun a)

Instance details

Defined in PlutusIR.Purity

Methods

(<>) :: EvalOrder tyname name uni fun a -> EvalOrder tyname name uni fun a -> EvalOrder tyname name uni fun a

sconcat :: NonEmpty (EvalOrder tyname name uni fun a) -> EvalOrder tyname name uni fun a

stimes :: Integral b => b -> EvalOrder tyname name uni fun a -> EvalOrder tyname name uni fun a

unEvalOrder :: EvalOrder tyname name uni fun a -> [EvalTerm tyname name uni fun a]

Get the evaluation order as a list of EvalTerms. Either terminates in a single Unknown, which means that we got to a point where evaluation continues but we don't know where; or terminates normally, in which case we actually got to the end of the evaluation order for the term.

data EvalTerm tyname name uni fun a

Either the "next" term to be evaluated, along with its Purity and WorkFreedom, or we don't know what comes next.

Constructors

Unknown

EvalTerm Purity WorkFreedom (Term tyname name uni fun a)