PlutusIR.Subst

Documentation

fvTerm :: Ord name => Traversal' (Term tyname name uni fun ann) name

Get all the free term variables in a PIR term.

ftvTerm :: Ord tyname => Traversal' (Term tyname name uni fun ann) tyname

Get all the free type variables in a PIR term.

fvBinding :: Ord name => Traversal' (Binding tyname name uni fun ann) name

Get all the free variables in a PIR single let-binding.

ftvBinding :: Ord tyname => Recursivity -> Traversal' (Binding tyname name uni fun ann) tyname

Get all the free type variables in a PIR single let-binding.

ftvTy :: Ord tyname => Traversal' (Type tyname uni ann) tyname

Get all the free type variables in a type.

vTerm :: Fold (Term tyname name uni fun ann) name

Get all the term variables in a term.

tvTerm :: Fold (Term tyname name uni fun ann) tyname

Get all the type variables in a term.

tvTy :: Fold (Type tyname uni ann) tyname

Get all the type variables in a type.