PlutusIR.TypeCheck

Description

Kindtype inferencechecking, mirroring PlutusCore.TypeCheck

Configuration.

newtype BuiltinTypes uni fun

Mapping from Builtins to their Normalized Kinds.

Constructors

BuiltinTypes

Fields

unBuiltinTypes :: Array fun (Dupable (Normalized (Type TyName uni ())))

data PirTCConfig uni fun

extending the plc typecheck config with AllowEscape

Constructors

PirTCConfig

Fields

_pirConfigTCConfig :: TypeCheckConfig uni fun

_pirConfigAllowEscape :: AllowEscape

tccBuiltinTypes :: HasTypeCheckConfig cfg uni fun => Lens' cfg (BuiltinTypes uni fun)

getDefTypeCheckConfig :: (MonadKindCheck err term uni fun ann m, Typecheckable uni fun) => ann -> m (PirTCConfig uni fun)

The default TypeCheckConfig.

Type checking, extending the plc typechecker

inferType :: MonadTypeCheckPir err uni fun ann m => PirTCConfig uni fun -> Term TyName Name uni fun ann -> m (Normalized (Type TyName uni ()))

Infer the type of a term. Note: The "inferred type" can escape its scope if YesEscape config is passed, see [PIR vs Paper Escaping Types Difference]

checkType :: MonadTypeCheckPir err uni fun ann m => PirTCConfig uni fun -> ann -> Term TyName Name uni fun ann -> Normalized (Type TyName uni ()) -> m ()

Check a term against a type. Infers the type of the term and checks that it's equal to the given type throwing a TypeError (annotated with the value of the ann argument) otherwise. Note: this may allow witnessing a type that escapes its scope, see [PIR vs Paper Escaping Types Difference]

inferTypeOfProgram :: MonadTypeCheckPir err uni fun ann m => PirTCConfig uni fun -> Program TyName Name uni fun ann -> m (Normalized (Type TyName uni ()))

Infer the type of a program. Note: The "inferred type" can escape its scope if YesEscape config is passed, see [PIR vs Paper Escaping Types Difference]

checkTypeOfProgram :: MonadTypeCheckPir err uni fun ann m => PirTCConfig uni fun -> ann -> Program TyName Name uni fun ann -> Normalized (Type TyName uni ()) -> m ()

Check a program against a type. Infers the type of the program and checks that it's equal to the given type throwing a TypeError (annotated with the value of the ann argument) otherwise. Note: this may allow witnessing a type that escapes its scope, see [PIR vs Paper Escaping Types Difference]