UntypedPlutusCore

Documentation

module PlutusCore.Name

substVarA :: Applicative f => (name -> f (Maybe (Term name uni fun ann))) -> Term name uni fun ann -> f (Term name uni fun ann)

Applicatively replace a variable using the given function.

substVar :: (name -> Maybe (Term name uni fun ann)) -> Term name uni fun ann -> Term name uni fun ann

Replace a variable using the given function.

termSubstNamesM :: Monad m => (name -> m (Maybe (Term name uni fun ann))) -> Term name uni fun ann -> m (Term name uni fun ann)

Naively monadically substitute names using the given function (i.e. do not substitute binders).

termSubstNames :: (name -> Maybe (Term name uni fun ann)) -> Term name uni fun ann -> Term name uni fun ann

Naively substitute names using the given function (i.e. do not substitute binders).

termSubstFreeNamesA :: (Applicative f, HasUnique name TermUnique) => (name -> f (Maybe (Term name uni fun ann))) -> Term name uni fun ann -> f (Term name uni fun ann)

Applicatively substitute *free* names using the given function.

termSubstFreeNames :: HasUnique name TermUnique => (name -> Maybe (Term name uni fun ann)) -> Term name uni fun ann -> Term name uni fun ann

Substitute *free* names using the given function.

termMapNames :: forall name name' uni fun ann. (name -> name') -> Term name uni fun ann -> Term name' uni fun ann

Completely replace the names with a new name type.

programMapNames :: forall name name' uni fun ann. (name -> name') -> Program name uni fun ann -> Program name' uni fun ann

uniquesTerm :: HasUniques (Term name uni fun ann) => Term name uni fun ann -> Set Unique

Get all the uniques in a term

vTerm :: Ord name => Term name uni fun ann -> Set name

Get all the term variables in a term.

termSize :: Term name uni fun ann -> Integer

Count the number of AST nodes in a term.

programSize :: Program name uni fun ann -> Integer

Count the number of AST nodes in a program.

serialisedSize :: Flat a => a -> Integer

Compute the size of the serialized form of a value.

simplifyTerm :: forall uni fun m a. (ToBuiltinMeaning uni fun, MonadQuote m) => SimplifyOpts a -> Term Name uni fun a -> m (Term Name uni fun a)

simplifyProgram :: forall uni fun m a. (ToBuiltinMeaning uni fun, MonadQuote m) => SimplifyOpts a -> Program Name uni fun a -> m (Program Name uni fun a)

data SimplifyOpts a

Constructors

SimplifyOpts

Fields

_soMaxSimplifierIterations :: Int

_soInlineHints :: InlineHints Name a

soMaxSimplifierIterations :: forall a. Lens' (SimplifyOpts a) Int

soInlineHints :: forall a a. Lens (SimplifyOpts a) (SimplifyOpts a) (InlineHints Name a) (InlineHints Name a)

defaultSimplifyOpts :: SimplifyOpts a

newtype InlineHints name a

Constructors

InlineHints

Fields

shouldInline :: a -> name -> Bool

module UntypedPlutusCore.DeBruijn

module UntypedPlutusCore.Core

data Term name uni fun ann

The type of Untyped Plutus Core terms. Mirrors the type of Typed Plutus Core terms except

all types are removed

IWrap and Unwrap are removed

type abstractions are replaced with Delay

type instantiations are replaced with Force

The latter two are due to the fact that we don't have value restriction in Typed Plutus Core and hence a computation can be stuck expecting only a single type argument for the computation to become unstuck. Therefore we can't just silently remove type abstractions and instantiations and need to replace them with something else that also blocks evaluation (in order for the semantics of an erased program to match with the semantics of the original typed one). Delay and Force serve exactly this purpose.

Constructors

Var !ann !name

LamAbs !ann !name !(Term name uni fun ann)

Apply !ann !(Term name uni fun ann) !(Term name uni fun ann)

Force !ann !(Term name uni fun ann)

Delay !ann !(Term name uni fun ann)

Constant !ann !(Some (ValueOf uni))

Builtin !ann !fun

Error !ann

data Program name uni fun ann

A Program is simply a Term coupled with a Version of the core language.

Constructors

Program

Fields

_progAnn :: ann

_progVer :: Version ann

_progTerm :: Term name uni fun ann

applyProgram :: Program name uni fun () -> Program name uni fun () -> Program name uni fun ()

Take one UPLC program and apply it to another.

parseScoped :: (MonadQuote (Either (ParseErrorBundle Text ParseError)), AsUniqueError (ParseErrorBundle Text ParseError) SourcePos) => ByteString -> Either (ParseErrorBundle Text ParseError) (Program Name DefaultUni DefaultFun SourcePos)

Parse and rewrite so that names are globally unique, not just unique within their scope.

data DefaultUni a

The universe used by default.

data DefaultFun

Default built-in functions.

When updating these, make sure to add them to the protocol version listing! See Note [New builtins and protocol versions]