UntypedPlutusCore

Documentation

module PlutusCore.Name.Unique

substVarA :: Applicative f => (name -> ann -> f (Maybe (Term name uni fun ann))) -> Term name uni fun ann -> f (Term name uni fun ann)

Applicatively replace a variable using the given function.

substVar :: (name -> ann -> Maybe (Term name uni fun ann)) -> Term name uni fun ann -> Term name uni fun ann

Replace a variable using the given function.

termSubstNamesM :: Monad m => (name -> ann -> m (Maybe (Term name uni fun ann))) -> Term name uni fun ann -> m (Term name uni fun ann)

Naively monadically substitute names using the given function (i.e. do not substitute binders).

termSubstNames :: (name -> ann -> Maybe (Term name uni fun ann)) -> Term name uni fun ann -> Term name uni fun ann

Naively substitute names using the given function (i.e. do not substitute binders).

termMapNames :: forall name name' uni fun ann. (name -> name') -> Term name uni fun ann -> Term name' uni fun ann

Completely replace the names with a new name type.

programMapNames :: forall name name' uni fun ann. (name -> name') -> Program name uni fun ann -> Program name' uni fun ann

substConstantA :: Applicative f => (ann -> Some (ValueOf uni) -> f (Maybe (Term name uni fun ann))) -> Term name uni fun ann -> f (Term name uni fun ann)

Applicatively replace a constant using the given function.

substConstant :: (ann -> Some (ValueOf uni) -> Maybe (Term name uni fun ann)) -> Term name uni fun ann -> Term name uni fun ann

Replace a constant using the given function.

termSubstConstantsM :: Monad m => (ann -> Some (ValueOf uni) -> m (Maybe (Term name uni fun ann))) -> Term name uni fun ann -> m (Term name uni fun ann)

Monadically substitute constants using the given function.

termSubstConstants :: (ann -> Some (ValueOf uni) -> Maybe (Term name uni fun ann)) -> Term name uni fun ann -> Term name uni fun ann

Substitute constants using the given function.

vTerm :: Fold (Term name uni fun ann) name

Get all the term variables in a term.

data SimplifyOpts name a

Constructors

SimplifyOpts

Fields

_soMaxSimplifierIterations :: Int

_soMaxCseIterations :: Int

_soCseWhichSubterms :: CseWhichSubterms

_soConservativeOpts :: Bool

_soInlineHints :: InlineHints name a

_soInlineConstants :: Bool

_soInlineCallsiteGrowth :: AstSize

_soPreserveLogging :: Bool

_soApplyToCase :: Bool

soApplyToCase :: forall name a. Lens' (SimplifyOpts name a) Bool

soMaxSimplifierIterations :: forall name a. Lens' (SimplifyOpts name a) Int

soMaxCseIterations :: forall name a. Lens' (SimplifyOpts name a) Int

soCseWhichSubterms :: forall name a. Lens' (SimplifyOpts name a) CseWhichSubterms

soInlineHints :: forall name a name a. Lens (SimplifyOpts name a) (SimplifyOpts name a) (InlineHints name a) (InlineHints name a)

soConservativeOpts :: forall name a. Lens' (SimplifyOpts name a) Bool

soInlineConstants :: forall name a. Lens' (SimplifyOpts name a) Bool

soInlineCallsiteGrowth :: forall name a. Lens' (SimplifyOpts name a) AstSize

soPreserveLogging :: forall name a. Lens' (SimplifyOpts name a) Bool

defaultSimplifyOpts :: SimplifyOpts name a

simplifyTerm :: forall name uni fun m a. Compiling m uni fun name a => SimplifyOpts name a -> BuiltinSemanticsVariant fun -> Term name uni fun a -> m (Term name uni fun a)

simplifyProgram :: forall name uni fun m a. Compiling m uni fun name a => SimplifyOpts name a -> BuiltinSemanticsVariant fun -> Program name uni fun a -> m (Program name uni fun a)

simplifyProgramWithTrace :: forall name uni fun m a. Compiling m uni fun name a => SimplifyOpts name a -> BuiltinSemanticsVariant fun -> Program name uni fun a -> m (Program name uni fun a, SimplifierTrace name uni fun a)

newtype InlineHints name a

Constructors

InlineHints

Fields

shouldInline :: a -> name -> Inline

data CseWhichSubterms

Which subterms should be considered as candidates?

Constructors

AllSubterms

ExcludeWorkFree

termSimplifier :: forall name uni fun m a. Compiling m uni fun name a => SimplifyOpts name a -> BuiltinSemanticsVariant fun -> Term name uni fun a -> SimplifierT name uni fun a m (Term name uni fun a)

module UntypedPlutusCore.Transform.Simplifier

module UntypedPlutusCore.DeBruijn

module UntypedPlutusCore.Core

module UntypedPlutusCore.Check.Scope

module UntypedPlutusCore.AstSize

data Term name uni fun ann

The type of Untyped Plutus Core terms. Mirrors the type of Typed Plutus Core terms except

all types are removed

IWrap and Unwrap are removed

type abstractions are replaced with Delay

type instantiations are replaced with Force

The latter two are due to the fact that we don't have value restriction in Typed Plutus Core and hence a computation can be stuck expecting only a single type argument for the computation to become unstuck. Therefore we can't just silently remove type abstractions and instantiations and need to replace them with something else that also blocks evaluation (in order for the semantics of an erased program to match with the semantics of the original typed one). Delay and Force serve exactly this purpose.

Constructors

Var !ann !name

LamAbs !ann !name !(Term name uni fun ann)

Apply !ann !(Term name uni fun ann) !(Term name uni fun ann)

Force !ann !(Term name uni fun ann)

Delay !ann !(Term name uni fun ann)

Constant !ann !(Some (ValueOf uni))

Builtin !ann !fun

Error !ann

Constr !ann !Word64 ![Term name uni fun ann]

Case !ann !(Term name uni fun ann) !(Vector (Term name uni fun ann)))))))))

type HasUniques (Term name uni fun ann)

Instance details

Defined in UntypedPlutusCore.Core.Type

type HasUniques (Term name uni fun ann) = HasUnique name TermUnique

type UniOf (Term name uni fun ann)

Instance details

Defined in UntypedPlutusCore.Core.Type

type UniOf (Term name uni fun ann) = uni

data Program name uni fun ann

A Program is simply a Term coupled with a Version of the core language.

Constructors

Program

Fields

_progAnn :: ann

_progVer :: Version

_progTerm :: Term name uni fun ann

applyProgram :: (MonadError ApplyProgramError m, Semigroup a) => Program name uni fun a -> Program name uni fun a -> m (Program name uni fun a)

Applies one program to another. Fails if the versions do not match and tries to merge annotations.

parseScoped :: (Text -> m (Program Name DefaultUni DefaultFun SrcSpan)

Parse and rewrite so that names are globally unique, not just unique within their scope.

data DefaultUni a

The universe used by default.

type ToHoles DefaultUni hole Vector a))

type IsBuiltin DefaultUni ([a] :: Type)

Instance details

Defined in PlutusCore.Default.Universe

type IsBuiltin DefaultUni ([a] :: Type) = IsBuiltin DefaultUni (ElaborateBuiltin DefaultUni [a])

type IsBuiltin DefaultUni ((a, b) :: Type)

Instance details

Defined in PlutusCore.Default.Universe

type IsBuiltin DefaultUni ((a, b) :: Type) = IsBuiltin DefaultUni (ElaborateBuiltin DefaultUni (a, b))

data DefaultFun

Default built-in functions.

When updating these, make sure to add them to the protocol version listing! See Note [New builtins/language versions and protocol versions]

Pretty DefaultFun

Instance details

Defined in PlutusCore.Default.Builtins

Methods

pretty :: DefaultFun -> Doc ann

prettyList :: [DefaultFun] -> Doc ann

uni ThunkInfo)
showTypeOf :: Proxy (BuiltinSemanticsVariant DefaultFun) -> String

Pretty (BuiltinSemanticsVariant DefaultFun)

Instance details

Defined in PlutusCore.Default.Builtins

Methods

pretty :: BuiltinSemanticsVariant DefaultFun -> Doc ann

prettyList :: [BuiltinSemanticsVariant DefaultFun] -> Doc ann

type Rep DefaultFun

Instance details

Defined in PlutusCore.Default.Builtins

type Rep DefaultFun = D1 ('MetaData "DefaultFun" "PlutusCore.Default.Builtins" "plutus-core-1.60.0.0-inplace" 'False) ((((((C1 ('MetaCons "AddInteger" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "SubtractInteger" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "MultiplyInteger" 'PrefixI 'False) (U1 :: Type -> Type))) :+: (C1 ('MetaCons "DivideInteger" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "QuotientInteger" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "RemainderInteger" 'PrefixI 'False) (U1 :: Type -> Type)))) :+: ((C1 ('MetaCons "ModInteger" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "EqualsInteger" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "LessThanInteger" 'PrefixI 'False) (U1 :: Type -> Type))) :+: (C1 ('MetaCons "LessThanEqualsInteger" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "AppendByteString" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "ConsByteString" 'PrefixI 'False) (U1 :: Type -> Type))))) :+: (((C1 ('MetaCons "SliceByteString" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "LengthOfByteString" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "IndexByteString" 'PrefixI 'False) (U1 :: Type -> Type))) :+: (C1 ('MetaCons "EqualsByteString" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "LessThanByteString" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "LessThanEqualsByteString" 'PrefixI 'False) (U1 :: Type -> Type)))) :+: ((C1 ('MetaCons "Sha2_256" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "Sha3_256" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "Blake2b_256" 'PrefixI 'False) (U1 :: Type -> Type))) :+: ((C1 ('MetaCons "VerifyEd25519Signature" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "VerifyEcdsaSecp256k1Signature" 'PrefixI 'False) (U1 :: Type -> Type)) :+: (C1 ('MetaCons "VerifySchnorrSecp256k1Signature" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "AppendString" 'PrefixI 'False) (U1 :: Type -> Type)))))) :+: ((((C1 ('MetaCons "EqualsString" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "EncodeUtf8" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "DecodeUtf8" 'PrefixI 'False) (U1 :: Type -> Type))) :+: (C1 ('MetaCons "IfThenElse" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "ChooseUnit" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "Trace" 'PrefixI 'False) (U1 :: Type -> Type)))) :+: ((C1 ('MetaCons "FstPair" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "SndPair" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "ChooseList" 'PrefixI 'False) (U1 :: Type -> Type))) :+: (C1 ('MetaCons "MkCons" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "HeadList" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "TailList" 'PrefixI 'False) (U1 :: Type -> Type))))) :+: (((C1 ('MetaCons "NullList" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "ChooseData" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "ConstrData" 'PrefixI 'False) (U1 :: Type -> Type))) :+: (C1 ('MetaCons "MapData" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "ListData" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "IData" 'PrefixI 'False) (U1 :: Type -> Type)))) :+: ((C1 ('MetaCons "BData" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "UnConstrData" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "UnMapData" 'PrefixI 'False) (U1 :: Type -> Type))) :+: ((C1 ('MetaCons "UnListData" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "UnIData" 'PrefixI 'False) (U1 :: Type -> Type)) :+: (C1 ('MetaCons "UnBData" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "EqualsData" 'PrefixI 'False) (U1 :: Type -> Type))))))) :+: (((((C1 ('MetaCons "SerialiseData" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "MkPairData" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "MkNilData" 'PrefixI 'False) (U1 :: Type -> Type))) :+: (C1 ('MetaCons "MkNilPairData" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "Bls12_381_G1_add" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "Bls12_381_G1_neg" 'PrefixI 'False) (U1 :: Type -> Type)))) :+: ((C1 ('MetaCons "Bls12_381_G1_scalarMul" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "Bls12_381_G1_equal" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "Bls12_381_G1_hashToGroup" 'PrefixI 'False) (U1 :: Type -> Type))) :+: (C1 ('MetaCons "Bls12_381_G1_compress" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "Bls12_381_G1_uncompress" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "Bls12_381_G2_add" 'PrefixI 'False) (U1 :: Type -> Type))))) :+: (((C1 ('MetaCons "Bls12_381_G2_neg" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "Bls12_381_G2_scalarMul" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "Bls12_381_G2_equal" 'PrefixI 'False) (U1 :: Type -> Type))) :+: (C1 ('MetaCons "Bls12_381_G2_hashToGroup" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "Bls12_381_G2_compress" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "Bls12_381_G2_uncompress" 'PrefixI 'False) (U1 :: Type -> Type)))) :+: ((C1 ('MetaCons "Bls12_381_millerLoop" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "Bls12_381_mulMlResult" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "Bls12_381_finalVerify" 'PrefixI 'False) (U1 :: Type -> Type))) :+: ((C1 ('MetaCons "Keccak_256" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "Blake2b_224" 'PrefixI 'False) (U1 :: Type -> Type)) :+: (C1 ('MetaCons "IntegerToByteString" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "ByteStringToInteger" 'PrefixI 'False) (U1 :: Type -> Type)))))) :+: ((((C1 ('MetaCons "AndByteString" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "OrByteString" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "XorByteString" 'PrefixI 'False) (U1 :: Type -> Type))) :+: (C1 ('MetaCons "ComplementByteString" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "ReadBit" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "WriteBits" 'PrefixI 'False) (U1 :: Type -> Type)))) :+: ((C1 ('MetaCons "ReplicateByte" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "ShiftByteString" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "RotateByteString" 'PrefixI 'False) (U1 :: Type -> Type))) :+: ((C1 ('MetaCons "CountSetBits" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "FindFirstSetBit" 'PrefixI 'False) (U1 :: Type -> Type)) :+: (C1 ('MetaCons "Ripemd_160" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "ExpModInteger" 'PrefixI 'False) (U1 :: Type -> Type))))) :+: (((C1 ('MetaCons "DropList" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "LengthOfArray" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "ListToArray" 'PrefixI 'False) (U1 :: Type -> Type))) :+: (C1 ('MetaCons "IndexArray" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "Bls12_381_G1_multiScalarMul" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "Bls12_381_G2_multiScalarMul" 'PrefixI 'False) (U1 :: Type -> Type)))) :+: ((C1 ('MetaCons "InsertCoin" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "LookupCoin" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "UnionValue" 'PrefixI 'False) (U1 :: Type -> Type))) :+: ((C1 ('MetaCons "ValueContains" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "ValueData" 'PrefixI 'False) (U1 :: Type -> Type)) :+: (C1 ('MetaCons "UnValueData" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "ScaleValue" 'PrefixI 'False) (U1 :: Type -> Type))))))))

data BuiltinSemanticsVariant DefaultFun

Instance details

Defined in PlutusCore.Default.Builtins

data BuiltinSemanticsVariant DefaultFun

= DefaultFunSemanticsVariantA

| DefaultFunSemanticsVariantB

| DefaultFunSemanticsVariantC

| DefaultFunSemanticsVariantD

| DefaultFunSemanticsVariantE

type CostingPart uni DefaultFun

Instance details

Defined in PlutusCore.Default.Builtins

type CostingPart uni DefaultFun = BuiltinCostModel

type Rep (BuiltinSemanticsVariant DefaultFun)

Instance details

Defined in PlutusCore.Default.Builtins

type Rep (BuiltinSemanticsVariant DefaultFun) = D1 ('MetaData "BuiltinSemanticsVariant" "PlutusCore.Default.Builtins" "plutus-core-1.60.0.0-inplace" 'False) ((C1 ('MetaCons "DefaultFunSemanticsVariantA" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "DefaultFunSemanticsVariantB" 'PrefixI 'False) (U1 :: Type -> Type)) :+: (C1 ('MetaCons "DefaultFunSemanticsVariantC" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "DefaultFunSemanticsVariantD" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "DefaultFunSemanticsVariantE" 'PrefixI 'False) (U1 :: Type -> Type))))