UntypedPlutusCore.Core

Documentation

module UntypedPlutusCore.Core.Type

termBinds :: Traversal' (Term name uni fun ann) name

Get all the direct child 'name a's of the given Term from LamAbses.

termVars :: Traversal' (Term name uni fun ann) name

Get all the direct child 'name a's of the given Term from Vars.

termUniques :: HasUniques (Term name uni fun ann) => Traversal' (Term name uni fun ann) Unique

Get all the direct child Uniques of the given Term.

termSubterms :: Traversal' (Term name uni fun ann) (Term name uni fun ann)

Get all the direct child Terms of the given Term.

termSubtermsDeep :: Fold (Term name uni fun ann) (Term name uni fun ann)

Get all the transitive child Terms of the given Term.

termUniquesDeep :: HasUniques (Term name uni fun ann) => Fold (Term name uni fun ann) Unique

Get all the transitive child Uniques of the given Term.

Base functors

data TermF (name :: Type) (uni :: Type -> Type) (fun :: Type) (ann :: Type) r

Constructors

VarF !ann !name

LamAbsF !ann !name !r

ApplyF !ann !r !r

ForceF !ann !r

DelayF !ann !r

ConstantF !ann !(Some (ValueOf uni))

BuiltinF !ann !fun

ErrorF !ann

termTagWidth :: NumBits

Using 4 bits to encode term tags.

encodeTermTag :: Word8 -> Encoding

decodeTermTag :: Get Word8

encodeTerm :: forall name uni fun ann. (Closed uni, uni `Everywhere` Flat, Flat fun, Flat ann, Flat name, Flat (Binder name)) => Term name uni fun ann -> Encoding

decodeTerm :: forall name uni fun ann. (Closed uni, uni `Everywhere` Flat, Flat fun, Flat ann, Flat name, Flat (Binder name)) => (fun -> Maybe String) -> Get (Term name uni fun ann)

sizeTerm :: forall name uni fun ann. (Closed uni, uni `Everywhere` Flat, Flat fun, Flat ann, Flat name, Flat (Binder name)) => Term name uni fun ann -> NumBits -> NumBits

decodeProgram :: forall name uni fun ann. (Closed uni, uni `Everywhere` Flat, Flat fun, Flat ann, Flat name, Flat (Binder name)) => (fun -> Maybe String) -> Get (Program name uni fun ann)