UntypedPlutusCore.Core

decodeProgram :: forall name uni fun ann. (Closed uni, uni `Everywhere` NumBits

newtype UnrestrictedProgram name uni fun ann

A program that can be serialized without any restrictions, e.g. on the set of allowable builtins or term constructs. It is generally safe to use this newtype for serializing, but it should only be used for deserializing in tests.

Constructors

UnrestrictedProgram

Fields

unUnrestrictedProgram :: Program name uni fun ann

UnrestrictedProgram name1 uni1 fun1 ann1 ~ t => Rewrapped (UnrestrictedProgram name2 uni2 fun2 ann2) t

Instance details

Defined in UntypedPlutusCore.Core.Instance.Flat

type Unwrapped (UnrestrictedProgram name uni fun ann)

Instance details

Defined in UntypedPlutusCore.Core.Instance.Flat

type Unwrapped (UnrestrictedProgram name uni fun ann) = Program name uni fun ann

splitParams :: Term name uni fun a -> ([(name, a)], Term name uni fun a)

Strips off lambda binders.

splitApplication :: Term name uni fun a -> (Term name uni fun a, [(a, Term name uni fun a)])

Strip off arguments