UntypedPlutusCore.Evaluation.Machine.Cek.CekMachineCosts

Documentation

type CekMachineCosts = CekMachineCostsBase Identity

Costs for evaluating AST nodes. Times should be specified in picoseconds, memory sizes in bytes.

data CekMachineCostsBase f

Constructors

CekMachineCostsBase

Fields

cekStartupCost :: f ExBudget

cekVarCost :: f ExBudget

cekConstCost :: f ExBudget

cekLamCost :: f ExBudget

cekDelayCost :: f ExBudget

cekForceCost :: f ExBudget

cekApplyCost :: f ExBudget

cekBuiltinCost :: f ExBudget

Just the cost of evaluating a Builtin node, not the builtin itself. There's no entry for Error since we'll be exiting anyway; also, what would happen if calling Error caused the budget to be exceeded?

cekConstrCost :: f ExBudget

cekCaseCost :: f ExBudget