UntypedPlutusCore.Evaluation.Machine.Cek.Internal

Description

The CEK machine. The CEK machine relies on variables having non-equal Uniques whenever they have non-equal string names. I.e. Uniques are used instead of string names. This is for efficiency reasons. The CEK machines handles name capture by design.

Documentation

data EvaluationResult a

The parameterized type of results various evaluation engines return. On the PLC side this becomes (via makeKnown) either a call to Error or a value of the PLC counterpart of type a.

Constructors

EvaluationSuccess !a

EvaluationFailure

data CekValue uni fun

Constructors

VCon !(Some (ValueOf uni))

VDelay !(Term NamedDeBruijn uni fun ()) !(CekValEnv uni fun)

VLamAbs !NamedDeBruijn !(Term NamedDeBruijn uni fun ()) !(CekValEnv uni fun)

VBuiltin
A partial builtin application, accumulating arguments for eventual full application. We don't need a CekValEnv here unlike in the other constructors, because VBuiltin values always store their corresponding Terms fully discharged, see the comments at the call sites (search for VBuiltin).

Fields

!fun

So that we know, for what builtin we're calculating the cost. We can sneak this into BuiltinRuntime, so that we don't need to store it here, but somehow doing so was consistently slowing evaluation down by half a percent. Might be noise, might be not, but at least we know that removing this fun is not helpful anyway. See this commit reversing the change: https://github.com/input-output-hk/plutus/pull/4778/commits/86a3e24ca3c671cc27c6f4344da2bcd14f961706

(Term NamedDeBruijn uni fun ())

This must be lazy. It represents the fully discharged partial application of the builtin function that we're going to run when it's fully saturated. We need the Term to be able to return it in case full saturation is never achieved and a partial application needs to be returned in the result. The laziness is important, because the arguments are discharged values and discharging is expensive, so we don't want to do it unless we really have to. Making this field strict resulted in a 3-4.5% slowdown at the time of writing.

!(BuiltinRuntime (CekValue uni fun))

The partial application and its costing function. Check the docs of BuiltinRuntime for details.

data CekUserError

Constructors

CekOutOfExError ExRestrictingBudget
The final overspent (i.e. negative) budget.

CekEvaluationFailure
Error has been called or a builtin application has failed

type CekEvaluationException name uni fun = EvaluationException CekUserError (MachineError fun) (Term name uni fun ())

The CEK machine-specific EvaluationException.

newtype CekBudgetSpender uni fun s

The CEK machine is parameterized over a spendBudget function. This makes the budgeting machinery extensible and allows us to separate budgeting logic from evaluation logic and avoid branching on the union of all possible budgeting state types during evaluation.

Constructors

CekBudgetSpender

Fields

unCekBudgetSpender :: ExBudgetCategory fun -> ExBudget -> CekM uni fun s ()

data ExBudgetInfo cost uni fun s

Runtime budgeting info.

Constructors

ExBudgetInfo

Fields

_exBudgetModeSpender :: !(CekBudgetSpender uni fun s)

A spending function.

_exBudgetModeGetFinal :: !(ST s cost)

For accessing the final state.

_exBudgetModeGetCumulative :: !(ST s ExBudget)

For accessing the cumulative budget.

newtype ExBudgetMode cost uni fun

A budgeting mode to execute the CEK machine in.

Constructors

ExBudgetMode

Fields

unExBudgetMode :: forall s. ST s (ExBudgetInfo cost uni fun s)

type CekEmitter uni fun s = DList Text -> CekM uni fun s ()

The CEK machine is parameterized over an emitter function, similar to CekBudgetSpender.

data CekEmitterInfo uni fun s

Runtime emitter info, similar to ExBudgetInfo.

Constructors

CekEmitterInfo

Fields

_cekEmitterInfoEmit :: !(CekEmitter uni fun s)

_cekEmitterInfoGetFinal :: !(ST s [Text])

newtype EmitterMode uni fun

An emitting mode to execute the CEK machine in, similar to ExBudgetMode.

Constructors

EmitterMode

Fields

unEmitterMode :: forall s. ST s ExBudget -> ST s (CekEmitterInfo uni fun s)

newtype CekM uni fun s a

The monad the CEK machine runs in.

Constructors

CekM

Fields

unCekM :: ST s a

data ErrorWithCause err cause

An error and (optionally) what caused it.

Constructors

ErrorWithCause

Fields

_ewcError :: err

_ewcCause :: Maybe cause

data EvaluationError user internal

The type of errors (all of them) which can occur during evaluation (some are used-caused, some are internal).

Constructors

InternalEvaluationError internal
Indicates bugs.

UserEvaluationError user
Indicates user errors.

data ExBudgetCategory fun

Constructors

BStep StepKind

BBuiltinApp fun

BStartup

data StepKind

Constructors

BConst

BVar

BLamAbs

BApply

BDelay

BForce

BBuiltin

type PrettyUni uni fun = (Pretty (SomeTypeIn uni), Closed uni, uni `Everywhere` PrettyConst, Pretty fun, Typeable uni, Typeable fun)

The set of constraints we need to be able to print things in universes, which we need in order to throw exceptions.

extractEvaluationResult :: Either (EvaluationException user internal term) a -> Either (ErrorWithCause internal term) (EvaluationResult a)

Turn any UserEvaluationError into an EvaluationFailure.

runCekDeBruijn :: (Ix fun, PrettyUni uni fun) => MachineParameters CekMachineCosts CekValue uni fun -> ExBudgetMode cost uni fun -> EmitterMode uni fun -> Term NamedDeBruijn uni fun () -> (Either (CekEvaluationException NamedDeBruijn uni fun) (Term NamedDeBruijn uni fun ()), cost, [Text])

Evaluate a term using the CEK machine and keep track of costing, logging is optional.

dischargeCekValue :: CekValue uni fun -> Term NamedDeBruijn uni fun ()

Convert a CekValue into a Term by replacing all bound variables with the terms they're bound to (which themselves have to be obtain by recursively discharging values).