UntypedPlutusCore.Evaluation.Machine.SteppableCek

Description

The API to the Steppable CEK machine. Provides the same interface to original CEK machine.

Running the machine

runCek :: ThrowableBuiltins uni fun => MachineParameters CekMachineCosts fun (CekValue uni fun ann) -> ExBudgetMode cost uni fun -> EmitterMode uni fun -> Term Name uni fun ann -> CekReport cost Name uni fun

Evaluate a term using the Steppable CEK machine with logging enabled and keep track of costing. A wrapper around the internal runCek to debruijn input and undebruijn output. *THIS FUNCTION IS PARTIAL if the input term contains free variables*

runCekDeBruijn :: ThrowableBuiltins uni fun => MachineParameters CekMachineCosts fun (CekValue uni fun ann) -> ExBudgetMode cost uni fun -> EmitterMode uni fun -> NTerm uni fun ann -> CekReport cost NamedDeBruijn uni fun

runCekNoEmit :: ThrowableBuiltins uni fun => MachineParameters CekMachineCosts fun (CekValue uni fun ann) -> ExBudgetMode cost uni fun -> Term Name uni fun ann -> (Either (CekEvaluationException Name uni fun) (Term Name uni fun ()), cost)

Evaluate a term using the Steppable CEK machine with logging disabled and keep track of costing. *THIS FUNCTION IS PARTIAL if the input term contains free variables*

evaluateCek :: ThrowableBuiltins uni fun => EmitterMode uni fun -> MachineParameters CekMachineCosts fun (CekValue uni fun ann) -> Term Name uni fun ann -> (Text])

Evaluate a term using the Steppable CEK machine with logging enabled. *THIS FUNCTION IS PARTIAL if the input term contains free variables*

evaluateCekNoEmit :: ThrowableBuiltins uni fun => MachineParameters CekMachineCosts fun (CekValue uni fun ann) -> Term Name uni fun ann -> Either (CekEvaluationException Name uni fun) (Term Name uni fun ())

Evaluate a term using the Steppable CEK machine with logging disabled. *THIS FUNCTION IS PARTIAL if the input term contains free variables*

data EvaluationResult a

The parameterized type of results various evaluation engines return.

Constructors

EvaluationSuccess !a

EvaluationFailure

Methods

typeAst :: Type0 tyname uni ()

(TypeError ('Text "Use \8216BuiltinResult\8217 instead of \8216EvaluationResult\8217") :: Constraint, uni ~ UniOf val) => MakeKnownIn uni val (EvaluationResult a)

Instance details

Defined in PlutusCore.Builtin.KnownType

Methods

makeKnown :: EvaluationResult a -> BuiltinResult val

(TypeError ('Text "Use \8216BuiltinResult\8217 instead of \8216EvaluationResult\8217") :: Constraint, uni ~ UniOf val) => ReadKnownIn uni val (EvaluationResult a)

Instance details

Defined in PlutusCore.Builtin.KnownType

Methods

readKnown :: val -> ReadKnownM (EvaluationResult a)

PrettyBy config a => PrettyBy config (EvaluationResult a)

Instance details

Defined in PlutusCore.Evaluation.Result

Methods

prettyBy :: config -> EvaluationResult a -> Doc ann

prettyListBy :: config -> [EvaluationResult a] -> Doc ann

Generic (EvaluationResult a)

Instance details

Defined in PlutusCore.Evaluation.Result

Associated Types

type Rep (EvaluationResult a) :: Type -> Type

Methods

from :: EvaluationResult a -> Rep (EvaluationResult a) x

to :: Rep (EvaluationResult a) x -> EvaluationResult a

Show a => Show (EvaluationResult a)

Instance details

Defined in PlutusCore.Evaluation.Result

Methods

showsPrec :: Int -> EvaluationResult a -> ShowS

show :: EvaluationResult a -> String

showList :: [EvaluationResult a] -> ShowS

NFData a => NFData (EvaluationResult a)

Instance details

Defined in PlutusCore.Evaluation.Result

Methods

rnf :: EvaluationResult a -> ()

Eq a => Eq (EvaluationResult a)

Instance details

Defined in PlutusCore.Evaluation.Result

Methods

(==) :: EvaluationResult a -> EvaluationResult a -> Bool

(/=) :: EvaluationResult a -> EvaluationResult a -> Bool

PrettyClassic a => Pretty (EvaluationResult a)

Instance details

Defined in PlutusCore.Evaluation.Result

Methods

pretty :: EvaluationResult a -> Doc ann

prettyList :: [EvaluationResult a] -> Doc ann

type ToBinds uni acc (EvaluationResult a :: Type)

Instance details

Defined in PlutusCore.Builtin.KnownTypeAst

type ToBinds uni acc (EvaluationResult a :: Type) = ToBinds uni acc a

type ToHoles uni _1 (EvaluationResult a :: Type)

Instance details

Defined in PlutusCore.Builtin.KnownTypeAst

type ToHoles uni _1 (EvaluationResult a :: Type) = '[TypeHole a :: Hole]

type IsBuiltin uni (EvaluationResult a :: Type)

Instance details

Defined in PlutusCore.Builtin.KnownTypeAst

type IsBuiltin uni (EvaluationResult a :: Type) = 'False

type Rep (EvaluationResult a)

Instance details

Defined in PlutusCore.Evaluation.Result

type Rep (EvaluationResult a) = D1 ('MetaData "EvaluationResult" "PlutusCore.Evaluation.Result" "plutus-core-1.60.0.0-inplace" 'False) (C1 ('MetaCons "EvaluationSuccess" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'SourceStrict 'DecidedStrict) (Rec0 a)) :+: C1 ('MetaCons "EvaluationFailure" 'PrefixI 'False) (U1 :: Type -> Type))

splitStructuralOperational :: Either (EvaluationException structural operational term) a -> Either (ErrorWithCause structural term) (EvaluationResult a)

Preserve the contents of an StructuralError as a Left and turn an OperationalError into a Right EvaluationFailure (thus erasing the content of the error in the latter case).

unsafeSplitStructuralOperational :: (PrettyPlc structural, PrettyPlc term, Typeable structural, Typeable term) => Either (EvaluationException structural operational term) a -> EvaluationResult a

Throw on a StructuralError and turn an OperationalError into an EvaluationFailure (thus erasing the content of the error in the latter case).

Errors

data CekUserError

Constructors

CekCaseBuiltinError Text)) :+: (C1 ('MetaCons "CekOutOfExError" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'SourceStrict 'DecidedStrict) (Rec0 ExRestrictingBudget)) :+: C1 ('MetaCons "CekEvaluationFailure" 'PrefixI 'False) (U1 :: Type -> Type)))

data ErrorWithCause err cause

An error and (optionally) what caused it.

Constructors

ErrorWithCause

Fields

_ewcError :: !err

_ewcCause :: !(Maybe cause)

type CekEvaluationException name uni fun = EvaluationException (MachineError fun) CekUserError (Term name uni fun ())

The CEK machine-specific EvaluationException.

data EvaluationError structural operational

The type of errors that can occur during evaluation. There are two kinds of errors:

Structural ones -- these are errors that are indicative of the _structure_ of the program being wrong. For example, a free variable was encountered during evaluation, a non-function was applied to an argument or tailList was applied to a non-list.

Operational ones -- these are errors that are indicative of the _logic_ of the program being wrong. For example, error was executed, tailList was applied to an empty list or evaluation ran out of gas.

On the chain both of these are just regular failures and we don't distinguish between them there: if a script fails, it fails, it doesn't matter what the reason was. However in the tests it does matter why the failure occurred: a structural error may indicate that the test was written incorrectly while an operational error may be entirely expected.

In other words, structural errors are "runtime type errors" and operational errors are regular runtime errors. Which means that evaluating an (erased) well-typed program should never produce a structural error, only an operational one. This creates a sort of "runtime type system" for UPLC and it would be great to stick to it and enforce in tests etc, but we currently don't.

Constructors

StructuralError !structural

OperationalError !operational

Costing

data ExBudgetCategory fun

Constructors

BStep StepKind

BBuiltinApp fun

BStartup

newtype CekBudgetSpender uni fun s

The CEK machine is parameterized over a spendBudget function. This makes the budgeting machinery extensible and allows us to separate budgeting logic from evaluation logic and avoid branching on the union of all possible budgeting state types during evaluation.

Constructors

CekBudgetSpender

Fields

unCekBudgetSpender :: ExBudgetCategory fun -> ExBudget -> CekM uni fun s ()

newtype ExBudgetMode cost uni fun

A budgeting mode to execute the CEK machine in.

Constructors

ExBudgetMode

Fields

unExBudgetMode :: forall s. ST s (ExBudgetInfo cost uni fun s)

data StepKind

Constructors

BConst

BVar

BLamAbs

BApply

BDelay

BForce

BBuiltin

BConstr

BCase

newtype CekExTally fun

For a detailed report on what costs how much + the same overall budget that Counting gives. The (derived) MonoidalHashMap (ExBudgetCategory fun) ExBudget))))

newtype CountingSt

For calculating the cost of execution by counting up using the Monoid instance of ExBudget.

Constructors

CountingSt ExBudget

data TallyingSt fun

Constructors

TallyingSt (CekExTally fun) ExBudget

newtype RestrictingSt

Constructors

RestrictingSt ExRestrictingBudget

type CekMachineCosts = CekMachineCostsBase Identity

Costs for evaluating AST nodes. Times should be specified in picoseconds, memory sizes in bytes.

Costing modes

counting :: ExBudgetMode CountingSt uni fun

For calculating the cost of execution.

tallying :: Hashable fun => ExBudgetMode (TallyingSt fun) uni fun

For a detailed report on what costs how much + the same overall budget that Counting gives.

restricting :: ThrowableBuiltins uni fun => ExRestrictingBudget -> ExBudgetMode RestrictingSt uni fun

For execution, to avoid overruns.

restrictingEnormous :: ThrowableBuiltins uni fun => ExBudgetMode RestrictingSt uni fun

restricting instantiated at enormousBudget.

enormousBudget :: ExRestrictingBudget

When we want to just evaluate the program we use the Restricting mode with an enormous budget, so that evaluation costs of on-chain budgeting are reflected accurately in benchmarks.

Emitter modes

noEmitter :: EmitterMode uni fun

No emitter.

logEmitter :: EmitterMode uni fun

Emits log only.

logWithTimeEmitter :: EmitterMode uni fun

Emits log with timestamp.

logWithBudgetEmitter :: EmitterMode uni fun

Emits log with the budget.

logWithCallTraceEmitter :: EmitterMode uni fun

Emits log and, when script evaluation fails, call trace.

This requires script to be compiled with `PlutusTx.Plugin:profile-all` turned on because this relies on compiler-generated trace calls that notifies entrance and exit of a function call. These traces that mark entrance and exit are ordinary traces like "-> rob:Example.hs:3:1-3:15" and "<- bob:Example.hs:1:1-1:13" with "->" and "<-" prefixies, where "bob" and "rob" is the name of the function with source span. If regular script with no entrance/exit marker is given, this emitter will behave identically to logEmitter.

When script evaluation fails, this emitter will give call trace of the functions that led to the evaluation failure. This is useful for pin-pointing specific area of the codebase that caused failure when debugging a script. When script evaluation passes, every trace message generated by `profile-all` flag will be removed, and this emitter will behave identically to logEmitter.

Misc

data CekValue uni fun ann

Constructors

VCon !(Some (ValueOf uni))

VDelay !(NTerm uni fun ann) !(CekValEnv uni fun ann)

VLamAbs !NamedDeBruijn !(NTerm uni fun ann) !(CekValEnv uni fun ann)

VBuiltin
A partial builtin application, accumulating arguments for eventual full application. We don't need a CekValEnv here unlike in the other constructors, because VBuiltin values always store their corresponding Terms fully discharged, see the comments at the call sites (search for VBuiltin).

Fields

!fun

So that we know, for what builtin we're calculating the cost. We can sneak this into BuiltinRuntime, so that we don't need to store it here, but somehow doing so was consistently slowing evaluation down by half a percent. Might be noise, might be not, but at least we know that removing this fun is not helpful anyway. See this commit reversing the change: https://github.com/IntersectMBO/plutus/pull/4778/commits/86a3e24ca3c671cc27c6f4344da2bcd14f961706

(NTerm uni fun ())

This must be lazy. It represents the fully discharged partial application of the builtin function that we're going to run when it's fully saturated. We need the Term to be able to return it in case full saturation is never achieved and a partial application needs to be returned in the result. The laziness is important, because the arguments are discharged values and discharging is expensive, so we don't want to do it unless we really have to. Making this field strict resulted in a 3-4.5% slowdown at the time of writing.

!(BuiltinRuntime (CekValue uni fun ann))

The partial application and its costing function. Check the docs of BuiltinRuntime for details. | A constructor value, including fully computed arguments and the tag.

VConstr !Word64 !(EmptyOrMultiStack uni fun ann)

readKnownCek :: (ThrowableBuiltins uni fun, ReadKnown (Term Name uni fun ()) a) => MachineParameters CekMachineCosts fun (CekValue uni fun ann) -> Term Name uni fun ann -> Either (CekEvaluationException Name uni fun) a

Unlift a value using the Steppable CEK machine. *THIS FUNCTION IS PARTIAL if the input term contains free variables*

class hashWithSalt.

Hashable is intended exclusively for use in in-memory data structures. . Hashable does not have a fixed standard. This allows it to improve over time. . Because it does not have a fixed standard, different computers or computers on different versions of the code will observe different hash values. As such, Hashable is not recommended for use other than in-memory datastructures. Specifically, Hashable is not intended for network use or in applications which persist hashed values. For stable hashing use named hashes: sha256, crc32, xxhash etc.

If you are looking for Hashable instance in time package, check time-compat

type ThrowableBuiltins uni fun = (PrettyUni uni, Pretty fun, Typeable uni, Typeable fun)

The set of constraints we need to be able to throw exceptions with things with built-in types and functions in them.